Assesses confounder balancing

Draws on functions from the cobalt package to quantify the relations between exposure and confounders at each exposure time point according to the guidelines from Jackson, 2016 on how to assess balance for time-varying exposures.

assessBalance(
  home_dir,
  data,
  exposure,
  exposure_time_pts,
  outcome,
  type,
  formulas,
  weights = NULL,
  balance_thresh = 0.1,
  imp_conf = NULL,
  verbose = TRUE,
  save.out = TRUE
)

Arguments

home_dir

(optional) path to home directory (required if save.out = TRUE)

data

data in wide format as: a data frame, list of imputed data frames, or mids object

exposure

name of exposure variable

exposure_time_pts

list of integers at which weights will be created/assessed that correspond to time points when exposure wass measured

outcome

name of outcome variable with ".timepoint" suffix

type

type of balance assessment; 'prebalance' or 'weighted'

formulas

list of balancing formulas at each time point output from createFormulas()

weights

list of IPTW weights output from createWeights, required for type 'weighted'

balance_thresh

(optional) one or two numbers between 0 and 1 indicating a single balancing threshold or thresholds for more and less important confounders, respectively (default = 0.1)

imp_conf

(optional) list of variable names reflecting important confounders, required if two balance thresholds are supplied

verbose

(optiona) TRUE or FALSE indicator for user output (default is TRUE)

save.out

(optional) TRUE or FALSE indicator to save output and intermediary output locally (default is TRUE)

Value

a data frame of balance statistics

See also

Examples

test <- data.frame(ID = 1:50,
                   A.1 = rnorm(n = 50),
                   A.2 = rnorm(n = 50),
                   A.3 = rnorm(n = 50),
                   B.1 = rnorm(n = 50),
                   B.2 = rnorm(n = 50),
                   B.3 = rnorm(n = 50),
                   C = rnorm(n = 50),
                   D.3 = rnorm(n = 50))
test[, c("A.1", "A.2", "A.3")] <- lapply(test[, c("A.1", "A.2", "A.3")], as.numeric)

f <- createFormulas(exposure = "A",
                    exposure_time_pts = c(1, 2, 3),
                    outcome = "D.3",
                    tv_confounders = c("A.1", "A.2", "A.3", "B.1", "B.2", "B.3"),
                    ti_confounders = "C",
                    type = "short",
                    save.out = FALSE)
#> USER ALERT: Please manually inspect the short balancing formula below that includes time-varying confounders at t-1 only:
#> The short formula for A-D.3 at A time point 1 is: 
#> A.1 ~ C
#> <environment: 0x561658d2ce60>
#> 
#> USER ALERT: Please manually inspect the short balancing formula below that includes time-varying confounders at t-1 only:
#> The short formula for A-D.3 at A time point 2 is: 
#> A.2 ~ A.1 + B.1 + C
#> <environment: 0x561658d2ce60>
#> 
#> USER ALERT: Please manually inspect the short balancing formula below that includes time-varying confounders at t-1 only:
#> The short formula for A-D.3 at A time point 3 is: 
#> A.3 ~ A.2 + B.2 + C
#> <environment: 0x561658d2ce60>
#> 

#Prebalance
b <- assessBalance(data = test,
                   exposure = "A",
                   exposure_time_pts = c(1, 2, 3),
                   outcome = "D.3",
                   type = "prebalance",
                   formulas = f,
                   save.out = FALSE)
#> USER ALERT: The following statistics display covariate imbalance at each exposure time point prior to weighting,
#>             using short formulas.
#> 
#> Error in model.frame.default(formula = std_bal_stats ~ name, data = f_stats): invalid type (list) for variable 'name'
b <- assessBalance(data = test,
                   exposure = "A",
                   exposure_time_pts = c(1, 2, 3),
                   outcome = "D.3",
                   type = "prebalance",
                   formulas = f,
                   balance_thresh = 0.2,
                   save.out = FALSE)
#> USER ALERT: The following statistics display covariate imbalance at each exposure time point prior to weighting,
#>             using short formulas.
#> 
#> Error in model.frame.default(formula = std_bal_stats ~ name, data = f_stats): invalid type (list) for variable 'name'
b <- assessBalance(data = test,
                   exposure = "A",
                   exposure_time_pts = c(1, 2, 3),
                   outcome = "D.3",
                   type = "prebalance",
                   formulas = f,
                   balance_thresh = c(0.1, 0.2),
                   imp_conf = "B.1",
                   save.out = FALSE)
#> USER ALERT: The following statistics display covariate imbalance at each exposure time point prior to weighting,
#>             using short formulas.
#> 
#> Error in model.frame.default(formula = std_bal_stats ~ name, data = f_stats): invalid type (list) for variable 'name'

# Weighted
w <- createWeights(data = test,
                   exposure = "A",
                   outcome = "D.3",
                   formulas = f,
                   save.out = FALSE)
#> Error in weightit2cbps.cont(covs = covs, treat = treat, subset = by.factor ==     i, s.weights = s.weights, missing = missing, verbose = verbose,     ...): The package "CBPS" is required.

b <- assessBalance(data = test,
                   exposure = "A",
                   exposure_time_pts = c(1, 2, 3),
                   outcome = "D.3",
                   type = "weighted",
                   weights = w,
                   formulas = f,
                   save.out = FALSE)
#> Error in eval(expr, envir, enclos): object 'w' not found
b <- assessBalance(data = test,
                   exposure = "A",
                   exposure_time_pts = c(1, 2, 3),
                   outcome = "D.3",
                   type = "weighted",
                   weights = w,
                   formulas = f,
                   balance_thresh = 0.2,
                   save.out = FALSE)
#> Error in eval(expr, envir, enclos): object 'w' not found
b <- assessBalance(data = test,
                   exposure = "A",
                   exposure_time_pts = c(1, 2, 3),
                   outcome = "D.3",
                   type = "weighted",
                   weights = w,
                   formulas = f,
                   balance_thresh = c(0.1, 0.2),
                   imp_conf = "B.1",
                   save.out = FALSE)
#> Error in eval(expr, envir, enclos): object 'w' not found