Create balancing formulas

Creates balancing formulas relating exposure to all relevant time-varying and time invariant confounders at each exposure time point to be used to create IPTW weights.

createFormulas(
  home_dir,
  exposure,
  exposure_time_pts,
  outcome,
  type,
  ti_confounders,
  tv_confounders = NULL,
  bal_stats = NULL,
  concur_conf = NULL,
  keep_conf = NULL,
  custom = NULL,
  verbose = TRUE,
  save.out = TRUE
)

Arguments

home_dir

path to home directory

exposure

name of exposure variable

exposure_time_pts

list of integers at which weights will be created/assessed that correspond to time points when exposure wass measured

outcome

name of outcome variable with ".timepoint" suffix

type

type of formula to create from 'full' (includes all lagged time-varying confounders), 'short' (includes time-varying confounders at t-1 lag only), or 'update' (adds to 'short' formulas any imbalanced time-varying confounders at lags great than t-1)

ti_confounders

(optional) list of time invariant confounders

tv_confounders

(optional) list of time-varying confounders with ".timepoint" suffix

bal_stats

list of balance statistics from assessBalance(), required for 'update' type

concur_conf

(optional) list of variable names reflecting time-varying confounders to retain in formulas contemporaneously (default is none)

keep_conf

(optional) list of variable names reflecting confounders to always retain in formulas (default depends on type)

custom

(optional) custom list of formulas at each exposure time point (default is to create automatically according to type)

verbose

(optional) TRUE or FALSE indicator for user output (default is TRUE)

save.out

(optional) TRUE or FALSE indicator to save output and intermediary output locally (default is TRUE)

Value

list of balancing formulas at each exposure time point

Examples

#Full Formulas
f <- createFormulas(exposure = "A",
                    exposure_time_pts = c(1, 2, 3),
                    outcome = "D.3",
                    tv_confounders = c("A.1", "A.2", "A.3", "B.1", "B.2", "B.3"),
                    ti_confounders = "C",
                    type = "full",
                    save.out = FALSE)
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A-D.3 at A time point 1 is: 
#> A.1 ~ C
#> <environment: 0x5616525e64c0>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A-D.3 at A time point 2 is: 
#> A.2 ~ A.1 + B.1 + C
#> <environment: 0x5616525e64c0>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A-D.3 at A time point 3 is: 
#> A.3 ~ A.1 + A.2 + B.1 + B.2 + C
#> <environment: 0x5616525e64c0>
#> 

#Short Formulas
f <- createFormulas(exposure = "A",
                    exposure_time_pts = c(1, 2, 3),
                    outcome = "D.3",
                    tv_confounders = c("A.1", "A.2", "A.3", "B.1", "B.2", "B.3"),
                    ti_confounders = "C",
                    type = "short",
                    save.out = FALSE)
#> USER ALERT: Please manually inspect the short balancing formula below that includes time-varying confounders at t-1 only:
#> The short formula for A-D.3 at A time point 1 is: 
#> A.1 ~ C
#> <environment: 0x5616526981e0>
#> 
#> USER ALERT: Please manually inspect the short balancing formula below that includes time-varying confounders at t-1 only:
#> The short formula for A-D.3 at A time point 2 is: 
#> A.2 ~ A.1 + B.1 + C
#> <environment: 0x5616526981e0>
#> 
#> USER ALERT: Please manually inspect the short balancing formula below that includes time-varying confounders at t-1 only:
#> The short formula for A-D.3 at A time point 3 is: 
#> A.3 ~ A.2 + B.2 + C
#> <environment: 0x5616526981e0>
#> 

#Update Formulas
test <- data.frame(ID = 1:50,
                   A.1 = rnorm(n = 50),
                   A.2 = rnorm(n = 50),
                   A.3 = rnorm(n = 50),
                   B.1 = rnorm(n = 50),
                   B.2 = rnorm(n = 50),
                   B.3 = rnorm(n = 50),
                   C = rnorm(n = 50),
                   D.3 = rnorm(n = 50))
test[, c("A.1", "A.2", "A.3")] <- lapply(test[, c("A.1", "A.2", "A.3")], as.numeric)

w <- createWeights(data = test,
                   exposure = "A",
                   outcome = "D.3",
                   formulas = f,
                   save.out = FALSE)
#> Error in weightit2cbps.cont(covs = covs, treat = treat, subset = by.factor ==     i, s.weights = s.weights, missing = missing, verbose = verbose,     ...): The package "CBPS" is required.

b <- assessBalance(data = test,
                   exposure = "A",
                   exposure_time_pts = c(1, 2, 3),
                   outcome = "D.3",
                   type = "weighted",
                   weights = w,
                   formulas = f,
                   save.out = FALSE)
#> Error in eval(expr, envir, enclos): object 'w' not found

f <- createFormulas(exposure = "A",
                    exposure_time_pts = c(1, 2, 3),
                    outcome = "D.3",
                    tv_confounders = c("A.1", "A.2", "A.3", "B.1", "B.2", "B.3"),
                    ti_confounders = "C",
                    type = "update",
                    bal_stats = b,
                    save.out = FALSE)
#> Error in eval(expr, envir, enclos): object 'b' not found