Creates IPTW balancing weights

Creates IPTW balancing weights at each user-specified exposure time point using balancing formulas that relate exposure at each time point to all relevant confounders.

createWeights(
  home_dir,
  data,
  exposure,
  outcome,
  formulas,
  method = "cbps",
  SL.library = NA,
  read_in_from_file = "no",
  verbose = TRUE,
  save.out = TRUE
)

Arguments

home_dir

path to home directory (required if 'save.out' = TRUE)

data

data in wide format as: a data frame, list of imputed data frames, or mids object

exposure

name of exposure variable

outcome

name of outcome variable with ".timepoint" suffix

formulas

list of balancing formulas at each time point output from createFormulas()

method

(optional) character string of WeightItMSM() balancing method abbreviation (default is Covariate Balancing Propensity Score "cbps")

SL.library

required for superLearner weighting method; see SuperLearner::listWrappers() for options

read_in_from_file

(optional) "yes" or "no" indicator to read in weights that have been previously run and saved locally (default is "no")

verbose

(optional) TRUE or FALSE indicator for user output (default is TRUE)

save.out

(optional) TRUE or FALSE indicator to save output and intermediary output locally (default is TRUE)

Value

list of IPTW balancing weights

See also

Examples

test <- data.frame(ID = 1:50,
                   A.1 = rnorm(n = 50),
                   A.2 = rnorm(n = 50),
                   A.3 = rnorm(n = 50),
                   B.1 = rnorm(n = 50),
                   B.2 = rnorm(n = 50),
                   B.3 = rnorm(n = 50),
                   C = rnorm(n = 50),
                   D.3 = rnorm(n = 50))
test[, c("A.1", "A.2", "A.3")] <- lapply(test[, c("A.1", "A.2", "A.3")], as.numeric)

f <- createFormulas(exposure = "A",
                    exposure_time_pts = c(1, 2, 3),
                    outcome = "D.3",
                    tv_confounders = c("A.1", "A.2", "A.3", "B.1", "B.2", "B.3"),,
                    ti_confounders = "C",
                    type = "short",
                    save.out = FALSE)
#> USER ALERT: Please manually inspect the short balancing formula below that includes time-varying confounders at t-1 only:
#> The short formula for A - D.3 at A time point 1 is: 
#> A.1 ~ C
#> <environment: 0x555707b69d40>
#> 
#> USER ALERT: Please manually inspect the short balancing formula below that includes time-varying confounders at t-1 only:
#> The short formula for A - D.3 at A time point 2 is: 
#> A.2 ~ A.1 + B.1 + C
#> <environment: 0x555707b69d40>
#> 
#> USER ALERT: Please manually inspect the short balancing formula below that includes time-varying confounders at t-1 only:
#> The short formula for A - D.3 at A time point 3 is: 
#> A.3 ~ A.2 + B.2 + C
#> <environment: 0x555707b69d40>
#> 

w <- createWeights(data = test,
                   exposure = "A",
                   outcome = "D.3",
                   formulas = f,
                   save.out = FALSE)
#> Error in weightit2cbps.cont(covs = covs, treat = treat, subset = by.factor ==     i, s.weights = s.weights, missing = missing, verbose = verbose,     ...): The package "CBPS" is required.

w <- createWeights(data = test,
                   exposure = "A",
                   outcome = "D.3",
                   formulas = f,
                   method = "cbps",
                   save.out = FALSE)
#> Error in weightit2cbps.cont(covs = covs, treat = treat, subset = by.factor ==     i, s.weights = s.weights, missing = missing, verbose = verbose,     ...): The package "CBPS" is required.