Visualize distribution of sample across exposure histories

Create customized, user-specified exposure histories and tables displaying sample distribution across them for user inspection.

eval_hist(
  data,
  exposure,
  epochs = NULL,
  time_pts,
  hi_lo_cut = NULL,
  ref = NA,
  comps = NULL,
  verbose = TRUE
)

Arguments

data

data in wide format as: a data frame, list of imputed data frames, or mids object

exposure

name of exposure variable

epochs

(optional) data frame of exposure epoch labels and values

time_pts

list of integers at which weights will be created/assessed that correspond to time points when exposure was measured

hi_lo_cut

list of two numbers indicating quantile values that reflect high and low values, respectively, for continuous exposure

ref

(optional) string of "-"-separated "l" and "h" values indicative of a reference exposure history to which to compare comparison, required if comparison is supplied

comps

(optional) list of one or more strings of "-"-separated "l" and "h" values indicative of comparison history/histories to compare to reference, required if reference is supplied

verbose

(optional) TRUE or FALSE indicator for user output (default is TRUE)

Value

none

Examples

test <- data.frame(ID = 1:50,
                   A.1 = rnorm(n = 50),
                   A.2 = rnorm(n = 50),
                   A.3 = rnorm(n = 50),
                   B.1 = rnorm(n = 50),
                   B.2 = rnorm(n = 50),
                   B.3 = rnorm(n = 50),
                   C = rnorm(n = 50),
                   D.3 = rnorm(n = 50))
test[, c("A.1", "A.2", "A.3")] <- lapply(test[, c("A.1", "A.2", "A.3")], as.numeric)

h <- eval_hist(data = test,
               exposure = "A",
               time_pts = c(1, 2, 3))
#> Summary of Exposure Main Effects: 
#> 
#> USER ALERT: Out of the total of 50 individuals in the sample, below is the distribution of the 48 (96%) individuals that fall into 8
#>                   out of the 8 total user-defined exposure histories created from median split values for low and high levels of exposure
#>                   A, respectively, across 1, 2, 3. 
#> USER ALERT: Please inspect the distribution of the sample across the following exposure histories and ensure there is
#>         sufficient spread to avoid extrapolation and low precision: 
#> 
#> Table: Summary of user-specified exposure A histories based on exposure main effects 1, 2, 3
#>                                containing time points 1, 2, 3:
#> 
#> |history |  n|
#> |:-------|--:|
#> |h-h-h   |  2|
#> |h-h-l   |  7|
#> |h-l-h   |  7|
#> |h-l-l   |  7|
#> |l-h-h   |  7|
#> |l-h-l   |  7|
#> |l-l-h   |  8|
#> |l-l-l   |  3|
#> 
h <- eval_hist(data = test,
               exposure = "A",
               time_pts = c(1, 2, 3),
               epochs = data.frame(epochs = c("Infancy", "Toddlerhood"),
                                   values = I(list(c(1, 2), c(3)))))
#> Summary of Exposure Main Effects: 
#> 
#> USER ALERT: Out of the total of 50 individuals in the sample, below is the distribution of the 48 (96%) individuals that fall into 4
#>                   out of the 4 total user-defined exposure histories created from median split values for low and high levels of exposure
#>                   A, respectively, across Infancy, Toddlerhood. 
#> USER ALERT: Please inspect the distribution of the sample across the following exposure histories and ensure there is
#>         sufficient spread to avoid extrapolation and low precision: 
#> 
#> Table: Summary of user-specified exposure A histories based on exposure main effects Infancy, Toddlerhood
#>                                containing time points c(1, 2), 3:
#> 
#> |history |  n|
#> |:-------|--:|
#> |h-h     | 10|
#> |h-l     | 13|
#> |l-h     | 14|
#> |l-l     | 11|
#> 
h <- eval_hist(data = test,
               exposure = "A",
               time_pts = c(1, 2, 3),
               hi_lo_cut = c(0.6, 0.3))
#> Summary of Exposure Main Effects: 
#> 
#> USER ALERT: Out of the total of 50 individuals in the sample, below is the distribution of the 19 (38%) individuals
#>                   that fall into 7 out of the 8 the total user-defined exposure histories created from
#>                   30th and 60th percentile values for low and high levels of exposure A, respectively, across 1, 2, 3. 
#> USER ALERT: Please inspect the distribution of the sample across the following exposure histories and ensure there is
#>         sufficient spread to avoid extrapolation and low precision: 
#> Warning: USER ALERT: There are no individuals in your sample that fall into l-l-l exposure history/histories.
#>                   You may wish to consider different high/low cutoffs (for continuous exposures), alternative epochs, or choose a different measure to avoid extrapolation.
#> 
#> 
#> Table: Summary of user-specified exposure A histories based on exposure main effects 1, 2, 3
#>                                containing time points 1, 2, 3:
#> 
#> |history |  n|
#> |:-------|--:|
#> |h-h-h   |  2|
#> |h-h-l   |  5|
#> |h-l-h   |  4|
#> |h-l-l   |  3|
#> |l-h-h   |  1|
#> |l-h-l   |  1|
#> |l-l-h   |  3|
#> 
h <- eval_hist(data = test,
               exposure = "A",
               time_pts = c(1, 2, 3),
               hi_lo_cut = c(0.6, 0.3),
               ref = "l-l-l",
               comps = "h-h-h")
#> Summary of Exposure Main Effects: 
#> 
#> USER ALERT: Out of the total of 50 individuals in the sample, below is the distribution of the 19 (38%) individuals
#>                   that fall into 7 out of the 2 the total user-defined exposure histories created from
#>                   30th and 60th percentile values for low and high levels of exposure A, respectively, across 1, 2, 3. 
#> USER ALERT: Please inspect the distribution of the sample across the following exposure histories and ensure there is
#>         sufficient spread to avoid extrapolation and low precision: 
#> Warning: USER ALERT: There are no individuals in your sample that fall into l-l-l exposure history/histories.
#>                   You may wish to consider different high/low cutoffs (for continuous exposures), alternative epochs, or choose a different measure to avoid extrapolation.
#> 
#> 
#> Table: Summary of user-specified exposure A histories based on exposure main effects 1, 2, 3
#>                                containing time points 1, 2, 3:
#> 
#> |history |  n|
#> |:-------|--:|
#> |h-h-h   |  2|
#> |h-h-l   |  5|
#> |h-l-h   |  4|
#> |h-l-l   |  3|
#> |l-h-h   |  1|
#> |l-h-l   |  1|
#> |l-l-h   |  3|
#>