Fit outcome model

Fits weighted marginal outcome model as a generalized linear model of the user's choosing, relating exposure main effects to outcome using IPTW weights.

Usage

fitModel(
  data,
  obj,
  weights,
  outcome,
  model = c("m0", "m1", "m2", "m3"),
  int_order = NA,
  covariates = NULL,
  family = NULL,
  link = NA,
  verbose = FALSE,
  save.out = FALSE,
  home_dir = NULL
)

# S3 method for devMSM_models
print(x, i = 1, ...)

Arguments

data

data in wide format as: a data frame, list of imputed data frames, or mids object from the mice package

obj

initialized MSM object from initMSM()

weights

list of IPTW weights output from createWeights()

outcome

name of outcome variable with ".timepoint" suffix. See initMSM() for details on suffix

model

character indicating one of the following outcome models:

"m0" (exposure main effects)

"m1" (exposure main effects & covariates)

"m2" (exposure main effects & their interactions)

"m3" (exposure main effects, their interactions, & covariates)

int_order

integer specification of highest order exposure main effects interaction, required for interaction models ("m2", "m3")

covariates

list of characters reflecting variable names of covariates, required for covariate models ("m1", "m3")

family

(optional) family function specification for WeightIt::glm_weightit() model

link

(optional) character link function specification for WeightIt::glm_weightit() model

verbose

(optional) TRUE or FALSE indicator for printing output to console (default is FALSE)

save.out

(optional) TRUE or FALSE indicator to save output and intermediary output locally (default is FALSE)

home_dir

path to home directory (required if save.out = TRUE)

x

devMSM_models object from fitModel

i

For multiply imputed datset, i selects which imputation to print results for. Default is i = 1. With i = TRUE, all imputed datasets will be looped over.

...

ignored

Value

list containing WeightIt::glm_weightit() model output. It is the length of the number of datasets (1 for a data.frame or the number of imputed datasets)

See also

Examples

library(devMSMs)
data <- data.frame(
  ID = 1:50,
  A.1 = rnorm(n = 50),
  A.2 = rnorm(n = 50),
  A.3 = rnorm(n = 50),
  B.1 = rnorm(n = 50),
  B.2 = rnorm(n = 50),
  B.3 = rnorm(n = 50),
  C = rnorm(n = 50),
  D.3 = rnorm(n = 50)
)
obj <- initMSM(
  data,
  exposure = c("A.1", "A.2", "A.3"),
  ti_conf = c("C"),
  tv_conf = c("B.1", "B.2", "B.3", "D.3")
)
f <- createFormulas(obj, type = "short")
w <- createWeights(data = data, obj = obj, formulas = f)

fit_m0 <- fitModel(
  data = data, obj = obj, weights = w, 
  outcome = "D.3", model = "m0"
)
print(fit_m0)
#> Please inspect the following likelihood ratio test to determine if the exposures collective predict significant variation in the outcome compared to a model without exposure terms.
#> We strongly suggest only conducting history comparisons if the likelihood ratio test is significant.
#> Analysis of Deviance Table
#> 
#> Model 1: D.3 ~ A.1 + A.2 + A.3
#> Model 2: D.3 ~ 1
#>   Resid. Df Resid. Dev Df Deviance      F Pr(>F)
#> 1        46     41.498                          
#> 2        49     43.421 -3  -1.9232 0.7106 0.5506
#> 
#> 
#> The marginal model, m0, is summarized below:
#> +-------------+--------+-----------------+-------+
#> |             | (1)                              |
#> +-------------+--------+-----------------+-------+
#> |             | Est.   | CI              | p     |
#> +=============+========+=================+=======+
#> | (Intercept) | -0.076 | [-0.341, 0.189] | 0.574 |
#> +-------------+--------+-----------------+-------+
#> | A.1         | 0.216  | [-0.038, 0.471] | 0.095 |
#> +-------------+--------+-----------------+-------+
#> | A.2         | -0.047 | [-0.317, 0.223] | 0.734 |
#> +-------------+--------+-----------------+-------+
#> | A.3         | -0.077 | [-0.388, 0.233] | 0.626 |
#> +-------------+--------+-----------------+-------+
#> | Num.Obs.    | 50     |                 |       |
#> +-------------+--------+-----------------+-------+
#> | R2          | 0.033  |                 |       |
#> +-------------+--------+-----------------+-------+ 

fit_m1 <- fitModel(
  data = data, obj = obj, weights = w, 
  outcome = "D.3", model = "m1", 
  covariates = c("C")
)
print(fit_m1)
#> Please inspect the following likelihood ratio test to determine if the exposures collective predict significant variation in the outcome compared to a model without exposure terms.
#> We strongly suggest only conducting history comparisons if the likelihood ratio test is significant.
#> Analysis of Deviance Table
#> 
#> Model 1: D.3 ~ A.1 + A.2 + A.3 + C
#> Model 2: D.3 ~ C
#>   Resid. Df Resid. Dev Df Deviance      F Pr(>F)
#> 1        45     39.583                          
#> 2        48     41.205 -3  -1.6225 0.6148  0.609
#> 
#> 
#> The marginal model, m1, is summarized below:
#> +-------------+--------+-----------------+-------+
#> |             | (1)                              |
#> +-------------+--------+-----------------+-------+
#> |             | Est.   | CI              | p     |
#> +=============+========+=================+=======+
#> | (Intercept) | -0.071 | [-0.326, 0.185] | 0.587 |
#> +-------------+--------+-----------------+-------+
#> | A.1         | 0.194  | [-0.065, 0.453] | 0.143 |
#> +-------------+--------+-----------------+-------+
#> | A.2         | -0.070 | [-0.329, 0.189] | 0.597 |
#> +-------------+--------+-----------------+-------+
#> | A.3         | -0.061 | [-0.366, 0.243] | 0.693 |
#> +-------------+--------+-----------------+-------+
#> | C           | -0.192 | [-0.418, 0.035] | 0.097 |
#> +-------------+--------+-----------------+-------+
#> | Num.Obs.    | 50     |                 |       |
#> +-------------+--------+-----------------+-------+
#> | R2          | 0.062  |                 |       |
#> +-------------+--------+-----------------+-------+ 

fit_m2 <- fitModel(
  data = data, obj = obj, weights = w, 
  outcome = "D.3", model = "m2", 
  int_order = 2
)
print(fit_m2)
#> Please inspect the following likelihood ratio test to determine if the exposures collective predict significant variation in the outcome compared to a model without exposure terms.
#> We strongly suggest only conducting history comparisons if the likelihood ratio test is significant.
#> Analysis of Deviance Table
#> 
#> Model 1: D.3 ~ A.1 + A.2 + A.3 + A.1 + A.2 + A.3 + A.1:A.2 + A.1:A.3 + 
#>     A.2:A.3
#> Model 2: D.3 ~ 1
#>   Resid. Df Resid. Dev Df Deviance      F Pr(>F)
#> 1        43     39.672                          
#> 2        49     43.421 -6   -3.749 0.6772 0.6686
#> 
#> 
#> The marginal model, m2, is summarized below:
#> +-------------+--------+-----------------+-------+
#> |             | (1)                              |
#> +-------------+--------+-----------------+-------+
#> |             | Est.   | CI              | p     |
#> +=============+========+=================+=======+
#> | (Intercept) | -0.048 | [-0.305, 0.209] | 0.715 |
#> +-------------+--------+-----------------+-------+
#> | A.1         | 0.095  | [-0.253, 0.442] | 0.593 |
#> +-------------+--------+-----------------+-------+
#> | A.2         | -0.114 | [-0.394, 0.165] | 0.422 |
#> +-------------+--------+-----------------+-------+
#> | A.3         | -0.210 | [-0.548, 0.128] | 0.224 |
#> +-------------+--------+-----------------+-------+
#> | A.1 × A.2   | 0.135  | [-0.155, 0.426] | 0.360 |
#> +-------------+--------+-----------------+-------+
#> | A.1 × A.3   | 0.208  | [-0.214, 0.629] | 0.334 |
#> +-------------+--------+-----------------+-------+
#> | A.2 × A.3   | 0.177  | [-0.151, 0.505] | 0.290 |
#> +-------------+--------+-----------------+-------+
#> | Num.Obs.    | 50     |                 |       |
#> +-------------+--------+-----------------+-------+
#> | R2          | 0.075  |                 |       |
#> +-------------+--------+-----------------+-------+ 

fit_m3 <- fitModel(
  data = data, obj = obj, weights = w, 
  outcome = "D.3", model = "m3",
  int_order = 2, covariates = c("C")
)
print(fit_m3)
#> Please inspect the following likelihood ratio test to determine if the exposures collective predict significant variation in the outcome compared to a model without exposure terms.
#> We strongly suggest only conducting history comparisons if the likelihood ratio test is significant.
#> Analysis of Deviance Table
#> 
#> Model 1: D.3 ~ A.1 + A.2 + A.3 + A.1 + A.2 + A.3 + A.1:A.2 + A.1:A.3 + 
#>     A.2:A.3 + C
#> Model 2: D.3 ~ C
#>   Resid. Df Resid. Dev Df Deviance      F Pr(>F)
#> 1        42     37.005                          
#> 2        48     41.205 -6  -4.2005 0.7946 0.5795
#> 
#> 
#> The marginal model, m3, is summarized below:
#> +-------------+--------+------------------+-------+
#> |             | (1)                               |
#> +-------------+--------+------------------+-------+
#> |             | Est.   | CI               | p     |
#> +=============+========+==================+=======+
#> | (Intercept) | -0.041 | [-0.288, 0.207]  | 0.747 |
#> +-------------+--------+------------------+-------+
#> | A.1         | 0.065  | [-0.292, 0.422]  | 0.720 |
#> +-------------+--------+------------------+-------+
#> | A.2         | -0.148 | [-0.407, 0.110]  | 0.261 |
#> +-------------+--------+------------------+-------+
#> | A.3         | -0.224 | [-0.544, 0.096]  | 0.170 |
#> +-------------+--------+------------------+-------+
#> | C           | -0.232 | [-0.450, -0.014] | 0.037 |
#> +-------------+--------+------------------+-------+
#> | A.1 × A.2   | 0.127  | [-0.161, 0.416]  | 0.386 |
#> +-------------+--------+------------------+-------+
#> | A.1 × A.3   | 0.228  | [-0.188, 0.645]  | 0.283 |
#> +-------------+--------+------------------+-------+
#> | A.2 × A.3   | 0.249  | [-0.075, 0.574]  | 0.132 |
#> +-------------+--------+------------------+-------+
#> | Num.Obs.    | 50     |                  |       |
#> +-------------+--------+------------------+-------+
#> | R2          | 0.122  |                  |       |
#> +-------------+--------+------------------+-------+