Formats long data

Formats long data

formatLongData(
  home_dir,
  data,
  exposure,
  exposure_time_pts,
  outcome,
  tv_confounders,
  time_var = NA,
  id_var = NA,
  missing = NA,
  factor_confounders = NULL,
  save.out = TRUE
)

Arguments

home_dir

path to home directory

data

dataframe in long format

exposure

name of exposure variable

exposure_time_pts

list of integers at which weights will be created/assessed that correspond to time points when exposure was measured

outcome

name of outcome variable with ".timepoint" suffix

tv_confounders

list of time-varying confounders with ".timepoint" suffix

time_var

(optional) variable name in original dataset demarcating time

id_var

(optional) variable name in original dataset demarcating ID

missing

(optional) indicator for missing data in original dataset

factor_confounders

(optional) list of variable names that are factors (default is numeric)

save.out

(optional) TRUE or FALSE indicator to save output and intermediary output locally (default is TRUE)

Value

formatted long dataset

Examples

test <- data.frame(ID = 1:50,
                   A.1 = rnorm(n = 50),
                   A.2 = rnorm(n = 50),
                   A.3 = rnorm(n = 50),
                   B.1 = rnorm(n = 50),
                   B.2 = rnorm(n = 50),
                   B.3 = rnorm(n = 50),
                   C = rnorm(n = 50),
                   D.3 = rnorm(n = 50))
test[, c("A.1", "A.2", "A.3")] <- lapply(test[, c("A.1", "A.2", "A.3")], as.numeric)

test_long <- stats::reshape(data = test,
                            idvar = "ID", #'list ID variable
                            varying = c("A.1", "A.2", "A.3", "B.1", "B.2", "B.3"),
                            direction = "long")

test_long_format <- formatLongData(data = test_long,
                                   exposure = "A",
                                   exposure_time_pts = c(1, 2, 3),
                                   outcome = "D.3",
                                   tv_confounders = c("A.1", "A.2", "A.3", "B.1", "B.2", "B.3"),
                                   time_var = "time",
                                   id_var = NA,
                                   missing = NA,
                                   factor_confounders = "C",
                                   save.out = FALSE)
#> Table: Summary of A Exposure Information
#> 
#> | WAVE|       mean|        sd|       min|      max|
#> |----:|----------:|---------:|---------:|--------:|
#> |    1|  0.0916316| 1.1073026| -2.279805| 2.493613|
#> |    2| -0.2911299| 0.9481285| -2.668422| 1.587347|
#> |    3|  0.0738718| 1.0769932| -2.528841| 2.381511|
#> Error in loadNamespace(x): there is no package called ‘psych’