Fits outcome model

Fits outcome model

getModel(
  d,
  exposure,
  exposure_time_pts,
  outcome,
  epochs,
  exp_epochs,
  int_order,
  model,
  fam,
  covariates,
  verbose
)

Arguments

d

wide data frame

exposure

name of exposure variable

exposure_time_pts

list of integers at which weights will be created/assessed that correspond to time points when exposure was measured

outcome

name of outcome variable with ".timepoint" suffix

epochs

data frame of exposure epoch labels and values

exp_epochs

list of exposure epochs

int_order

integer specification of highest order exposure main effects interaction for interaction models

model

character indicating one of the following outcome models:

"m0" (exposure main effects)

"m1" (exposure main effects & covariates)

"m2" (exposure main effects & their interactions)

"m3" (exposure main effects, their interactions, & covariates)

"covs" (covariate only model)

"int" (intercept only model)

fam

function specification for svyglm model

covariates

list of characters reflecting variable names of covariates for covariate models

verbose

TRUE or FALSE indicator for user output (default is TRUE)

Value

list of fitted model(s)

Examples

test <- data.frame(ID = 1:50,
                   A.1 = rnorm(n = 50),
                   A.2 = rnorm(n = 50),
                   A.3 = rnorm(n = 50),
                   B.1 = rnorm(n = 50),
                   B.2 = rnorm(n = 50),
                   B.3 = rnorm(n = 50),
                   C = rnorm(n = 50),
                   D.3 = rnorm(n = 50))
test[, c("A.1", "A.2", "A.3")] <- lapply(test[, c("A.1", "A.2", "A.3")], as.numeric)

f <- createFormulas(exposure = "A",
                    exposure_time_pts = c(1, 2, 3),
                    outcome = "D.3",
                    tv_confounders = c("A.1", "A.2", "A.3", "B.1", "B.2", "B.3"),
                    ti_confounders = "C",
                    type = "full",
                    save.out = FALSE)
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A - D.3 at A time point 1 is: 
#> A.1 ~ C
#> <environment: 0x55eef19b28b8>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A - D.3 at A time point 2 is: 
#> A.2 ~ A.1 + B.1 + C
#> <environment: 0x55eef19b28b8>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A - D.3 at A time point 3 is: 
#> A.3 ~ A.1 + A.2 + B.1 + B.2 + C
#> <environment: 0x55eef19b28b8>
#> 

w <- createWeights(data = test,
                   exposure = "A",
                   outcome = "D.3",
                   formulas = f,
                   save.out = FALSE)
#> Error in weightit2cbps.cont(covs = covs, treat = treat, subset = by.factor ==     i, s.weights = s.weights, missing = missing, verbose = verbose,     ...): The package "CBPS" is required.

epochs = data.frame(epochs = c("Infancy", "Toddlerhood"),
                    values = I(list(c(1, 2), c(3))))
e <- apply(expand.grid("A", as.character(epochs[, 1])), 1,
           paste, sep = "", collapse = ".")
test$weights <- w[[1]]$weights
#> Error in eval(expr, envir, enclos): object 'w' not found

g <- getModel(d = test,
              exposure = "A",
              exposure_time_pts = c(1, 2, 3),
              outcome = "D.3",
              epochs = epochs,
              exp_epochs = e,
              fam = gaussian,
              model = "m0")
#> Error in model.frame.default(~weights, data = structure(list(ID = 1:50,     A.1 = c(-0.747561716383354, 1.64806378586186, 0.955114329975459,     1.28699314902885, -0.114266489962407, 2.10501704651085, 0.176146033889081,     0.539101850898798, 0.641919790397721, 0.758882332795139,     -0.608727637140833, -0.717235272214663, 0.00617685315040539,     -0.391481970608912, 0.28415608107183, -0.136934655779507,     1.485457093172, 0.822790283542882, -1.55913528760522, -0.116082005832342,     0.35188438750634, -0.980035398123721, 0.346888242274675,     1.20421117207689, -0.792867510761413, 0.0612718021407875,     0.0550197936705209, 1.37920053444169, -0.699595112278956,     -0.111994800282821, -0.518372227686208, 0.0549180222187502,     -0.475844663766105, -0.153289051580788, 1.20253026951676,     -0.492567768673965, -1.97559785192818, -1.2133250429042,     -1.60367350391608, 1.36359098916481, 0.3961646540344, -0.18078264888672,     -0.586666043021248, -1.60315094426458, 0.143210898010546,     -1.0683564354958, 1.58652697799477, 1.50638133483242, 0.895174765777266,     0.21420515346599), A.2 = c(-1.31749491331094, -0.0274910114822239,     -0.347754057592497, -0.0731369401524488, -0.0240838584665583,     -0.857154804563274, -2.76196071154553, 0.371097763253969,     0.163433483265024, -0.30171905621295, 0.76383378413822, -0.249333778237109,     -0.606271029374412, 0.0992854500191257, -0.581665675108691,     -0.28378053597578, -1.80676220431367, 1.53431548491657, -2.29886937441612,     -1.99591576601417, -0.124937516026693, -0.103608169608654,     -0.712546827553852, -0.352162161919822, -0.130583023681016,     -1.27474547917307, 0.541749359234357, -1.13597448041985,     0.0150638844125439, 0.156158197051525, 0.0362785901011277,     0.947063247088099, -0.41406284033421, -0.131662415320732,     -1.42681922928886, -1.7120873360947, 0.671909212932614, 0.182671000023944,     0.315215709181341, -0.150630570545466, 0.978305216902416,     2.32289346984791, 1.07697151627867, 0.722522120906756, 0.450545324879515,     -0.0595912404235378, 1.30633506225077, -0.303151076300435,     -1.82604814623298, -0.275656807836192), A.3 = c(1.02356809432066,     0.358615525632914, 0.474891691850671, -0.639926839093334,     0.670863645516353, 0.753393170617611, -1.45018027327184,     -0.245408076446433, -2.05453875624042, -0.818598602383382,     -1.48384381952996, -0.44152996477205, -1.07573161303059,     -1.7570734962745, 0.327931992559763, -1.1174578350861, -0.884683014038859,     -0.551720723756673, -0.287381241651399, -0.350412417657962,     0.301668087372932, 0.821694926787016, -0.803562942712331,     0.581226483856061, -0.967881751344047, 0.988997176774624,     0.345419299276186, -1.15237273310748, 0.125665980404271,     0.183939102609659, -0.228844305540251, -0.100356663658847,     -1.47351719576572, -0.574945134098843, -0.956381861253519,     -0.894209749636448, -0.234500232593119, -0.274320825845292,     1.27731036285258, -2.13535020143429, 0.911288771000118, 0.675635919436521,     0.513570571305943, -0.0166335211558308, -1.0013759041769,     0.497509179057425, -1.95052889620838, -0.461764945989694,     -0.791512045673544, 1.10094663398126), B.1 = c(-1.42076244045144,     0.537443223247468, 0.221406335414841, -0.185293341722248,     0.270499835223389, 0.482259038833677, -1.22251812747326,     0.839946981044409, -1.64216781474427, 0.488186782277989,     -0.726029085038374, 0.011760796360565, -1.61224675369893,     -0.732854616451234, 0.502934824620935, -0.531793303409868,     1.72014903629104, -2.26192952307069, 0.333755403767623, -0.983828083708662,     0.618361969781919, -0.744860452342839, -1.1305392193163,     0.767130387578949, 1.59496632850079, -1.48864473705503, 0.800280612018164,     -3.14268845736716, 0.054415871019726, -0.946590167964698,     0.893226371127006, -0.412571193735262, 2.77584161246475,     -0.216498163863216, 0.652627391775333, 0.202565790335694,     0.779328582296469, 0.713469748678563, -0.79673731295256,     -0.14750744494286, 0.968827449488177, -0.910124240787513,     -0.500331724876508, -0.671613134473562, -1.37908256177635,     1.8995359321747, 0.180238447607777, -1.20496843305771, -0.95423076338411,     -0.369538712355749), B.2 = c(0.546422149962971, -0.225491384678108,     -0.0870577201917232, 1.35391401242084, -1.57892344250324,     1.07381675178133, 0.407005719702597, -2.91989470380116, -0.29992646040183,     -1.51281608014954, -0.154828861459123, -1.29674363829109,     -0.224931261418355, 0.516877850834978, 0.0304341312663208,     1.85274578069425, 0.688949977802636, -0.687905762763597,     1.13761740793213, -0.977922918833842, 0.362502685366038,     1.38032858382555, -1.16897845114476, 3.18352560277988, -0.206760419811207,     -1.09394147311766, -0.439529252701944, -0.0231056633774306,     1.55269853995557, 0.53787799007007, 0.322922128515472, -0.245985891069853,     -0.434395595930004, 1.43058376694689, 0.0358119507983857,     -0.815056457622175, 1.7442115786181, 0.568886134671572, 1.36149859600274,     2.02366722835854, 0.0781207317138845, 1.23352079500716, 0.897636384943683,     0.685060470310726, 0.141929722990043, 0.096406822282814,     1.06275554725623, -0.0367993368397109, -0.0805543867416027,     0.902679358149746), B.3 = c(-0.289016263065824, 0.0964497208371347,     -0.308306410688526, -0.00968004286634292, -0.674972710922599,     -1.22024404194255, -0.103726531802598, -0.418167411600272,     0.172128535597389, 0.225055960165651, -0.946878619513069,     -0.491057902773052, 0.212963490412072, 0.159915158841372,     -1.42506744942798, 0.916411873662817, -0.370702038154651,     0.235570251169964, -0.799452213908074, 0.589393738247662,     -0.544365299950436, -0.855461147010064, -0.519234670812131,     -0.386233011500345, 0.444315738160471, -0.512301982776863,     0.216522373111052, -2.17238019247438, 0.294113500979096,     1.22915905943622, 0.877068560027936, -0.325620401142022,     -0.862699286371259, 1.02616781612562, -0.548247320199136,     0.639457511043848, -0.122039220283344, -0.262569572534231,     0.872885753350408, -0.508381466514671, 1.42131780262367,     -1.08876071376561, -0.929659739511179, -1.06505567613805,     1.65526393724974, 1.40789528920272, 0.0482380770327222, 1.576812904565,     0.886041209501155, 1.67833210314168), C = c(0.249281432094083,     -0.75060678066242, 1.96663829512242, 1.07311148307014, -1.16001391612946,     0.982805970577121, -0.276075525739205, -1.51361904811996,     -0.76025192204235, -0.0711963381202679, -0.509744604107613,     1.14773027660966, -0.214966662501713, -1.30021509025219,     0.26771685449963, 1.52729008845314, 0.16152592014128, -1.16163993121707,     -0.728271579346013, 2.18552316158483, 3.2267202172321, -0.371951911411176,     0.22679218579872, -0.454673354209094, -1.03221271591231,     0.000525801209002479, -0.779950861575267, -0.212275730508459,     -0.958723955391695, 0.597603022712479, 0.783059485690152,     -0.672838314923719, 0.270569482965948, -1.17175599797298,     1.3481819702168, 0.092048082587236, -0.112300470624316, -0.0351407243172009,     -1.02822522093621, 0.132709699770899, 0.0274134925249082,     -0.0628956586925408, 1.22330274534736, 0.147546292423498,     -0.560545777675606, 2.48324625259345, 1.06539341377306, -0.552199826153987,     -0.830758992188295, 0.981101996664634), D.3 = c(-0.30738769311648,     -0.0955324035182882, 0.662494977483462, -0.32917475187351,     0.473669672767266, -1.70988262190955, -0.258874498378276,     1.30260677697798, 0.0831171435737973, -0.725151535273343,     0.286998990891197, 1.65587312988346, 0.74502356753718, -0.132896755189255,     0.270454986983101, -0.58487644410089, -0.358568514873609,     -0.442558820091117, -0.646375870404918, 1.55753640133878,     -0.721094010853735, -1.13103345214381, 0.349307534821004,     -0.00269156806152906, -1.97645429805896, 0.205281018728229,     -0.756567391932105, 0.562973338363062, -0.849666998843892,     -0.890717062226568, -1.50408723711045, 0.0662914063775097,     -0.943575624942303, 0.969699333341892, -0.816848924743825,     1.92697670527135, -0.709455478880648, -0.832222042530713,     1.62281245312759, 0.253654812080715, 1.21565015410517, 0.758201212677142,     -0.568645400017312, 0.807764912685715, 0.708118683432948,     1.07664847964614, 0.856064311192142, -0.375098606051673,     -0.414043294840855, 2.34490588705741), A.Infancy = c(-1.03252831484715,     0.810286387189818, 0.303680136191481, 0.606928104438201,     -0.0691751742144826, 0.623931120973788, -1.29290733882822,     0.455099807076384, 0.402676636831372, 0.228581638291094,     0.0775530734986935, -0.483284525225886, -0.300047088112003,     -0.146098260294893, -0.148754797018431, -0.210357595877644,     -0.160652555570835, 1.17855288422973, -1.92900233101067,     -1.05599888592326, 0.113473435739823, -0.541821783866188,     -0.182829292639588, 0.426024505078534, -0.461725267221215,     -0.606736838516141, 0.298384576452439, 0.12161302701092,     -0.342265613933206, 0.022081698384352, -0.24104681879254,     0.500990634653425, -0.444953752050158, -0.14247573345076,     -0.11214447988605, -1.10232755238433, -0.651844319497783,     -0.515327021440128, -0.64422889736737, 0.606480209309672,     0.687234935468408, 1.07105541048059, 0.245152736628711, -0.440314411678912,     0.296878111445031, -0.563973837959669, 1.44643102012277,     0.601615129265992, -0.465436690227857, -0.030725827185101    ), A.Toddlerhood = c(1.02356809432066, 0.358615525632914,     0.474891691850671, -0.639926839093334, 0.670863645516353,     0.753393170617611, -1.45018027327184, -0.245408076446433,     -2.05453875624042, -0.818598602383382, -1.48384381952996,     -0.44152996477205, -1.07573161303059, -1.7570734962745, 0.327931992559763,     -1.1174578350861, -0.884683014038859, -0.551720723756673,     -0.287381241651399, -0.350412417657962, 0.301668087372932,     0.821694926787016, -0.803562942712331, 0.581226483856061,     -0.967881751344047, 0.988997176774624, 0.345419299276186,     -1.15237273310748, 0.125665980404271, 0.183939102609659,     -0.228844305540251, -0.100356663658847, -1.47351719576572,     -0.574945134098843, -0.956381861253519, -0.894209749636448,     -0.234500232593119, -0.274320825845292, 1.27731036285258,     -2.13535020143429, 0.911288771000118, 0.675635919436521,     0.513570571305943, -0.0166335211558308, -1.0013759041769,     0.497509179057425, -1.95052889620838, -0.461764945989694,     -0.791512045673544, 1.10094663398126)), row.names = c("1", "2", "3", "4", "5", "6", "7", "8", "9", "10", "11", "12", "13", "14", "15", "16", "17", "18", "19", "20", "21", "22", "23", "24", "25", "26", "27", "28", "29", "30", "31", "32", "33", "34", "35", "36", "37", "38", "39", "40", "41", "42", "43", "44", "45", "46", "47", "48", "49", "50"), class = "data.frame"), na.action = function (object,     ...) {    if (NCOL(object) == 0)         object    else {        ok <- complete.cases(object)        if (all(ok))             object        else stop(message)    }}): object is not a matrix