Imputes dataset so there is no missing at each time point using parallel processing to speed up

Imputes dataset so there is no missing at each time point using parallel processing to speed up

imputeData(
  data,
  m = 5,
  method = "rf",
  home_dir = NA,
  exposure,
  outcome,
  para_proc = TRUE,
  read_imps_from_file = "no",
  save.out = TRUE
)

Arguments

data

data in wide format

m

(optional) integer number of imputed datasets (default is 5)

method

(optional) character string of imputation method from mice() (default is random forest "rf")

home_dir

path to home directory

exposure

name of exposure variable

outcome

name of outcome variable with ".timepoint" suffix

para_proc

(optional) TRUE/FALSE whether to do parallel processing using multiple cores to speed up process (default = TRUE)

read_imps_from_file

(optional) "yes" or "no" indicatorto read in weights that have been previously run and saved locally (default is "no")

save.out

(optional) TRUE or FALSE indicator to save output and intermediary output locally (default is TRUE)

Value

mice object of m imputed datasets

See also

Examples

test <- data.frame(ID = 1:50,
                   A.1 = rnorm(n = 50),
                   A.2 = rnorm(n = 50),
                   A.3 = rnorm(n = 50),
                   B.1 = rnorm(n = 50),
                   B.2 = rnorm(n = 50),
                   B.3 = rnorm(n = 50),
                   C = rnorm(n = 50),
                   D.3 = rnorm(n = 50))
test[, c("A.1", "A.2", "A.3")] <- lapply(test[, c("A.1", "A.2", "A.3")], as.numeric)

test_miss <- missMethods::delete_MAR_1_to_x(as.data.frame(test), p = 0.20,
                                            cols_mis = c("A.1", "B.2", "C"),
                                            cols_ctrl = c("B.1", "B.1", "B.1"), 3)
test_i <- imputeData(data = test_miss,
                     m = 3,
                     method = "rf",
                     exposure = "A",
                     outcome = "D.3",
                     para_proc = TRUE,
                     read_imps_from_file = "no",
                     save.out = FALSE)
#> Creating 3 imputed datasets using the rf imputation method in mice. This may take some time to run.
#> ### Using 2 cores
#> ### Using doParallelMC as the backend
#> Loading required package: foreach
#> Loading required package: rngtools
#> USER ALERT: Please view any logged events from the imputation below: 
#> Table: Logged Events from mice
#> 
#>