Inspect long/wide/imputed data

Inspect long/wide/imputed data

inspectData(
  data,
  home_dir,
  exposure,
  exposure_time_pts,
  outcome,
  tv_confounders,
  ti_confounders,
  epochs = NULL,
  hi_lo_cut = NULL,
  reference = NA,
  comparison = NULL,
  verbose = TRUE,
  save.out = TRUE
)

Arguments

data

data in wide format as: a data frame, list of imputed data frames, or mids object

home_dir

(optional) path to home directory (required if save.out = TRUE)

exposure

name of exposure variable

exposure_time_pts

list of integers at which weights will be created/assessed that correspond to time points when exposure was measured

outcome

name of outcome variable with ".timepoint" suffix

tv_confounders

list of time-varying confounders with ".timepoint" suffix

ti_confounders

list of time invariant confounders

epochs

(optional) data frame of exposure epoch labels and values

hi_lo_cut

(optional) list of two numbers indicating quantile values that reflect high and low values, respectively, for continuous exposure (default is median split)

reference

(optional)string of "-"-separated "l" and "h" values indicative of a reference exposure history to which to compare comparison, required if comparison is specified

comparison

(optional)list of one or more strings of "-"-separated "l" and "h" values indicative of comparison history/histories to compare to reference, required if reference is specified

verbose

(optional) TRUE or FALSE indicator for user output (default is TRUE)

save.out

(optional) TRUE or FALSE indicator to save output and intermediary output locally (default is TRUE)

Value

none

Examples

test <- data.frame(ID = 1:50,
                   A.1 = rnorm(n = 50),
                   A.2 = rnorm(n = 50),
                   A.3 = rnorm(n = 50),
                   B.1 = rnorm(n = 50),
                   B.2 = rnorm(n = 50),
                   B.3 = rnorm(n = 50),
                   C = rnorm(n = 50),
                   D.3 = rnorm(n = 50))
test[, c("A.1", "A.2", "A.3")] <- lapply(test[, c("A.1", "A.2", "A.3")], as.numeric)

inspectData(data = test,
            exposure = "A",
            exposure_time_pts = c(1, 2, 3),
            outcome = "D.3",
            tv_confounders = c("A.1", "A.2", "A.3", "B.1", "B.2", "B.3"),
            ti_confounders = "C",
            save.out = FALSE)
#> 
#> The following variables are designated as numeric: 
#> [1] "A.1, A.2, A.3, B.1, B.2, B.3, C, D.3"
#> 
#> The following variables are designated as factors: 
#> [1] ""
#> 
#> *temp: please inspect the levels of your factors below. at present, excluding ID, the code can only accept 2-level factors. set rest to numeric 
#> named list()
#> Table: Other variable types
#> 
#> |   |variable |type    |
#> |:--|:--------|:-------|
#> |ID |ID       |integer |
#> 
#> Error in loadNamespace(x): there is no package called ‘psych’
inspectData(data = test,
            exposure = "A",
            exposure_time_pts = c(1, 2, 3),
            outcome = "D.3",
            tv_confounders = c("A.1", "A.2", "A.3", "B.1", "B.2", "B.3"),
            ti_confounders = "C",
            hi_lo_cut = c(0.8, 0.2),
            save.out = FALSE)
#> 
#> The following variables are designated as numeric: 
#> [1] "A.1, A.2, A.3, B.1, B.2, B.3, C, D.3"
#> 
#> The following variables are designated as factors: 
#> [1] ""
#> 
#> *temp: please inspect the levels of your factors below. at present, excluding ID, the code can only accept 2-level factors. set rest to numeric 
#> named list()
#> Table: Other variable types
#> 
#> |   |variable |type    |
#> |:--|:--------|:-------|
#> |ID |ID       |integer |
#> 
#> Error in loadNamespace(x): there is no package called ‘psych’
inspectData(data = test,
            exposure = "A",
            exposure_time_pts = c(1, 2, 3),
            outcome = "D.3",
            tv_confounders = c("A.1", "A.2", "A.3", "B.1", "B.2", "B.3"),
            ti_confounders = "C",
            hi_lo_cut = c(0.8, 0.2),
            reference = "l-l-l",
            comparison = "h-h-h",
            save.out = FALSE)
#> 
#> The following variables are designated as numeric: 
#> [1] "A.1, A.2, A.3, B.1, B.2, B.3, C, D.3"
#> 
#> The following variables are designated as factors: 
#> [1] ""
#> 
#> *temp: please inspect the levels of your factors below. at present, excluding ID, the code can only accept 2-level factors. set rest to numeric 
#> named list()
#> Table: Other variable types
#> 
#> |   |variable |type    |
#> |:--|:--------|:-------|
#> |ID |ID       |integer |
#> 
#> Error in loadNamespace(x): there is no package called ‘psych’
inspectData(data = test,
            exposure = "A",
            exposure_time_pts = c(1, 2, 3),
            outcome = "D.3",
            tv_confounders = c("A.1", "A.2", "A.3", "B.1", "B.2", "B.3"),
            ti_confounders = "C",
            epochs = data.frame(epochs = c("Infancy", "Toddlerhood"),
                                values = I(list(c(1, 2), c(3)))),
            save.out = FALSE)
#> 
#> The following variables are designated as numeric: 
#> [1] "A.1, A.2, A.3, B.1, B.2, B.3, C, D.3"
#> 
#> The following variables are designated as factors: 
#> [1] ""
#> 
#> *temp: please inspect the levels of your factors below. at present, excluding ID, the code can only accept 2-level factors. set rest to numeric 
#> named list()
#> Table: Other variable types
#> 
#> |   |variable |type    |
#> |:--|:--------|:-------|
#> |ID |ID       |integer |
#> 
#> Error in loadNamespace(x): there is no package called ‘psych’