Create love plots showing balancing statistics

Create love plots showing balancing statistics

make_love_plot(
  home_dir,
  folder,
  exposure,
  exposure_time_pt,
  exposure_type,
  k = 0,
  form_name,
  balance_stats,
  data_type,
  balance_thresh,
  weights_method,
  imp_conf,
  verbose,
  save.out
)

Arguments

home_dir

path to home directory

folder

folder path for saving

exposure

name of exposure variable

exposure_time_pt

exposure time point integer

exposure_type

character indicating binary or continuous exposure type

k

imputation number

form_name

formula name

balance_stats

data frame of balance statistics

data_type

single or imputed data type

balance_thresh

one or two numbers between 0 and 1 indicating a single balancing threshold or thresholds for more and less important confounders, respectively

weights_method

method character string of WeightItMSM() balancing method abbreviation

imp_conf

list of variable names reflecting important confounders

verbose

TRUE or FALSE indicator for user output (default is TRUE)

save.out

TRUE or FALSE indicator to save output and intermediary output locally

Value

none

Examples

f <- createFormulas(exposure = "A",
                    exposure_time_pts = c(1, 2, 3),
                    outcome = "D.3",
                    tv_confounders = c("A.1", "A.2", "A.3", "B.1", "B.2", "B.3"),
                    ti_confounders = "C",
                    type = "full",
                    save.out = FALSE)
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A - D.3 at A time point 1 is: 
#> A.1 ~ C
#> <environment: 0x555708bb58a0>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A - D.3 at A time point 2 is: 
#> A.2 ~ A.1 + B.1 + C
#> <environment: 0x555708bb58a0>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A - D.3 at A time point 3 is: 
#> A.3 ~ A.1 + A.2 + B.1 + B.2 + C
#> <environment: 0x555708bb58a0>
#> 

test <- data.frame(ID = 1:50,
                   A.1 = rnorm(n = 50),
                   A.2 = rnorm(n = 50),
                   A.3 = rnorm(n = 50),
                   B.1 = rnorm(n = 50),
                   B.2 = rnorm(n = 50),
                   B.3 = rnorm(n = 50),
                   C = rnorm(n = 50),
                   D.3 = rnorm(n = 50))
test[, c("A.1", "A.2", "A.3")] <- lapply(test[, c("A.1", "A.2", "A.3")], as.numeric)

w <- createWeights(data = test,
                   exposure = "A",
                   outcome = "D.3",
                   formulas = f,
                   save.out = FALSE)
#> Error in weightit2cbps.cont(covs = covs, treat = treat, subset = by.factor ==     i, s.weights = s.weights, missing = missing, verbose = verbose,     ...): The package "CBPS" is required.

b <- assessBalance(data = test,
                   exposure = "A",
                   exposure_time_pts = c(1, 2, 3),
                   outcome = "D.3",
                   type = "weighted",
                   weights = w,
                   formulas = f,
                   save.out = FALSE)
#> Error in eval(expr, envir, enclos): object 'w' not found

p <- make_love_plot(folder = "prebalance/",
                    exposure = "A",
                    exposure_time_pt = 1,
                    exposure_type = "continuous",
                    form_name = "form_name",
                    balance_stats = b,
                    data_type = "single",
                    balance_thresh = 0.1,
                    imp_conf = NULL,
                    weights_method = w[[1]]$method,
                    save.out = FALSE,
                    verbose = TRUE)
#> Error in eval(expr, envir, enclos): object 'b' not found
p <- make_love_plot(folder = "weighted/",
                    exposure = "A",
                    exposure_time_pt = 2,
                    exposure_type = "continuous",
                    form_name = "form_name",
                    balance_stats = b,
                    data_type = "single",
                    balance_thresh = c(0.05, 0.1),
                    imp_conf = "A.2",
                    weights_method = w[[1]]$method,
                    save.out = FALSE,
                    verbose = TRUE)
#> Error in eval(expr, envir, enclos): object 'b' not found