Wide complete data (continuous exposure) These data are simulated based on data from the Family Life Project (FLP), a longitudinal study following 1,292 families representative of two geographic areas (three counties in North Carolina and three counties in Pennsylvania) with high rural child poverty (Vernon-Feagans et al., 2013; Burchinal et al., 2008). These data contain economic strain (ESEATA1) as a continuously distributed variable.

Wide complete data (continuous exposure) These data are simulated based on data from the Family Life Project (FLP), a longitudinal study following 1,292 families representative of two geographic areas (three counties in North Carolina and three counties in Pennsylvania) with high rural child poverty (Vernon-Feagans et al., 2013; Burchinal et al., 2008). These data contain economic strain (ESEATA1) as a continuously distributed variable.

Usage

sim_data_wide

Format

A wide data frame of 1,292 observations There are 36 measured variables.

"ID" person identifier

"WAVE" age (in months) when data were collected

"ESETA1" is the continuous exposure of economic strain

"StrDif_Tot.58" is the continuous outcome of behavioral problems

"InRatioCor" is the income-to-needs ratio

"PmEd2" is the parent's education level

"state" is the family's state of residence

"TcBlac2" is the family's race (1 = Black?, 0 = White?)

"bioDadInHH2" is whether the biological father lives with the family (insert coding)

"HomeOwnd" indicator of whether family owns home (insert coding)

"KFASTScr"

"PmBlac2" primary careigver race (insert coding)

"PmAge2"

"PmMrSt2"

"RMomAgeU"

"RHealth"

"RHasSO"

"SmokTotl"

"caregiv_health"

"peri_health"

"SWghtLB"

"SurpPreg"

"DrnkFreq"

"gov_assist"

"ALI_LE"

"B18Raw"

"CORTB"

"EARS_TJo"

"fscore"

"HOMEETA1"

"IBRAttn"

"LESMnNeg"

"LESMnPos"

"MDI"

"RHAsSO"

"SAAmylase"

"WndNbrhood"

References

DeJoseph, M. L., Sifre, R. D., Raver, C. C., Blair, C. B., & Berry, D. (2021). Capturing Environmental Dimensions of Adversity and Resources in the Context of Poverty Across Infancy Through Early Adolescence: A Moderated Nonlinear Factor Model. Child Development, n/a(n/a). https://doi.org/10.1111/cdev.13504

Burchinal, M., Howes, C., Pianta, R., Bryant, D., Early, D., Clifford, R., & Barbarin, O. (2008). Predicting Child Outcomes at the End of Kindergarten from the Quality of Pre-Kindergarten Teacher–Child Interactions and Instruction. Applied Developmental Science, 12(3), 140–153. https://doi.org/10.1080/10888690802199418

Vernon-Feagans, L., Cox, M., Willoughby, M., Burchinal, M., Garrett-Peters, P., Mills-Koonce, R., Garrett-Peiers, P., Conger, R. D., & Bauer, P. J. (2013). The Family Life Project: An Epidemiological and Developmental Study of Young Children Living in Poor Rural Communities. Monographs of the Society for Research in Child Development, 78(5), i–150.