Trim IPTW balancing weights

Trims IPTW balancing weights with heavy right tails by populating all weight values above a given quantile with the weight value of that quantile.

trimWeights(
  home_dir,
  exposure,
  outcome,
  weights,
  quantile = 0.95,
  verbose = TRUE,
  save.out = TRUE
)

Arguments

home_dir

path to home directory

exposure

name of exposure variable

outcome

name of outcome variable with ".timepoint" suffix

weights

list of IPTW weights output from createWeights()

quantile

(optional) numeric value between 0 and 1 of quantile value at which to trim weights (default is 0.95)

verbose

(optional) TRUE or FALSE indicator for user output (default is TRUE)

save.out

(optional) TRUE or FALSE indicator to save output and intermediary output locally (default is TRUE)

Value

list of model output with trimmed weights

See also

Examples

f <- createFormulas(exposure = "A",
                    exposure_time_pts = c(1, 2, 3),
                    outcome = "D.3",
                    tv_confounders = c("B.1", "B.2", "B.3"),
                    ti_confounders = "C",
                    type = "full",
                    save.out = FALSE)
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A-D.3 at A time point 1 is: 
#> A.1 ~ C
#> <environment: 0x55b1c5293ad8>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A-D.3 at A time point 2 is: 
#> A.2 ~ B.1 + C
#> <environment: 0x55b1c5293ad8>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A-D.3 at A time point 3 is: 
#> A.3 ~ B.1 + B.2 + C
#> <environment: 0x55b1c5293ad8>
#> 

test <- data.frame(ID = 1:50,
                   A.1 = rnorm(n = 50),
                   A.2 = rnorm(n = 50),
                   A.3 = rnorm(n = 50),
                   B.1 = rnorm(n = 50),
                   B.2 = rnorm(n = 50),
                   B.3 = rnorm(n = 50),
                   C = rnorm(n = 50),
                   D.3 = rnorm(n = 50))
test[, c("A.1", "A.2", "A.3")] <- lapply(test[, c("A.1", "A.2", "A.3")], as.numeric)

w <- createWeights(data = test,
                   exposure = "A",
                   outcome = "D.3",
                   formulas = f,
                   save.out = FALSE)
#> Error in weightit2cbps.cont(covs = covs, treat = treat, subset = by.factor ==     i, s.weights = s.weights, missing = missing, verbose = verbose,     ...): The package "CBPS" is required.

t <- trimWeights(exposure = "A",
                 outcome = "D.3",
                 weights = w,
                 save.out = FALSE)
#> Error in eval(expr, envir, enclos): object 'w' not found
t <- trimWeights(exposure = "A",
                 outcome = "D.3",
                 weights = w,
                 quantile = 0.75,
                 save.out = FALSE)
#> Error in eval(expr, envir, enclos): object 'w' not found