Estimate, compare, and visualize exposure histories

Takes fitted model output to created predicted values for user-specified histories (pooling for imputed data), before conducting contrast comparisons (pooling for imputed data), correcting for multiple comparisons, and then plotting results.

compareHistories(
  home_dir,
  exposure,
  exposure_time_pts,
  outcome,
  tv_confounders,
  model,
  epochs = NULL,
  hi_lo_cut = NULL,
  reference = NA,
  comparison = NULL,
  mc_comp_method = "BH",
  dose_level = "h",
  exp_lab = NA,
  out_lab = NA,
  colors = "Dark2",
  verbose = TRUE,
  save.out = TRUE
)

Arguments

home_dir

path to home directory (required if 'save.out' = TRUE)

exposure

name of exposure variable

exposure_time_pts

list of integers at which weights will be created/assessed that correspond to time points when exposure wass measured

outcome

name of outcome variable with ".timepoint" suffix

tv_confounders

list of time-varying confounders with ".timepoint" suffix

model

list of model outputs from fitModel()

epochs

(optional) data frame of exposure epoch labels and values

hi_lo_cut

(optional) list of two numbers indicating quantile values that reflect high and low values, respectively, for continuous exposure (default is median split)

reference

(optional) string of "-"-separated "l" and "h" values indicative of a reference exposure history to which to compare comparison, required if comparison is supplied

comparison

(optional) list of one or more strings of "-"-separated "l" and "h" values indicative of comparison history/histories to compare to reference, required if reference is supplied

mc_comp_method

(optional) character abbreviation for multiple comparison correction method for stats::p.adjust, default is Benjamini-Hochburg ("BH")

dose_level

(optional) "l" or "h" indicating whether low or high doses should be tallied in tables and plots (default is high "h")

exp_lab

(optional) character label for exposure variable in plots (default is variable name)

out_lab

(optional) character label for outcome variable in plots (default is variable name)

colors

(optional) character specifying Brewer palette or list of colors (n(epochs)+1) for plotting (default is "Dark2" palette)

verbose

(optional) TRUE or FALSE indicator for user output (default is TRUE)

save.out

(optional) TRUE or FALSE indicator to save output and intermediary output locally (default is TRUE)

Value

data frame of history comparisons

See also

Examples

f <- createFormulas(exposure = "A",
                    exposure_time_pts = c(1, 2, 3),
                    outcome = "D.3",
                    tv_confounders = c("A.1", "A.2", "A.3", "B.1", "B.2", "B.3"),
                    ti_confounders = "C",
                    type = "full",
                    save.out = FALSE)
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A-D.3 at A time point 1 is: 
#> A.1 ~ C
#> <environment: 0x55e7860f21e0>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A-D.3 at A time point 2 is: 
#> A.2 ~ A.1 + B.1 + C
#> <environment: 0x55e7860f21e0>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A-D.3 at A time point 3 is: 
#> A.3 ~ A.1 + A.2 + B.1 + B.2 + C
#> <environment: 0x55e7860f21e0>
#> 

test <- data.frame(ID = 1:50,
                   A.1 = rnorm(n = 50),
                   A.2 = rnorm(n = 50),
                   A.3 = rnorm(n = 50),
                   B.1 = rnorm(n = 50),
                   B.2 = rnorm(n = 50),
                   B.3 = rnorm(n = 50),
                   C = rnorm(n = 50),
                   D.3 = rnorm(n = 50))
test[, c("A.1", "A.2", "A.3")] <- lapply(test[, c("A.1", "A.2", "A.3")], as.numeric)

w <- createWeights(data = test,
                   exposure = "A",
                   outcome = "D.3",
                   formulas = f,
                   save.out = FALSE)
#> Error in weightit2cbps.cont(covs = covs, treat = treat, subset = by.factor ==     i, s.weights = s.weights, missing = missing, verbose = verbose,     ...): The package "CBPS" is required.

m <- fitModel(data = test,
              weights = w,
              exposure = "A",
              exposure_time_pts = c(1, 2, 3),
              outcome = "D.3",
              model = "m0",
              save.out = FALSE)
#> Error in eval(expr, envir, enclos): object 'w' not found

r <- compareHistories(exposure = "A",
                      exposure_time_pts = c(1, 2, 3),
                      outcome = "D.3",
                      tv_confounders = c("A.1", "A.2", "A.3", "B.1", "B.2", "B.3"),
                      model = m,
                      save.out = FALSE)
#> Error in eval(expr, envir, enclos): object 'm' not found
r <- compareHistories(exposure = "A",
                      exposure_time_pts = c(1, 2, 3),
                      outcome = "D.3",
                      tv_confounders = c("A.1", "A.2", "A.3", "B.1", "B.2", "B.3"),
                      model = m,
                      reference = "l-l-l",
                      comparison = "h-h-h",
                      save.out = FALSE)
#> Error in eval(expr, envir, enclos): object 'm' not found
r <- compareHistories(exposure = "A",
                      exposure_time_pts = c(1, 2, 3),
                      outcome = "D.3",
                      tv_confounders = c("A.1", "A.2", "A.3", "B.1", "B.2", "B.3"),
                      model = m,
                      reference = "l-l-l",
                      comparison = c("h-h-h", "h-l-l"),
                      save.out = FALSE)
#> Error in eval(expr, envir, enclos): object 'm' not found