Fit outcome model

Fits weighted marginal outcome model as a generalized linear model of the user's choosing, relating exposure main effects to outcome using IPTW weights.

Usage

fitModel(
  data,
  obj,
  weights,
  outcome,
  model = c("m0", "m1", "m2", "m3"),
  int_order = NA,
  covariates = NULL,
  family = NULL,
  link = NA,
  verbose = FALSE,
  save.out = FALSE
)

# S3 method for devMSM_models
print(x, i = 1, save.out = FALSE, ...)

Arguments

data

data in wide format as: a data frame, list of imputed data frames, or mids object from the mice package

obj

initialized MSM object from initMSM()

weights

list of IPTW weights output from createWeights()

outcome

name of outcome variable with ".timepoint" suffix. See initMSM() for details on suffix

model

character indicating one of the following outcome models:

"m0" (exposure main effects)

"m1" (exposure main effects & covariates)

"m2" (exposure main effects & their interactions)

"m3" (exposure main effects, their interactions, & covariates)

int_order

integer specification of highest order exposure main effects interaction, required for interaction models ("m2", "m3")

covariates

list of characters reflecting variable names of covariates, required for covariate models ("m1", "m3")

family

(optional) family function specification for WeightIt::glm_weightit() model

link

(optional) character link function specification for WeightIt::glm_weightit() model

verbose

(optional) TRUE or FALSE indicator for printing output to console. default is FALSE.

save.out

(optional) Either logical or a character string. If TRUE, it will output the result to a default file name within home_dir set in initMSM(). You can load the data with x <- readRDS(file). To use a non-default file name, specify a character string with the file name. It will save relative to home_dir. There might be naming conflicts where two objects get saved to the same file. In these cases, users should specify a custom name. default is FALSE.

x

devMSM_models object from fitModel

i

For multiply imputed datset, i selects which imputation to print results for. Default is i = 1. With i = TRUE, all imputed datasets will be looped over. With i = NULL, will average over all imputed datasets and summarize that.

...

ignored

Value

list containing WeightIt::glm_weightit() model output. It is the length of the number of datasets (1 for a data.frame or the number of imputed datasets)

See also

Examples

library(devMSMs)
data <- data.frame(
  ID = 1:50,
  A.1 = rnorm(n = 50),
  A.2 = rnorm(n = 50),
  A.3 = rnorm(n = 50),
  B.1 = rnorm(n = 50),
  B.2 = rnorm(n = 50),
  B.3 = rnorm(n = 50),
  C = rnorm(n = 50),
  D.3 = rnorm(n = 50)
)
obj <- initMSM(
  data,
  exposure = c("A.1", "A.2", "A.3"),
  ti_conf = c("C"),
  tv_conf = c("B.1", "B.2", "B.3", "D.3")
)
f <- createFormulas(obj, type = "short")
w <- createWeights(data = data, obj = obj, formulas = f)

fit_m0 <- fitModel(
  data = data, obj = obj, weights = w, 
  outcome = "D.3", model = "m0"
)
print(fit_m0)
#> Please inspect the following likelihood ratio test to determine if the exposures collective predict significant variation in the outcome compared to a model without exposure terms.
#> We strongly suggest only conducting history comparisons if the likelihood ratio test is significant.
#> Analysis of Deviance Table
#> 
#> Model 1: D.3 ~ A.1 + A.2 + A.3
#> Model 2: D.3 ~ 1
#>   Resid. Df Resid. Dev Df Deviance      F Pr(>F)
#> 1        46     42.114                          
#> 2        49     43.992 -3  -1.8776 0.6836 0.5666
#> 
#> 
#> The marginal model, m0, is summarized below:
#> +-------------+--------+-----------------+-------+
#> |             | (1)                              |
#> +-------------+--------+-----------------+-------+
#> |             | Est.   | CI              | p     |
#> +=============+========+=================+=======+
#> | (Intercept) | -0.056 | [-0.326, 0.215] | 0.687 |
#> +-------------+--------+-----------------+-------+
#> | A.1         | 0.207  | [-0.044, 0.458] | 0.106 |
#> +-------------+--------+-----------------+-------+
#> | A.2         | -0.081 | [-0.358, 0.196] | 0.568 |
#> +-------------+--------+-----------------+-------+
#> | A.3         | -0.050 | [-0.371, 0.271] | 0.759 |
#> +-------------+--------+-----------------+-------+
#> | Num.Obs.    | 50     |                 |       |
#> +-------------+--------+-----------------+-------+
#> | R2          | 0.029  |                 |       |
#> +-------------+--------+-----------------+-------+ 

fit_m1 <- fitModel(
  data = data, obj = obj, weights = w, 
  outcome = "D.3", model = "m1", 
  covariates = c("C")
)
print(fit_m1)
#> Please inspect the following likelihood ratio test to determine if the exposures collective predict significant variation in the outcome compared to a model without exposure terms.
#> We strongly suggest only conducting history comparisons if the likelihood ratio test is significant.
#> Analysis of Deviance Table
#> 
#> Model 1: D.3 ~ A.1 + A.2 + A.3 + C
#> Model 2: D.3 ~ C
#>   Resid. Df Resid. Dev Df Deviance      F Pr(>F)
#> 1        45     39.728                          
#> 2        48     41.355 -3  -1.6266 0.6141 0.6094
#> 
#> 
#> The marginal model, m1, is summarized below:
#> +-------------+--------+-----------------+-------+
#> |             | (1)                              |
#> +-------------+--------+-----------------+-------+
#> |             | Est.   | CI              | p     |
#> +=============+========+=================+=======+
#> | (Intercept) | -0.050 | [-0.306, 0.207] | 0.703 |
#> +-------------+--------+-----------------+-------+
#> | A.1         | 0.180  | [-0.074, 0.434] | 0.165 |
#> +-------------+--------+-----------------+-------+
#> | A.2         | -0.102 | [-0.362, 0.159] | 0.444 |
#> +-------------+--------+-----------------+-------+
#> | A.3         | -0.034 | [-0.344, 0.275] | 0.827 |
#> +-------------+--------+-----------------+-------+
#> | C           | -0.215 | [-0.440, 0.011] | 0.062 |
#> +-------------+--------+-----------------+-------+
#> | Num.Obs.    | 50     |                 |       |
#> +-------------+--------+-----------------+-------+
#> | R2          | 0.058  |                 |       |
#> +-------------+--------+-----------------+-------+ 

fit_m2 <- fitModel(
  data = data, obj = obj, weights = w, 
  outcome = "D.3", model = "m2", 
  int_order = 2
)
print(fit_m2)
#> Please inspect the following likelihood ratio test to determine if the exposures collective predict significant variation in the outcome compared to a model without exposure terms.
#> We strongly suggest only conducting history comparisons if the likelihood ratio test is significant.
#> Analysis of Deviance Table
#> 
#> Model 1: D.3 ~ A.1 + A.2 + A.3 + A.1 + A.2 + A.3 + A.1:A.2 + A.1:A.3 + 
#>     A.2:A.3
#> Model 2: D.3 ~ 1
#>   Resid. Df Resid. Dev Df Deviance     F Pr(>F)
#> 1        43     39.584                         
#> 2        49     43.992 -6  -4.4077 0.798 0.5768
#> 
#> 
#> The marginal model, m2, is summarized below:
#> +-------------+--------+-----------------+-------+
#> |             | (1)                              |
#> +-------------+--------+-----------------+-------+
#> |             | Est.   | CI              | p     |
#> +=============+========+=================+=======+
#> | (Intercept) | -0.030 | [-0.284, 0.225] | 0.820 |
#> +-------------+--------+-----------------+-------+
#> | A.1         | 0.072  | [-0.261, 0.404] | 0.672 |
#> +-------------+--------+-----------------+-------+
#> | A.2         | -0.155 | [-0.426, 0.116] | 0.263 |
#> +-------------+--------+-----------------+-------+
#> | A.3         | -0.187 | [-0.528, 0.154] | 0.282 |
#> +-------------+--------+-----------------+-------+
#> | A.1 × A.2   | 0.172  | [-0.095, 0.439] | 0.207 |
#> +-------------+--------+-----------------+-------+
#> | A.1 × A.3   | 0.213  | [-0.206, 0.632] | 0.319 |
#> +-------------+--------+-----------------+-------+
#> | A.2 × A.3   | 0.194  | [-0.128, 0.516] | 0.237 |
#> +-------------+--------+-----------------+-------+
#> | Num.Obs.    | 50     |                 |       |
#> +-------------+--------+-----------------+-------+
#> | R2          | 0.071  |                 |       |
#> +-------------+--------+-----------------+-------+ 

fit_m3 <- fitModel(
  data = data, obj = obj, weights = w, 
  outcome = "D.3", model = "m3",
  int_order = 2, covariates = c("C")
)
print(fit_m3)
#> Please inspect the following likelihood ratio test to determine if the exposures collective predict significant variation in the outcome compared to a model without exposure terms.
#> We strongly suggest only conducting history comparisons if the likelihood ratio test is significant.
#> Analysis of Deviance Table
#> 
#> Model 1: D.3 ~ A.1 + A.2 + A.3 + A.1 + A.2 + A.3 + A.1:A.2 + A.1:A.3 + 
#>     A.2:A.3 + C
#> Model 2: D.3 ~ C
#>   Resid. Df Resid. Dev Df Deviance      F Pr(>F)
#> 1        42     36.455                          
#> 2        48     41.355 -6  -4.8999 0.9409 0.4763
#> 
#> 
#> The marginal model, m3, is summarized below:
#> +-------------+--------+------------------+-------+
#> |             | (1)                               |
#> +-------------+--------+------------------+-------+
#> |             | Est.   | CI               | p     |
#> +=============+========+==================+=======+
#> | (Intercept) | -0.023 | [-0.263, 0.217]  | 0.852 |
#> +-------------+--------+------------------+-------+
#> | A.1         | 0.043  | [-0.295, 0.381]  | 0.805 |
#> +-------------+--------+------------------+-------+
#> | A.2         | -0.182 | [-0.426, 0.061]  | 0.142 |
#> +-------------+--------+------------------+-------+
#> | A.3         | -0.202 | [-0.519, 0.115]  | 0.212 |
#> +-------------+--------+------------------+-------+
#> | C           | -0.251 | [-0.465, -0.036] | 0.022 |
#> +-------------+--------+------------------+-------+
#> | A.1 × A.2   | 0.155  | [-0.105, 0.415]  | 0.243 |
#> +-------------+--------+------------------+-------+
#> | A.1 × A.3   | 0.237  | [-0.173, 0.647]  | 0.257 |
#> +-------------+--------+------------------+-------+
#> | A.2 × A.3   | 0.261  | [-0.052, 0.574]  | 0.102 |
#> +-------------+--------+------------------+-------+
#> | Num.Obs.    | 50     |                  |       |
#> +-------------+--------+------------------+-------+
#> | R2          | 0.120  |                  |       |
#> +-------------+--------+------------------+-------+