Fit outcome model

Fits weighted marginal outcome model as a generalized linear model of the user's choosing, relating exposure main effects to outcome using IPTW weights.

fitModel(
  home_dir,
  data,
  weights,
  exposure,
  exposure_time_pts,
  outcome,
  model,
  family = gaussian,
  link = "identity",
  int_order = NA,
  covariates = NULL,
  epochs = NULL,
  verbose = TRUE,
  save.out = TRUE
)

Arguments

home_dir

path to home directory (required if 'save.out' = TRUE)

data

data in wide format as: a data frame, list of imputed data frames, or mids object

weights

list of IPTW weights output from createWeights()

exposure

name of exposure variable

exposure_time_pts

list of integers at which weights will be created/assessed that correspond to time points when exposure was measured

outcome

name of outcome variable with ".timepoint" suffix

model

character indicating one of the following outcome models:

"m0" (exposure main effects)

"m1" (exposure main effects & covariates)

"m2" (exposure main effects & their interactions)

"m3" (exposure main effects, their interactions, & covariates)

family

(optional) family function specification for svyglm model

link

(optional) character link function specification for svyglm model

int_order

integer specification of highest order exposure main effects interaction, required for interaction models ("m2", "m3")

covariates

list of characters reflecting variable names of covariates, required for covariate models ("m1", "m3")

epochs

(optional) data frame of exposure epoch labels and values

verbose

(optional) TRUE or FALSE indicator for user output (default is TRUE)

save.out

(optional) TRUE or FALSE indicator to save output and intermediary output locally (default is TRUE)

Value

list of svyglm model output

See also

Examples

f <- createFormulas(exposure = "A",
                    exposure_time_pts = c(1, 2, 3),
                    outcome = "D.3",
                    tv_confounders = c("A.1", "A.2", "A.3", "B.1", "B.2", "B.3"),
                    ti_confounders = "C",
                    type = "full",
                    save.out = FALSE)
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A - D.3 at A time point 1 is: 
#> A.1 ~ C
#> <environment: 0x556dc67618a8>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A - D.3 at A time point 2 is: 
#> A.2 ~ A.1 + B.1 + C
#> <environment: 0x556dc67618a8>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A - D.3 at A time point 3 is: 
#> A.3 ~ A.1 + A.2 + B.1 + B.2 + C
#> <environment: 0x556dc67618a8>
#> 

test <- data.frame(ID = 1:50,
                   A.1 = rnorm(n = 50),
                   A.2 = rnorm(n = 50),
                   A.3 = rnorm(n = 50),
                   B.1 = rnorm(n = 50),
                   B.2 = rnorm(n = 50),
                   B.3 = rnorm(n = 50),
                   C = rnorm(n = 50),
                   D.3 = rnorm(n = 50))
test[, c("A.1", "A.2", "A.3")] <- lapply(test[, c("A.1", "A.2", "A.3")], as.numeric)

w <- createWeights(data = test,
                   exposure = "A",
                   outcome = "D.3",
                   formulas = f,
                   save.out = FALSE)
#> Error in weightit2cbps.cont(covs = covs, treat = treat, subset = by.factor ==     i, s.weights = s.weights, missing = missing, verbose = verbose,     ...): The package "CBPS" is required.

m <- fitModel(data = test,
              weights = w,
              exposure = "A",
              exposure_time_pts = c(1, 2, 3),
              outcome = "D.3",
              model = "m0",
              save.out = FALSE)
#> Error in eval(expr, envir, enclos): object 'w' not found
m <- fitModel(data = test,
              weights = w,
              exposure = "A",
              exposure_time_pts = c(1, 2, 3),
              outcome = "D.3",
              model = "m0",
              family = gaussian,
              link = "identity",
              epochs = data.frame(epochs = c("Infancy", "Toddlerhood"),
                                  values = I(list(c(1, 2), c(3)))),
              save.out = FALSE)
#> Error in eval(expr, envir, enclos): object 'w' not found
m <- fitModel(data = test,
              weights = w,
              exposure = "A",
              exposure_time_pts = c(1, 2, 3),
              outcome = "D.3",
              model = "m1",
              covariates = "C",
              save.out = FALSE)
#> Error in eval(expr, envir, enclos): object 'w' not found
m <- fitModel(data = test,
              weights = w,
              exposure = "A",
              exposure_time_pts = c(1, 2, 3),
              outcome = "D.3",
              model = "m2",
              int_order = 3,
              save.out = FALSE)
#> Error in eval(expr, envir, enclos): object 'w' not found
m <- fitModel(data = test,
              weights = w,
              exposure = "A",
              exposure_time_pts = c(1, 2, 3),
              outcome = "D.3",
              model = "m3",
              int_order = 3,
              covariates = "C",
              save.out = FALSE)
#> Error in eval(expr, envir, enclos): object 'w' not found