Fits outcome model

Fits outcome model

getModel(
  d,
  exposure,
  exposure_time_pts,
  outcome,
  epochs,
  exp_epochs,
  int_order,
  model,
  fam,
  covariates,
  verbose
)

Arguments

d

wide data frame

exposure

name of exposure variable

exposure_time_pts

list of integers at which weights will be created/assessed that correspond to time points when exposure was measured

outcome

name of outcome variable with ".timepoint" suffix

epochs

data frame of exposure epoch labels and values

exp_epochs

list of exposure epochs

int_order

integer specification of highest order exposure main effects interaction for interaction models

model

character indicating one of the following outcome models:

"m0" (exposure main effects)

"m1" (exposure main effects & covariates)

"m2" (exposure main effects & their interactions)

"m3" (exposure main effects, their interactions, & covariates)

"covs" (covariate only model)

"int" (intercept only model)

fam

function specification for svyglm model

covariates

list of characters reflecting variable names of covariates for covariate models

verbose

TRUE or FALSE indicator for user output (default is TRUE)

Value

list of fitted model(s)

Examples

test <- data.frame(ID = 1:50,
                   A.1 = rnorm(n = 50),
                   A.2 = rnorm(n = 50),
                   A.3 = rnorm(n = 50),
                   B.1 = rnorm(n = 50),
                   B.2 = rnorm(n = 50),
                   B.3 = rnorm(n = 50),
                   C = rnorm(n = 50),
                   D.3 = rnorm(n = 50))
test[, c("A.1", "A.2", "A.3")] <- lapply(test[, c("A.1", "A.2", "A.3")], as.numeric)

f <- createFormulas(exposure = "A",
                    exposure_time_pts = c(1, 2, 3),
                    outcome = "D.3",
                    tv_confounders = c("A.1", "A.2", "A.3", "B.1", "B.2", "B.3"),
                    ti_confounders = "C",
                    type = "full",
                    save.out = FALSE)
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A - D.3 at A time point 1 is: 
#> A.1 ~ C
#> <environment: 0x55570d2e9d98>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A - D.3 at A time point 2 is: 
#> A.2 ~ A.1 + B.1 + C
#> <environment: 0x55570d2e9d98>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A - D.3 at A time point 3 is: 
#> A.3 ~ A.1 + A.2 + B.1 + B.2 + C
#> <environment: 0x55570d2e9d98>
#> 

w <- createWeights(data = test,
                   exposure = "A",
                   outcome = "D.3",
                   formulas = f,
                   save.out = FALSE)
#> Error in weightit2cbps.cont(covs = covs, treat = treat, subset = by.factor ==     i, s.weights = s.weights, missing = missing, verbose = verbose,     ...): The package "CBPS" is required.

epochs = data.frame(epochs = c("Infancy", "Toddlerhood"),
                    values = I(list(c(1, 2), c(3))))
e <- apply(expand.grid("A", as.character(epochs[, 1])), 1,
           paste, sep = "", collapse = ".")
test$weights <- w[[1]]$weights
#> Error in eval(expr, envir, enclos): object 'w' not found

g <- getModel(d = test,
              exposure = "A",
              exposure_time_pts = c(1, 2, 3),
              outcome = "D.3",
              epochs = epochs,
              exp_epochs = e,
              fam = gaussian,
              model = "m0")
#> Error in model.frame.default(~weights, data = structure(list(ID = 1:50,     A.1 = c(1.78525988578754, 0.9489244017148, -0.825248394102053,     -1.19408227255501, 1.66951722513618, -0.0886432398044239,     1.09777457822444, -1.54088679502851, -1.29101544090044, -2.27223907621134,     -1.77442915747792, -0.179817369930716, 0.396319941259102,     0.285836683553629, 1.06709459699988, 1.93388174160854, -1.26240640606372,     -0.115049072845119, -0.614590925014932, -0.839043220064391,     -0.719616529572742, -0.0957573209651568, 0.872639610749317,     0.174253666865034, 0.812318781979611, 1.61726223844255, -0.0231787291058135,     -0.924341412380484, 1.66845877057525, 1.20869276539124, 0.608586398323056,     -0.228580075542, -1.63220959731886, 0.108302233268565, -0.402824213865403,     -1.59798290244421, -0.0420596135532244, 1.43587939058093,     0.330254494321955, -0.733373115013391, -0.162190684570611,     0.084171540355275, 0.209024246806683, 0.318676831753315,     -0.108860760357649, 0.36033562891414, 0.0313516229796133,     0.749650788422051, -0.0663258419441518, 0.956268966416241    ), A.2 = c(0.282892127694941, -0.515260959830844, 0.326952413815629,     1.15593682472475, -1.56356774100278, 0.174597161281479, 0.260117400938963,     -1.53917138088692, 0.694935671283907, -0.205185236993553,     0.0529065220534898, 1.50017377392088, 1.78670754486974, 0.322297414489165,     2.0961045457776, -0.774907946350006, 0.788898413767666, 0.072915854266946,     -0.579523910401368, 0.404013435489335, 1.04354207920016,     1.50780166942599, -0.30225816570277, -0.483911899550385,     2.20283577369593, -0.138076962307243, -1.38115625541858,     -0.380122680653679, 0.242483082699578, -0.51903254164068,     1.21909691258779, -1.3839324189662, -0.522163964739677, 0.294468889638143,     0.973505494609302, -1.79625580104088, 0.544498437550469,     -0.0771163291062608, 0.0977961133049498, -0.624895048079557,     0.87296225461869, -0.933643186074115, -0.998039236018376,     1.94825630766305, -0.219968644310959, -0.804320823454139,     0.806962126205402, 2.14260785933368, -0.472807412562524,     1.27532263326232), A.3 = c(-0.585292053546965, -0.299673272897062,     -0.138745162298936, 0.405927844471914, -0.441919062358705,     0.704218763205377, 0.471371630790983, -1.2718153033427, -0.867831687291905,     0.325264978854213, -0.137213895765171, -1.46220339384539,     -1.28578631384167, 1.46582173030441, -0.73364261419217, 0.328417968404614,     2.85593854290605, -1.75220285024006, -1.15736805355881, 0.98489127193365,     -0.683931526036577, -1.44611567802682, 0.129663021415484,     0.62957305056915, 1.31209766810394, 1.31901982110119, 0.10728132300516,     -1.85291147418205, 3.10077765446411, 0.598063141570573, 1.24554655722094,     -0.425526454431276, -0.0707474034152256, 1.08455685887526,     0.901477190455546, 0.343609213771726, 0.235480098645658,     -0.0321328598382742, -0.0117130649754238, -0.173058720788833,     1.38322442979857, 1.2420294421769, -1.51641585897269, -0.461319593521075,     -0.580579044093066, 0.981380288318276, -2.13857997366828,     0.123089746853917, -0.214399696379507, -0.0812409265957216    ), B.1 = c(-1.06407542844018, -2.07754991898091, 0.279067516710441,     -0.187513496759143, 0.0768080051820252, 0.678854452048914,     0.138934438984159, -0.986822622282893, -2.08763373594768,     -1.39152947485899, 0.214773795006982, -1.42134068030754,     0.0892191903195769, -1.29018340854526, -0.391111498168038,     1.33964135924085, -0.533696167791358, -0.214555852033691,     0.930192425768249, -1.17403108050427, 0.846047093235125,     1.4551300295797, 0.272570274676554, -0.273221508015907, 0.0217491382398025,     1.11155279515363, 1.40185193400671, -0.0911328485616693,     -1.43152004658976, -0.845139320385046, 0.459788730781098,     2.01482262190916, -1.80626259751529, 0.543576964252969, -0.666499187004516,     -0.412830532544394, 0.0396326214040429, 1.33857647975083,     2.70438337437898, 1.13892505172441, -0.766329379700949, -0.205130280553624,     0.93939298195596, 0.946092388135021, -0.959517886538973,     -0.232195017899252, 0.956545439303501, -1.23918182654884,     -2.77219356572879, 0.0678417167909742), B.2 = c(1.42335523581288,     -2.63489328806884, 1.29018689413253, -1.45924080532874, -0.581586994367315,     1.03076430864745, -1.0126710899456, 0.457461608132175, 0.59124207970731,     0.133041844382133, -0.654056792047571, 0.650569280768358,     0.0335505668979686, -1.03473952887288, 1.51060945428126,     1.30962220094422, -0.174454927075456, 0.683637145456271,     -2.94047119259278, 0.340742740361874, 0.449730486864183,     -0.239305586811494, -1.80212571495331, -1.21790625780196,     0.345952621807202, 1.23641594186153, 0.833379775164103, -0.439427650257516,     -0.235712128545342, -0.120927229776749, 0.0438939009367547,     -0.693744038815913, -1.03380535591261, 1.2658069510038, -0.00472455366602131,     0.122185527063563, 0.39245176404058, -0.22320517799101, -1.33703992803864,     0.704021526131557, 1.93712160903052, -1.63732536511903, 0.00365398050388543,     -1.88976745124766, -1.34254522751488, -0.332335326357284,     0.390620933128735, -0.325089231426904, -2.91060068763481,     0.857052593216491), B.3 = c(-1.14632012069916, 0.198451093019923,     -0.0500472980476262, -2.17081109482018, -3.08985275266331,     0.254822584398835, -0.719354911527551, -0.348990881558434,     -0.682465702102004, 0.11518068209578, -0.193073336355654,     0.468387460904565, 1.02955608351985, -0.19351892998716, -0.234021040102489,     -0.283699152702446, 0.339051286684164, -0.250284252819901,     -0.762052583338819, 0.0807952898148608, 0.927659121173905,     1.07670369409258, -0.252383730411463, 1.99826719434186, -1.12926497646415,     0.698445382306308, -2.12319119571834, 0.343204675105542,     -0.75823543173307, 0.376087162437538, 1.03212844466526, -1.08699893740469,     -2.43309240334931, -0.675048668113913, -0.0341857874459949,     0.43984442116543, -2.01727902181277, -0.528248076725183,     0.484306084317086, -1.32335730515861, 0.0521414443552804,     0.707646965382606, 0.220839553449628, 0.0101560265579283,     0.768533740031274, 0.115016237011826, -0.819252136318841,     -0.286859993135301, 1.58316657841229, 0.557884723713563),     C = c(-0.188879330647135, -0.43756563155601, 0.460461948435193,     -1.07133614368509, -2.7037954921404, 1.13769241861855, 0.471194550855602,     0.569820743807863, 0.0550721042985288, -0.440614403193196,     -0.274023584711912, 0.0240392601722047, 1.49635038491161,     -0.0770128543130996, -0.880007066842148, -1.19364938689514,     1.60741650691366, -0.915053620837002, -0.779996214241383,     0.321749741822709, -1.20960939833003, -0.871042721031388,     0.10930190476833, -0.990751678616211, 0.631148253734195,     -1.1191294638681, 0.162811963288142, -2.17953930868191, 0.74155226552357,     -0.205228144509286, -2.37057512699712, -0.84056403894772,     0.748178844723114, -0.031406174396153, 0.729413695305438,     -2.19862107100925, -0.271693526928433, 1.9910662413353, -0.868961333373115,     2.97821158808206, 0.256920600086935, 0.159676816674984, 1.26834886252105,     -0.168398357152699, 0.413691209884712, 0.2802432976181, -2.31605732752417,     0.450932975161564, -0.240478408914112, -0.194384842635901    ), D.3 = c(1.53315725077824, -0.222611479757563, -1.37929672650971,     0.58545110681693, 0.030835572009497, -0.108902294778676,     0.332038197079068, 1.10048498280604, -0.66062149390152, -0.86463345648057,     -0.409233933890466, 1.46370838346369, 0.291136824439838,     0.330924767353028, 1.19759549223534, 0.766455320365527, 0.441321878696321,     -1.16449107179834, 0.443006395573607, 0.117045492508907,     0.440579901562025, -0.546748056271159, -1.98859851337047,     -1.42396054797616, 0.0893681623185425, -0.462875636750476,     1.25910468628951, 0.438308697223657, 0.340945221299071, -0.816112371297965,     -1.03627407614083, 0.00876274985409656, -0.473102693257904,     -0.2693706067746, 0.628572023063246, -0.597216535677192,     -0.0666942475974821, 0.987157932583161, 0.29400537609768,     0.270549070543243, 1.31716526209578, -1.60554903754294, -1.22795527959431,     -1.52088827149672, 0.669973876597963, 0.770476806518992,     -0.218935615237216, 0.52307523305318, 0.936955294059841,     0.148663229398397), A.Infancy = c(1.03407600674124, 0.216831720941978,     -0.249147990143212, -0.01907272391513, 0.0529747420667001,     0.0429769607385276, 0.678945989581701, -1.54002908795771,     -0.298039884808266, -1.23871215660245, -0.860761317712215,     0.660178201995082, 1.09151374306442, 0.304067049021397, 1.58159957138874,     0.579486897629267, -0.236753996148027, -0.0210666092890865,     -0.59705741770815, -0.217514892287528, 0.161962774813709,     0.706022174230417, 0.285190722523273, -0.154829116342675,     1.50757727783777, 0.739592638067653, -0.702167492262197,     -0.652232046517081, 0.955470926637414, 0.34483011187528,     0.913841655455423, -0.8062562472541, -1.07718678102927, 0.201385561453354,     0.28534064037195, -1.69711935174255, 0.251219411998622, 0.679381530737335,     0.214025303813452, -0.679134081546474, 0.35538578502404,     -0.42473582285942, -0.394507494605846, 1.13346656970818,     -0.164414702334304, -0.22199259727, 0.419156874592508, 1.44612932387787,     -0.269566627253338, 1.11579579983928), A.Toddlerhood = c(-0.585292053546965,     -0.299673272897062, -0.138745162298936, 0.405927844471914,     -0.441919062358705, 0.704218763205377, 0.471371630790983,     -1.2718153033427, -0.867831687291905, 0.325264978854213,     -0.137213895765171, -1.46220339384539, -1.28578631384167,     1.46582173030441, -0.73364261419217, 0.328417968404614, 2.85593854290605,     -1.75220285024006, -1.15736805355881, 0.98489127193365, -0.683931526036577,     -1.44611567802682, 0.129663021415484, 0.62957305056915, 1.31209766810394,     1.31901982110119, 0.10728132300516, -1.85291147418205, 3.10077765446411,     0.598063141570573, 1.24554655722094, -0.425526454431276,     -0.0707474034152256, 1.08455685887526, 0.901477190455546,     0.343609213771726, 0.235480098645658, -0.0321328598382742,     -0.0117130649754238, -0.173058720788833, 1.38322442979857,     1.2420294421769, -1.51641585897269, -0.461319593521075, -0.580579044093066,     0.981380288318276, -2.13857997366828, 0.123089746853917,     -0.214399696379507, -0.0812409265957216)), row.names = c("1", "2", "3", "4", "5", "6", "7", "8", "9", "10", "11", "12", "13", "14", "15", "16", "17", "18", "19", "20", "21", "22", "23", "24", "25", "26", "27", "28", "29", "30", "31", "32", "33", "34", "35", "36", "37", "38", "39", "40", "41", "42", "43", "44", "45", "46", "47", "48", "49", "50"), class = "data.frame"), na.action = function (object,     ...) {    if (NCOL(object) == 0)         object    else {        ok <- complete.cases(object)        if (all(ok))             object        else stop(message)    }}): object is not a matrix