Fits outcome model

Fits outcome model

getModel(
  d,
  exposure,
  exposure_time_pts,
  outcome,
  epochs,
  exp_epochs,
  int_order,
  model,
  fam,
  covariates,
  verbose
)

Arguments

d

wide data frame

exposure

name of exposure variable

exposure_time_pts

list of integers at which weights will be created/assessed that correspond to time points when exposure was measured

outcome

name of outcome variable with ".timepoint" suffix

epochs

data frame of exposure epoch labels and values

exp_epochs

list of exposure epochs

int_order

integer specification of highest order exposure main effects interaction for interaction models

model

character indicating one of the following outcome models:

"m0" (exposure main effects)

"m1" (exposure main effects & covariates)

"m2" (exposure main effects & their interactions)

"m3" (exposure main effects, their interactions, & covariates)

"covs" (covariate only model)

"int" (intercept only model)

fam

function specification for svyglm model

covariates

list of characters reflecting variable names of covariates for covariate models

verbose

TRUE or FALSE indicator for user output (default is TRUE)

Value

list of fitted model(s)

Examples

test <- data.frame(ID = 1:50,
                   A.1 = rnorm(n = 50),
                   A.2 = rnorm(n = 50),
                   A.3 = rnorm(n = 50),
                   B.1 = rnorm(n = 50),
                   B.2 = rnorm(n = 50),
                   B.3 = rnorm(n = 50),
                   C = rnorm(n = 50),
                   D.3 = rnorm(n = 50))
test[, c("A.1", "A.2", "A.3")] <- lapply(test[, c("A.1", "A.2", "A.3")], as.numeric)

f <- createFormulas(exposure = "A",
                    exposure_time_pts = c(1, 2, 3),
                    outcome = "D.3",
                    tv_confounders = c("A.1", "A.2", "A.3", "B.1", "B.2", "B.3"),
                    ti_confounders = "C",
                    type = "full",
                    save.out = FALSE)
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A - D.3 at A time point 1 is: 
#> A.1 ~ C
#> <environment: 0x556dcdde2d28>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A - D.3 at A time point 2 is: 
#> A.2 ~ A.1 + B.1 + C
#> <environment: 0x556dcdde2d28>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A - D.3 at A time point 3 is: 
#> A.3 ~ A.1 + A.2 + B.1 + B.2 + C
#> <environment: 0x556dcdde2d28>
#> 

w <- createWeights(data = test,
                   exposure = "A",
                   outcome = "D.3",
                   formulas = f,
                   save.out = FALSE)
#> Error in weightit2cbps.cont(covs = covs, treat = treat, subset = by.factor ==     i, s.weights = s.weights, missing = missing, verbose = verbose,     ...): The package "CBPS" is required.

epochs = data.frame(epochs = c("Infancy", "Toddlerhood"),
                    values = I(list(c(1, 2), c(3))))
e <- apply(expand.grid("A", as.character(epochs[, 1])), 1,
           paste, sep = "", collapse = ".")
test$weights <- w[[1]]$weights
#> Error in eval(expr, envir, enclos): object 'w' not found

g <- getModel(d = test,
              exposure = "A",
              exposure_time_pts = c(1, 2, 3),
              outcome = "D.3",
              epochs = epochs,
              exp_epochs = e,
              fam = gaussian,
              model = "m0")
#> Error in model.frame.default(~weights, data = structure(list(ID = 1:50,     A.1 = c(-0.816633015956666, 0.657807652299718, -2.00351155355848,     -0.257874110162669, 1.63502856977382, -0.142835888086233,     -0.415666001817787, -1.95593383782711, -0.40054893315927,     2.52629041697892, -0.821914253095121, 0.435477064803542,     0.0370844394964263, 1.1736900262828, -1.44129829497411, 0.238182825774531,     -0.133855545453, -0.662589092560129, 1.76055075550704, 1.82700510414681,     1.11881564399042, -0.41571912234949, -0.32497261654157, -0.0976721038589318,     -0.0612941519151219, 1.28770707642616, 0.853634930806566,     0.239521654777232, 0.150378758480531, 0.184436425266491,     0.917424636564748, -0.80697525907501, 0.763549257666317,     -0.724859832656978, 0.0223019384025433, 1.35152474222839,     -0.198459217000384, 0.458554613284639, -1.61824310646171,     -0.96229395433735, 0.412024225537481, -0.640078376324447,     1.59577474571203, -3.56644042555283, 0.193172039335395, 0.924740881448532,     -0.756923207474755, -0.138695991393196, 1.61499302766459,     -1.47367833898804), A.2 = c(-1.77863428996696, -0.224383230845185,     1.72028201298172, 1.27999292153033, -0.250047717601376, 0.476708224254363,     0.634733166053785, 0.703420751515458, -0.30983423527841,     0.888362583318518, 1.86733032880183, -0.0881055063523346,     0.802819636735432, 0.542558350453895, -0.287688225128415,     -0.330387189966089, -0.627791902948248, 1.12888299190098,     0.564604775889367, 0.719184110038441, -0.275678941330166,     0.673429806419874, 0.180387444894318, -1.89236330008863,     -1.66622441984064, -0.261335445716624, -0.107922660853366,     -0.578416507388466, 0.318048175941962, -0.337132949207044,     -1.45616016685236, 0.802020447758407, -1.30487625339252,     1.70425704912583, -0.390065322938457, -0.355100189672148,     -0.118900646201898, 0.545688458399091, -1.38605183849124,     0.674334323192093, 2.4857136461355, 1.20490526173112, 1.12327129346605,     -0.978901895818738, -0.579718174387515, 0.428021858787462,     -1.0856149887303, 0.691075205738759, 0.139545985831849, 0.397454111830186    ), A.3 = c(-2.35734057712114, -0.650578766370228, -0.303565669729049,     -1.09645989673152, 1.12926380765685, -0.870860306471085,     0.663907357392106, 0.660445820911702, 0.576350087172987,     2.16143125378795, -1.6062875513006, 1.01841056994766, -1.45759967152152,     0.365321499595321, -1.15014160337901, -1.2877948018874, -0.463640021838757,     -0.671530450569942, -1.02198178213771, 1.09493860743599,     -0.75736052715919, -0.177621745270575, 1.11775851826356,     0.991663055110058, 0.131426065436693, 0.419057505125303,     -0.532179454822446, 0.799049568170068, 1.82760577464747,     0.651656287061297, -0.679464060218095, -0.447374195599269,     0.472401367266829, 1.13849729814436, 0.740950919145356, -2.33344514406585,     0.946316153159592, -0.53309795291642, -0.307092041667615,     -1.45244803117353, -0.351092526298523, -0.703384686168277,     -0.445225012088468, 0.157127579597557, 1.71205962176568,     0.633373922229737, 0.367719880571251, -2.07962026061515,     1.76340361387574, 2.26925479066318), B.1 = c(-0.256248103952529,     -0.641072372977024, -0.239454358125214, 0.299779301948975,     -0.994114471740135, -2.08501273156141, 0.211733021713705,     -1.81965980315443, -0.446548417083521, -0.281629258732573,     -0.387725615396291, 0.0520373940399275, -0.669775159520064,     0.70347442548736, -0.936186847405574, -0.159139500644831,     -0.957662292268044, -1.02276851819617, -0.799883454452965,     -0.322780458976663, 1.06383872275199, -0.398893612983895,     1.10195999629017, -1.37093628493821, 0.400387860838441, -1.91086151531339,     0.96116948646415, -1.4340574094532, 0.305495184715463, 0.57980974542361,     -0.584205100819545, 0.288055597053542, -0.764600280379145,     -0.980093733512453, 0.861835472553753, -1.68789134876313,     -1.7820212838088, 0.791141017443998, 0.503998844747654, -0.0877404350651968,     0.586787707849082, -0.827111650230891, -0.0760797102809321,     -0.394864574556658, -0.743618192174624, 0.805533796137002,     -0.505628938620339, -1.19007327916304, 0.656352721196876,     -0.440119616779287), B.2 = c(0.113995498396657, 0.548962157667213,     -0.715398487332887, 1.78804248453557, -0.346279637535285,     1.09256857221158, -0.7413578337947, 0.277579958651529, -1.20668078167523,     1.08659348804375, 0.250538415810572, 1.3271115262965, -0.721388724291772,     -1.16049477133239, -0.180542496995062, -1.40110641782689,     1.86676721363349, 0.870784425692042, 0.355758933552907, -0.629472003566239,     -1.96614150149457, -1.12381340193, 1.26514352356467, -1.50348976800617,     0.961773198143141, -0.837124334152472, -0.954365586989638,     -0.0929834938215895, 1.24570475029569, 1.19807695898369,     0.482397692353237, -0.513303236857199, -0.729513172313512,     0.0551239466112334, 0.584660423767127, -0.563595839692887,     0.131443859491843, 0.892094224467992, -0.138427734416959,     1.6041543527741, 0.567558193544278, 0.804788586047674, -1.71121737752329,     -1.43537824793793, 0.653903751716844, -1.26844761628517,     0.316926291565746, 1.06200615053777, 0.135478216447318, 0.424146503726079    ), B.3 = c(0.612775932666924, -1.67549356924992, -0.285846354432756,     0.216590421842186, -0.0494134917274552, 0.209627658714565,     1.45214889805275, 0.459868706735072, 1.16568545821559, -0.41315016423372,     0.929628345751797, 0.96521768287686, 0.0711160691100984,     -0.295291695789221, -0.206377284918072, -1.23624657377962,     -0.778367348293608, -0.349315484476653, 0.798119169386686,     3.21080978921648, -1.35548439834668, -1.14818310240556, 0.421573699971629,     -0.700374611641294, 0.0171033290093861, -1.02503623710737,     0.0704457245689849, 1.49763043148939, -1.65600166185924,     0.713049537460627, -0.336144381805695, -0.922020406010471,     0.986233001114492, 0.701790251286427, -0.715810450977496,     -0.567437359817809, 1.09511317032304, -1.79502910899034,     1.7537958757808, 1.84113764310434, 0.870372870145728, 1.26336104686756,     -0.109727707912439, 0.229897550532512, 0.287699554400583,     -0.453241007750766, -2.48370749349018, 0.904564220899669,     1.09827748988026, -1.10242136875885), C = c(1.15421003003815,     -0.488683016595535, -0.779473369956346, 0.947803679785697,     0.105322806676673, -0.718061373811655, 0.375449454349387,     -1.19401320865852, 0.381940252963916, -0.0486870289557342,     -1.62686339044557, -0.571310742542943, -0.359964644752723,     -0.0369827748001824, 0.661760638059278, 1.86315684768746,     1.23467836080991, -0.0559826811194968, 0.807463210805293,     -0.565076770801173, 0.519687412404046, -0.522667661812544,     0.910206684533475, -1.04251989608389, -0.633550112091349,     0.0748003473830658, -1.02786549644388, 1.37417490575499,     0.4929309178472, 1.62592952538774, -2.09313765929926, 0.61414290828417,     -0.677828015530801, 0.892482682492328, 0.0477264289420484,     -0.400935932258443, -0.00348146094164529, 1.28655312140333,     -0.717322975508928, -0.943351180074433, 0.327971974802986,     0.744939103675975, -1.36260257503738, -1.88966259378976,     0.0985007944761119, 0.686869581096317, 0.662171188046189,     0.236210743356841, 2.41171587253624, -0.135937048668909),     D.3 = c(0.509576825562294, 1.90346761042887, 0.211423240477883,     -0.6894264034822, 0.358737060624844, -0.586695471826919,     0.204340659216242, -2.00772135821658, -0.194224677239498,     -0.955401010363882, 0.451664856924412, 1.00444027463199,     1.20310113680182, 1.47168575350325, 0.0605097683046114, -0.711748923290628,     1.17803042792097, 1.84645574992668, 1.80268585386381, -0.586797064956585,     -0.910770348517665, 0.732804545591888, 0.878832683927531,     0.0672552091114446, 1.11327953944365, 0.0992349443547568,     -1.01675869779133, -0.671030797526141, -0.848558897371024,     -1.02243845258757, 0.0362957445330971, 0.0429181581758843,     -0.839145640369215, -0.379204233351838, 0.314673130011033,     -0.720452148667216, -0.801145479891883, 0.112122078057399,     -1.04867020057311, 0.39369517132327, 1.39487039777737, 0.103245262430698,     0.17413273233256, 0.735848472280457, -0.385198680113631,     -1.39297220187687, 1.07759737423451, 1.36785646607029, 1.57938611089995,     -1.57266564248848), A.Infancy = c(-1.29763365296181, 0.216712210727267,     -0.14161477028838, 0.511059405683831, 0.692490426086222,     0.166936168084065, 0.109533582117999, -0.626256543155826,     -0.35519158421884, 1.70732650014872, 0.522708037853354, 0.173685779225604,     0.419952038115929, 0.858124188368347, -0.864493260051262,     -0.046102182095779, -0.380823724200624, 0.233146949670425,     1.1625777656982, 1.27309460709263, 0.421568351330127, 0.128855342035192,     -0.072292585823626, -0.995017701973781, -0.863759285877881,     0.513185815354768, 0.3728561349766, -0.169447426305617, 0.234213467211247,     -0.0763482619702765, -0.269367765143806, -0.00247740565830151,     -0.270663497863102, 0.489698608234426, -0.183881692267957,     0.498212276278121, -0.158679931601141, 0.502121535841865,     -1.50214747247647, -0.143979815572629, 1.44886893583649,     0.282413442703337, 1.35952301958904, -2.27267116068578, -0.19327306752606,     0.676381370117997, -0.921269098102528, 0.276189607172782,     0.87726950674822, -0.538112113578927), A.Toddlerhood = c(-2.35734057712114,     -0.650578766370228, -0.303565669729049, -1.09645989673152,     1.12926380765685, -0.870860306471085, 0.663907357392106,     0.660445820911702, 0.576350087172987, 2.16143125378795, -1.6062875513006,     1.01841056994766, -1.45759967152152, 0.365321499595321, -1.15014160337901,     -1.2877948018874, -0.463640021838757, -0.671530450569942,     -1.02198178213771, 1.09493860743599, -0.75736052715919, -0.177621745270575,     1.11775851826356, 0.991663055110058, 0.131426065436693, 0.419057505125303,     -0.532179454822446, 0.799049568170068, 1.82760577464747,     0.651656287061297, -0.679464060218095, -0.447374195599269,     0.472401367266829, 1.13849729814436, 0.740950919145356, -2.33344514406585,     0.946316153159592, -0.53309795291642, -0.307092041667615,     -1.45244803117353, -0.351092526298523, -0.703384686168277,     -0.445225012088468, 0.157127579597557, 1.71205962176568,     0.633373922229737, 0.367719880571251, -2.07962026061515,     1.76340361387574, 2.26925479066318)), row.names = c("1", "2", "3", "4", "5", "6", "7", "8", "9", "10", "11", "12", "13", "14", "15", "16", "17", "18", "19", "20", "21", "22", "23", "24", "25", "26", "27", "28", "29", "30", "31", "32", "33", "34", "35", "36", "37", "38", "39", "40", "41", "42", "43", "44", "45", "46", "47", "48", "49", "50"), class = "data.frame"), na.action = function (object,     ...) {    if (NCOL(object) == 0)         object    else {        ok <- complete.cases(object)        if (all(ok))             object        else stop(message)    }}): object is not a matrix