Fits outcome model

Fits outcome model

getModel(
  d,
  exposure,
  exposure_time_pts,
  outcome,
  epochs,
  exp_epochs,
  int_order,
  model,
  fam,
  covariates,
  verbose
)

Arguments

d

wide data frame

exposure

name of exposure variable

exposure_time_pts

list of integers at which weights will be created/assessed that correspond to time points when exposure was measured

outcome

name of outcome variable with ".timepoint" suffix

epochs

data frame of exposure epoch labels and values

exp_epochs

list of exposure epochs

int_order

integer specification of highest order exposure main effects interaction for interaction models

model

character indicating one of the following outcome models:

"m0" (exposure main effects)

"m1" (exposure main effects & covariates)

"m2" (exposure main effects & their interactions)

"m3" (exposure main effects, their interactions, & covariates)

"covs" (covariate only model)

"int" (intercept only model)

fam

function specification for svyglm model

covariates

list of characters reflecting variable names of covariates for covariate models

verbose

TRUE or FALSE indicator for user output (default is TRUE)

Value

list of fitted model(s)

Examples

test <- data.frame(ID = 1:50,
                   A.1 = rnorm(n = 50),
                   A.2 = rnorm(n = 50),
                   A.3 = rnorm(n = 50),
                   B.1 = rnorm(n = 50),
                   B.2 = rnorm(n = 50),
                   B.3 = rnorm(n = 50),
                   C = rnorm(n = 50),
                   D.3 = rnorm(n = 50))
test[, c("A.1", "A.2", "A.3")] <- lapply(test[, c("A.1", "A.2", "A.3")], as.numeric)

f <- createFormulas(exposure = "A",
                    exposure_time_pts = c(1, 2, 3),
                    outcome = "D.3",
                    tv_confounders = c("A.1", "A.2", "A.3", "B.1", "B.2", "B.3"),
                    ti_confounders = "C",
                    type = "full",
                    save.out = FALSE)
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A-D.3 at A time point 1 is: 
#> A.1 ~ C
#> <environment: 0x556a9c833ed0>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A-D.3 at A time point 2 is: 
#> A.2 ~ A.1 + B.1 + C
#> <environment: 0x556a9c833ed0>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A-D.3 at A time point 3 is: 
#> A.3 ~ A.1 + A.2 + B.1 + B.2 + C
#> <environment: 0x556a9c833ed0>
#> 

w <- createWeights(data = test,
                   exposure = "A",
                   outcome = "D.3",
                   formulas = f,
                   save.out = FALSE)
#> Error in weightit2cbps.cont(covs = covs, treat = treat, subset = by.factor ==     i, s.weights = s.weights, missing = missing, verbose = verbose,     ...): The package "CBPS" is required.

epochs = data.frame(epochs = c("Infancy", "Toddlerhood"),
                    values = I(list(c(1, 2), c(3))))
e <- apply(expand.grid("A", as.character(epochs[, 1])), 1,
           paste, sep = "", collapse = ".")
test$weights <- w[[1]]$weights
#> Error in eval(expr, envir, enclos): object 'w' not found

g <- getModel(d = test,
              exposure = "A",
              exposure_time_pts = c(1, 2, 3),
              outcome = "D.3",
              epochs = epochs,
              exp_epochs = e,
              fam = gaussian,
              model = "m0")
#> Error in model.frame.default(~weights, data = structure(list(ID = 1:50,     A.1 = c(0.314264943375281, 0.947618033692379, -0.693960348285558,     0.444022341105509, -0.369209342286739, 0.638684190158849,     1.11914481795072, -0.231366558658226, 0.502484634680268,     -0.497288124555821, -1.40929644797547, 0.344041922341579,     -0.0876918188342045, 0.418724494016794, -1.34126158450405,     -0.853607376800867, 1.43654374993903, -0.867748320311063,     0.525461779710956, -0.569106292248567, 1.75094236317704,     -1.08740036034499, 1.07965753713629, 1.44406499212204, -0.260561049757115,     1.74912175130402, 2.48927887124934, 1.61039018222586, 3.16548645182325,     0.298559062071203, 1.65079448909119, -0.298119104705968,     -0.389663734814854, -0.174198203072985, 0.978746137545588,     -0.425802892521342, -0.87831214017956, 0.0205941136703221,     0.4519573465658, 0.107597757702521, 0.0656243308907501, -1.20014006374726,     -0.0994518552101053, 0.0919230396906143, 0.739562415443318,     -1.02141992158901, 2.57730178445909, -0.500044469607077,     0.999333422403327, 0.691021958992356), A.2 = c(-0.0810273980056784,     1.02019070143435, 0.476305299564901, -0.111277303599691,     0.0130917460979778, -1.35041078588239, 0.96485880794021,     0.586953788752325, 1.57787533368413, 0.412642838929063, 3.61004775741485,     0.227945140582888, -0.26495304491238, -0.315675242256046,     0.222379600848689, 0.735648729515093, 0.0301332027763013,     1.0482565707902, 0.424533474890526, -0.612126482996835, -1.5486421267813,     1.26787746794113, 0.932810932996685, -2.33775226656931, 0.551203903424337,     0.186730328150755, 1.71201688479261, 0.373022396159777, -0.627577211505485,     -0.111283332395017, -1.9098511544714, 0.287057384688783,     -0.76687414841993, 0.01448910454905, -1.43366587648743, 1.63718877790083,     -0.496763088519548, 1.54758954588817, -0.550576455847568,     -0.536800306017357, -0.065790076678193, 1.18594673886543,     0.132561809012787, 0.609872962288068, -0.601195893017301,     0.60278697952619, -0.605402932330508, -1.29788507310703,     0.363209250044862, 0.562844281903413), A.3 = c(-0.655380003089107,     1.41281184799849, -1.30538984192475, 0.675755533657954, -0.236933298514275,     -0.532959584000621, 0.066067941712647, 0.524365658963086,     1.48215813602541, 1.14360442191943, -0.110066878800688, -0.245230516435395,     -0.119857271496354, 0.664928392965796, 1.46631508773975,     -1.24236718742883, 0.156045038599505, 0.325829506272068,     -0.477791451194344, -0.492732574913167, -0.0526317058256597,     0.26840901380344, -1.03408495184637, -0.943528917513111,     -1.77479772052448, 0.889941758069214, -1.89953392915917,     -0.406879049142618, 0.244374193162048, -0.319534130399239,     -0.165141608194062, 0.381286418962181, 0.462893871793746,     0.628841575280653, 1.00714757086669, 1.01345020599023, 0.255810897562266,     -0.544458645675202, -0.47839954288242, -0.322819116183573,     -1.88683995677618, 0.661679176810943, -0.980920848710406,     -1.80922677040781, 0.857798612403642, 0.513590135226914,     -1.36128115266634, 1.7197942430227, 0.534036990839069, -0.794646596512534    ), B.1 = c(0.255531539617997, 0.226279327676296, 1.20409553096468,     -2.52694645247944, -1.44231914021025, -1.19151526409184,     0.657295350769517, 0.97082855014071, 1.77428934107447, 0.606797344196159,     -0.548387054608916, 2.08573262774884, -1.19840038452947,     -0.457072539987638, -1.2725294428973, 2.57187259558582, 0.819676584353534,     -0.143052235477219, 1.99931135527586, -0.148369662030347,     -0.924594699961147, -0.790810533645348, -0.706375670454195,     0.00128547677538509, -1.39044640829202, -0.00428940368942703,     0.784602795858216, 0.101830429163976, -0.498776302247702,     -1.52225235194629, -0.757957731134675, 0.0863701592000017,     -2.3570400202639, 1.20564979827352, 0.221849002642949, -0.917046210057974,     1.47335904999179, 0.0149998215216897, 0.648039262301149,     -1.0969324079629, 0.529113531241385, -0.122276397390269,     0.764568913086512, 0.0105618943105659, -0.755497284894977,     -1.10883595696003, 0.390683593620729, -0.576384254453287,     0.989221554883567, -0.30064742413304), B.2 = c(1.37176568997994,     -1.01257076419609, -0.454355889583126, 1.18531987159112,     -0.337200738081423, -0.588081823692104, 0.565979899164246,     -0.854757427321977, -0.0272386042625929, -1.29869844801657,     -0.388089740440768, -0.63925977589945, -0.69606123452543,     1.41282712975019, 0.250059901042361, -0.212698815963515,     0.101217094113434, 1.32074904589113, -0.818555519553134,     0.207826225241793, -0.0638038829303174, -0.283251106614033,     -1.55432251116679, 2.89730157987633, 0.0387776605069816,     0.118730479168488, 0.26673249363504, -0.191992892458593,     -0.0489203763779952, 0.526789503701099, 0.110646449955307,     2.1053109964403, 0.930165745927096, -0.41327361647868, 0.439938107859046,     -1.20368687012389, 1.00961628083874, 1.10527058387121, 0.581158479896806,     -1.37158196943184, 0.125143253449268, 0.233206794365932,     0.19022075894356, 0.324130444046336, 0.346432468533853, 0.0811038624951404,     -0.734994013671809, -0.347734493570021, -0.450247376440086,     0.00882110324185214), B.3 = c(-2.4903610023623, -1.60974576433012,     2.85035450345433, -0.348681393218442, 0.403332383345367,     0.710741182681513, -0.985588251352825, 0.265778730882853,     0.212863408331048, 1.02389065187704, -0.353276761963208,     -0.460205819357411, -0.802021938925472, -1.25428385726971,     -0.842583589528921, -0.374347424148574, -1.46915836954983,     0.721989169616493, 1.24114007878353, 1.16659607313751, 1.59293768002091,     0.645069819529732, 0.675686807574642, -0.612711192333351,     0.425278521457652, -1.93707954722846, -0.0706568726922178,     -0.869830068867696, -0.460370864736858, -2.4010642314392,     0.343690520070116, -0.488982826375872, -0.374962935441453,     -0.884477938590409, -0.464864951465448, -1.68662126417685,     1.02875623279699, -0.321842247511532, 0.914637859852495,     -0.19886436894072, 1.20977825509385, 1.5077361191587, -0.394291335649916,     -0.133925591403908, 1.53550225483165, -0.16817107399039,     0.283183110730082, -0.348109914665892, -0.0448908370747505,     1.80956747545256), C = c(1.51493007964502, 1.03729594917494,     -0.545123820224646, -0.628891294481158, -0.116582421160253,     -0.147875340010428, -1.49713319548303, -0.0734224172056018,     -0.966709031183907, 0.669990663383067, 0.879770607842342,     0.602573939726198, 0.223238801413043, 0.053455902374922,     -0.733157798000334, 2.04590570069298, 0.317580935131115,     -1.08282512461163, 1.07437965440523, 0.44667734501986, -2.21029074111121,     0.929711785161978, -0.659635660320964, -1.26017163283741,     -1.26480156937977, -0.375217765832383, -0.0860977259521369,     -0.0651468131398467, 0.323702338263862, -1.70525002042542,     0.64511758370501, -1.29831665979042, 1.20638817237629, 0.844314111974891,     -0.0636375861457515, 0.671281186000464, 1.59491446903093,     -0.409032648999581, 0.774637562813559, -1.45392511604895,     -0.815848452713876, -0.765291271644164, 0.214789325409845,     1.11223028801227, -1.10052898780165, 1.47665878473129, 0.434142626819089,     -0.728554484419323, 0.255036917797344, -0.454188675970312    ), D.3 = c(0.129635969604044, 0.477696230936746, -0.0850930509142858,     -0.161887167567656, 0.12635595320992, 0.773514483168459,     -1.11845707380062, 0.555807651510042, -0.549572472919134,     -1.28886360313572, -0.659887424510109, -0.491815055344374,     1.12080940876791, -1.24898758905897, 0.521754574278553, -0.424829139334728,     0.291386500596075, -0.129538799128093, -2.29653927215967,     -0.398087025530304, -0.185517527011284, 0.707016116315493,     0.783883333554803, 0.142361539199591, -0.54505535554137,     -0.414935391596879, 0.528173503413425, -0.764202676089733,     -0.530566780837671, 1.31678195171574, 1.45102365720237, 0.0877293457776188,     1.22187000338712, 0.0541715742454939, -1.29502197627655,     0.170821631109879, -0.77738724519996, 1.06488207704398, -0.351490845325606,     0.0310127454321992, 0.378758489281302, -0.896958934174957,     1.56269918895452, -0.335883074616336, -0.575505760664726,     1.37923264598338, -1.02956514905121, -0.880647272841834,     0.159040573543652, -1.6459379060931), A.Infancy = c(0.116618772684801,     0.983904367563365, -0.108827524360328, 0.166372518752909,     -0.178058798094381, -0.35586329786177, 1.04200181294546,     0.177793615047049, 1.0401799841822, -0.042322642813379, 1.10037565471969,     0.285993531462233, -0.176322431873292, 0.051524625880374,     -0.559440991827681, -0.058979323642887, 0.733338476357666,     0.0902541252395686, 0.474997627300741, -0.590616387622701,     0.10115011819787, 0.09023855379807, 1.00623423506649, -0.446843637223635,     0.145321426833611, 0.967926039727388, 2.10064787802097, 0.991706289192818,     1.26895462015888, 0.093637864838093, -0.129528332690105,     -0.00553086000859249, -0.578268941617392, -0.0798545492619675,     -0.227459869470921, 0.605692942689744, -0.687537614349554,     0.784091829779246, -0.049309554640884, -0.214601274157418,     -8.28728937214557e-05, -0.007096662440915, 0.0165549769013408,     0.350898000989341, 0.0691832612130085, -0.20931647103141,     0.985949426064291, -0.898964771357053, 0.681271336224095,     0.626933120447885), A.Toddlerhood = c(-0.655380003089107,     1.41281184799849, -1.30538984192475, 0.675755533657954, -0.236933298514275,     -0.532959584000621, 0.066067941712647, 0.524365658963086,     1.48215813602541, 1.14360442191943, -0.110066878800688, -0.245230516435395,     -0.119857271496354, 0.664928392965796, 1.46631508773975,     -1.24236718742883, 0.156045038599505, 0.325829506272068,     -0.477791451194344, -0.492732574913167, -0.0526317058256597,     0.26840901380344, -1.03408495184637, -0.943528917513111,     -1.77479772052448, 0.889941758069214, -1.89953392915917,     -0.406879049142618, 0.244374193162048, -0.319534130399239,     -0.165141608194062, 0.381286418962181, 0.462893871793746,     0.628841575280653, 1.00714757086669, 1.01345020599023, 0.255810897562266,     -0.544458645675202, -0.47839954288242, -0.322819116183573,     -1.88683995677618, 0.661679176810943, -0.980920848710406,     -1.80922677040781, 0.857798612403642, 0.513590135226914,     -1.36128115266634, 1.7197942430227, 0.534036990839069, -0.794646596512534    )), row.names = c(NA, -50L), class = "data.frame"), na.action = function (object,     ...) {    if (NCOL(object) == 0)         object    else {        ok <- complete.cases(object)        if (all(ok))             object        else stop(message)    }}): object is not a matrix