Fits outcome model

Fits outcome model

getModel(
  d,
  exposure,
  exposure_time_pts,
  outcome,
  epochs,
  exp_epochs,
  int_order,
  model,
  fam,
  covariates,
  verbose
)

Arguments

d

wide data frame

exposure

name of exposure variable

exposure_time_pts

list of integers at which weights will be created/assessed that correspond to time points when exposure was measured

outcome

name of outcome variable with ".timepoint" suffix

epochs

data frame of exposure epoch labels and values

exp_epochs

list of exposure epochs

int_order

integer specification of highest order exposure main effects interaction for interaction models

model

character indicating one of the following outcome models:

"m0" (exposure main effects)

"m1" (exposure main effects & covariates)

"m2" (exposure main effects & their interactions)

"m3" (exposure main effects, their interactions, & covariates)

"covs" (covariate only model)

"int" (intercept only model)

fam

function specification for svyglm model

covariates

list of characters reflecting variable names of covariates for covariate models

verbose

TRUE or FALSE indicator for user output (default is TRUE)

Value

list of fitted model(s)

Examples

test <- data.frame(ID = 1:50,
                   A.1 = rnorm(n = 50),
                   A.2 = rnorm(n = 50),
                   A.3 = rnorm(n = 50),
                   B.1 = rnorm(n = 50),
                   B.2 = rnorm(n = 50),
                   B.3 = rnorm(n = 50),
                   C = rnorm(n = 50),
                   D.3 = rnorm(n = 50))
test[, c("A.1", "A.2", "A.3")] <- lapply(test[, c("A.1", "A.2", "A.3")], as.numeric)

f <- createFormulas(exposure = "A",
                    exposure_time_pts = c(1, 2, 3),
                    outcome = "D.3",
                    tv_confounders = c("A.1", "A.2", "A.3", "B.1", "B.2", "B.3"),
                    ti_confounders = "C",
                    type = "full",
                    save.out = FALSE)
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A-D.3 at A time point 1 is: 
#> A.1 ~ C
#> <environment: 0x556e3dbfd390>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A-D.3 at A time point 2 is: 
#> A.2 ~ A.1 + B.1 + C
#> <environment: 0x556e3dbfd390>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A-D.3 at A time point 3 is: 
#> A.3 ~ A.1 + A.2 + B.1 + B.2 + C
#> <environment: 0x556e3dbfd390>
#> 

w <- createWeights(data = test,
                   exposure = "A",
                   outcome = "D.3",
                   formulas = f,
                   save.out = FALSE)
#> Error in weightit2cbps.cont(covs = covs, treat = treat, subset = by.factor ==     i, s.weights = s.weights, missing = missing, verbose = verbose,     ...): The package "CBPS" is required.

epochs = data.frame(epochs = c("Infancy", "Toddlerhood"),
                    values = I(list(c(1, 2), c(3))))
e <- apply(expand.grid("A", as.character(epochs[, 1])), 1,
           paste, sep = "", collapse = ".")
test$weights <- w[[1]]$weights
#> Error in eval(expr, envir, enclos): object 'w' not found

g <- getModel(d = test,
              exposure = "A",
              exposure_time_pts = c(1, 2, 3),
              outcome = "D.3",
              epochs = epochs,
              exp_epochs = e,
              fam = gaussian,
              model = "m0")
#> Error in model.frame.default(~weights, data = structure(list(ID = 1:50,     A.1 = c(0.880229666476174, 0.610350840347189, -0.436805751050938,     0.108722551812645, 0.421236595003309, 0.257171274721728,     -0.423867646130371, 2.05648090898808, -0.986980828643108,     0.26151326714478, -0.528610433022993, 2.31092433629325, 0.982245194841156,     0.419003752559624, -1.2498399613331, 0.574730331746979, 1.48751502609011,     -0.262716051528016, 0.328693182317326, -0.106244873045524,     0.977907241729108, -0.0899524900875678, 0.220325127005496,     -2.6182616026148, -0.213386285688914, -0.840622025924537,     -0.973892422056174, -0.261741529798455, -0.742354101183116,     1.81184414361507, -2.27414471419988, -0.718924068420997,     0.764124743053279, 1.52109314768188, 0.4134977915951, -0.594637053262596,     0.699695483947714, 1.48409779679795, -0.823695623913896,     0.107404589161615, -0.138673895708932, 1.10462104563041,     -0.883411121837181, -1.34986864828576, -1.13667893630471,     -0.49669500762309, -0.823502485884763, -0.784565162379218,     0.752146228509466, -0.0413792518861548), A.2 = c(0.507709537233715,     0.39750453422716, -0.417938028720009, 0.37273140401417, 0.251599021537327,     0.045015874405299, 0.253789566724414, -0.101413308499331,     -1.71221671334926, 1.35347832146007, 0.87770165566751, 1.72873455468163,     -0.0461742740673254, -0.364019314411407, -1.09639086212805,     1.16386098922434, -1.11686794657945, 1.41071870839284, 0.530132719026427,     0.111279389911995, -1.32254974693451, 1.533684069859, -0.671146685580416,     -1.15971175828938, 1.63720137977462, 1.71688034987862, -1.02299614971071,     -0.553590911694472, -0.542298234775454, 1.4729926240731,     2.27571027808829, 0.277972929870463, -0.185052435414356,     1.03617245145783, 1.42355094987275, -0.661993688620721, -0.794482965901634,     1.15773852880718, -1.24003549010991, 2.04385127055406, 0.332081690832403,     -1.83487964601418, 1.25716367797381, 0.730881479418287, 1.31981516736373,     -0.0115647508837404, -0.690897998098789, -0.068450419233328,     -0.087716875161975, -1.14934122234966), A.3 = c(-0.602381637471607,     1.74186900492755, 0.317433137213049, 0.91230415121334, -0.281627946551685,     -0.461499707347741, 0.519960625469771, 1.20516955620998,     -0.00643964563019493, -0.960395868198456, 0.847071190606425,     -0.0611738610199305, 0.546851477215917, 0.95838201276413,     -0.0308980470929579, -0.507836555583225, 0.290663099980682,     1.82334379763579, 0.244492943523155, -1.46443395166553, -0.774192000906254,     1.97347910200689, -1.66010474167948, -0.139601240433156,     -2.16178578210347, -0.789153033550317, -1.06126741903938,     -1.3871410251396, 1.49904159912067, 0.924400373549812, 0.504051778043338,     -0.695730181186703, 0.543956304814683, 2.21887468061734,     0.682274992289392, 1.00464553380553, 0.313941471558116, 0.24822683628142,     -1.11983024488595, -0.549862573131475, -0.823138692702099,     -1.21586848384435, 1.17274850383192, 1.14137656944447, -0.107231886435014,     -0.289508753054704, -0.262172223404473, 0.46351239423928,     -1.24751829582634, -1.53911829635762), B.1 = c(-1.04837146242752,     -0.56534350135067, 0.612593174884114, 1.32431906438993, 0.68483110771615,     -1.20742502036011, -1.03728178742916, -1.89896545914034,     -1.9361772722903, 1.8722288517409, 0.591642118042683, -0.075491229059956,     -0.797893062002709, 0.493171705145826, -0.655374889970293,     0.135040798756679, -1.10862510046683, 0.56558449070577, -2.03617417504078,     -0.299920960531505, 1.59787286376679, 1.74505863643405, 1.29702325065144,     -3.02235937427464, -0.00632452016202612, 1.30030878405858,     -0.0624416482670965, -0.0311688154692063, 0.595791614336655,     -1.77423245295655, 0.526604947175985, -0.843041770120168,     0.212027465113572, 0.436917494855823, -0.610354009127131,     -1.19700696978584, -0.0549360137904636, 0.175177536599373,     -0.847458702778159, -0.0463313752637009, 1.62702842944622,     2.08722540265649, -0.214039285011796, -0.169571248827656,     0.303619702201986, -0.138763242795458, 1.90070608945347,     -0.409868112492997, -1.3953307459631, -0.328197429147393),     B.2 = c(2.24607266302852, -0.0309047720281355, -1.2149236129215,     0.13217101217673, 0.274881480296276, 2.28229713637739, 1.83787508564824,     -0.548764677800593, 0.473355776684122, 0.0263186110855783,     0.0993629280981354, -1.17457356270706, -1.07127368544779,     1.77870332725628, -1.34741796816374, -0.0796715675548103,     0.786068302190084, 0.691328276927952, -0.948544709383826,     -0.702829768518326, -0.917280967278455, -1.7869822988751,     -1.12826411241897, 1.74254494367803, -0.663331084066731,     -0.991570973413089, 0.801598310718314, 0.894890760673441,     0.00950946809042452, -1.65213808142631, -0.374174473301704,     0.662141967551157, 1.29599684284031, -0.310634005170562,     -0.62900510668863, -0.98116524653599, 1.07783128809719, 1.24731445461389,     1.65283267403587, -1.21771414543805, -2.86358836149384, 0.137249961142083,     0.640667771312467, -0.316679351753319, 0.460262770779627,     1.25797949636449, 1.77886402000648, -0.144013393097669, -0.0993040531802585,     0.236040712892172), B.3 = c(0.191209967632841, 0.42835684233735,     -0.797946479954535, 0.871975453610669, -0.240944774169275,     -0.698770268640435, 1.2286596972171, -0.960396628008911,     1.22134849495083, -0.738664522070711, -1.09167690727004,     1.36234121654341, 0.891451007767803, -0.663845945011405,     0.992560700899632, -0.406582038188456, 0.121332730322338,     2.18939285420691, 0.0748739461366914, 0.776352760998471,     0.988696706091126, 2.73188391083157, 0.763685573571282, -0.103834126879183,     1.40331684631902, 0.629215106621439, -0.278859456397801,     0.540895751188265, -0.516406810447407, 1.19358930653858,     0.197090436449509, -1.77297087227711, -0.60608748772378,     1.17203229635847, -0.0262715605583832, -1.34519111538238,     -2.04292520952335, -1.34276220603911, -1.11197009833498,     0.986856371459049, -1.24936531606083, 1.49375481858327, -0.252508845499413,     0.197307562008107, -0.98893881838249, 1.98313250129101, 0.0317084714335494,     -1.28236318235023, -0.829362779770574, -1.31406810986186),     C = c(-0.375133380952056, -0.767929472061099, -0.130349523436809,     -1.02404921731049, 0.682747803708795, -0.0641442368402656,     1.09887865496145, 0.204253975201291, -0.401395969963601,     -0.100549000584188, 0.658876233420671, -0.0347573945401171,     -1.23695490614847, -0.201214883610075, -1.78053431495501,     0.871335900816196, -1.56821623950724, 1.07032089723341, -1.74320432219729,     0.854326981006201, 0.612824938891492, 1.01062586854409, -0.766692361164632,     -0.951292022579451, -0.159826766390671, -2.88002829994384,     -0.135659372790935, -1.64895542389896, -1.06639405237682,     0.686226482847427, 0.273490542803117, -1.00693601183008,     -1.19853512842869, -2.37286903137462, -0.307542630666458,     1.1432427529513, -1.94710355389259, 0.0971373417291922, 0.222908172676513,     1.02312717605153, -1.07875332810701, -0.875739881269348,     -0.517188218198344, -1.57712570833453, -0.229358290000446,     -0.203926394836537, 3.77122066455853, -0.498736047541879,     1.57566014418179, -1.77785221904858), D.3 = c(0.178434849639681,     -0.856344161962275, 0.406127298417318, 1.35060607979799,     2.04747234759657, -1.25245425617911, 0.183093526476885, -0.0990604559035016,     0.103270859357583, -0.851570639354791, 1.69591128285155,     -0.412416137388296, 0.266281969935292, -1.15241338956277,     1.32321395822052, -0.423243216803365, 0.852813286836738,     0.829318403255303, -1.29754032249377, -1.58677393405353,     0.528715724973326, -0.0481949352643565, 0.958283489811363,     1.15851928594646, 0.125106406110324, -0.572113702726414,     -0.818921968374486, -0.677042027928203, 2.32259006207939,     2.05694270285371, 1.00256184305204, -0.274041720703495, -0.547134855847293,     0.890251984837462, -0.234988710555946, 1.85413432035186,     0.02596306517641, -0.676197178507812, 0.108481814795398,     1.6255828586444, -0.907584710052402, 0.467305476416633, 0.298302541030666,     1.1744283547238, -0.152017490701535, 2.66343391836318, -1.30543220548538,     -0.677144302071771, -0.0375412272074293, -0.490568981952253    ), A.Infancy = c(0.693969601854944, 0.503927687287175, -0.427371889885473,     0.240726977913407, 0.336417808270318, 0.151093574563514,     -0.0850390397029785, 0.977533800244375, -1.34959877099618,     0.807495794302425, 0.174545611322258, 2.01982944548744, 0.468035460386915,     0.0274922190741085, -1.17311541173058, 0.869295660485659,     0.18532353975533, 0.574001328432412, 0.429412950671877, 0.0025172584332355,     -0.172321252602701, 0.721865789885716, -0.22541077928746,     -1.88898668045209, 0.711907547042853, 0.438129161977041,     -0.998444285883442, -0.407666220746464, -0.642326167979285,     1.64241838384409, 0.000782781944205002, -0.220475569275267,     0.289536153819462, 1.27863279956986, 0.918524370733925, -0.628315370941658,     -0.04739374097696, 1.32091816280257, -1.0318655570119, 1.07562792985784,     0.0967038975617355, -0.365129300191885, 0.186876278068315,     -0.309493584433737, 0.0915681155295099, -0.254129879253415,     -0.757200241991776, -0.426507790806273, 0.332214676673745,     -0.595360237117907), A.Toddlerhood = c(-0.602381637471607,     1.74186900492755, 0.317433137213049, 0.91230415121334, -0.281627946551685,     -0.461499707347741, 0.519960625469771, 1.20516955620998,     -0.00643964563019493, -0.960395868198456, 0.847071190606425,     -0.0611738610199305, 0.546851477215917, 0.95838201276413,     -0.0308980470929579, -0.507836555583225, 0.290663099980682,     1.82334379763579, 0.244492943523155, -1.46443395166553, -0.774192000906254,     1.97347910200689, -1.66010474167948, -0.139601240433156,     -2.16178578210347, -0.789153033550317, -1.06126741903938,     -1.3871410251396, 1.49904159912067, 0.924400373549812, 0.504051778043338,     -0.695730181186703, 0.543956304814683, 2.21887468061734,     0.682274992289392, 1.00464553380553, 0.313941471558116, 0.24822683628142,     -1.11983024488595, -0.549862573131475, -0.823138692702099,     -1.21586848384435, 1.17274850383192, 1.14137656944447, -0.107231886435014,     -0.289508753054704, -0.262172223404473, 0.46351239423928,     -1.24751829582634, -1.53911829635762)), row.names = c(NA, -50L), class = "data.frame"), na.action = function (object, ...) {    if (NCOL(object) == 0)         object    else {        ok <- complete.cases(object)        if (all(ok))             object        else stop(message)    }}): object is not a matrix