Fits outcome model

Fits outcome model

getModel(
  d,
  exposure,
  exposure_time_pts,
  outcome,
  epochs,
  exp_epochs,
  int_order,
  model,
  fam,
  covariates,
  verbose
)

Arguments

d

wide data frame

exposure

name of exposure variable

exposure_time_pts

list of integers at which weights will be created/assessed that correspond to time points when exposure was measured

outcome

name of outcome variable with ".timepoint" suffix

epochs

data frame of exposure epoch labels and values

exp_epochs

list of exposure epochs

int_order

integer specification of highest order exposure main effects interaction for interaction models

model

character indicating one of the following outcome models:

"m0" (exposure main effects)

"m1" (exposure main effects & covariates)

"m2" (exposure main effects & their interactions)

"m3" (exposure main effects, their interactions, & covariates)

"covs" (covariate only model)

"int" (intercept only model)

fam

function specification for svyglm model

covariates

list of characters reflecting variable names of covariates for covariate models

verbose

TRUE or FALSE indicator for user output (default is TRUE)

Value

list of fitted model(s)

Examples

test <- data.frame(ID = 1:50,
                   A.1 = rnorm(n = 50),
                   A.2 = rnorm(n = 50),
                   A.3 = rnorm(n = 50),
                   B.1 = rnorm(n = 50),
                   B.2 = rnorm(n = 50),
                   B.3 = rnorm(n = 50),
                   C = rnorm(n = 50),
                   D.3 = rnorm(n = 50))
test[, c("A.1", "A.2", "A.3")] <- lapply(test[, c("A.1", "A.2", "A.3")], as.numeric)

f <- createFormulas(exposure = "A",
                    exposure_time_pts = c(1, 2, 3),
                    outcome = "D.3",
                    tv_confounders = c("A.1", "A.2", "A.3", "B.1", "B.2", "B.3"),
                    ti_confounders = "C",
                    type = "full",
                    save.out = FALSE)
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A-D.3 at A time point 1 is: 
#> A.1 ~ C
#> <environment: 0x56339ca1c9e0>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A-D.3 at A time point 2 is: 
#> A.2 ~ A.1 + B.1 + C
#> <environment: 0x56339ca1c9e0>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A-D.3 at A time point 3 is: 
#> A.3 ~ A.1 + A.2 + B.1 + B.2 + C
#> <environment: 0x56339ca1c9e0>
#> 

w <- createWeights(data = test,
                   exposure = "A",
                   outcome = "D.3",
                   formulas = f,
                   save.out = FALSE)
#> Error in weightit2cbps.cont(covs = covs, treat = treat, subset = by.factor ==     i, s.weights = s.weights, missing = missing, verbose = verbose,     ...): The package "CBPS" is required.

epochs = data.frame(epochs = c("Infancy", "Toddlerhood"),
                    values = I(list(c(1, 2), c(3))))
e <- apply(expand.grid("A", as.character(epochs[, 1])), 1,
           paste, sep = "", collapse = ".")
test$weights <- w[[1]]$weights
#> Error in eval(expr, envir, enclos): object 'w' not found

g <- getModel(d = test,
              exposure = "A",
              exposure_time_pts = c(1, 2, 3),
              outcome = "D.3",
              epochs = epochs,
              exp_epochs = e,
              fam = gaussian,
              model = "m0")
#> Error in model.frame.default(~weights, data = structure(list(ID = 1:50,     A.1 = c(-0.959178284414148, -0.344572560485694, -0.353586741512312,     -3.14695311936398, -1.03501916879966, -0.257235174490716,     0.750793058902121, -0.698376819100523, -0.713749633123377,     -0.0648593798664712, 0.736235439283136, -0.51789080720145,     0.361417878586087, -1.43928326524219, -0.212808281641557,     -0.641530973140392, 1.68850754829722, -0.772034622248759,     1.16200881771387, -2.12057614006998, 0.99633037728299, -1.36273598241008,     0.120341536250197, 0.476481336834866, -0.119614926198441,     1.5217416158383, -0.983596437596931, -0.0863649504740911,     2.77581864202004, 1.08180409872695, 0.0861107332423, -1.58385214878161,     -1.56830186601282, -0.948464518320629, 0.229632286352552,     -0.124868744752883, -0.699552638695826, 0.157061153279808,     -0.716418594086713, -0.585638683300616, 1.44258900483174,     1.62469759929399, -0.325542451614919, -0.406160660160792,     0.151575959075558, 1.83173099342848, -0.113686468412095,     -1.03765759543754, 1.17090788407107, 0.320697131640352),     A.2 = c(0.712562077547777, -0.36915597183024, 0.895847437528041,     -0.595627000940859, 1.02058679939521, 0.391136766976401,     -1.26116500351854, -2.26132016307821, -2.37592111665449,     1.21609108727261, 0.796287509377843, -1.97058823098847, 0.924404767177973,     -0.785902095727585, -1.11022367641035, -1.44839619998385,     -1.17355855682059, 0.705153986910406, 1.78518842624446, 0.388174497026432,     -0.643490790276525, -1.52721773667788, -2.8461367620691,     -0.332859572139974, -0.318859341150629, -0.102163813837145,     -2.225275606062, -1.42246967960532, -1.77125869067483, -2.2460922275099,     0.169751781194465, -0.309453292690901, -0.360064977263428,     0.189016953323689, 0.077426467110261, 0.431570361774743,     0.520804634801927, 0.0410072370227668, -0.904182648128348,     1.12362315190356, -0.110788343931107, 0.651315819823687,     -0.282101420625614, 1.027462658867, -0.536482454411534, 0.395063886423672,     0.685926472375798, 0.12769502224265, 0.748002196627931, 0.509605372082963    ), A.3 = c(0.341742196474228, -0.521392150013809, -0.517536072183286,     0.852469228623431, -1.04591619442821, 0.320146989600097,     -0.555899010798018, 1.17651564263507, 1.33561682055792, 0.130650602638632,     1.35234320488492, 0.542733857622772, 0.143064289606842, 0.0871015498733637,     1.17582602766745, 1.99987066820685, -0.592174926861781, 1.27705305952224,     1.45152385081465, -1.58419658818574, -0.652904760892918,     2.60706295965568, -1.07045225041049, -0.620823010914776,     1.22607418813111, 0.028089542814643, 1.54368197688243, -1.67073194627987,     -0.124828071834947, 1.02969211342394, 0.868686674946237,     0.373222557824687, 1.26920628222727, -0.742588906903473,     -1.77779731659949, 0.476012305190347, -1.02212708578187,     0.955373327040163, -1.34256735159531, -1.95661823570911,     0.972302149457726, 0.505019033137127, 1.15334509995887, -0.549500909234674,     0.296461363696025, 0.665942645671326, 0.0390239304280915,     -1.04413065638998, 0.718892312409745, -1.33806650754645),     B.1 = c(0.379245773290567, 0.989381013858169, 0.334046820623707,     -1.12304073074012, 0.819505985593038, 0.477296122578835,     -0.462197509439014, 0.342019271948034, 0.96683644129569,     0.112240633545889, -1.60797765369934, -0.862242797953142,     -0.908588858007145, 0.533045472738885, 0.12093249650726,     1.7544273992435, -0.976974203631928, -1.03728819868612, -0.489354064333907,     -1.68347433109025, -0.499834764300024, -1.19892481489715,     0.675462466855994, 2.02196554236307, -0.945349216999145,     0.990169152344249, -0.735763940353952, 0.205700592820653,     1.10781973828781, -0.646282237360272, -1.02658961118846,     1.41199003553823, -0.67021035485433, -0.508535238980147,     -1.57422825973394, 1.47222249162912, 1.48943349741118, 1.15705592869676,     0.493374717400478, 1.44377196864697, -0.986608283710051,     -0.300386563292312, -1.3937305461936, -1.62338553162839,     0.381067453995783, 0.170840328578402, 0.444386495506425,     -0.0299702159639961, 1.1018808021716, -0.642210182371262),     B.2 = c(-1.11243664331411, 0.310942713686836, -1.01837657809334,     -0.678961600121079, 0.076221098823115, -0.455685122695582,     -1.0904581982572, -0.297223745934149, -0.519098431857722,     0.371361650398151, -1.58696927286146, 0.0308821539544463,     1.006235810091, 1.07820476421865, -0.36444021892721, -0.33224027285304,     -2.55064790469303, 0.0685227319018729, -0.844501549208418,     -0.458178060198892, -0.550403433743411, -0.826119744635628,     0.356901824734508, 0.0497999154478436, 0.482865819841538,     1.58483231183148, 0.498341275913848, -0.435955925571016,     -1.08031629208305, -0.723516268894845, 0.265847231237074,     0.48493063604155, -0.618189042137945, -2.2654370914645, 1.19811574552997,     0.042816531348534, -0.13460237773457, -1.54441983797158,     -0.697307336152476, -1.59846229569322, -0.650900331905784,     0.804665591128999, -0.745766285021161, 0.335422228621634,     0.342096640359711, -1.60430740467576, -1.32764802948259,     -0.655487224603245, -0.690494140266918, 1.94682998183481),     B.3 = c(-1.105543212741, 1.04421150642247, -0.182297501639781,     -0.159915339253201, 1.85758496536438, -1.10252985148676,     -0.916528718432512, -0.249373277961517, -1.56317528394741,     0.678576934331333, 0.45631332788022, -0.671019942444433,     -0.725201158857781, 0.0216101010115252, 0.379550429805014,     0.231345556297217, 0.443217774238362, -0.749547713858461,     0.407287370228294, 0.195763953538303, -0.380698087913916,     0.651512737931277, -1.46819013991627, -1.19202779095913,     -1.41403148761533, -0.116380647188062, 0.260163818981671,     0.432288833540144, -1.3973551192045, -1.48610278731871, 0.409702546432508,     -0.295385547265452, -0.92702587148998, -0.717025082259378,     0.590191020864883, -0.753721054703788, 0.234423879174453,     -0.637034950133797, -0.588901963199455, -0.00246154857123374,     0.778431767967784, 2.84161392461564, -0.133179341273258,     -0.299278281811781, -0.107307693969859, -0.748324661198462,     -0.543650508079289, 1.17564426045281, 0.657914719187124,     -1.54187092344023), C = c(-0.0196282251446571, -0.608094987776006,     -0.104629079778135, 0.243556770172841, 0.100738373851175,     0.626709949973057, -0.261686917172044, 0.653722338553335,     0.364341163297751, -0.365393238868813, 0.253557485412269,     -1.09548885407805, -0.56682801434722, -0.892698291035217,     -2.69364515634371, 1.08964468860013, 0.213439972418765, 1.32181748152518,     -0.452866398686231, -0.254099891818169, 0.499355291863323,     -1.9348532630792, 1.19557313879732, -0.287628661712733, 0.485319443747821,     -0.583379786239393, 0.899950804885611, 0.0539545420368654,     -1.98460057770249, -1.4011547540093, 1.58028107048546, 0.164942634015424,     -0.0589561725300945, 0.491610150093611, -0.974476811093363,     1.73412644057726, -0.309488569254542, -0.205837876546423,     -0.147281069805799, -0.00490430667435454, 0.14765912167522,     -0.413785458724144, -0.386045250359692, 0.702244111868944,     0.00898393969347266, 1.06736521801689, -0.765397487161658,     -1.18024307161814, 1.06706183827731, -0.871701298095344),     D.3 = c(0.404817227824457, 0.904700186522429, -0.993544469025426,     0.0940255973270813, -0.274298120351041, 0.30808658924782,     -0.0628734313574088, -0.0409981095972546, 0.920003558560965,     -0.215915694378629, 0.686775958855456, 0.609357733695384,     -0.790633513177101, 0.199860888340363, -0.343181177833134,     -0.812609878035965, 1.86031718514514, -1.43124062546393,     2.77736573489381, 0.615185008785972, 1.61234280918263, -0.648681770284068,     0.484555832535666, 0.155752986293748, -0.195241736263377,     -0.436953954889413, -0.496199166735571, -0.0693831170727364,     0.386538405961782, 0.690807388083996, -0.232769540652469,     -0.0148848233660243, -0.21759659496871, -1.98193920354862,     -0.62998561945588, 0.163471804506272, -0.641925983218546,     -0.210904763756554, 0.0121823617926417, -1.78992704083047,     0.303337700622244, 0.240683587861177, -0.361920786264095,     0.9582151877886, -2.03228128927787, -0.122134844649675, 0.76453499479238,     1.71832981976418, 1.29176675080275, 0.787438376282101), A.Infancy = c(-0.123308103433185,     -0.356864266157967, 0.271130348007865, -1.87129006015242,     -0.00721618470222496, 0.0669507962428425, -0.255185972308209,     -1.47984849108937, -1.54483537488893, 0.575615853703069,     0.766261474330489, -1.24423951909496, 0.64291132288203, -1.11259268048489,     -0.661515979025953, -1.04496358656212, 0.257474495738315,     -0.0334403176691765, 1.47359862197917, -0.866200821521774,     0.176419793503233, -1.44497685954398, -1.36289761290945,     0.071810882347446, -0.219237133674535, 0.709788901000577,     -1.60443602182947, -0.754417315039706, 0.502279975672605,     -0.582144064391475, 0.127931257218382, -0.946652720736256,     -0.964183421638124, -0.37972378249847, 0.153529376731406,     0.15335080851093, -0.0893740019469495, 0.0990341951512874,     -0.810300621107531, 0.268992234301472, 0.665900330450316,     1.13800670955884, -0.303821936120267, 0.310650999353104,     -0.192453247667988, 1.11339743992608, 0.286120001981852,     -0.454981286597445, 0.959455040349501, 0.415151251861657),     A.Toddlerhood = c(0.341742196474228, -0.521392150013809,     -0.517536072183286, 0.852469228623431, -1.04591619442821,     0.320146989600097, -0.555899010798018, 1.17651564263507,     1.33561682055792, 0.130650602638632, 1.35234320488492, 0.542733857622772,     0.143064289606842, 0.0871015498733637, 1.17582602766745,     1.99987066820685, -0.592174926861781, 1.27705305952224, 1.45152385081465,     -1.58419658818574, -0.652904760892918, 2.60706295965568,     -1.07045225041049, -0.620823010914776, 1.22607418813111,     0.028089542814643, 1.54368197688243, -1.67073194627987, -0.124828071834947,     1.02969211342394, 0.868686674946237, 0.373222557824687, 1.26920628222727,     -0.742588906903473, -1.77779731659949, 0.476012305190347,     -1.02212708578187, 0.955373327040163, -1.34256735159531,     -1.95661823570911, 0.972302149457726, 0.505019033137127,     1.15334509995887, -0.549500909234674, 0.296461363696025,     0.665942645671326, 0.0390239304280915, -1.04413065638998,     0.718892312409745, -1.33806650754645)), row.names = c(NA, -50L), class = "data.frame"), na.action = function (object, ...) {    if (NCOL(object) == 0)         object    else {        ok <- complete.cases(object)        if (all(ok))             object        else stop(message)    }}): object is not a matrix