Fits outcome model

Fits outcome model

getModel(
  d,
  exposure,
  exposure_time_pts,
  outcome,
  epochs,
  exp_epochs,
  int_order,
  model,
  fam,
  covariates,
  verbose
)

Arguments

d

wide data frame

exposure

name of exposure variable

exposure_time_pts

list of integers at which weights will be created/assessed that correspond to time points when exposure was measured

outcome

name of outcome variable with ".timepoint" suffix

epochs

data frame of exposure epoch labels and values

exp_epochs

list of exposure epochs

int_order

integer specification of highest order exposure main effects interaction for interaction models

model

character indicating one of the following outcome models:

"m0" (exposure main effects)

"m1" (exposure main effects & covariates)

"m2" (exposure main effects & their interactions)

"m3" (exposure main effects, their interactions, & covariates)

"covs" (covariate only model)

"int" (intercept only model)

fam

function specification for svyglm model

covariates

list of characters reflecting variable names of covariates for covariate models

verbose

TRUE or FALSE indicator for user output (default is TRUE)

Value

list of fitted model(s)

Examples

test <- data.frame(ID = 1:50,
                   A.1 = rnorm(n = 50),
                   A.2 = rnorm(n = 50),
                   A.3 = rnorm(n = 50),
                   B.1 = rnorm(n = 50),
                   B.2 = rnorm(n = 50),
                   B.3 = rnorm(n = 50),
                   C = rnorm(n = 50),
                   D.3 = rnorm(n = 50))
test[, c("A.1", "A.2", "A.3")] <- lapply(test[, c("A.1", "A.2", "A.3")], as.numeric)

f <- createFormulas(exposure = "A",
                    exposure_time_pts = c(1, 2, 3),
                    outcome = "D.3",
                    tv_confounders = c("A.1", "A.2", "A.3", "B.1", "B.2", "B.3"),
                    ti_confounders = "C",
                    type = "full",
                    save.out = FALSE)
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A-D.3 at A time point 1 is: 
#> A.1 ~ C
#> <environment: 0x556fc2f51560>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A-D.3 at A time point 2 is: 
#> A.2 ~ A.1 + B.1 + C
#> <environment: 0x556fc2f51560>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A-D.3 at A time point 3 is: 
#> A.3 ~ A.1 + A.2 + B.1 + B.2 + C
#> <environment: 0x556fc2f51560>
#> 

w <- createWeights(data = test,
                   exposure = "A",
                   outcome = "D.3",
                   formulas = f,
                   save.out = FALSE)
#> Error in weightit2cbps.cont(covs = covs, treat = treat, subset = by.factor ==     i, s.weights = s.weights, missing = missing, verbose = verbose,     ...): The package "CBPS" is required.

epochs = data.frame(epochs = c("Infancy", "Toddlerhood"),
                    values = I(list(c(1, 2), c(3))))
e <- apply(expand.grid("A", as.character(epochs[, 1])), 1,
           paste, sep = "", collapse = ".")
test$weights <- w[[1]]$weights
#> Error in eval(expr, envir, enclos): object 'w' not found

g <- getModel(d = test,
              exposure = "A",
              exposure_time_pts = c(1, 2, 3),
              outcome = "D.3",
              epochs = epochs,
              exp_epochs = e,
              fam = gaussian,
              model = "m0")
#> Error in model.frame.default(~weights, data = structure(list(ID = 1:50,     A.1 = c(-0.346553922959993, -0.622769855118892, -0.806530294204445,     -1.07433692032261, -0.410434044679758, 0.441067892562704,     1.00938047664618, 0.674588955316057, 0.65223609966396, 1.35095150264699,     0.384860531798744, 0.316721225343258, 1.59009101309922, -0.292164116161295,     -1.09525757275089, 2.19715193539106, -0.741245432979331,     -0.0401234424262566, 0.47449691513813, -1.43713989880213,     0.728949054624062, 0.566777513031112, 0.829688599144483,     1.25776155174735, 1.35850828721776, 0.90211730707699, -0.105988537719833,     0.333198602007681, -0.049614143222191, -1.79760276598169,     -0.870001375059833, -1.64564591627206, 0.889636110001417,     1.77014107039323, -1.57828696790559, -2.51931090802765, -0.131828124277959,     1.80361322051875, -1.94951739128338, -0.127533994209029,     1.76252715289902, -0.783679782668031, -0.790725468634383,     0.0839417274268609, 0.166780216891599, 0.71203904426456,     0.801361144530686, -1.03385263328979, -2.02111664677834,     0.749119870375818), A.2 = c(1.36638903106489, 1.70879014510705,     -0.062955214296611, 0.2517909964966, 0.305366517323834, -0.116750671278097,     -1.05250386589551, -0.909873626241322, 1.3125397529873, -2.11604873735685,     0.112193424737728, -0.151892683964199, -1.37517277028802,     -1.10000194110109, 0.601155025335372, -0.247132402376824,     -0.101696514106485, 1.55827210956589, 1.15529735128196, -0.584617071031756,     -1.08026526826995, -0.330327121583261, 0.220340852791157,     -0.478708054730003, -0.291433937925048, -0.939089528207953,     -1.0203485125259, -0.682227862432228, 0.131617972331356,     0.332587200848658, 0.0133032683710555, 0.48063954550635,     0.477628852621495, -0.769990681879596, 0.237290516181762,     -0.252635802861172, 1.71582998034446, -0.512751177827676,     -0.0976805376397656, -0.662013028568398, -1.05407262906514,     -1.16470138542863, 0.425036991609906, -0.0924680978380959,     1.02826955814597, 0.498375271394302, 0.256379647432872, -0.495926414453371,     0.255799484079764, 0.267847020697811), A.3 = c(1.13676899233306,     -1.10613750833316, 0.231331584770472, -0.168608024432754,     1.38850701136477, 0.27656965320257, 0.733965508458296, 0.958693766378354,     0.304537054887779, -2.40853740673755, 1.121109827496, 2.25717015025848,     0.563795218194217, -1.39247433475275, -0.401356513911052,     -0.3228714252591, -1.78216904255307, 3.11322945801592, 0.419247980924459,     2.12500777033793, 0.260601871216142, -0.299740867521143,     -1.03000887978437, 0.382003120456902, -1.00817897746004,     -0.0533688670725329, -0.433490552186331, -0.7160497403291,     0.333609584925598, 0.304058672334642, 0.269491605792624,     0.945264017533158, 0.204859510663772, -0.358590723314749,     1.87472825176886, 2.59731276919754, -1.23619819148838, -0.0349743942465159,     -2.35489826792904, 0.9769226704518, 0.344371182792477, 0.721889442458662,     -1.93340467192212, 0.252930571953711, 1.02765560244002, 0.413293127909248,     -0.453062308683513, 0.347822433342483, -0.277621686094269,     0.825751341118705), B.1 = c(-0.0126197547066312, -0.232194984911964,     -0.846468749232957, 1.26713456439139, 0.18877855204032, -0.446529577681946,     0.566601068772055, -0.703530947412144, -0.400770345365911,     -1.71149044110614, 0.214602686657672, 0.0972474721529545,     -0.740773474729163, 0.221842873583689, 0.174254028946507,     -1.05488536497321, 0.175483614257887, -0.380871108512374,     -0.403189365602523, -1.99953987841345, 0.744024593929851,     0.832624486374903, 1.6799126471563, -1.05224418126077, -0.689953927047499,     1.40771902444718, -0.692406729049356, 0.753310543877445,     -0.376284134035344, 1.6833001436836, 0.324340876840986, 0.167459667866946,     -1.08488490805163, -1.22317115085786, -1.90159191610442,     0.886082209156287, -0.10042785619544, 0.0075477715964708,     0.587415738419682, -1.34235511485509, 0.936625570077086,     0.514650690695098, -0.154308887363526, -0.338257738635429,     -0.0995322219499938, 1.91284135207762, -0.698191783817355,     -0.403805830584098, -0.170310025317179, 0.185794414577448    ), B.2 = c(0.888935354871692, 0.783811805779887, 1.31047322058215,     0.763205733081674, 0.321240393370365, -0.270367525345955,     -0.204642237184303, 1.62201809931046, -1.36302804894889,     0.293938607014751, -0.238340799232713, 1.03563824002337,     0.117783088915621, 0.108951245933997, 1.52065736179184, -1.50496203661997,     1.98953502414853, -1.47321836090711, -0.160803903109471,     0.967560061235086, 1.09158722287847, 1.46135452987845, 0.458561439139773,     1.77487611694261, -0.331830291530132, 3.33734394810175, 0.931222715361263,     -0.686916976289242, 1.3459479183012, -0.950206349670716,     2.79675950407185, 0.244639624727528, 0.228270782985004, 0.356691841266576,     0.775677892900527, 0.379098619992055, -0.907057730935508,     1.65732070166765, 0.399418169068634, 0.423213118302789, 2.41363494910669,     -1.77172850456754, -0.166785866932805, 1.64289106333814,     0.149346552912953, 0.618913292989817, -1.65033579965307,     0.630378343750512, 1.16514517885977, 0.950841070594502),     B.3 = c(-0.183121007795301, -1.38316202905326, 0.251272547166223,     1.94735012301349, 1.75442714749554, -1.32942412586362, 0.143708023083659,     -0.0950321517502263, -1.3847591250147, -0.787190282235961,     -0.435621560430781, -0.478137093916653, -1.31001942038679,     -0.0407611298782115, 1.02928385015626, 0.998553371985706,     -1.22186750852912, -0.902736996479268, 0.0658784014757282,     -0.12781434800775, 0.976642929772087, 1.26252403933602, -0.18500370531008,     0.0568561292970978, 1.13051423590015, 1.9322456080797, 2.26820630806047,     -0.826834034940548, -0.722990126956056, 0.420534627641491,     0.180929727642159, -0.111373535367124, -1.44702079957161,     -1.55700374766989, -1.24172826389387, 1.4537702716933, 1.60789394628144,     -0.773161642937442, -0.978981821460647, -0.524473790675127,     0.917399838975794, -1.87529924122014, -0.0544079400324088,     -1.26213392902138, -1.29455901094236, 0.416348589348425,     0.975291942879328, -0.0972190713260442, 0.0658307159370669,     -0.77971066910546), C = c(-0.648955007044037, -1.37421915324919,     0.446232038636144, -0.0424430023300944, 0.185520429832656,     0.319900980531844, -1.66506401844196, -0.406120160077306,     0.0379332908634042, -1.78130034744066, 0.202473482356945,     -1.13441423833942, -0.151657770705066, 0.345412902273998,     -2.07934526747841, -1.36162379674133, -0.1308813798102, 0.601838553959515,     -1.50149157858786, 0.328891465901555, 0.754279192027253,     1.67442485680985, -0.144582415547587, 0.437494229316643,     -0.341238728554894, 0.384570458421903, -0.931104963680175,     0.418518632910363, 0.325578310203487, -0.221129937916503,     -1.22027382309079, -0.482537455002693, -0.0683493167174857,     0.35818445630664, 0.874354904943934, -1.8894652793933, -0.297700798912566,     -1.3189965908784, 0.00465252058520191, 1.17098444955376,     -0.314138292925124, -0.790443839171136, 1.7214611166727,     0.354950216998035, -1.46280497885971, 1.05768179729643, -0.360581443553879,     0.546956068081258, 0.646474506424557, -0.64699465916303),     D.3 = c(0.46501837630421, -0.371211391673225, -0.643698608567112,     1.19050365801033, 0.572941136194939, -0.496621053749136,     1.1214146136187, -0.996450539028823, -0.161392880950108,     -0.0542832708881005, 0.242604067926738, 0.0300498983538038,     -0.284295101193018, -0.668152346754555, 0.545768914689046,     -0.427123909444461, -1.7954594614618, 1.10977252974256, 0.0336403485908864,     -0.473106275028348, 0.0937749326179459, -1.33133754168288,     0.194662782496826, -0.375191005408775, -1.16735896416576,     0.0788192675472849, -0.524425572901323, -0.126224503037416,     -1.70111050779848, 0.327923966201917, 0.184679775803943,     -0.722767524359618, -0.00972976849848632, 0.523942386711165,     2.13792175522199, -0.166670607740017, 0.152692746500754,     -1.4959394835978, -0.475521844098498, 0.199229244801089,     -0.386504890589694, -1.26721568883617, 2.11454389571811,     0.394706876241036, -0.16478095455639, 0.840019014156548,     -0.0618606567784986, 1.12209477315627, 1.17133634716409,     1.08116928234569), A.Infancy = c(0.509917554052449, 0.543010144994079,     -0.434742754250528, -0.411272961913005, -0.052533763677962,     0.162158610642303, -0.021561694624665, -0.117642335462632,     0.98238792632563, -0.38254861735493, 0.248526978268236, 0.0824142706895295,     0.1074591214056, -0.696083028631193, -0.247051273707759,     0.975009766507118, -0.421470973542908, 0.759074333569817,     0.814897133210045, -1.01087848491694, -0.175658106822944,     0.118225195723926, 0.52501472596782, 0.389526748508673, 0.533537174646356,     -0.0184861105654815, -0.563168525122867, -0.174514630212274,     0.0410019145545825, -0.732507782566516, -0.428349053344389,     -0.582503185382855, 0.683632481311456, 0.500075194256817,     -0.670498225861914, -1.38597335544441, 0.792000928033251,     0.645431021345537, -1.02359896446157, -0.394773511388714,     0.35422726191694, -0.97419058404833, -0.182844238512238,     -0.0042631852056175, 0.597524887518785, 0.605207157829431,     0.528870395981779, -0.76488952387158, -0.882658581349288,     0.508483445536814), A.Toddlerhood = c(1.13676899233306, -1.10613750833316,     0.231331584770472, -0.168608024432754, 1.38850701136477,     0.27656965320257, 0.733965508458296, 0.958693766378354, 0.304537054887779,     -2.40853740673755, 1.121109827496, 2.25717015025848, 0.563795218194217,     -1.39247433475275, -0.401356513911052, -0.3228714252591,     -1.78216904255307, 3.11322945801592, 0.419247980924459, 2.12500777033793,     0.260601871216142, -0.299740867521143, -1.03000887978437,     0.382003120456902, -1.00817897746004, -0.0533688670725329,     -0.433490552186331, -0.7160497403291, 0.333609584925598,     0.304058672334642, 0.269491605792624, 0.945264017533158,     0.204859510663772, -0.358590723314749, 1.87472825176886,     2.59731276919754, -1.23619819148838, -0.0349743942465159,     -2.35489826792904, 0.9769226704518, 0.344371182792477, 0.721889442458662,     -1.93340467192212, 0.252930571953711, 1.02765560244002, 0.413293127909248,     -0.453062308683513, 0.347822433342483, -0.277621686094269,     0.825751341118705)), row.names = c(NA, -50L), class = "data.frame"),     na.action = function (object, ...)     {        if (NCOL(object) == 0)             object        else {            ok <- complete.cases(object)            if (all(ok))                 object            else stop(message)        }    }): object is not a matrix