Fits outcome model

Fits outcome model

getModel(
  d,
  exposure,
  exposure_time_pts,
  outcome,
  epochs,
  exp_epochs,
  int_order,
  model,
  fam,
  covariates,
  verbose
)

Arguments

d

wide data frame

exposure

name of exposure variable

exposure_time_pts

list of integers at which weights will be created/assessed that correspond to time points when exposure was measured

outcome

name of outcome variable with ".timepoint" suffix

epochs

data frame of exposure epoch labels and values

exp_epochs

list of exposure epochs

int_order

integer specification of highest order exposure main effects interaction for interaction models

model

character indicating one of the following outcome models:

"m0" (exposure main effects)

"m1" (exposure main effects & covariates)

"m2" (exposure main effects & their interactions)

"m3" (exposure main effects, their interactions, & covariates)

"covs" (covariate only model)

"int" (intercept only model)

fam

function specification for svyglm model

covariates

list of characters reflecting variable names of covariates for covariate models

verbose

TRUE or FALSE indicator for user output (default is TRUE)

Value

list of fitted model(s)

Examples

test <- data.frame(ID = 1:50,
                   A.1 = rnorm(n = 50),
                   A.2 = rnorm(n = 50),
                   A.3 = rnorm(n = 50),
                   B.1 = rnorm(n = 50),
                   B.2 = rnorm(n = 50),
                   B.3 = rnorm(n = 50),
                   C = rnorm(n = 50),
                   D.3 = rnorm(n = 50))
test[, c("A.1", "A.2", "A.3")] <- lapply(test[, c("A.1", "A.2", "A.3")], as.numeric)

f <- createFormulas(exposure = "A",
                    exposure_time_pts = c(1, 2, 3),
                    outcome = "D.3",
                    tv_confounders = c("A.1", "A.2", "A.3", "B.1", "B.2", "B.3"),
                    ti_confounders = "C",
                    type = "full",
                    save.out = FALSE)
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A-D.3 at A time point 1 is: 
#> A.1 ~ C
#> <environment: 0x55871c77dab8>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A-D.3 at A time point 2 is: 
#> A.2 ~ A.1 + B.1 + C
#> <environment: 0x55871c77dab8>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A-D.3 at A time point 3 is: 
#> A.3 ~ A.1 + A.2 + B.1 + B.2 + C
#> <environment: 0x55871c77dab8>
#> 

w <- createWeights(data = test,
                   exposure = "A",
                   outcome = "D.3",
                   formulas = f,
                   save.out = FALSE)
#> Error in weightit2cbps.cont(covs = covs, treat = treat, subset = by.factor ==     i, s.weights = s.weights, missing = missing, verbose = verbose,     ...): The package "CBPS" is required.

epochs = data.frame(epochs = c("Infancy", "Toddlerhood"),
                    values = I(list(c(1, 2), c(3))))
e <- apply(expand.grid("A", as.character(epochs[, 1])), 1,
           paste, sep = "", collapse = ".")
test$weights <- w[[1]]$weights
#> Error in eval(expr, envir, enclos): object 'w' not found

g <- getModel(d = test,
              exposure = "A",
              exposure_time_pts = c(1, 2, 3),
              outcome = "D.3",
              epochs = epochs,
              exp_epochs = e,
              fam = gaussian,
              model = "m0")
#> Error in model.frame.default(~weights, data = structure(list(ID = 1:50,     A.1 = c(1.18695222210549, 0.807749918831592, 0.949344713260945,     1.18410521044554, 1.69977404389233, 0.590430873110255, -1.24407982761097,     -0.538778451211194, -0.388181148312388, 0.989649118539514,     -0.180431405110872, -0.235052116858415, -0.814531088308716,     -0.737663415953795, 0.74739845098463, -0.188358649585239,     -1.14438031388762, 0.237587326974365, 1.30537762400925, 0.686522238518508,     0.223057497702277, 0.565935546616263, -0.335928745807053,     0.594134793589883, 1.17254969764533, -0.752391775397338,     1.13523279223611, -0.736218520961245, -1.63928104440177,     0.202553875145203, 0.283274938806048, -0.701108937418706,     0.696923024598, 0.579518019425608, 0.153540370245625, 2.71635771297409,     2.20634670775491, -0.410128096314115, -0.286315080098774,     -0.740656861236601, -1.48064476512418, 0.300965236775746,     -1.05700896278496, 0.512821991639982, -0.0507647182109789,     0.0791520284144984, 0.651770781961091, -1.83315332824165,     -0.0681968231832358, -0.40147181148324), A.2 = c(1.05379756152695,     -0.580769612814275, 0.994954037734561, 0.796249841854399,     1.59428906160806, 1.96198595950774, -0.723689193857447, 1.6341234507569,     -1.76470039766319, 0.156428997219354, -1.31934174309598,     0.14047097744187, -0.564612495298621, -0.44382104727486,     0.183254568117309, 1.05003118952323, -0.365214217795718,     0.499461177790741, -1.0779742597065, -0.474732211895079,     1.5899122566636, -1.33588593686943, -1.05463307541943, -1.44973793886376,     0.901926054838716, 2.42301765386721, -0.301525927883518,     0.427634678539434, 1.54103173380874, 2.06164936755543, 0.0855640778991488,     0.144224601532455, 0.223709620730882, -0.651621657830288,     1.0788338627779, -1.16375026667997, 0.527983552256871, 0.218389276549327,     1.34539533527994, -0.0165178818936453, 3.20492704389849,     -0.959419228831211, -0.453834008138689, -0.565875952391763,     1.74175928569412, 1.26707434096491, 1.70539108775297, 0.157517239147237,     0.399763337362502, 0.15204998217353), A.3 = c(-0.750106998843856,     1.64140511673752, 0.0865000162482888, -0.639100909221846,     1.02368857554925, 0.922225024964584, -0.118650547014026,     -0.737676531215581, -0.62275423119201, 0.249745572987292,     1.21280021127817, -1.74994776121318, -1.18717125195556, -1.01544294008452,     1.82555339886061, -1.76951513105537, -1.03242096426707, -1.36036668679704,     0.677269252180403, -0.429347327109977, -0.665445488971177,     1.31616451071973, -0.237876042506639, -1.47506173923184,     -1.00408043821922, -0.527841948608461, 0.797962758829814,     -0.838258984833987, -0.532512494751816, -0.457737002087205,     -0.315473382104133, 1.36138456357406, -0.374037711190001,     0.643752901186531, -0.151103629297967, -0.62028402225094,     -0.942972367757903, -1.51055910951721, 0.149849873240316,     -0.423857659798822, -1.09215314739752, -1.49469929866363,     0.694576336932461, 1.38546881060818, -0.248265942209845,     0.682202496072833, -0.233912848288393, -0.515709577356009,     0.670826196978497, 0.378438302731969), B.1 = c(0.829351670017716,     1.10506836937273, -1.04176549584342, -0.332380811381242,     -2.4422412948075, -0.615312030180608, -0.401518795217517,     1.76769471192309, 0.520546163536354, -2.36368853368674, -0.802268302818618,     -1.46831130277485, 0.0532485670983976, -0.339016961081544,     0.894352960914054, -0.502300906654063, -0.447909131370347,     -0.950558280148694, 0.413311529049411, 0.224294131494573,     -0.444457606758555, 0.0425921791803123, -0.167385233362808,     -1.46311206516435, -1.86041232394877, -0.224651116368653,     1.11345255294394, -0.57024239001688, -0.485261944706893,     -0.138863823978952, 0.791615589596668, -1.41020268862977,     0.266160234773238, 1.61632290338268, 2.07708219193236, -0.664618979676815,     -1.24407764225221, -0.00828481311226451, -1.34311351132182,     -0.349788745740523, 1.15404272943271, -0.886205647315426,     1.06838071777213, -0.135829417903211, 0.937126556972367,     1.55045143622043, -0.235151759059573, 0.126485086277476,     0.476907443978726, 0.13936660319683), B.2 = c(-0.0529401476493045,     -0.309947568992898, 2.08509275087045, -0.076454723087738,     1.52032475692957, 1.57539193839222, 0.0931817243291067, 0.299092469985603,     -0.1230826364499, -0.586112145501621, -1.71591573917482,     1.01648759177537, -1.2117334045582, 3.12219717711716, 0.918031858394006,     0.504097393312782, -1.89122587282136, -0.522080309761985,     0.0779136101723419, -0.10662602286071, -0.186028645232224,     -0.495855692848327, 0.026834310329884, -0.391550495969195,     -0.923647344088814, -1.08397349588441, 0.497396598246156,     0.456994204748768, -1.26528785446691, 0.111226661620802,     0.357473782702603, -0.961990163687466, 0.180383358489902,     -0.234667229637495, -0.785433778072847, -1.21790245155467,     -2.85122827020647, -0.464658764438303, 0.261418909468837,     -0.139664529272918, 0.74694445719619, 2.22615833168145, 1.09752282492358,     -0.473455668245834, 0.338816114743265, 0.171666029251037,     -0.718650580754792, -1.7270497539998, 0.601312815417297,     0.852444750922077), B.3 = c(0.752315911599887, 0.326289201711061,     -3.66503610043553, -1.40061206439067, 0.340082069721473,     -1.02718205143056, -0.992968970123602, 1.14170725300544,     -0.236033354613437, -0.53742980745675, 0.0956788990638515,     1.61628489960614, -0.0957018089786848, 2.06962058674661,     -0.498071812848172, 0.246241454710674, 0.618285173810648,     0.846593935547014, 0.267156069029193, 0.414752135817002,     2.03096153152208, 0.85071643539616, 0.0898327355424327, -0.52581920386668,     -1.83261718867727, -1.88266916522925, -1.33102288272767,     0.194389806320942, 0.418837198139565, 0.622966616309784,     -0.0861211744662678, 1.11725899022309, -1.56111592646192,     0.667474049510809, -0.688015946494827, -1.36704956291344,     1.47558514363301, -0.200832799611173, 1.03864839694265, -0.228293702059337,     -0.734757799919904, -1.22457866920737, -1.69741048583756,     -1.20488768328951, -1.33049321830635, 0.863022498270815,     0.0776584314371942, 1.18390002880834, -1.45954239713659,     1.37753542814591), C = c(-0.0134877681245929, -0.739433067797364,     0.323574714313103, -2.43970847504039, -0.913156202849795,     -1.00290393687019, -0.599294881364083, 0.21694563155243,     1.18054959109108, -2.1861943581486, -1.01541427777928, -1.04829725109419,     -1.21298946083238, -0.949982618141253, 1.80760833734348,     0.238434480808944, 0.256665736729629, 0.864977044340758,     -0.323204528304618, 1.44536103942355, -0.115313794333965,     0.600945596140874, -0.0245234257686585, -0.0838473329277865,     -1.17307454464258, -0.00189308573734753, 0.230202966239565,     -1.07557855295253, 0.902975748310891, 0.134277377217152,     0.964246669085334, 0.243292992188617, 0.278773272113563,     -0.652131250761869, 0.43319118018156, 0.37812083132195, 1.18959265298997,     1.18818191667027, 0.323210477313557, -0.404408322837717,     -1.01462179932866, -0.261556673823405, -0.932029578240124,     -1.43208483979198, -0.755304453261165, -1.21711448955568,     1.50940455681691, 0.0315103205751204, -0.596952624496374,     0.223807465087062), D.3 = c(-0.956558507496997, -0.036814668163844,     -1.04918953649503, -0.467605980386396, -1.19230159438666,     1.68147788707639, 0.0720589659995475, -0.759579920772702,     0.591257958255505, 0.0629595282784773, -3.56624406204723,     1.36520127715979, -0.820208894626262, 0.733845751868385,     0.233584498851352, 0.120672508207761, 0.251650311522319,     0.688445364011904, 1.3683825724012, -0.49944730117842, 0.575459633763925,     1.57212470109013, 1.29006350288578, 1.39387019858134, 1.66839132060664,     -0.389649911913403, -1.34074758931077, -0.540848010918564,     0.233229762850806, 1.01730693216781, 0.756891426305519, 0.987894060777167,     0.0940851909280677, -0.404496325850744, 0.252579089921721,     -0.894240917873712, 0.494464717380507, -0.910033483066738,     0.415810618163943, -0.761883573538111, 0.172650761618341,     -0.163021920330769, -0.594319543199226, -1.09668032211909,     -0.0709093358806486, 0.946356794324938, -1.34905793104207,     -0.142193009410774, 1.08437146059873, 0.52219000226392),     A.Infancy = c(1.12037489181622, 0.113490153008659, 0.972149375497753,     0.990177526149969, 1.6470315527502, 1.276208416309, -0.983884510734208,     0.547672499772853, -1.07644077298779, 0.573039057879434,     -0.749886574103426, -0.0472905697082725, -0.689571791803669,     -0.590742231614327, 0.465326509550969, 0.430836269968995,     -0.754797265841669, 0.368524252382553, 0.113701682151375,     0.105895013311714, 0.906484877182938, -0.384975195126584,     -0.695280910613241, -0.427801572636938, 1.03723787624202,     0.835312939234936, 0.416853432176296, -0.154291921210905,     -0.049124655296515, 1.13210162135032, 0.184419508352598,     -0.278442167943125, 0.460316322664441, -0.03605181920234,     0.616187116511763, 0.77630372314706, 1.36716513000589, -0.095869409882394,     0.529540127590583, -0.378587371565123, 0.862141139387155,     -0.329226996027733, -0.755421485461824, -0.0265269803758905,     0.845497283741571, 0.673113184689704, 1.17858093485703, -0.837818044547206,     0.165783257089633, -0.124710914654855), A.Toddlerhood = c(-0.750106998843856,     1.64140511673752, 0.0865000162482888, -0.639100909221846,     1.02368857554925, 0.922225024964584, -0.118650547014026,     -0.737676531215581, -0.62275423119201, 0.249745572987292,     1.21280021127817, -1.74994776121318, -1.18717125195556, -1.01544294008452,     1.82555339886061, -1.76951513105537, -1.03242096426707, -1.36036668679704,     0.677269252180403, -0.429347327109977, -0.665445488971177,     1.31616451071973, -0.237876042506639, -1.47506173923184,     -1.00408043821922, -0.527841948608461, 0.797962758829814,     -0.838258984833987, -0.532512494751816, -0.457737002087205,     -0.315473382104133, 1.36138456357406, -0.374037711190001,     0.643752901186531, -0.151103629297967, -0.62028402225094,     -0.942972367757903, -1.51055910951721, 0.149849873240316,     -0.423857659798822, -1.09215314739752, -1.49469929866363,     0.694576336932461, 1.38546881060818, -0.248265942209845,     0.682202496072833, -0.233912848288393, -0.515709577356009,     0.670826196978497, 0.378438302731969)), row.names = c("1", "2", "3", "4", "5", "6", "7", "8", "9", "10", "11", "12", "13", "14", "15", "16", "17", "18", "19", "20", "21", "22", "23", "24", "25", "26", "27", "28", "29", "30", "31", "32", "33", "34", "35", "36", "37", "38", "39", "40", "41", "42", "43", "44", "45", "46", "47", "48", "49", "50"), class = "data.frame"), na.action = function (object,     ...) {    if (NCOL(object) == 0)         object    else {        ok <- complete.cases(object)        if (all(ok))             object        else stop(message)    }}): object is not a matrix