Fits outcome model

Fits outcome model

getModel(
  d,
  exposure,
  exposure_time_pts,
  outcome,
  epochs,
  exp_epochs,
  int_order,
  model,
  fam,
  covariates,
  verbose
)

Arguments

d

wide data frame

exposure

name of exposure variable

exposure_time_pts

list of integers at which weights will be created/assessed that correspond to time points when exposure was measured

outcome

name of outcome variable with ".timepoint" suffix

epochs

data frame of exposure epoch labels and values

exp_epochs

list of exposure epochs

int_order

integer specification of highest order exposure main effects interaction for interaction models

model

character indicating one of the following outcome models:

"m0" (exposure main effects)

"m1" (exposure main effects & covariates)

"m2" (exposure main effects & their interactions)

"m3" (exposure main effects, their interactions, & covariates)

"covs" (covariate only model)

"int" (intercept only model)

fam

function specification for svyglm model

covariates

list of characters reflecting variable names of covariates for covariate models

verbose

TRUE or FALSE indicator for user output (default is TRUE)

Value

list of fitted model(s)

Examples

test <- data.frame(ID = 1:50,
                   A.1 = rnorm(n = 50),
                   A.2 = rnorm(n = 50),
                   A.3 = rnorm(n = 50),
                   B.1 = rnorm(n = 50),
                   B.2 = rnorm(n = 50),
                   B.3 = rnorm(n = 50),
                   C = rnorm(n = 50),
                   D.3 = rnorm(n = 50))
test[, c("A.1", "A.2", "A.3")] <- lapply(test[, c("A.1", "A.2", "A.3")], as.numeric)

f <- createFormulas(exposure = "A",
                    exposure_time_pts = c(1, 2, 3),
                    outcome = "D.3",
                    tv_confounders = c("A.1", "A.2", "A.3", "B.1", "B.2", "B.3"),
                    ti_confounders = "C",
                    type = "full",
                    save.out = FALSE)
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A-D.3 at A time point 1 is: 
#> A.1 ~ C
#> <environment: 0x55cf865f4c90>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A-D.3 at A time point 2 is: 
#> A.2 ~ A.1 + B.1 + C
#> <environment: 0x55cf865f4c90>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A-D.3 at A time point 3 is: 
#> A.3 ~ A.1 + A.2 + B.1 + B.2 + C
#> <environment: 0x55cf865f4c90>
#> 

w <- createWeights(data = test,
                   exposure = "A",
                   outcome = "D.3",
                   formulas = f,
                   save.out = FALSE)
#> Error in weightit2cbps.cont(covs = covs, treat = treat, subset = by.factor ==     i, s.weights = s.weights, missing = missing, verbose = verbose,     ...): The package "CBPS" is required.

epochs = data.frame(epochs = c("Infancy", "Toddlerhood"),
                    values = I(list(c(1, 2), c(3))))
e <- apply(expand.grid("A", as.character(epochs[, 1])), 1,
           paste, sep = "", collapse = ".")
test$weights <- w[[1]]$weights
#> Error in eval(expr, envir, enclos): object 'w' not found

g <- getModel(d = test,
              exposure = "A",
              exposure_time_pts = c(1, 2, 3),
              outcome = "D.3",
              epochs = epochs,
              exp_epochs = e,
              fam = gaussian,
              model = "m0")
#> Error in model.frame.default(~weights, data = structure(list(ID = 1:50,     A.1 = c(-0.0400610278374667, -0.753102879360501, -0.118370548810848,     1.07657669095276, 1.05581193548678, 0.0594460061193935, -0.882516224096029,     0.414156889446986, -1.39168112836105, 0.733687210961904,     -1.35228102417146, 0.396917405027109, 0.615608403625386,     0.880873782667106, 1.29945511656465, 1.43784785554118, 1.34020474545729,     -0.400920936293805, 0.393333947514562, -0.288783200757707,     -0.701739326231301, 0.363460143356197, 0.744257803557274,     -0.209127106183862, -0.0100921777935358, -0.91178578412542,     1.62383939683864, 2.13793178430634, 1.963572398331, -0.0914878537519621,     -1.24636873778429, -0.658274728168395, -0.730082173971122,     1.44061189877704, -0.171529047990211, 0.302795683811923,     -0.71229782007933, 1.5916082229211, -1.26330627918042, -0.0950594470833334,     0.615580187554063, -0.556102296722466, 1.06030461541736,     -2.64615247665989, 0.9762015881251, -0.00672327281792632,     1.11149664815071, -0.827632119113731, 1.10883567800802, 0.877188118250129    ), A.2 = c(1.01141830410004, 0.240294167891023, -0.103624854227065,     0.546477070592841, 0.284820067332506, -0.860244278122326,     -0.521559629645129, 1.26745972005191, -0.804764313289251,     1.77793303233398, 0.380961332980915, -0.103205408798577,     0.551033054740853, 0.0499949831795145, -0.273520355189503,     -0.321451763574057, 0.311144997660891, 0.761266958308041,     -0.86245767423609, -0.0811984454633089, 0.843788127845661,     -0.196540569602557, -0.736117794322849, -1.69287409993696,     -0.105551614945235, 0.623409384467476, -1.01390489374164,     -0.0284876867659018, 0.751818789883015, -0.490329013624545,     -0.365050858731065, 0.20488954813878, 0.17437823087902, 1.5891149274561,     -0.239724735543444, 1.66198334835221, 0.489632119269515,     0.00409574974078615, -0.385812544358659, 0.603873534173338,     0.701178728316481, 0.951643201013969, -1.10366608694582,     0.624924257064433, -1.11590442168562, 0.352745185849065,     0.363707760948617, 0.783692382873854, 0.190349257103378,     -0.919843710856639), A.3 = c(-0.0121801523948759, -0.970790210900303,     -1.56057452819449, 0.391619810737679, 0.0873858583569035,     0.0144266135442932, 0.986518894594654, -1.4502220072186,     -0.273877722468976, 0.0840805167582077, 2.06838774050663,     -0.0326351434657489, -1.44693338017919, -1.32305792800674,     0.12461726402245, 0.554954113340506, -1.02340581805054, 0.256105291824228,     1.06008723435079, -0.86517772549475, -0.442615725212612,     -1.36266026097688, 0.644543365340282, -0.166959608167107,     0.114785918298058, 0.625745135681049, 0.224073122585652,     -0.143896990733609, -0.249052471759193, -1.23608943797537,     1.13286628781217, -0.223283660750511, 0.180610736031202,     0.354284679638171, 1.02928837386481, -0.0975769493286014,     1.88066752585066, -0.449418763082696, 0.621187605173345,     1.72265458427993, -1.3591576104453, -0.348250688360225, 0.616053140245495,     1.33553426406305, -1.61749976233146, -0.802272347880136,     0.810821321851327, -0.26258392069423, 0.608941848240805,     -0.780340451152022), B.1 = c(0.278891565614189, 0.998448713402665,     -0.374236391996688, -0.904195600997224, -1.59242867490434,     0.740793996514348, -0.608039602118728, -0.25688889518765,     -1.74191565593413, 0.425325020812971, 0.999209323390832,     -0.540809259348764, -1.82551037781476, 0.0844661424079019,     0.779003542520002, -0.654785963092171, 0.846886693010609,     1.00069598360768, 0.275619013104535, -0.877453625882026,     1.04100453251466, 0.0765080719700702, -0.4045781671745, -0.152021447956687,     -0.807041311240021, -1.12774957486937, -1.92212253064588,     1.23010970239265, 0.0695622150600765, 0.0581122527132414,     -0.462770830506113, -1.15945959622628, -1.64671727095704,     -0.926458557267774, -0.764034703594565, 0.303955610040873,     0.782852983720455, 0.030211910023298, -0.216106508391452,     -1.33334153564256, 0.325374722006003, 0.26409853251691, -0.0321900826889887,     -0.260307288400946, 0.0121789626691298, 1.03927960611895,     0.833156062284502, 0.771837291903221, -1.73759066200391,     -1.78877853311476), B.2 = c(-1.43015594793368, 0.932359984097735,     -0.856125392474498, 0.520424766778777, 0.104326202964549,     -0.203913812560565, -0.0199444668233845, -1.23274741853202,     -0.0622408096298512, -0.590548883111608, 0.546109344413351,     -0.407904225707714, 0.0445805475625551, -0.927741886459765,     1.00592619588014, -1.85326035596206, 1.60174569712751, 0.238671947805481,     -0.286419763149584, 1.98767597024441, 0.722990486293774,     0.743776286796527, 1.09188806167183, -0.164739256775241,     1.05855732043242, -1.01837046366378, -0.0137270305973808,     -0.307748354667669, 2.35472558778082, -0.240118748561481,     -0.569502437705067, 0.269691623791565, 0.232243036896234,     -1.18565258841698, -0.632861185457484, 0.14731710031052,     0.915429892448387, 0.872741475490794, -0.906095386758674,     -0.0103970470086457, -0.629255413781423, -1.9399669892836,     0.302819366776766, -0.544045419556707, -0.845136123223181,     -0.229319317069214, 0.383468147897266, -0.670220950391419,     -1.39606596051654, -1.05466385059927), B.3 = c(-1.06781351547211,     0.0128940681449891, -0.275856358487608, -2.56092423859333,     -0.341587406528286, -0.604835469514252, 0.136455962246885,     -0.466399575510122, 0.324454437083034, -1.43318851068493,     0.833205048744565, -0.0620686405801605, 0.79755328591691,     1.32302443022261, -0.292255707584419, -0.587030320133059,     0.534359271168263, 0.96788453072463, 0.094165421564679, 0.229301454443218,     0.975625235968203, 0.435514659863022, 0.416669255120802,     0.287793301373243, -0.274554806188366, 1.19460654016503,     1.76979912741752, 0.854069562893364, -2.2666140050856, 2.34393855293323,     -0.486630160237567, -0.635133901550477, 0.346062723877987,     2.16717541477353, -1.44285711861107, 1.34442763347089, -2.48267898447114,     1.17848004167527, -0.641925676815678, 0.0128894433372174,     -0.172108850557581, -0.0511070099471468, -1.49056316545189,     0.986554226498241, 1.39478105630843, 0.772100955493786, -1.65627676744124,     -0.0306668914815045, -0.579258740631335, -0.989306383953057    ), C = c(-0.965798928052189, 1.75131593942003, -0.837263156036298,     -0.00673119611646814, 2.27605472262162, 0.327203775776181,     -0.0645837541942786, -0.943170618380061, 1.68863806072552,     0.47713795005375, -0.680959936482436, 2.36151298776827, 0.140596816133105,     -0.12356300789282, 0.0731512069336309, 1.33201552426219,     -0.484910619875239, -0.00490132196707641, 0.230960764234305,     1.32299682937527, 0.497425651365118, -2.33259349821011, -0.467203552399836,     -1.13588043081049, -0.279342230828677, 0.153188894652804,     -0.810547090791017, 0.447937344390438, 0.35280433680086,     -1.78713483552817, -0.380599886304277, 1.28508200663951,     -0.589035702137231, 2.0111888923475, 0.505548113619959, -1.24058050056741,     -1.20369613584757, -1.70696372833554, 0.0159730684843683,     -1.14929007830819, 0.225578952624363, -1.13890095007449,     -0.342494071805414, 1.20352066221481, 0.467321218067806,     -1.2273922050196, 0.417639593721892, -1.21349974661611, 0.201130074143511,     0.00448956378588508), D.3 = c(2.61013949465971, 1.18427015474261,     1.28120207134245, -0.22808185951598, 1.41542840857263, -0.744021520216692,     2.40911020383606, 0.955753013285075, -0.870388917270502,     -3.41024492713534, 1.49759949725592, -1.47249736561976, -0.144944315192386,     -1.10601478167732, -0.0850121467144648, 1.40561000712675,     -1.32646638264651, 0.90256621698278, 0.0404433444687302,     0.708040856872477, -0.62486802692335, -0.419284816838039,     -0.812967877123321, 1.47072747948775, 0.168365227624377,     1.12767996299442, 0.492404549388032, -1.50728384081848, -0.398820574331153,     1.42128964573822, -1.91866862506515, -0.13147591698495, -0.246597442487109,     0.531860936407251, -0.968246735912903, 0.453972262166776,     1.7552110435505, -0.0586345414055642, 1.61711018984513, -0.584121108319629,     -0.169324209172918, -1.63050658626376, -1.70620747917373,     1.03813717858197, -1.34491603046522, -0.784721541919188,     -0.590703622840071, -1.12447732649111, 0.283152137410666,     0.676286857903939), A.Infancy = c(0.485678638131287, -0.256404355734739,     -0.110997701518956, 0.811526880772801, 0.670316001409643,     -0.400399136001466, -0.702037926870579, 0.840808304749448,     -1.09822272082515, 1.25581012164794, -0.485659845595273,     0.146855998114266, 0.58332072918312, 0.46543438292331, 0.512967380687573,     0.558198045983561, 0.82567487155909, 0.180173011007118, -0.234561863360764,     -0.184990823110508, 0.07102440080718, 0.08345978687682, 0.00407000461721252,     -0.951000603060411, -0.0578218963693854, -0.144188199828972,     0.3049672515485, 1.05472204877022, 1.35769559410701, -0.290908433688254,     -0.805709798257678, -0.226692590014808, -0.277851971546051,     1.51486341311657, -0.205626891766827, 0.982389516082066,     -0.111332850404907, 0.797851986330943, -0.82455941176954,     0.254407043545002, 0.658379457935272, 0.197770452145752,     -0.0216807357642299, -1.01061410979773, -0.06985141678026,     0.173010956515569, 0.737602204549664, -0.0219698681199385,     0.649592467555699, -0.021327796303255), A.Toddlerhood = c(-0.0121801523948759,     -0.970790210900303, -1.56057452819449, 0.391619810737679,     0.0873858583569035, 0.0144266135442932, 0.986518894594654,     -1.4502220072186, -0.273877722468976, 0.0840805167582077,     2.06838774050663, -0.0326351434657489, -1.44693338017919,     -1.32305792800674, 0.12461726402245, 0.554954113340506, -1.02340581805054,     0.256105291824228, 1.06008723435079, -0.86517772549475, -0.442615725212612,     -1.36266026097688, 0.644543365340282, -0.166959608167107,     0.114785918298058, 0.625745135681049, 0.224073122585652,     -0.143896990733609, -0.249052471759193, -1.23608943797537,     1.13286628781217, -0.223283660750511, 0.180610736031202,     0.354284679638171, 1.02928837386481, -0.0975769493286014,     1.88066752585066, -0.449418763082696, 0.621187605173345,     1.72265458427993, -1.3591576104453, -0.348250688360225, 0.616053140245495,     1.33553426406305, -1.61749976233146, -0.802272347880136,     0.810821321851327, -0.26258392069423, 0.608941848240805,     -0.780340451152022)), row.names = c(NA, -50L), class = "data.frame"),     na.action = function (object, ...)     {        if (NCOL(object) == 0)             object        else {            ok <- complete.cases(object)            if (all(ok))                 object            else stop(message)        }    }): object is not a matrix