Fits outcome model

Fits outcome model

getModel(
  d,
  exposure,
  exposure_time_pts,
  outcome,
  epochs,
  exp_epochs,
  int_order,
  model,
  fam,
  covariates,
  verbose
)

Arguments

d

wide data frame

exposure

name of exposure variable

exposure_time_pts

list of integers at which weights will be created/assessed that correspond to time points when exposure was measured

outcome

name of outcome variable with ".timepoint" suffix

epochs

data frame of exposure epoch labels and values

exp_epochs

list of exposure epochs

int_order

integer specification of highest order exposure main effects interaction for interaction models

model

character indicating one of the following outcome models:

"m0" (exposure main effects)

"m1" (exposure main effects & covariates)

"m2" (exposure main effects & their interactions)

"m3" (exposure main effects, their interactions, & covariates)

"covs" (covariate only model)

"int" (intercept only model)

fam

function specification for svyglm model

covariates

list of characters reflecting variable names of covariates for covariate models

verbose

TRUE or FALSE indicator for user output (default is TRUE)

Value

list of fitted model(s)

Examples

test <- data.frame(ID = 1:50,
                   A.1 = rnorm(n = 50),
                   A.2 = rnorm(n = 50),
                   A.3 = rnorm(n = 50),
                   B.1 = rnorm(n = 50),
                   B.2 = rnorm(n = 50),
                   B.3 = rnorm(n = 50),
                   C = rnorm(n = 50),
                   D.3 = rnorm(n = 50))
test[, c("A.1", "A.2", "A.3")] <- lapply(test[, c("A.1", "A.2", "A.3")], as.numeric)

f <- createFormulas(exposure = "A",
                    exposure_time_pts = c(1, 2, 3),
                    outcome = "D.3",
                    tv_confounders = c("A.1", "A.2", "A.3", "B.1", "B.2", "B.3"),
                    ti_confounders = "C",
                    type = "full",
                    save.out = FALSE)
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A - D.3 at A time point 1 is: 
#> A.1 ~ C
#> <environment: 0x55a8c42f0458>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A - D.3 at A time point 2 is: 
#> A.2 ~ A.1 + B.1 + C
#> <environment: 0x55a8c42f0458>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A - D.3 at A time point 3 is: 
#> A.3 ~ A.1 + A.2 + B.1 + B.2 + C
#> <environment: 0x55a8c42f0458>
#> 

w <- createWeights(data = test,
                   exposure = "A",
                   outcome = "D.3",
                   formulas = f,
                   save.out = FALSE)
#> Error in weightit2cbps.cont(covs = covs, treat = treat, subset = by.factor ==     i, s.weights = s.weights, missing = missing, verbose = verbose,     ...): The package "CBPS" is required.

epochs = data.frame(epochs = c("Infancy", "Toddlerhood"),
                    values = I(list(c(1, 2), c(3))))
e <- apply(expand.grid("A", as.character(epochs[, 1])), 1,
           paste, sep = "", collapse = ".")
test$weights <- w[[1]]$weights
#> Error in eval(expr, envir, enclos): object 'w' not found

g <- getModel(d = test,
              exposure = "A",
              exposure_time_pts = c(1, 2, 3),
              outcome = "D.3",
              epochs = epochs,
              exp_epochs = e,
              fam = gaussian,
              model = "m0")
#> Error in model.frame.default(~weights, data = structure(list(ID = 1:50,     A.1 = c(-0.998180269695054, -0.832743722764948, 0.682041622791173,     1.26027179390602, -2.71232821233125, -1.00777056065818, -1.11066319506939,     -0.412786612680756, -0.967240475918654, 0.145312524843576,     0.382903632857379, 0.405040564933637, -1.07917268687908,     0.88360048106622, -0.339439374677199, -1.5156950729516, -0.845628930554663,     0.435690215526374, -0.0714609553663585, 0.530660518984383,     0.44550544845583, 1.18302259213871, -1.05067232654034, -1.33329696154771,     1.51709572868388, 1.72198624285223, -0.582981561092715, 0.989753934099912,     0.241848198908502, -0.276516439668964, 0.0927271463632328,     0.915633392118762, 0.665993363885946, 0.255790251521208,     1.81357598452179, -0.250253384991214, 1.34463714537779, -1.12679463946121,     -0.765066749206802, -1.07196215321187, 0.807632333943485,     -1.60945413763589, 0.176000546296459, -1.05230183226608,     0.574451847771349, -0.709260823753894, -0.188330419522677,     -1.91946922500151, -1.1960572950653, 0.185480079079473),     A.2 = c(0.589122686824734, 0.696065214219478, 0.0163954295305824,     0.884624324667453, 2.54572984951618, 1.34719241544077, 0.134968149504334,     -2.23516038650885, 0.0485174083515975, 0.264576009024729,     -0.998900602332022, -0.86462836458228, 1.02152261705287,     3.31097234099274, -0.014231233364871, 0.599816836288446,     2.30551305006297, 1.87877998669197, -1.05062171996938, -0.0677868836890717,     1.91193918560314, -0.737877138108634, -0.207058759664748,     0.285323754319013, 0.491636166942439, 0.671716883222183,     0.744223721877386, -0.336863898061003, 0.779819092133886,     0.300655113459717, -1.3422523181664, 1.83628916332437, -0.914036859758762,     1.50745579120332, 0.415827769915339, -0.614812488387839,     -0.581710952044077, 0.337879146512858, 0.202191724505779,     -1.44287199718438, -0.738232712944632, 0.706425730867783,     0.281961506152745, 0.884270475641617, -0.936449796883926,     0.130448420209638, -1.05096295264053, -0.66179202405569,     0.986573314220653, 0.241648527392136), A.3 = c(0.349187694617661,     -0.200269939879905, 1.9334516186281, 2.57549801558247, -0.582610083746661,     -0.968073009985471, -0.315899083846711, -0.372592546357032,     0.144910080351062, 0.926164182629391, -1.62752410068022,     -0.0972976064461776, -0.538444716058197, -0.436016362364983,     0.37175498263121, 0.950487277614436, -1.09031634761367, 1.5888109783921,     0.255916611498599, 1.1524233600222, 0.223727459020932, -1.26196741023199,     0.263999254932218, -0.773980581297019, 0.338949202172037,     0.678129243103169, -0.327259046004216, 0.542414932141102,     0.270619622093286, -1.09993566322367, 1.26782208816988, -0.388024322923136,     -1.10924452173061, -0.98584614616617, 0.541570598425265,     -0.852345020979757, 0.644917135618728, 0.625296296899878,     1.08208984550568, 0.159118854767889, 0.0228410265636957,     0.33005444391753, -2.25992176725694, 0.8350427995447, -0.434131895885335,     0.576486925573514, 0.579378437674327, 0.853311311806999,     -2.04365783472946, 1.90758088247735), B.1 = c(-1.44809641826974,     -0.190961724995855, -0.214579578436757, 0.808300098941123,     0.281001500007386, -1.09902511238473, 2.43625643484535, -0.313789960400356,     0.600914140881858, -0.0185420585301925, -0.794860939006536,     0.692976053404247, -0.794819849610314, -0.388176619950679,     1.17495864387428, -0.686200715842348, -0.808750620366424,     1.13378906678335, -0.470257762667313, -0.803807769033277,     -0.994074825419225, -1.08147280165548, -2.59991309961861,     0.655039474436534, -1.61568896597303, 1.43513464144556, -0.443233815793768,     -0.519370090384748, -1.04644058902034, 0.138626966854454,     0.678529227512875, -0.351966966183755, 0.80512669419569,     -0.0750567741100231, -0.00996860898082233, 1.54139639023827,     -2.25058451591508, 0.235144035646932, 0.780929687249864,     -0.00847753277745351, 0.35272547857523, 1.54951983411632,     -1.69530704135571, -0.0431060247944099, -0.0811851709973076,     0.876276778246331, -1.19025443671701, 0.000120977864303642,     1.38905687606462, -2.08906639562131), B.2 = c(-0.0710936816549833,     0.823601817282546, 0.603812800023062, 1.19016873893716, -0.858782832971751,     -0.990203325429115, 0.159052470298418, 1.39893317268245,     1.16148345949841, 1.25836027965436, -0.0863307500980649,     1.37397261692203, 0.606382272411104, -1.02389906843033, -0.865727546362955,     -0.584871763379115, -0.282620453955197, -0.225039239636567,     -0.247450088180841, 0.209216653333453, 1.54720845676341,     -0.277502656501065, -0.0917048869077266, 0.258856636642159,     1.77880099162888, -0.405006340722202, 0.249876758126572,     -0.544191253047051, -1.07539357026201, 1.12886825380667,     2.04861393909392, -0.289041020267688, 1.10095835751119, -0.0616801680537725,     -0.0272677642929879, 0.686447553019403, 0.478324660570492,     -0.282164891546668, 0.614602653730874, 1.62037349577514,     0.385154753715854, -0.573598767661919, -1.35345861578016,     -0.763921110751842, -0.629596582054612, 1.9446137701989,     1.07494477666314, 0.113165311513689, -1.2695320111717, -1.05971896328688    ), B.3 = c(-0.501779869080295, -0.604996902286456, 0.031671103538679,     -0.708852067490348, -0.217284690162645, -0.405297643983328,     0.771346236269251, 0.503424707534631, -1.2818720725201, 0.972359888817865,     0.912534192070219, 0.395122073119418, 0.904026064429089,     1.40020795506936, 0.015026130129608, -0.690910943226312,     -0.734779344602374, 0.160134772026132, 0.0256267082778891,     0.494170575504134, -1.10171375060667, -2.32460099946174,     -0.164254814805174, 0.700468517732962, 0.264427241078577,     -0.984107788753814, 0.361165651141314, 1.24759955748486,     0.496528038047416, -0.730917724261011, 0.0225764997903536,     0.987313962090172, 0.490116952859266, 0.0565643580053785,     -1.20242293914663, 0.95334701570473, 0.76284850044171, -0.262850904377495,     0.0366612387788148, 0.334553019468737, -0.903149157618236,     -1.55782486496439, 0.09117164430957, 0.0837381006181216,     -0.985043458371368, -0.173461014836435, -0.342604964613995,     -0.159211360545016, 0.547760320310976, -0.51186038629266),     C = c(1.82511585786374, 1.29880850898546, -0.726952447752023,     -0.0491801773356927, -0.564237610316413, 0.615951372557494,     1.69481289135554, -1.1083340685368, 0.700353678116454, 0.0877751713467383,     0.134979645172892, -1.29803238898606, 0.7077556034125, 1.07013948612116,     -0.311886692922934, -1.66423572514902, -0.420099402404442,     -1.19804478096769, 1.33735738256518, 1.53222461682263, 0.977266571457463,     0.271358240213696, 1.16280336284204, -0.845126671281965,     -1.09075545899249, 0.8888377139458, -0.932624831711189, 0.963785486253406,     -0.464940406547315, -2.08673299113778, -0.68158297917022,     0.416279387568796, -0.513659900263312, -0.397201964115321,     0.10045763518596, -1.10032843484778, 0.650635729194379, -0.0783555994077153,     -0.71337992676382, 0.112690801586243, 0.394957073898473,     0.693929881282776, 0.625003543011512, -0.13411820607563,     0.38483228570943, -0.387141360425332, -1.12894761884803,     -0.0653638755232528, 0.624023783421861, 1.96945582211818),     D.3 = c(-0.974618414571545, -1.99041447841755, -0.794977902950695,     1.50787270288178, 2.0689897900973, 1.32488637222132, 0.0564308306168786,     0.169483946888011, 0.190630125129928, -0.820487765584966,     -0.895787308570053, 0.0210846324431603, -0.0125642979003035,     0.00604494261669225, 0.599419245276747, 0.452708472393879,     -0.229107201643837, 0.244953927577573, 0.474938422317423,     1.62985367993653, -0.0598670985978366, 1.06346465933516,     -0.41056496195677, 0.679270969373588, -0.258854143221463,     1.65503500522211, -0.345023532339008, -0.0394748543953658,     0.732029973136291, 1.09390397382273, 1.16584811618478, -1.05431654606883,     0.25061015741187, 1.27816831470701, -0.967743155631101, 0.196479638702842,     -1.20741655606837, 0.650513333520336, 0.809780966574015,     0.156970058932593, 1.10874211633106, -1.46166931969261, -0.787885680294151,     -1.02780570429843, 0.136791845973682, 0.402065554964939,     -0.338047299564921, -1.24637504536545, 0.190203253528841,     -0.345527670777181), A.Infancy = c(-0.20452879143516, -0.068339254272735,     0.349218526160878, 1.07244805928674, -0.083299181407535,     0.169710927391295, -0.487847522782528, -1.3239734995948,     -0.459361533783528, 0.204944266934153, -0.307998484737321,     -0.229793899824321, -0.0288250349131051, 2.09728641102948,     -0.176835304021035, -0.457939118331577, 0.729942059754153,     1.15723510110917, -0.561041337667869, 0.231436817647656,     1.17872231702948, 0.222572727015038, -0.628865543102544,     -0.523986603614348, 1.00436594781316, 1.19685156303721, 0.0806210803923355,     0.326445018019454, 0.510833645521194, 0.0120693368953765,     -0.624762585901584, 1.37596127772157, -0.124021747936408,     0.881623021362264, 1.11470187721856, -0.432532936689527,     0.381463096666857, -0.394457746474176, -0.281437512350512,     -1.25741707519812, 0.0346998104994265, -0.451514203384053,     0.228981026224602, -0.0840156783122315, -0.180998974556289,     -0.289406201772128, -0.619646686081603, -1.2906306245286,     -0.104741990422323, 0.213564303235804), A.Toddlerhood = c(0.349187694617661,     -0.200269939879905, 1.9334516186281, 2.57549801558247, -0.582610083746661,     -0.968073009985471, -0.315899083846711, -0.372592546357032,     0.144910080351062, 0.926164182629391, -1.62752410068022,     -0.0972976064461776, -0.538444716058197, -0.436016362364983,     0.37175498263121, 0.950487277614436, -1.09031634761367, 1.5888109783921,     0.255916611498599, 1.1524233600222, 0.223727459020932, -1.26196741023199,     0.263999254932218, -0.773980581297019, 0.338949202172037,     0.678129243103169, -0.327259046004216, 0.542414932141102,     0.270619622093286, -1.09993566322367, 1.26782208816988, -0.388024322923136,     -1.10924452173061, -0.98584614616617, 0.541570598425265,     -0.852345020979757, 0.644917135618728, 0.625296296899878,     1.08208984550568, 0.159118854767889, 0.0228410265636957,     0.33005444391753, -2.25992176725694, 0.8350427995447, -0.434131895885335,     0.576486925573514, 0.579378437674327, 0.853311311806999,     -2.04365783472946, 1.90758088247735)), row.names = c("1", "2", "3", "4", "5", "6", "7", "8", "9", "10", "11", "12", "13", "14", "15", "16", "17", "18", "19", "20", "21", "22", "23", "24", "25", "26", "27", "28", "29", "30", "31", "32", "33", "34", "35", "36", "37", "38", "39", "40", "41", "42", "43", "44", "45", "46", "47", "48", "49", "50"), class = "data.frame"), na.action = function (object,     ...) {    if (NCOL(object) == 0)         object    else {        ok <- complete.cases(object)        if (all(ok))             object        else stop(message)    }}): object is not a matrix