Fits outcome model

Fits outcome model

getModel(
  d,
  exposure,
  exposure_time_pts,
  outcome,
  epochs,
  exp_epochs,
  int_order,
  model,
  fam,
  covariates,
  verbose
)

Arguments

d

wide data frame

exposure

name of exposure variable

exposure_time_pts

list of integers at which weights will be created/assessed that correspond to time points when exposure was measured

outcome

name of outcome variable with ".timepoint" suffix

epochs

data frame of exposure epoch labels and values

exp_epochs

list of exposure epochs

int_order

integer specification of highest order exposure main effects interaction for interaction models

model

character indicating one of the following outcome models:

"m0" (exposure main effects)

"m1" (exposure main effects & covariates)

"m2" (exposure main effects & their interactions)

"m3" (exposure main effects, their interactions, & covariates)

"covs" (covariate only model)

"int" (intercept only model)

fam

function specification for svyglm model

covariates

list of characters reflecting variable names of covariates for covariate models

verbose

TRUE or FALSE indicator for user output (default is TRUE)

Value

list of fitted model(s)

Examples

test <- data.frame(ID = 1:50,
                   A.1 = rnorm(n = 50),
                   A.2 = rnorm(n = 50),
                   A.3 = rnorm(n = 50),
                   B.1 = rnorm(n = 50),
                   B.2 = rnorm(n = 50),
                   B.3 = rnorm(n = 50),
                   C = rnorm(n = 50),
                   D.3 = rnorm(n = 50))
test[, c("A.1", "A.2", "A.3")] <- lapply(test[, c("A.1", "A.2", "A.3")], as.numeric)

f <- createFormulas(exposure = "A",
                    exposure_time_pts = c(1, 2, 3),
                    outcome = "D.3",
                    tv_confounders = c("A.1", "A.2", "A.3", "B.1", "B.2", "B.3"),
                    ti_confounders = "C",
                    type = "full",
                    save.out = FALSE)
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A-D.3 at A time point 1 is: 
#> A.1 ~ C
#> <environment: 0x558c5a620eb0>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A-D.3 at A time point 2 is: 
#> A.2 ~ A.1 + B.1 + C
#> <environment: 0x558c5a620eb0>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A-D.3 at A time point 3 is: 
#> A.3 ~ A.1 + A.2 + B.1 + B.2 + C
#> <environment: 0x558c5a620eb0>
#> 

w <- createWeights(data = test,
                   exposure = "A",
                   outcome = "D.3",
                   formulas = f,
                   save.out = FALSE)
#> Error in weightit2cbps.cont(covs = covs, treat = treat, subset = by.factor ==     i, s.weights = s.weights, missing = missing, verbose = verbose,     ...): The package "CBPS" is required.

epochs = data.frame(epochs = c("Infancy", "Toddlerhood"),
                    values = I(list(c(1, 2), c(3))))
e <- apply(expand.grid("A", as.character(epochs[, 1])), 1,
           paste, sep = "", collapse = ".")
test$weights <- w[[1]]$weights
#> Error in eval(expr, envir, enclos): object 'w' not found

g <- getModel(d = test,
              exposure = "A",
              exposure_time_pts = c(1, 2, 3),
              outcome = "D.3",
              epochs = epochs,
              exp_epochs = e,
              fam = gaussian,
              model = "m0")
#> Error in model.frame.default(~weights, data = structure(list(ID = 1:50,     A.1 = c(-0.128308897641881, -1.0858001997289, -1.02612046019531,     -0.385208609812239, 0.928172057514051, -0.24501073110909,     -0.252383421614813, -0.432000593038222, 1.56770668757261,     -0.368707081544697, -0.334103507690447, -0.0156103592312692,     -1.28973676652603, -1.24672077382691, 0.822543674829224,     1.32849779049248, 0.0484832666735409, -1.07458836534714,     0.0777535150796239, 0.796445319716741, 2.95558437349811,     1.20442943121228, 0.192427506064848, -1.71995376945393, -0.172640749913231,     -0.499071777700849, -0.445347728335282, 0.516753077573648,     -0.83538063708846, -0.974865690566125, -1.66862839365614,     -0.195208803088512, -0.774319814092007, 0.22457412984773,     -1.02758753198892, -0.474313335510937, -0.0346099173474399,     1.24876415621838, 0.161338057482011, 0.873751772562972, -1.36927162264147,     -0.843448339860069, 0.839681055416256, -1.02433429023379,     -1.65795307684781, -0.0753505675072796, -0.384610443397682,     0.353704936245668, 1.33201847546493, -0.312269845114168),     A.2 = c(-0.143212292602582, 2.0430451158106, -0.158614707802498,     0.1964667883247, 0.680487152992169, -0.320066367119442, 0.583220488801594,     -0.607101040901512, 0.226817221811864, -0.244222962317924,     1.96174517581895, 0.753858105138214, 0.196168527640052, -0.0995231266195181,     -2.73583009948172, 0.882251819471324, -0.991331226789606,     1.75368158871871, -0.129584614463541, -1.03561072165369,     -0.564532825832055, -0.172622794048933, -0.040216196631122,     -0.86386114608479, 1.20260975391644, -0.0134922028153214,     -0.195143621104894, -0.0475039636260007, 1.27701151122762,     -1.34337377808164, -0.310380299730446, 0.187686052270565,     0.514489529381039, -1.65667166153127, -0.538575436113109,     1.55930915620719, 0.635252365518648, -0.730496040534794,     0.0166069424873051, 0.396836349164832, 0.524413617361251,     0.580509306396905, -0.174533853417402, 1.99282024766124,     1.10983355462775, 0.544792272430435, -0.684778709971904,     -1.47569527525187, 1.56702861986152, -1.15268964940507),     A.3 = c(-1.10550342339737, -1.00872493014948, 0.951426834586491,     -1.46809181975145, 0.393049591492504, -0.385711343782592,     0.0989669913431898, 0.246562565162158, 0.623388567477292,     -0.916275720927245, 0.312464168080061, -0.443375539518351,     0.133827080635848, 0.16629940022454, -1.94400261122621, -1.16886403887867,     -0.561895379568893, 2.56478931577004, -1.14057300330968,     0.663444292810713, -0.451348111295471, 1.45062400651166,     -0.489246251387616, -0.587992240186818, 0.923643329163157,     -0.9950665904289, 1.22384236425698, 0.88987959465783, -0.331343638992347,     -0.181781560559585, 0.883119363761548, -0.274166956269007,     -0.210219540194802, 1.29309543093285, 0.473949271414512,     0.682814502884572, 0.0651542427354941, 0.0406670782940565,     -0.193019936956229, 1.95052080668884, -1.25994744256099,     0.783187273108865, -0.424877816642998, -0.0100223089923874,     -1.07725361567255, -0.867565684845788, -0.0262202543130873,     -0.496995234554467, -1.76972296271339, 0.421258427576249),     B.1 = c(-0.120745203188767, -0.996606231220559, -0.388678020768697,     0.140473647027753, -1.3376667063094, 0.956847415349004, 0.113458170247414,     0.750354894524749, -0.509749801268746, 0.629782689817139,     0.290829216284356, 1.67319156830424, -0.0998866076060593,     0.348168091579682, 0.0803596712092673, -0.379151330367835,     -0.721366565037865, 0.799753141327868, -0.216009470686033,     0.401742895792608, -0.292410664082679, -1.03573803495145,     -0.746057331917428, -0.330897614325857, -0.838279871601864,     -0.0364276672086343, -0.125137255383909, 0.790467943924229,     -0.654135678796144, 1.97981672754796, 0.0604868900705832,     0.975364476585064, -1.53015553115475, -0.872915526673813,     -0.0166506334503518, -1.27559888922029, 0.870209176006466,     1.35550138693398, 0.739521398637561, 1.17375119487275, 1.63430581417361,     -0.828377658887483, -0.71368006654331, -1.04917783712565,     -1.05698426099032, -0.824073459647233, 0.958202253652363,     -0.628606195984644, 0.314697719994631, -1.24976114658357),     B.2 = c(-0.532216893070408, -1.28347525623574, -0.0478591752533972,     1.38289292395439, 1.12218783318909, -0.805270520494042, -0.471660497616076,     -3.54663511110284, 0.60578449377428, 0.37258567515695, -0.137973067230821,     0.521506183823194, -0.371468803493874, 0.657767522912654,     1.24281652459942, -0.751628055385643, -0.0507717755531494,     -0.329780844337879, -1.16448224736263, -0.915923533159872,     0.403274031493114, 0.163624331094302, -0.112395234784166,     0.841764618333542, 2.01496407918433, -1.65283876957065, 1.4994555108281,     -1.30863670316778, -0.898019652440045, -0.646873771193177,     0.466859910588797, 2.44737626255852, -1.70830803325211, 0.412072180224031,     -0.436985050972009, -0.832942172306372, -0.695328731493346,     -0.934673995180135, -0.565703844255549, 0.310297968490697,     -0.168010916173511, 2.17214909410838, 1.23830388133965, 1.0728302046627,     0.0351991028021629, 0.252796199674705, -0.5154651410183,     0.465614485021484, -0.571653579573618, -0.279301383170547    ), B.3 = c(0.41504417649236, 0.511059439065875, 1.88876776146756,     0.170868114951523, -0.591540418477291, -0.0992611151459835,     -1.00443235960192, 0.509061336128745, -1.7614416625814, -0.346831156487243,     -0.198910337010389, 0.361664504808108, 0.620460746539421,     0.263306349772873, 0.674505088200632, -1.41601142798686,     -0.0366333282409011, -0.906040307062128, 0.610956639757964,     -1.4861429862079, 0.445631609015616, -0.10507908161388, 0.502158662549578,     0.742595796931578, -1.21824607982635, 0.955912246062191,     0.446737960637607, 0.714823706995723, 0.86545182717586, 0.551217920882586,     -0.201387314315574, 0.0124675701300321, -0.150907075466897,     1.18956323018592, -0.888918523678105, -1.61964253426491,     -0.386741690400879, 0.874780316945498, -0.964605392242192,     1.18479099593829, 1.52196655742564, -0.410731223157343, 0.466218241707988,     0.626949294419978, 0.115993480107628, 1.12254294257623, 1.16788195407949,     -0.554228327786456, -1.92930763964212, 1.59184586060487),     C = c(-0.428145314309538, 0.328107594252671, 0.101558025607514,     1.33526668160634, -0.510036045002973, 0.044351819163786,     1.72710360340054, -0.832577310198287, -0.167715133605583,     0.455112298410678, -0.718838935957452, -1.02192751094122,     -0.808398508175321, -1.21092274733365, -0.520628374378239,     0.386481839528367, 0.39799446921419, -1.74352851156946, 0.755089859489746,     1.63911807486452, -1.42414004330443, 0.706027493535128, -1.27003533159862,     0.0350296896373089, -0.569345099731588, -0.197478225355884,     1.02803366260559, 0.629578565441061, -2.53221680570412, -1.06007599093339,     -0.991227696966027, 0.276476934331861, -2.41662537440682,     -0.681941403323693, -0.50555972563873, -0.947682829253158,     -0.565589827311291, 0.681329560478898, -0.81943806727359,     2.4083995144039, -1.41238256862628, 0.0911569402552428, 0.382052089369633,     0.224712160015003, 0.221879857299438, 0.155753857594054,     0.229927538851537, -0.83999074401369, 1.28616538177855, 1.31524449523465    ), D.3 = c(-1.03581599579622, 0.262949652757751, -1.07707716999342,     -0.0105601042180673, 0.890758532771825, 0.304935243056958,     -0.863079167690277, 0.0705384846755313, 1.75433239065706,     -0.288181259870113, 0.560768356686459, -0.155043644728235,     0.796540358280569, 0.57100360323949, -1.64708688031278, -1.47610391383105,     1.16523863477583, -1.71327510479477, 0.488269211833382, 0.742810932883404,     -0.163544973948093, 0.576817038988468, 0.204879069591729,     0.587249065593644, -0.118804912366526, -0.794401164753918,     -0.865936002285489, -0.618716607010779, -1.64757442103843,     0.553312547571814, 1.3403289714656, -0.997962824117948, -0.708647824602231,     0.274324353390716, -0.73188351322768, -1.75873266343962,     -0.283775722365199, -0.0972541504114972, -0.687876944642005,     1.26414429015645, -1.14117433923926, -1.50168342666594, -1.21017343938161,     1.80164643422559, 1.39017429907629, -0.119955953226368, -0.824959080661774,     -0.560432190271646, 0.359571100046046, 1.31550373089824),     A.Infancy = c(-0.135760595122231, 0.47862245804085, -0.592367583998904,     -0.0943709107437695, 0.80432960525311, -0.282538549114266,     0.165418533593391, -0.519550816969867, 0.897261954692237,     -0.306465021931311, 0.813820834064251, 0.369123872953472,     -0.546784119442989, -0.673121950223214, -0.956643212326248,     1.1053748049819, -0.471423980058033, 0.339546611685785, -0.0259155496919585,     -0.119582700968475, 1.19552577383303, 0.515903318581673,     0.076105654716863, -1.29190745776936, 0.514984502001605,     -0.256281990258085, -0.320245674720088, 0.234624556973824,     0.22081543706958, -1.15911973432388, -0.989504346693293,     -0.00376137540897349, -0.129915142355484, -0.71604876584177,     -0.783081484051015, 0.542497910348126, 0.300321224085604,     0.259134057841793, 0.088972499984658, 0.635294060863902,     -0.42242900264011, -0.131469516731582, 0.332573600999427,     0.484242978713725, -0.27405976111003, 0.234720852461578,     -0.534694576684793, -0.560995169503101, 1.44952354766323,     -0.732479747259619), A.Toddlerhood = c(-1.10550342339737,     -1.00872493014948, 0.951426834586491, -1.46809181975145,     0.393049591492504, -0.385711343782592, 0.0989669913431898,     0.246562565162158, 0.623388567477292, -0.916275720927245,     0.312464168080061, -0.443375539518351, 0.133827080635848,     0.16629940022454, -1.94400261122621, -1.16886403887867, -0.561895379568893,     2.56478931577004, -1.14057300330968, 0.663444292810713, -0.451348111295471,     1.45062400651166, -0.489246251387616, -0.587992240186818,     0.923643329163157, -0.9950665904289, 1.22384236425698, 0.88987959465783,     -0.331343638992347, -0.181781560559585, 0.883119363761548,     -0.274166956269007, -0.210219540194802, 1.29309543093285,     0.473949271414512, 0.682814502884572, 0.0651542427354941,     0.0406670782940565, -0.193019936956229, 1.95052080668884,     -1.25994744256099, 0.783187273108865, -0.424877816642998,     -0.0100223089923874, -1.07725361567255, -0.867565684845788,     -0.0262202543130873, -0.496995234554467, -1.76972296271339,     0.421258427576249)), row.names = c(NA, -50L), class = "data.frame"),     na.action = function (object, ...)     {        if (NCOL(object) == 0)             object        else {            ok <- complete.cases(object)            if (all(ok))                 object            else stop(message)        }    }): object is not a matrix