Fits outcome model

Fits outcome model

getModel(
  d,
  exposure,
  exposure_time_pts,
  outcome,
  epochs,
  exp_epochs,
  int_order,
  model,
  fam,
  covariates,
  verbose
)

Arguments

d

wide data frame

exposure

name of exposure variable

exposure_time_pts

list of integers at which weights will be created/assessed that correspond to time points when exposure was measured

outcome

name of outcome variable with ".timepoint" suffix

epochs

data frame of exposure epoch labels and values

exp_epochs

list of exposure epochs

int_order

integer specification of highest order exposure main effects interaction for interaction models

model

character indicating one of the following outcome models:

"m0" (exposure main effects)

"m1" (exposure main effects & covariates)

"m2" (exposure main effects & their interactions)

"m3" (exposure main effects, their interactions, & covariates)

"covs" (covariate only model)

"int" (intercept only model)

fam

function specification for svyglm model

covariates

list of characters reflecting variable names of covariates for covariate models

verbose

TRUE or FALSE indicator for user output (default is TRUE)

Value

list of fitted model(s)

Examples

test <- data.frame(ID = 1:50,
                   A.1 = rnorm(n = 50),
                   A.2 = rnorm(n = 50),
                   A.3 = rnorm(n = 50),
                   B.1 = rnorm(n = 50),
                   B.2 = rnorm(n = 50),
                   B.3 = rnorm(n = 50),
                   C = rnorm(n = 50),
                   D.3 = rnorm(n = 50))
test[, c("A.1", "A.2", "A.3")] <- lapply(test[, c("A.1", "A.2", "A.3")], as.numeric)

f <- createFormulas(exposure = "A",
                    exposure_time_pts = c(1, 2, 3),
                    outcome = "D.3",
                    tv_confounders = c("A.1", "A.2", "A.3", "B.1", "B.2", "B.3"),
                    ti_confounders = "C",
                    type = "full",
                    save.out = FALSE)
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A - D.3 at A time point 1 is: 
#> A.1 ~ C
#> <environment: 0x558d8e180438>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A - D.3 at A time point 2 is: 
#> A.2 ~ A.1 + B.1 + C
#> <environment: 0x558d8e180438>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A - D.3 at A time point 3 is: 
#> A.3 ~ A.1 + A.2 + B.1 + B.2 + C
#> <environment: 0x558d8e180438>
#> 

w <- createWeights(data = test,
                   exposure = "A",
                   outcome = "D.3",
                   formulas = f,
                   save.out = FALSE)
#> Error in weightit2cbps.cont(covs = covs, treat = treat, subset = by.factor ==     i, s.weights = s.weights, missing = missing, verbose = verbose,     ...): The package "CBPS" is required.

epochs = data.frame(epochs = c("Infancy", "Toddlerhood"),
                    values = I(list(c(1, 2), c(3))))
e <- apply(expand.grid("A", as.character(epochs[, 1])), 1,
           paste, sep = "", collapse = ".")
test$weights <- w[[1]]$weights
#> Error in eval(expr, envir, enclos): object 'w' not found

g <- getModel(d = test,
              exposure = "A",
              exposure_time_pts = c(1, 2, 3),
              outcome = "D.3",
              epochs = epochs,
              exp_epochs = e,
              fam = gaussian,
              model = "m0")
#> Error in model.frame.default(~weights, data = structure(list(ID = 1:50,     A.1 = c(-0.102637035460888, -0.59351671894775, -1.27673423044699,     0.616743873585362, 0.155601459480244, -0.0195073633430896,     -0.624474241814916, -0.436820032811721, -0.149916680357223,     0.0889370828876759, -0.163095062210323, 0.797472758779313,     -0.559217810913308, -0.722540427740575, -0.392556472913571,     -0.425575109329466, 1.6508741544005, 0.58237993342624, 0.660644216169913,     1.4316974595881, 0.450058435799833, 0.731699666728636, -0.460728380194781,     -0.394296553026897, 0.411763583181747, -1.21147386558875,     -1.18672162245862, 1.14675745663006, -0.0394811968239069,     -0.29400776105399, 0.609416088949853, -0.212087289496225,     -0.761077604917367, -0.23442538324097, 0.213996503526746,     -0.603281740588753, -0.65461852344995, -0.971037659308809,     0.286975395311971, -0.255830580389476, -0.121038629514754,     0.389062214145422, 0.208978117782583, 0.689108577275735,     1.24748404183647, -0.341027348824265, 1.35824176565684, -0.510255437735953,     0.77277619000852, 0.13437959661549), A.2 = c(-0.528290646033751,     -0.173503158274175, -0.59879022757677, -1.2069484145026,     0.618237399043898, -0.190600548812005, 0.507533209565743,     -0.553244491171283, 0.664562840708163, 0.636778596567325,     -0.599450995432888, -1.81291542514415, 1.44644179204453,     0.366722579523181, -0.829893544495478, 0.740603023235457,     -0.737152077201771, -0.402177731368099, 0.309800424882883,     -0.623886440479107, 0.456354067303077, 1.19531646652029,     1.25263875775506, -0.593594180095996, -0.387067334036168,     0.0926412982555505, -1.07393347932245, -0.0797018826632438,     -0.34434927236442, -1.8401452286639, -1.09383987653988, 0.262850323495049,     -0.958087183552176, -1.41541793607257, -0.516700985125238,     1.27592016710392, -1.05995704836771, -1.02266995195879, -0.0197606095449391,     0.713865897023999, 0.27933975984704, 0.514118119666202, 0.545370369627648,     -1.53721089879543, 0.899418612822497, -2.17193159710921,     1.3978624440297, 1.25286229315042, 0.698347758235958, 0.810149043063989    ), A.3 = c(1.0706339507806, 0.0646409003041202, 0.476121513064278,     0.151132678108515, 1.7751947988724, 0.161549818053353, 0.75098682889789,     -0.263286573905514, -0.681117543320213, -1.1818116992078,     0.399378441652052, 0.561840160999938, -2.18722182912461,     0.361613105991902, 1.50907258319308, 0.129041284963085, 0.389957260055628,     -0.257195368777844, -1.13825414193485, -0.532537284796118,     0.625352834095896, -0.228556834011216, 0.726221863705752,     0.965891703469072, 1.00907064723735, -0.463941760493106,     0.146368185255155, 0.0247787175136598, 0.560700395488458,     -0.554297866281841, -0.987295988109119, -0.117336003096579,     0.198394581101648, -0.906739574865674, 0.121802337390529,     0.206226308817399, 1.16567949609508, 1.7910979245331, -2.40143241941735,     -0.109903901099592, 0.651250875541028, 0.1130789001852, -0.644329107728942,     -0.724460717643338, 0.81650586013871, -1.83790727479, -1.35183709892932,     -0.600193402597982, 0.186863204988443, 0.519790880039375),     B.1 = c(-0.225279956615807, -0.408090190230315, 1.90572527899065,     -2.48079571804452, -0.0120922031372846, -1.15891372082353,     0.30743899273776, 1.97452482131895, -1.19731730909226, 0.920063419522276,     -0.382464053722545, -0.77367741864387, -0.92122312588343,     -0.1357047981379, -1.64281461472819, 1.135989481531, 1.49808565941735,     -0.0995953624376518, -1.68808967471796, -1.43477151476512,     0.027228434791587, -1.492897333871, 0.257573830856414, 1.53899676570454,     0.362230420061141, -0.638630178007132, -0.129570355403188,     0.974678672637497, 0.159836432028483, -0.348235330694585,     0.360161886167877, 0.511479522952875, -0.0956898292035481,     -0.878466064864325, -0.120344844954953, -0.159444966490292,     -1.95173762689741, -0.7720854481966, -1.03550252692571, 1.42140835545115,     0.143912664853363, -1.62649762320684, 0.533742352875339,     0.414667045178255, 1.47091367762527, -0.990446251877872,     -0.047064016959916, 0.188948732186127, -0.320230605155177,     0.696256370979128), B.2 = c(-1.17501857275032, -0.455139438450494,     1.25012024071461, 0.803546734683621, -2.01380089826351, -0.644979939598432,     -2.76940729532883, 1.95907063864353, 1.26023887331691, 2.09994985059296,     -0.765904132209588, -0.385197313364974, -0.907861361230831,     0.409036205297491, -1.16287762164421, 0.212271944611593,     1.17466690944652, -0.547158317374855, -0.399012376940586,     -0.414486647956946, 1.30393132216948, 0.417222241805551,     0.00655915615682177, -2.00752909928355, -1.44218785969448,     0.331435757999897, 0.632581947879864, -1.06760392352181,     -1.38825014622004, 1.60815357378592, 0.312434897282775, -0.0411414134716972,     1.93197180765599, 0.588873002163526, -0.626399688387328,     0.410631521693622, -0.289236917264007, 1.14389926569605,     0.60381828174067, 0.156224712277466, -0.60864976383572, -1.15136419920065,     -1.19598265779737, -0.237004727752185, 0.2134661183531, 0.93317541100825,     -1.45254087916432, -0.257084430149236, -0.097183594748193,     -1.72382391382251), B.3 = c(1.36667893874389, -0.504099321945439,     -0.499212953470562, 0.433692821558687, -1.06297452438849,     -0.945835166589437, -0.415107468037888, 0.0661663872785276,     1.48657822475004, -0.96771801918734, 1.34243724191617, -0.0714584951638034,     2.56767332498748, 0.0879522184395555, 0.998482762923006,     -0.88248787942622, 1.30545198428492, -0.770126236446875,     1.79813327295279, 0.644478843752734, -0.700781590296288,     1.01549966685199, 1.06525218426234, 1.22056789079703, -0.334615826028504,     1.3648598578814, 0.520583936686458, 2.41851849322217, 2.1270103693148,     1.84985071923137, -0.758550036309206, -2.2899201035846, -0.947572260961531,     0.926120417859089, 0.192978881902151, -1.51590726561252,     0.147704596943778, -1.49088441973854, -0.802800524084848,     -0.821711888179896, -2.04983309372621, -0.115607906407123,     0.162962164084463, 0.641991426844229, -1.47062703411693,     1.8841152407786, 2.16130562373267, 0.259629512857066, 0.753258214044873,     1.11887569113393), C = c(-1.78880777946395, -1.00822760578111,     1.2833499792965, 1.05455767890905, -0.0314732603093212, -0.701741092107541,     -0.128028106984972, -0.123725044750463, -0.317062476844026,     1.219804865765, 0.522237864272988, 0.584476558002354, -2.62417229909955,     0.123469560211153, 0.376627840297917, 1.47173648420394, 0.584150814094616,     0.347933581128043, -0.937787993067639, -2.04935601455228,     0.683261697164211, 0.740273443724355, 0.416660176786263,     -0.00560842000103727, -1.17654178671655, 0.579868563732075,     0.573166641915777, -0.990841290240936, -0.022991046864821,     0.264701240060005, -0.867982318411106, -0.827612338653098,     -0.930980911020999, 0.587043978749159, -0.431493690175139,     -0.748624718102465, 0.630338805951549, 0.325906994441475,     -1.50853085010902, -0.207529092107634, -0.964081293413337,     -0.547434348329436, 1.50037442792921, -1.59776127309661,     -1.91662424425588, 0.884761687630734, 0.252001819860521,     1.41221623295843, -0.455153652036739, 1.68104117389584),     D.3 = c(-1.68468679281407, -0.570765122288637, -1.57049874604484,     1.7200699044003, -1.53076811540294, 1.12937857916405, 0.155295538280947,     -1.48740499596116, 0.940964358021124, 0.148192418170373,     0.524583339614653, 0.779513873003333, -0.950706983047059,     0.228599960200492, 1.20010179006661, 1.67631145142701, 0.116233549794953,     0.507796462006127, -1.87008269922148, -0.27695348266223,     -0.874287010663949, -1.41098330079547, 0.442101897592839,     -1.49257399565292, 0.460183680763713, 0.79513045401137, -1.51843245023764,     -1.34278539690478, -0.0204640469116197, 0.325020547480881,     0.131741287845062, -1.71557955842983, -0.70928439560703,     0.86835112727145, 0.961961806710016, -0.460710885882032,     -0.220481239108868, -1.18699074109258, -1.08816298075235,     0.890612075120588, -0.545477688276601, -0.0660146553132164,     0.335164526004038, 0.68700605011087, -0.401447552006398,     1.21835712771246, -1.00841764546926, -0.196133912153206,     -0.446043700501197, -0.385428921287237), A.Infancy = c(-0.315463840747319,     -0.383509938610962, -0.93776222901188, -0.295102270458619,     0.386919429262071, -0.105053956077547, -0.0584705161245865,     -0.495032261991502, 0.25732308017547, 0.3628578397275, -0.381273028821606,     -0.507721333182418, 0.443611990565611, -0.177908924108697,     -0.611225008704525, 0.157513956952995, 0.456861038599364,     0.0901011010290705, 0.485222320526398, 0.403905509554497,     0.453206251551455, 0.963508066624463, 0.395955188780139,     -0.493945366561446, 0.0123481245727895, -0.5594162836666,     -1.13032755089053, 0.533527786983408, -0.191915234594163,     -1.06707649485895, -0.242211893795014, 0.025381516999412,     -0.859582394234772, -0.82492165965677, -0.151352240799246,     0.336319213257584, -0.85728778590883, -0.9968538056338, 0.133607392883516,     0.229017658317262, 0.079150565166143, 0.451590166905812,     0.377174243705115, -0.424051160759848, 1.07345132732948,     -1.25647947296674, 1.37805210484327, 0.371303427707234, 0.735561974122239,     0.47226431983974), A.Toddlerhood = c(1.0706339507806, 0.0646409003041202,     0.476121513064278, 0.151132678108515, 1.7751947988724, 0.161549818053353,     0.75098682889789, -0.263286573905514, -0.681117543320213,     -1.1818116992078, 0.399378441652052, 0.561840160999938, -2.18722182912461,     0.361613105991902, 1.50907258319308, 0.129041284963085, 0.389957260055628,     -0.257195368777844, -1.13825414193485, -0.532537284796118,     0.625352834095896, -0.228556834011216, 0.726221863705752,     0.965891703469072, 1.00907064723735, -0.463941760493106,     0.146368185255155, 0.0247787175136598, 0.560700395488458,     -0.554297866281841, -0.987295988109119, -0.117336003096579,     0.198394581101648, -0.906739574865674, 0.121802337390529,     0.206226308817399, 1.16567949609508, 1.7910979245331, -2.40143241941735,     -0.109903901099592, 0.651250875541028, 0.1130789001852, -0.644329107728942,     -0.724460717643338, 0.81650586013871, -1.83790727479, -1.35183709892932,     -0.600193402597982, 0.186863204988443, 0.519790880039375)), row.names = c("1", "2", "3", "4", "5", "6", "7", "8", "9", "10", "11", "12", "13", "14", "15", "16", "17", "18", "19", "20", "21", "22", "23", "24", "25", "26", "27", "28", "29", "30", "31", "32", "33", "34", "35", "36", "37", "38", "39", "40", "41", "42", "43", "44", "45", "46", "47", "48", "49", "50"), class = "data.frame"), na.action = function (object,     ...) {    if (NCOL(object) == 0)         object    else {        ok <- complete.cases(object)        if (all(ok))             object        else stop(message)    }}): object is not a matrix