Fits outcome model

Fits outcome model

getModel(
  d,
  exposure,
  exposure_time_pts,
  outcome,
  epochs,
  exp_epochs,
  int_order,
  model,
  fam,
  covariates,
  verbose
)

Arguments

d

wide data frame

exposure

name of exposure variable

exposure_time_pts

list of integers at which weights will be created/assessed that correspond to time points when exposure was measured

outcome

name of outcome variable with ".timepoint" suffix

epochs

data frame of exposure epoch labels and values

exp_epochs

list of exposure epochs

int_order

integer specification of highest order exposure main effects interaction for interaction models

model

character indicating one of the following outcome models:

"m0" (exposure main effects)

"m1" (exposure main effects & covariates)

"m2" (exposure main effects & their interactions)

"m3" (exposure main effects, their interactions, & covariates)

"covs" (covariate only model)

"int" (intercept only model)

fam

function specification for svyglm model

covariates

list of characters reflecting variable names of covariates for covariate models

verbose

TRUE or FALSE indicator for user output (default is TRUE)

Value

list of fitted model(s)

Examples

test <- data.frame(ID = 1:50,
                   A.1 = rnorm(n = 50),
                   A.2 = rnorm(n = 50),
                   A.3 = rnorm(n = 50),
                   B.1 = rnorm(n = 50),
                   B.2 = rnorm(n = 50),
                   B.3 = rnorm(n = 50),
                   C = rnorm(n = 50),
                   D.3 = rnorm(n = 50))
test[, c("A.1", "A.2", "A.3")] <- lapply(test[, c("A.1", "A.2", "A.3")], as.numeric)

f <- createFormulas(exposure = "A",
                    exposure_time_pts = c(1, 2, 3),
                    outcome = "D.3",
                    tv_confounders = c("A.1", "A.2", "A.3", "B.1", "B.2", "B.3"),
                    ti_confounders = "C",
                    type = "full",
                    save.out = FALSE)
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A-D.3 at A time point 1 is: 
#> A.1 ~ C
#> <environment: 0x5633ed1bcda8>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A-D.3 at A time point 2 is: 
#> A.2 ~ A.1 + B.1 + C
#> <environment: 0x5633ed1bcda8>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A-D.3 at A time point 3 is: 
#> A.3 ~ A.1 + A.2 + B.1 + B.2 + C
#> <environment: 0x5633ed1bcda8>
#> 

w <- createWeights(data = test,
                   exposure = "A",
                   outcome = "D.3",
                   formulas = f,
                   save.out = FALSE)
#> Error in weightit2cbps.cont(covs = covs, treat = treat, subset = by.factor ==     i, s.weights = s.weights, missing = missing, verbose = verbose,     ...): The package "CBPS" is required.

epochs = data.frame(epochs = c("Infancy", "Toddlerhood"),
                    values = I(list(c(1, 2), c(3))))
e <- apply(expand.grid("A", as.character(epochs[, 1])), 1,
           paste, sep = "", collapse = ".")
test$weights <- w[[1]]$weights
#> Error in eval(expr, envir, enclos): object 'w' not found

g <- getModel(d = test,
              exposure = "A",
              exposure_time_pts = c(1, 2, 3),
              outcome = "D.3",
              epochs = epochs,
              exp_epochs = e,
              fam = gaussian,
              model = "m0")
#> Error in model.frame.default(~weights, data = structure(list(ID = 1:50,     A.1 = c(-1.58930264777601, 2.17152133890946, 0.517697714400325,     -1.56483011982357, -0.306891014360125, 0.284939445285292,     -0.00703483503681466, -1.4144241335195, -0.763013139895808,     -0.930978127596624, 2.11527436742354, 1.15153766717214, 0.740299896396292,     -1.51811454494651, -0.84534155537573, -1.18749581664396,     -0.444904562222988, 0.180179257546698, -0.68162477953018,     0.0375585833814797, -0.0418530104106721, 0.940437548978913,     0.0953912457280553, 0.315218276467958, -0.361024721491304,     -0.767349016597063, 0.719238471676506, 0.34597646859207,     0.890587086974619, 0.226826560399017, 1.45368165102102, -0.595214913750289,     -1.56404772914581, 1.50438693800371, -0.816340474369459,     0.181034861986421, 0.59935070171957, -1.09818901953632, 0.828348343750846,     1.33988695330647, -0.657294848619263, 0.80142087204754, 1.43822783210193,     0.230133729589985, -0.897711344847957, -0.988841484578266,     0.824155274169267, -0.0124888840391744, 0.0832393438363656,     0.285832433652665), A.2 = c(0.869034525788524, -0.641661428429105,     2.77346358881505, 0.873536738758949, 0.325097678034912, -1.05813854517495,     -1.18750143329137, -0.443568890322371, -0.958719105343107,     -1.01830641174806, -1.92139251697287, 0.415344011859461,     -2.6261996268242, -1.06054891970052, 0.251707168292936, 1.44279363461155,     1.38865322635189, -1.60596049314668, 1.71583678067365, 0.5573391012548,     -0.0880085330393013, 0.650558821317929, 0.42146511344932,     -0.382724771544729, 0.653130645105271, -0.276843158603629,     -0.526558481407698, 1.89670982698632, -0.991766195119829,     -0.255878287169243, -1.54634037753972, 1.53988691758055,     0.567040389670396, -0.83675188278461, -0.709520794866483,     -0.441964871809317, 0.210154757003725, -0.626606787454812,     0.139465218782327, -0.0307540373410604, -1.11083933844469,     1.78802875674019, 0.806144041366328, -1.26196683459028, -2.86813143494505,     -0.678264882835125, 1.71220295836921, 1.24734035404833, 0.635286104276745,     -0.445396542792636), A.3 = c(-1.94582952666754, 0.2458758686613,     -0.114843323204434, -0.653326453193117, -0.521334163619601,     0.0895518387065466, -0.20766693631036, 1.52607590428074,     -0.4987728922839, 0.829371039510892, 0.654933684378242, 0.0407738301543971,     -0.184345135014995, 0.730522224660397, 0.0401616043276377,     0.138284707505347, -0.51161240952806, -0.420395851407496,     -0.77930210132708, -0.304127588171866, -0.206330927797828,     1.68396072996797, -0.195872608025302, -0.363655846288592,     -0.362644976479584, -1.40690446251939, 0.708187421507729,     -1.09054015489071, -1.62878223261469, 0.272939538254605,     0.455398563639628, -2.19236687363479, -1.23693548997951,     -0.339295084097027, 0.954588547568362, -1.24975404516971,     -1.89842880961347, -1.71919073302738, -1.12534405492328,     1.19032480263027, 0.243592304735778, -0.0456213363096137,     1.08140120135091, -2.63850288654585, 0.288548232937203, -0.591811628043639,     0.0856511672691946, 0.207549734450574, 1.33945552426082,     0.446491015138253), B.1 = c(-0.317852786534519, 1.02439178548001,     -1.32906636736432, 0.454221367129611, 1.22011041825047, -0.55861849496638,     0.756435588168867, -1.363022060985, -0.278909070631817, -0.503086336010277,     -0.606383147049464, 0.777794722225535, 0.891458203494815,     0.937117171802079, 0.52500268648661, 0.941615092003982, -0.571394227017386,     -1.72947456871503, 1.97049041993099, 0.617007314609249, 1.34405646739502,     0.277307332220028, -0.454728593239323, -0.399555770924559,     -0.388622869367602, 0.81973732052521, -0.203498223886707,     1.13101267011144, -0.139935017129929, -0.0644357954340664,     -0.459004059334197, -0.0879244393932382, 1.06031263460978,     0.589356918078947, -0.396444251526719, 1.6248367787596, -1.81343239497288,     1.5668282783938, 0.833849197934213, 0.568001808725748, -1.26814236049239,     -0.590120207360856, -0.3131088094231, -1.39727036491014,     1.55593401642105, 0.844230911028594, 0.930515796506696, -0.427356335793333,     -0.71045344991989, 0.341340605996876), B.2 = c(1.13605687116686,     -1.80669877999738, -0.386192677134792, -1.47428580066612,     -1.17103737846907, -0.274693381320639, 1.13484341954393,     -0.627940352339303, -2.39191018817606, -0.151204648491011,     -0.528380533240842, -0.114440614022184, -1.59007498216188,     -0.0906528595996188, 0.039565430813309, -0.443299489367471,     -0.0458488505327129, 0.694158435056777, -0.844268179315586,     -0.44120244475108, -0.36013895086631, -0.194715323128069,     1.06830989575419, -0.463381418486509, 0.824203771290852,     0.689074441258521, -0.053558230514437, 1.41588191205498,     0.475635568108296, -0.396110360402234, -0.591531744054196,     0.191755551082868, 0.104510889910309, 0.406390121472505,     -0.665110328382664, -0.126883585419766, -0.409345237393278,     0.431687595672872, 0.0773677885599168, -0.730021360980453,     0.65123816118206, 1.56414334225457, 0.587012566553526, 0.956071805848133,     -0.0219818811164311, 2.55507153124453, -0.117235319657568,     -0.035219996451382, 0.212047577721991, 0.382577247751971),     B.3 = c(0.248106018214656, 1.03656958119965, -1.22680880519097,     -0.188432520628738, -0.345466661493722, -0.697914319676059,     -0.580927160623479, 0.322133188817325, -1.13726516457543,     -0.564622058025671, -1.78511687375659, 1.33043447905513,     -0.43899981331098, -0.918502830098314, 0.40137376708888,     -1.00704070274794, 0.422343885088959, -2.14654871289307,     -1.65468640811699, -1.30307289152652, 0.435385714554553,     -0.365630715595706, -0.12681436305572, 1.11647213277023,     1.72355335413075, 0.28044935766135, -0.88794705175374, -0.190474707606376,     0.0290892928447505, 0.143929630614789, 0.211916326745955,     0.130769164561504, 0.310972496728816, -1.40063202550852,     0.900095561556837, 1.20800762532962, 0.188255894234301, 0.435677227240123,     0.337060663438342, 0.978560896645866, 1.8387337708594, 1.43555658505401,     0.43631833121976, -0.90362793932741, -0.395438609936904,     0.500077915834642, -0.0975603767118152, 0.437260465759578,     -1.52669400509797, -0.925798383943106), C = c(1.43519408224274,     1.39268445257108, -0.10753443636343, -0.453379192248388,     -0.885856252461505, 0.361109904578232, 0.490587576097009,     -2.29077361790779, 1.47859073995405, -0.843931912596761,     0.7831420825408, -2.53906827391446, 1.37182487010611, -1.19449709428233,     -0.34445266590111, 0.391908932236161, -1.12267909409288,     0.000625730761797975, 1.42525403833402, 0.818734649135335,     -2.15373448821375, 0.753140895262399, -0.676643002866142,     0.383294897411139, -0.651907311072084, 0.624122326844538,     -0.538818079568166, 0.383510032906455, 0.702954718256283,     0.433776192406573, -0.675886769841932, -0.0327861837692213,     0.0779238507333116, 0.320013641876986, -0.749438321233022,     0.266603772414301, 0.291414586657942, 0.0980259296274954,     0.97139424598124, 0.729561104323665, -0.253961962427695,     0.378327652896213, -0.379961129767786, 1.21363906428031,     -0.526863070369324, -0.335220825339059, -0.153266102699078,     0.0939334274976991, -0.0256954232497301, -0.150891206681365    ), D.3 = c(-0.436176186141253, 0.754412681019542, 0.57294533576882,     -0.920477893439581, 0.158750912808567, 1.12773791281286,     -1.14491182959722, 0.374866355179756, 1.42795492677676, 1.54222875167436,     0.0178633765630573, -0.683658388552031, -1.08264349932822,     0.655174103286336, -1.18962446797954, 0.919285933265075,     -1.86689946225185, 0.447953773825712, -0.122957944588354,     0.322878945915332, -0.704419872413747, 1.97260373036256,     0.466496667003915, 0.393016606179011, 0.612991804814677,     -2.2101971198475, 0.0942537187781635, -0.846324303039202,     -2.85316666453926, 1.54906174033897, 0.599968631481332, -2.77286094646519,     0.620633660427943, 0.170893487654993, 1.45872710956761, -0.972687832598359,     0.387162636352437, -1.67116078299894, 1.33588968407173, 1.50744866375286,     -2.65824754189334, 0.354025348475639, 0.291530493693034,     1.94005983714237, 0.126717756970619, -1.42606193971012, -1.08363078465794,     0.532997470387884, -2.3058004768165, -0.993155769978749),     A.Infancy = c(-0.360134060993743, 0.764929955240177, 1.64558065160769,     -0.34564669053231, 0.00910333183739351, -0.386599549944829,     -0.597268134164092, -0.928996511920936, -0.860866122619457,     -0.974642269672342, 0.0969409252253349, 0.783440839515801,     -0.942949865213954, -1.28933173232351, -0.296817193541397,     0.127648908983795, 0.471874332064451, -0.712890617799991,     0.517106000571735, 0.29744884231814, -0.0649307717249867,     0.795498185148421, 0.258428179588688, -0.0337532475383855,     0.146052961806983, -0.522096087600346, 0.096339995134404,     1.1213431477892, -0.050589554072605, -0.014525863385113,     -0.04632936325935, 0.472336001915131, -0.498503669737707,     0.33381752760955, -0.762930634617971, -0.130465004911448,     0.404752729361647, -0.862397903495566, 0.483906781266586,     0.654566457982705, -0.884067093531977, 1.29472481439386,     1.12218593673413, -0.515916552500147, -1.8829213898965, -0.833553183706696,     1.26817911626924, 0.617425735004578, 0.359262724056555, -0.0797820545699855    ), A.Toddlerhood = c(-1.94582952666754, 0.2458758686613,     -0.114843323204434, -0.653326453193117, -0.521334163619601,     0.0895518387065466, -0.20766693631036, 1.52607590428074,     -0.4987728922839, 0.829371039510892, 0.654933684378242, 0.0407738301543971,     -0.184345135014995, 0.730522224660397, 0.0401616043276377,     0.138284707505347, -0.51161240952806, -0.420395851407496,     -0.77930210132708, -0.304127588171866, -0.206330927797828,     1.68396072996797, -0.195872608025302, -0.363655846288592,     -0.362644976479584, -1.40690446251939, 0.708187421507729,     -1.09054015489071, -1.62878223261469, 0.272939538254605,     0.455398563639628, -2.19236687363479, -1.23693548997951,     -0.339295084097027, 0.954588547568362, -1.24975404516971,     -1.89842880961347, -1.71919073302738, -1.12534405492328,     1.19032480263027, 0.243592304735778, -0.0456213363096137,     1.08140120135091, -2.63850288654585, 0.288548232937203, -0.591811628043639,     0.0856511672691946, 0.207549734450574, 1.33945552426082,     0.446491015138253)), row.names = c(NA, -50L), class = "data.frame"),     na.action = function (object, ...)     {        if (NCOL(object) == 0)             object        else {            ok <- complete.cases(object)            if (all(ok))                 object            else stop(message)        }    }): object is not a matrix