Fits outcome model

Fits outcome model

getModel(
  d,
  exposure,
  exposure_time_pts,
  outcome,
  epochs,
  exp_epochs,
  int_order,
  model,
  fam,
  covariates,
  verbose
)

Arguments

d

wide data frame

exposure

name of exposure variable

exposure_time_pts

list of integers at which weights will be created/assessed that correspond to time points when exposure was measured

outcome

name of outcome variable with ".timepoint" suffix

epochs

data frame of exposure epoch labels and values

exp_epochs

list of exposure epochs

int_order

integer specification of highest order exposure main effects interaction for interaction models

model

character indicating one of the following outcome models:

"m0" (exposure main effects)

"m1" (exposure main effects & covariates)

"m2" (exposure main effects & their interactions)

"m3" (exposure main effects, their interactions, & covariates)

"covs" (covariate only model)

"int" (intercept only model)

fam

function specification for svyglm model

covariates

list of characters reflecting variable names of covariates for covariate models

verbose

TRUE or FALSE indicator for user output (default is TRUE)

Value

list of fitted model(s)

Examples

test <- data.frame(ID = 1:50,
                   A.1 = rnorm(n = 50),
                   A.2 = rnorm(n = 50),
                   A.3 = rnorm(n = 50),
                   B.1 = rnorm(n = 50),
                   B.2 = rnorm(n = 50),
                   B.3 = rnorm(n = 50),
                   C = rnorm(n = 50),
                   D.3 = rnorm(n = 50))
test[, c("A.1", "A.2", "A.3")] <- lapply(test[, c("A.1", "A.2", "A.3")], as.numeric)

f <- createFormulas(exposure = "A",
                    exposure_time_pts = c(1, 2, 3),
                    outcome = "D.3",
                    tv_confounders = c("A.1", "A.2", "A.3", "B.1", "B.2", "B.3"),
                    ti_confounders = "C",
                    type = "full",
                    save.out = FALSE)
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A - D.3 at A time point 1 is: 
#> A.1 ~ C
#> <environment: 0x5567f0c11d18>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A - D.3 at A time point 2 is: 
#> A.2 ~ A.1 + B.1 + C
#> <environment: 0x5567f0c11d18>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A - D.3 at A time point 3 is: 
#> A.3 ~ A.1 + A.2 + B.1 + B.2 + C
#> <environment: 0x5567f0c11d18>
#> 

w <- createWeights(data = test,
                   exposure = "A",
                   outcome = "D.3",
                   formulas = f,
                   save.out = FALSE)
#> Error in weightit2cbps.cont(covs = covs, treat = treat, subset = by.factor ==     i, s.weights = s.weights, missing = missing, verbose = verbose,     ...): The package "CBPS" is required.

epochs = data.frame(epochs = c("Infancy", "Toddlerhood"),
                    values = I(list(c(1, 2), c(3))))
e <- apply(expand.grid("A", as.character(epochs[, 1])), 1,
           paste, sep = "", collapse = ".")
test$weights <- w[[1]]$weights
#> Error in eval(expr, envir, enclos): object 'w' not found

g <- getModel(d = test,
              exposure = "A",
              exposure_time_pts = c(1, 2, 3),
              outcome = "D.3",
              epochs = epochs,
              exp_epochs = e,
              fam = gaussian,
              model = "m0")
#> Error in model.frame.default(~weights, data = structure(list(ID = 1:50,     A.1 = c(-0.71822146215588, -0.390279890137747, 0.230052801925252,     0.31064026933713, 0.0843887060791787, 0.119073893571363,     1.05880176458008, -0.0791666844769881, -0.104170671336315,     0.533117873713797, -0.976981570314124, -1.23123931015425,     -1.53171740654773, 1.78469780443042, 0.802623192998084, 0.516339290441256,     -0.830012419717406, 0.125244393748474, 0.178508421724611,     0.576108758086658, -0.0935270702974297, -1.44877375490052,     0.716099449839637, 0.455574720933644, -0.141187259399309,     -0.063433413566157, 0.994477614736271, -0.326457669862465,     -1.29787830263515, 0.863063630470252, 0.192288868278887,     0.028068603992056, -0.288372429156353, -0.375840553976302,     0.173338380875896, -0.275044308686091, 0.873971791199811,     -0.54502366754338, -0.0213787690569233, 0.321349779912695,     -1.16589820009467, 1.6056619660075, -0.238480506446945, 0.0645128073887915,     -0.0345730580532317, 0.615429748659675, -0.554681169041105,     3.05551307584236, -0.114118138664693, 1.0833495799451), A.2 = c(2.24486294402687,     -1.09217179991462, -1.30315843200603, -1.16573652824457,     0.896304583564253, -0.9742883577993, 0.122684997475901, 1.11887709381198,     0.194866120981129, 1.31422375121059, 1.25064615845208, -1.37617664974954,     0.0825611359344156, 0.0295382453650936, -0.770254758119335,     0.111551977636813, -0.830153892158839, 1.46000058263977,     -0.807228990956141, -1.47463243380357, -2.505540982712, -1.65583515581984,     0.0368800611475063, 0.331175840955928, -0.756208751102887,     -0.479307787377095, -0.775883139622981, 0.255438316642491,     -1.0507135243, 0.222013913732361, 1.14374394485742, 0.110735364823672,     -0.996077026093433, -1.13978552780913, 0.139592699089921,     0.0861587502774576, -0.558379183233432, -0.613666606300556,     -2.49828411161364, 1.23922920135845, 1.3028359744714, 0.576845301080062,     -0.0347416454803678, -0.284448051147672, -0.831501802830797,     0.307625219779057, -0.746723252788113, -0.522578949261711,     -0.169697686083012, -0.265642989402487), A.3 = c(0.985251749690742,     -1.2213339913053, 1.61640930632188, -0.462374268360862, -0.117119776428056,     0.418888013308646, -0.656566747685426, 0.910408176713732,     -0.336537881793951, -0.213488217714087, 0.0628159484775363,     1.07581480229741, -0.483090520148175, 0.722662378564422,     0.473639665495944, 1.15075727653119, -1.84823791005774, 0.236665690689444,     -1.34357209703286, 0.5435665626903, -0.908111973543882, 2.29666014362046,     0.208856738235531, -1.22464794305497, 0.546427688879778,     -0.547221044731733, 0.234199447185805, -1.21298153691362,     -1.93404956679306, -0.71032414427559, 0.254957423045172,     -0.427498373547509, -0.807150336918104, -1.34228992785145,     -1.33666257083395, 0.307900844262373, -1.686295905958, -0.416630594730761,     0.266658235900349, -1.54145863788498, -0.69221001259025,     0.23073456825039, -0.146181212537745, 2.12029342944085, -0.166009338719369,     1.08102580788106, -0.847155880607225, 0.47061941753423, 0.94362606418964,     -0.994286108567972), B.1 = c(0.499625342746726, 0.124848503257135,     -0.414008158266919, 1.82291297331188, -0.685107803396851,     1.14105970155572, 0.0355296544400026, 0.830617384038838,     0.659879482899589, -0.0402834631269908, 0.53605243295179,     0.724409403611809, -0.118540467326732, -2.02964326907409,     2.79794522760781, -2.55215638655232, -0.519401486648107,     0.219988190603005, -1.33662375881305, 0.431786218202553,     -0.814304065278271, 0.923011966457998, -0.11984545714027,     -0.962064009938864, -0.871298026424854, 0.905119170832026,     -0.0408016866316948, -0.429933858382479, 0.131138878162761,     -0.36807580415199, -0.142328806142857, 0.567504189459049,     -0.00804341876803395, 0.733808341703043, -0.506636453202043,     -0.400445540306131, -1.43212038631751, -0.195563307624992,     1.03241117246686, -0.406542767820047, -2.23933029062978,     -0.688552696711565, -0.896758342646166, 0.660251853966625,     -1.01098592119195, 0.440982319598472, 0.582762886173871,     -0.0231710932039146, -0.606818116466398, -0.976360991825274    ), B.2 = c(-1.10479070039983, 1.07050557711582, -0.482855981217853,     0.952345867451765, 0.987638211327589, -0.62254310705436,     -1.54922674421376, -1.36820304372511, 1.02184769554136, -1.16193958324947,     -1.24734650676834, -0.276325658616318, 0.468546343686471,     -0.355517998480885, -0.971215092715313, 1.49224200460644,     0.0219149254532183, -0.246969173090097, -0.748156951307877,     0.181444938736048, -0.95782919839857, 0.0905118096552797,     0.863813347076986, -0.709253410753403, 0.684336002745537,     0.318771818887773, 0.786681527695656, 1.12351745309091, -1.30361225425927,     -1.4195570331543, -0.0824796787728111, -0.68347368491628,     -0.676958611484469, -0.586460263449756, 0.217182346362349,     -3.16984240770566, 1.83613620112185, -0.912830314161305,     0.66155136761429, -0.15401706647849, -1.40881810796539, 0.739288221572852,     -0.804863539114542, -0.894599568034294, -0.636827856127264,     0.688223073869278, 0.817166764384272, -0.987589389230249,     -0.18432458080175, 1.06928831835193), B.3 = c(1.60178586804678,     0.0413830234216079, 0.499140045968412, -0.0379658802295267,     -0.657190799895532, 1.66000991123343, -0.449895576284541,     0.870093265412995, -0.6533270720189, 1.47244970426417, 1.36631387057429,     0.864099687763595, 0.578152451848765, -0.119092272863518,     -0.987108654303423, -0.352811340300497, -1.99413371760035,     0.581313205624451, -0.283788562656681, -1.86860177506364,     0.521491831556354, -1.09016029515487, 0.766688875563858,     0.236832843333486, 1.36324329190126, 0.534500371904132, -0.681281553540725,     -1.19170975942285, 0.467651748045946, -0.0871353138774369,     -0.832227473092922, -0.809695768086143, -0.490960907115174,     0.636150645318457, -0.694390660248389, -0.356732951515705,     -0.388261228077102, 1.69833891316117, -0.72754841743999,     -0.734813880566919, -1.65409923642236, 1.89569199809554,     -0.457280066645549, 0.388343709316219, -1.05046236765944,     -2.06675460936864, -0.774985249225673, -1.56235651002979,     0.96849661800255, 0.527499454133168), C = c(0.189732124710899,     0.543988403184804, 0.460561279961054, -0.76555478994886,     0.246119118805332, -0.363746850018926, -0.448652649813029,     -0.113356270820839, 0.37804851277289, -1.34568020021049,     0.667082115337596, 0.0771895448081333, 0.348808278826359,     -0.726557300358924, -0.22740275640577, 1.75611634997848,     -0.624308128877267, -0.891813564259317, -1.41438837152303,     0.831651119209064, -2.00208241134183, 0.894679319681052,     0.239961790117692, 0.246010469697857, -1.41266451722566,     0.991039931718784, -0.351361934335882, 1.04444893336536,     1.18551553029991, -0.420631785523438, 1.03092450851513, -0.363248080318024,     -0.808037406491942, 0.308876229998868, -1.36966431970379,     -0.951572450728035, 0.850958625046453, -0.503702920247694,     -0.839569716063438, 0.510909056131847, -1.48986444274182,     -0.493836143931812, 0.527635597681528, -0.702825738564029,     1.17226220289158, 0.294692630060449, 0.772029069684453, 0.36639828412897,     1.57116484769951, 0.984408787450517), D.3 = c(0.185125461143592,     -1.10081637765185, -0.812132257190509, 1.30109189633862,     0.311307653484614, -0.184563723122741, -0.736133411840207,     -0.848589388087681, -1.31552703836623, -0.208927060399796,     0.0171239241482413, -0.0779026994962409, -0.0815348040307668,     -1.24668226875646, 0.462621018346684, 1.57070703210079, 0.331711930550096,     0.208250259878513, 0.276953648214616, 0.265697163633945,     -1.091193345914, 1.41554012036687, 0.552058235225744, 0.249229311277726,     0.621995799510123, 3.07249702600753, -0.936491920410764,     -1.00765754892582, -1.8896149830508, 1.14853238759135, 0.261003333639429,     1.46080019619546, -0.997439224001743, 0.79313595250157, 1.07134563553645,     -0.324145810477416, 0.340568259537503, -0.089912410452825,     0.0369378414067512, 0.75533074691891, -0.477283270544946,     -2.46931612113115, -1.04107180499712, -1.12019125197193,     1.14111316757294, -1.45436412114569, 0.616137862072754, -1.09544586551611,     -0.439425751508588, 0.979062404309659), A.Infancy = c(0.763320740935495,     -0.741225845026183, -0.536552815040389, -0.42754812945372,     0.490346644821716, -0.427607232113969, 0.590743381027991,     0.519855204667496, 0.045347724822407, 0.923670812462194,     0.136832294068978, -1.30370797995189, -0.724578135306657,     0.907118024897757, 0.0161842174393745, 0.313945634039035,     -0.830083155938123, 0.792622488194122, -0.314360284615765,     -0.449261837858456, -1.29953402650471, -1.55230445536018,     0.376489755493572, 0.393375280944786, -0.448698005251098,     -0.271370600471626, 0.109297237556645, -0.035509676609987,     -1.17429591346758, 0.542538772101307, 0.668016406568154,     0.069401984407864, -0.642224727624893, -0.757813040892716,     0.156465539982909, -0.0944427792043167, 0.15779630398319,     -0.579345136921968, -1.25983144033528, 0.780289490635573,     0.068468887188365, 1.09125363354378, -0.136611075963656,     -0.10996762187944, -0.433037430442014, 0.461527484219366,     -0.650702210914609, 1.26646706329032, -0.141907912373853,     0.408853295271307), A.Toddlerhood = c(0.985251749690742,     -1.2213339913053, 1.61640930632188, -0.462374268360862, -0.117119776428056,     0.418888013308646, -0.656566747685426, 0.910408176713732,     -0.336537881793951, -0.213488217714087, 0.0628159484775363,     1.07581480229741, -0.483090520148175, 0.722662378564422,     0.473639665495944, 1.15075727653119, -1.84823791005774, 0.236665690689444,     -1.34357209703286, 0.5435665626903, -0.908111973543882, 2.29666014362046,     0.208856738235531, -1.22464794305497, 0.546427688879778,     -0.547221044731733, 0.234199447185805, -1.21298153691362,     -1.93404956679306, -0.71032414427559, 0.254957423045172,     -0.427498373547509, -0.807150336918104, -1.34228992785145,     -1.33666257083395, 0.307900844262373, -1.686295905958, -0.416630594730761,     0.266658235900349, -1.54145863788498, -0.69221001259025,     0.23073456825039, -0.146181212537745, 2.12029342944085, -0.166009338719369,     1.08102580788106, -0.847155880607225, 0.47061941753423, 0.94362606418964,     -0.994286108567972)), row.names = c("1", "2", "3", "4", "5", "6", "7", "8", "9", "10", "11", "12", "13", "14", "15", "16", "17", "18", "19", "20", "21", "22", "23", "24", "25", "26", "27", "28", "29", "30", "31", "32", "33", "34", "35", "36", "37", "38", "39", "40", "41", "42", "43", "44", "45", "46", "47", "48", "49", "50"), class = "data.frame"), na.action = function (object, ...) {    if (NCOL(object) == 0)         object    else {        ok <- complete.cases(object)        if (all(ok))             object        else stop(message)    }}): object is not a matrix