Fits outcome model

Fits outcome model

getModel(
  d,
  exposure,
  exposure_time_pts,
  outcome,
  epochs,
  exp_epochs,
  int_order,
  model,
  fam,
  covariates,
  verbose
)

Arguments

d

wide data frame

exposure

name of exposure variable

exposure_time_pts

list of integers at which weights will be created/assessed that correspond to time points when exposure was measured

outcome

name of outcome variable with ".timepoint" suffix

epochs

data frame of exposure epoch labels and values

exp_epochs

list of exposure epochs

int_order

integer specification of highest order exposure main effects interaction for interaction models

model

character indicating one of the following outcome models:

"m0" (exposure main effects)

"m1" (exposure main effects & covariates)

"m2" (exposure main effects & their interactions)

"m3" (exposure main effects, their interactions, & covariates)

"covs" (covariate only model)

"int" (intercept only model)

fam

function specification for svyglm model

covariates

list of characters reflecting variable names of covariates for covariate models

verbose

TRUE or FALSE indicator for user output (default is TRUE)

Value

list of fitted model(s)

Examples

test <- data.frame(ID = 1:50,
                   A.1 = rnorm(n = 50),
                   A.2 = rnorm(n = 50),
                   A.3 = rnorm(n = 50),
                   B.1 = rnorm(n = 50),
                   B.2 = rnorm(n = 50),
                   B.3 = rnorm(n = 50),
                   C = rnorm(n = 50),
                   D.3 = rnorm(n = 50))
test[, c("A.1", "A.2", "A.3")] <- lapply(test[, c("A.1", "A.2", "A.3")], as.numeric)

f <- createFormulas(exposure = "A",
                    exposure_time_pts = c(1, 2, 3),
                    outcome = "D.3",
                    tv_confounders = c("A.1", "A.2", "A.3", "B.1", "B.2", "B.3"),
                    ti_confounders = "C",
                    type = "full",
                    save.out = FALSE)
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A - D.3 at A time point 1 is: 
#> A.1 ~ C
#> <environment: 0x5604736c05b0>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A - D.3 at A time point 2 is: 
#> A.2 ~ A.1 + B.1 + C
#> <environment: 0x5604736c05b0>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A - D.3 at A time point 3 is: 
#> A.3 ~ A.1 + A.2 + B.1 + B.2 + C
#> <environment: 0x5604736c05b0>
#> 

w <- createWeights(data = test,
                   exposure = "A",
                   outcome = "D.3",
                   formulas = f,
                   save.out = FALSE)
#> Error in weightit2cbps.cont(covs = covs, treat = treat, subset = by.factor ==     i, s.weights = s.weights, missing = missing, verbose = verbose,     ...): The package "CBPS" is required.

epochs = data.frame(epochs = c("Infancy", "Toddlerhood"),
                    values = I(list(c(1, 2), c(3))))
e <- apply(expand.grid("A", as.character(epochs[, 1])), 1,
           paste, sep = "", collapse = ".")
test$weights <- w[[1]]$weights
#> Error in eval(expr, envir, enclos): object 'w' not found

g <- getModel(d = test,
              exposure = "A",
              exposure_time_pts = c(1, 2, 3),
              outcome = "D.3",
              epochs = epochs,
              exp_epochs = e,
              fam = gaussian,
              model = "m0")
#> Error in model.frame.default(~weights, data = structure(list(ID = 1:50,     A.1 = c(-1.23632363692355, 0.247577484832373, -1.6384768559844,     -1.26178807406908, -2.20993552262735, 0.878688855409351,     0.182736169610843, 0.0894624496643239, 0.657966989354133,     0.772282776783564, 0.692008427219044, 0.264253428229933,     -0.99960576421238, 1.63050518255553, 0.795588660479148, 1.05572265962182,     0.09119138731052, -1.47472594546412, 0.513047783281506, -0.0269371752866384,     -0.406834120016845, -0.318441595593997, 0.42687399879028,     0.905278823064258, -1.35636215682845, -0.239407311863747,     0.937159297542609, 0.212548656903684, -0.833708166407356,     0.693677441455175, 0.205992085389505, -0.111993573252581,     0.296170449059652, -0.0859284135924114, 0.223889464425742,     -0.251128130411983, 0.533948543596629, -0.13374492756711,     0.32807715472786, 1.93277003466311, 0.731555053244811, 1.26603997025699,     1.40026532224811, -0.905441692948924, 1.34600035202826, -1.42599756745962,     0.658978142847378, 0.866275328589768, 0.689086036871524,     -0.137496033059538), A.2 = c(-0.894737874919155, -1.89767024706302,     -2.30023244154071, 1.83787132001796, -0.177556373211712,     -2.20247048410703, 1.2825278575156, -0.832887847925305, 0.241024718910464,     -0.624402202204002, -0.454554498900731, -1.08658052549559,     0.535628365264791, -1.59508006817273, 0.46587907261643, -0.0589022589432833,     0.442341107385808, 0.315739646912655, 0.529774957498727,     -0.0534321066731523, -2.07914578100812, -1.20411883409366,     -0.581864307472464, -0.252743282989816, -0.963659016304327,     -1.13294759157763, 0.696246081890572, 0.770053096601236,     -1.03654197357402, 1.45579177298045, -0.701731104558343,     2.64329936167562, -0.637824538584812, -0.240535880276716,     0.273156743322397, -2.62405939253362, 0.232194133456682,     1.36570460599451, -0.220802921508316, 1.0213718952219, 1.67539645854616,     -0.973420992879481, -0.311503219275969, 0.72299099857563,     -1.13583199528073, -1.45994698408916, -0.243050999768989,     -0.189341856905609, -0.156358417622691, -0.0101236160615761    ), A.3 = c(0.842693703165071, -0.84971916408893, 0.0280509304678829,     -0.997112654133754, 0.317581952561549, -0.694475028688701,     0.769093327902, -0.337347297040366, -0.753566338158385, -2.20608759817339,     -0.419272639196061, -1.29430489980313, -1.22732878109957,     -0.0821252044577607, -0.520328206714134, -1.59987633005624,     -1.14393663666292, -0.966515790776583, 0.131537976691662,     3.01495998881351, -0.91499100291357, 1.49899028785135, -0.761554746064041,     0.570302819773339, 2.19573295741834, 0.50813326546513, 0.959927497459825,     -1.32548070925419, -2.361577473601, 0.413540648242751, -0.353988550109803,     -0.6239563475501, 0.42274960427219, 1.05005230354565, -0.207548552343016,     -0.764729831788114, 1.4174171827115, 0.633336852702993, 0.00799456632681162,     0.597298818332023, 1.03707463598981, -1.33773039403693, 1.05483359961787,     -1.78873541683603, 1.09070778143266, 1.12633075742762, 0.52872847104189,     -0.555541119732609, 0.460458812026714, -2.56212059564247),     B.1 = c(-0.869685173336798, 1.88125446374121, 0.151203825955667,     0.713565764123254, 0.613449513116052, 0.350197524054996,     -1.07500633900385, 0.330833506374945, -0.128474441912643,     -1.067938342724, -0.0591234429246148, 0.85323335526522, 1.32787177652934,     0.690630761400235, -0.17489631721708, -1.49133670338102,     -0.0972425382006603, 2.38195339334397, -1.40594322798788,     -1.12646250052774, -0.0311648096972706, 0.361392350697876,     -1.35729848408822, 1.18284363022644, 1.58822180014895, 0.0457025213247133,     -0.0752339623837264, -0.572070976771062, 0.212933900111873,     -0.557084803046848, -0.812728190771691, 1.07733812836361,     -2.06713216576002, 0.936174356731915, -1.31968654394543,     -1.6139958036063, -0.190754814717412, 1.42919677005124, -1.70101155511065,     0.0647405973934586, 0.200473806060377, 0.518130969547634,     1.5391961997666, -1.3183260219726, -0.183600741410775, -0.608669695409674,     0.767954761220891, 0.51342065756471, 0.0204049332617537,     -1.51624783026896), B.2 = c(-0.120674794784162, -1.89803688578071,     -0.574477824943496, -0.0677941151272495, 0.0223074048477253,     0.92786882778415, -1.82207600064941, -0.113024290767033,     0.307908366653715, 0.891017264610484, 0.892880725129714,     1.25447241294461, -1.29295370301057, 0.00897280481986122,     -0.392685153065384, 0.786023875841717, 7.38005453580726e-05,     1.97361213931928, -0.34654501470343, 1.00724504510199, 0.689072779560351,     0.383639951842869, -0.258798013819445, 0.393497321209153,     0.013898592036566, 0.528124671119813, -1.47103350278486,     -0.943577842194099, -0.228212866235649, -0.140073308224068,     -0.697897616639145, 0.122006751286091, -0.422566049643754,     0.13934433559095, -0.591000738934735, 1.17363327312349, 0.910294562773251,     1.05256520100638, -0.0592975262353345, 1.6848781665953, -1.33516925220288,     1.33588014941071, 0.0517545599332573, 1.17783387424118, 0.708544644808356,     -2.00190540330929, -0.168587653420468, 0.199137063157653,     -0.312050481139737, 0.208782642246406), B.3 = c(-1.0579549459049,     -0.886871080520652, -0.550649243331584, 0.871206623119866,     0.0551714418130442, -0.337149505959164, -1.82388984860882,     1.91573371490999, -0.989732279812145, -1.41811495837428,     -0.738631974881962, -1.12490701157698, -1.04409614845404,     0.600254287689165, 1.4896415452404, -0.768992525410863, -0.420361757686989,     1.68768760549155, 0.0203465309108261, -0.123762878591175,     -2.19328690156577, 0.55745729999226, 2.04173306286761, 0.899798704035246,     -0.995843564565732, 0.672344254512453, -0.0545648907269237,     -1.71826021414866, 0.28353145010592, -0.434052458172983,     0.336312881278186, -1.20204506556096, 0.0909391882902116,     -0.194650649033901, 0.517125375784911, -0.468057190934849,     1.20441005318323, 0.0363003111787611, -1.11996663409686,     -0.299574410265684, -0.818778247017113, -1.52452333714181,     1.69199222909092, -0.433243333343198, -0.145594232788213,     0.213031222322179, 0.258116112790695, -1.89097854415742,     1.23392764872357, -0.310593761819311), C = c(-0.939796921140447,     1.13567115650552, 0.131408183466425, 0.85064532146451, 0.311085933356254,     0.229199997103115, -1.47695062767003, -0.336295524647032,     -1.00303363557902, -1.19865330777476, -0.0825690590000784,     0.0893218654854725, -1.49672300948427, -0.545285379499234,     0.0657843783372396, -0.506012897797104, 1.27503689347206,     0.991986481387566, -0.0391960030425431, 0.300645442397679,     -1.87458945146368, 0.889277932284549, -0.472789757664678,     0.981542331489142, 0.707282925795517, -0.150123664785009,     0.837371530056269, -0.99015956537662, 0.836272207958176,     0.254996819604471, 1.86533421320224, -0.354962143880382,     -0.670846708314277, -0.680807495378458, -1.31322977643748,     0.37767699942318, -0.243626255348687, 1.51575418501208, -0.952003327821878,     1.72328362056261, 0.453029498633862, 0.114879522010399, -1.0422372554849,     0.405799107422667, -0.802221093652041, 1.13518443779172,     0.247029612234101, 0.390386805861227, 1.62048329902561, -0.738063739755787    ), D.3 = c(-0.918893843062895, -0.790791593763891, -0.420852194286523,     -0.641208607390565, -0.679619336461118, 0.220934596349973,     0.950738753715083, -0.82944415855387, 0.687518728644898,     0.221254556286751, -1.05931583636356, -0.703360854469497,     -0.299428053457267, -0.192167769355092, 0.603345959227564,     -0.630526877682119, 2.05729507449579, 0.103631081119824,     0.202685629471161, -0.840323293928212, 0.129220945842934,     -0.228772478756195, -0.243139675332324, 0.610948231618641,     0.0518173146288983, 1.06099660453406, -1.79460855081227,     1.20387821490831, -1.78349244619006, 0.834536790149021, -0.837469777252212,     0.636927853891879, 0.259356959982021, -0.0390820863340939,     0.214328366263321, 0.539196963473558, -1.15816369980174,     -0.0216931184845184, 0.5332270273807, -1.43663305504417,     1.14805328684799, 0.869723049121515, 0.618933276728182, 0.0302690199366779,     -0.98404467908931, -0.0901002665315253, -0.851142966393139,     -0.957681887324194, -1.70403334508204, -1.09808805202146),     A.Infancy = c(-1.06553075592135, -0.825046381115323, -1.96935464876256,     0.28804162297444, -1.19374594791953, -0.661890814348839,     0.732632013563221, -0.371712699130491, 0.449495854132299,     0.073940287289781, 0.118726964159156, -0.411163548632829,     -0.231988699473795, 0.0177125571914, 0.630733866547789, 0.498410200339268,     0.266766247348164, -0.579493149275732, 0.521411370390117,     -0.0401846409798954, -1.24298995051248, -0.761280214843828,     -0.077495154341092, 0.326267770037221, -1.16001058656639,     -0.686177451720688, 0.816702689716591, 0.49130087675246,     -0.935125069990688, 1.07473460721781, -0.247869509584419,     1.26565289421152, -0.17082704476258, -0.163232146934564,     0.24852310387407, -1.4375937614728, 0.383071338526656, 0.6159798392137,     0.053637116609772, 1.4770709649425, 1.20347575589549, 0.146309488688754,     0.544381051486071, -0.091225347186647, 0.105084178373765,     -1.44297227577439, 0.207963571539194, 0.338466735842079,     0.266363809624417, -0.0738098245605571), A.Toddlerhood = c(0.842693703165071,     -0.84971916408893, 0.0280509304678829, -0.997112654133754,     0.317581952561549, -0.694475028688701, 0.769093327902, -0.337347297040366,     -0.753566338158385, -2.20608759817339, -0.419272639196061,     -1.29430489980313, -1.22732878109957, -0.0821252044577607,     -0.520328206714134, -1.59987633005624, -1.14393663666292,     -0.966515790776583, 0.131537976691662, 3.01495998881351,     -0.91499100291357, 1.49899028785135, -0.761554746064041,     0.570302819773339, 2.19573295741834, 0.50813326546513, 0.959927497459825,     -1.32548070925419, -2.361577473601, 0.413540648242751, -0.353988550109803,     -0.6239563475501, 0.42274960427219, 1.05005230354565, -0.207548552343016,     -0.764729831788114, 1.4174171827115, 0.633336852702993, 0.00799456632681162,     0.597298818332023, 1.03707463598981, -1.33773039403693, 1.05483359961787,     -1.78873541683603, 1.09070778143266, 1.12633075742762, 0.52872847104189,     -0.555541119732609, 0.460458812026714, -2.56212059564247)), row.names = c("1", "2", "3", "4", "5", "6", "7", "8", "9", "10", "11", "12", "13", "14", "15", "16", "17", "18", "19", "20", "21", "22", "23", "24", "25", "26", "27", "28", "29", "30", "31", "32", "33", "34", "35", "36", "37", "38", "39", "40", "41", "42", "43", "44", "45", "46", "47", "48", "49", "50"), class = "data.frame"), na.action = function (object,     ...) {    if (NCOL(object) == 0)         object    else {        ok <- complete.cases(object)        if (all(ok))             object        else stop(message)    }}): object is not a matrix