Fits outcome model

Fits outcome model

getModel(
  d,
  exposure,
  exposure_time_pts,
  outcome,
  epochs,
  exp_epochs,
  int_order,
  model,
  fam,
  covariates,
  verbose
)

Arguments

d

wide data frame

exposure

name of exposure variable

exposure_time_pts

list of integers at which weights will be created/assessed that correspond to time points when exposure was measured

outcome

name of outcome variable with ".timepoint" suffix

epochs

data frame of exposure epoch labels and values

exp_epochs

list of exposure epochs

int_order

integer specification of highest order exposure main effects interaction for interaction models

model

character indicating one of the following outcome models:

"m0" (exposure main effects)

"m1" (exposure main effects & covariates)

"m2" (exposure main effects & their interactions)

"m3" (exposure main effects, their interactions, & covariates)

"covs" (covariate only model)

"int" (intercept only model)

fam

function specification for svyglm model

covariates

list of characters reflecting variable names of covariates for covariate models

verbose

TRUE or FALSE indicator for user output (default is TRUE)

Value

list of fitted model(s)

Examples

test <- data.frame(ID = 1:50,
                   A.1 = rnorm(n = 50),
                   A.2 = rnorm(n = 50),
                   A.3 = rnorm(n = 50),
                   B.1 = rnorm(n = 50),
                   B.2 = rnorm(n = 50),
                   B.3 = rnorm(n = 50),
                   C = rnorm(n = 50),
                   D.3 = rnorm(n = 50))
test[, c("A.1", "A.2", "A.3")] <- lapply(test[, c("A.1", "A.2", "A.3")], as.numeric)

f <- createFormulas(exposure = "A",
                    exposure_time_pts = c(1, 2, 3),
                    outcome = "D.3",
                    tv_confounders = c("A.1", "A.2", "A.3", "B.1", "B.2", "B.3"),
                    ti_confounders = "C",
                    type = "full",
                    save.out = FALSE)
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A-D.3 at A time point 1 is: 
#> A.1 ~ C
#> <environment: 0x561652fd0b98>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A-D.3 at A time point 2 is: 
#> A.2 ~ A.1 + B.1 + C
#> <environment: 0x561652fd0b98>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A-D.3 at A time point 3 is: 
#> A.3 ~ A.1 + A.2 + B.1 + B.2 + C
#> <environment: 0x561652fd0b98>
#> 

w <- createWeights(data = test,
                   exposure = "A",
                   outcome = "D.3",
                   formulas = f,
                   save.out = FALSE)
#> Error in weightit2cbps.cont(covs = covs, treat = treat, subset = by.factor ==     i, s.weights = s.weights, missing = missing, verbose = verbose,     ...): The package "CBPS" is required.

epochs = data.frame(epochs = c("Infancy", "Toddlerhood"),
                    values = I(list(c(1, 2), c(3))))
e <- apply(expand.grid("A", as.character(epochs[, 1])), 1,
           paste, sep = "", collapse = ".")
test$weights <- w[[1]]$weights
#> Error in eval(expr, envir, enclos): object 'w' not found

g <- getModel(d = test,
              exposure = "A",
              exposure_time_pts = c(1, 2, 3),
              outcome = "D.3",
              epochs = epochs,
              exp_epochs = e,
              fam = gaussian,
              model = "m0")
#> Error in model.frame.default(~weights, data = structure(list(ID = 1:50,     A.1 = c(-3.01417120203218, -0.709479511396879, -2.10058230694861,     1.39963537568644, 0.152644052008996, 0.851906386746958, -0.231568462738449,     0.906622214555922, -0.137836764265414, -1.26156493009137,     -0.494477156392904, 0.128760473014243, -0.42786964595634,     -1.62471318540572, 0.955413850320851, 2.14254885143944, -1.10576715044288,     -1.2342385874093, 0.508605302632191, -1.34917560095693, -0.798572847852018,     0.10338905351475, -0.504567584523054, -1.24259030693167,     -1.3016276471688, -0.297171298231211, 0.836748856526425,     0.310300032179193, 1.03111967362923, -0.719679393596545,     -0.451710508221201, 0.772435299267611, 0.498073519394335,     0.770977660821068, -0.365842119389758, -1.52297188350773,     -0.571414908323238, -0.766359356400891, 0.135211225243478,     0.678288440245289, 0.653403536055341, -1.53114908001886,     -0.475822537077443, 0.0433592134638738, 0.18474069667828,     -0.482963745232658, -0.36269139954842, 0.245834070612026,     1.63739195779665, -1.70824732506584), A.2 = c(0.643546231066451,     0.396296795251258, 0.158439607949876, 0.734965785625072,     -1.44913176637221, 1.48113786013558, 0.7421419101983, -0.286622747275014,     -1.05149008917341, 1.80573000475777, -0.193304385496569,     0.360587710221758, -0.486021090545948, -1.45374491250225,     -1.02713219170897, 1.95809273788413, 1.20146743931027, 0.907750712390812,     0.224023474497806, 1.34745860860183, 0.445804522686004, -0.421954085616641,     -0.0209594689595273, -1.8717516250258, 1.8658763275915, 0.199812758847234,     1.78730814503909, -1.01135296170921, 0.239626353079458, 0.243011688123141,     -0.365195559008551, 2.58888322721272, 0.675782305917327,     -0.0123686730912537, -0.0352243759229873, -1.92206416224004,     0.1293359059573, -0.40396183982188, 1.02028366778817, 1.19021644915381,     0.215863983668646, -0.1199728655708, 0.360091502348275, 0.177293218324632,     0.770663648939704, -0.851711972324653, -0.496471538402024,     1.64212908610359, -1.57403380277498, -0.832365981232359),     A.3 = c(1.80044313580809, 0.000283432420265667, 0.548325820591647,     -1.14027294709641, 2.03941848422542, -0.4481257268348, 1.37782490304415,     0.567947485404128, -0.599436189327334, -1.11631425200835,     -0.352941536040201, 0.0449073597751439, -0.456178095525502,     -0.200260145072023, -0.719069883007849, 0.559103030586489,     -0.0638069038230313, -2.11729138227766, -0.705881551164213,     2.21072267754564, 1.08823329078015, 0.334818864660859, -0.797366832863065,     -1.13423354755659, -1.63289381916704, 0.855818202104051,     -0.386998992528597, 1.32128398803406, 1.00751441459459, -1.13563673402384,     -0.104901753378271, -0.12706024032561, -1.01117218077636,     -0.06639169441511, -0.753455817300122, -1.23438224430335,     -0.800688718219462, 2.23501444230896, -0.565481864425259,     -0.0191377168850629, -0.619980232693181, -0.799433047763583,     -0.264979446897048, 1.39353343882279, -0.351795424097662,     0.177676305022425, 0.130875341025922, -0.552101503921946,     -1.72241330261535, -2.33988439860123), B.1 = c(-0.117870504147702,     0.892713323072952, 0.527805837759018, 2.15661001653761, 0.896867998151497,     1.38252810450973, -0.319977360420102, -0.00596099486976494,     0.662655690911407, -2.48052153716073, 0.21752286166956, -0.302040299985901,     -0.0130795060403754, 1.55578443110867, -0.270510600148484,     1.51536490914149, -0.0326819928106209, -1.50117289121175,     -1.14699680420684, 0.460563876868581, 0.61436966913312, -1.1714435351955,     -1.603814582204, -0.784641019257819, 0.92460907196748, 0.489281344254898,     -0.0789671596423739, -0.865532847536881, 0.835617027307121,     1.02769833545639, 0.500944955486219, -0.997809940462165,     0.0624956685518915, -1.01652853213119, 0.0772671982288261,     -1.61436684976253, 1.88333158032199, -0.018258725616328,     -0.546028039423845, -1.2358622926458, 0.654828042916562,     0.664189165140216, 1.35349536046658, 0.441322287239152, -0.689156309596493,     -0.561142415386529, 0.266553066168183, -0.744002653977083,     1.60158134101999, -0.306942689918088), B.2 = c(0.784600373082796,     0.307939281142414, -0.371488528754817, -1.52561070249789,     0.406039056380432, -1.89757729466062, 1.36089878835393, -0.049463416891489,     -0.0425401496188881, -0.102498610110252, 0.456043997436182,     1.05361149266464, -0.331724219300738, -0.309969734841848,     -0.416678719467905, 0.587674467141001, -0.89324114001486,     -1.73679191880592, -0.229093020352182, 0.190487429878567,     -1.04829918917211, -0.530935778766187, -1.44387301739695,     -0.065182592443476, 0.729925870674186, 0.423402266326648,     0.446805934911859, -0.153019133852254, -0.174411471290704,     -0.261230425637194, -0.0662578484080478, 0.636992183209925,     0.580551292473339, -1.45648169575115, -0.662162747906733,     0.474840980458745, 1.73485086001106, -0.19909172236003, 0.317428906847615,     -0.50375784988531, 1.86174462512516, 0.046864343681687, -0.242461176437335,     1.59967839579964, 0.103472038084919, 0.332914562277353, 0.133191405059476,     -1.50893314448657, -1.23408469773393, -0.472617680285499),     B.3 = c(0.467931234974293, -0.309315048836755, -0.00181534509211089,     -0.181853564493715, -1.6448136582659, -0.458431699886326,     2.06417697597216, 1.50729879061963, -1.22032934145413, -0.0397107139531794,     -0.577403614252552, 0.0804903967831527, 0.706543722001411,     0.114545443475758, -0.720962209319175, -0.774828586097234,     -1.48110833306203, -0.693747738310044, 1.08294401171435,     0.73507784701252, -0.139866643724738, 0.225972368342377,     -0.357260148595984, -1.82934142847894, 0.546289493708867,     -0.550884992315534, -1.52752133130236, 1.12144249112549,     -1.13779262000489, -0.0599530892956226, -0.173827765996545,     0.174666928745824, -0.228968956701705, -1.23137265258063,     -0.450703695130588, -1.98017900218867, 0.768186417145278,     1.69130214529218, -0.264173960004334, -0.578418917673289,     -0.290994324093358, -0.892998078856321, -1.25372773603295,     -0.520269353942939, -0.273976637867181, -0.660451022542488,     0.531117930987065, 1.08608141910666, 0.527700761912959, 1.03336537996551    ), C = c(0.423320524819464, -1.76832659686676, 1.3539021654538,     1.14472520238505, -0.206855337760601, 1.28778474354933, -2.43379072478243,     0.971016182493316, 0.000169307189066599, 0.0958705955640823,     0.767049652760775, 0.286456552317082, 0.625784504920801,     0.987377141160143, -0.733971176795504, 0.388083491033607,     -1.13352923218095, 0.197289284253208, -1.49826072655474,     -0.145059706619928, -1.93351473895216, 1.42801322810042,     -1.35477240182294, 1.0695291094397, -0.832532510950475, 0.924687709775212,     -1.52798555769831, -0.392433470342023, -1.1102741738243,     0.639850233182215, -0.129102432231047, -0.132677082135567,     -0.380275271820333, -1.39687052582027, -2.24837534984139,     1.74217204761827, -1.34471886370847, 0.875075608825621, -0.718604044767392,     0.135450047626409, -0.523732842158685, 0.405755920954302,     -0.778316071705631, -1.09158760285508, -0.652849984162044,     -0.086911792515381, 1.49237019955233, 0.822760416266383,     -1.28029100197961, -0.498372749925622), D.3 = c(0.330494056993302,     0.773029595478747, -0.943027526285626, -1.07995951127834,     -0.412032872968875, -0.35221630576588, -0.196804129071446,     0.0520147656967217, -0.551200084966327, 1.42288165033196,     -1.98297951204916, 0.964634833408539, -0.339437429109144,     -0.927721010823302, -0.57441308262044, -0.313152555700258,     -0.397407878266654, -0.792132463107826, 1.01711587202065,     -1.5282222437615, 1.74921240926591, -1.11252069264656, 0.64375048921488,     1.11916916255931, 1.0282903083326, 0.19924957128088, -0.604729146980453,     -0.812490312784339, 1.59029758958699, 1.78691850700294, 0.522624706501009,     0.154145151774597, 0.536768411570388, 0.0541776063994176,     0.991221201130573, -1.65264345166711, -2.04784461465861,     -0.573510257266332, -0.138447771134536, -0.259888857976365,     -0.737889081033251, -0.488519138292062, 1.36072957195735,     -1.74480014036604, 0.284240787114348, 0.247813017892559,     0.486770501291457, -0.344789395115811, -0.738550337545034,     -1.46676192570182), A.Infancy = c(-1.18531248548286, -0.156591358072811,     -0.971071349499367, 1.06730058065576, -0.648243857181607,     1.16652212344127, 0.255286723729926, 0.309999733640454, -0.594663426719412,     0.2720825373332, -0.343890770944737, 0.244674091618001, -0.456945368251144,     -1.53922904895399, -0.0358591706940595, 2.05032079466178,     0.047850144433695, -0.163243937509244, 0.366314388564998,     -0.000858496177549961, -0.176384162583007, -0.159282516050946,     -0.262763526741291, -1.55717096597873, 0.28212434021135,     -0.0486792696919885, 1.31202850078276, -0.350526464765009,     0.635373013354344, -0.238333852736702, -0.408453033614876,     1.68065926324017, 0.586927912655831, 0.379304493864907, -0.200533247656373,     -1.72251802287389, -0.221039501182969, -0.585160598111386,     0.577747446515824, 0.934252444699549, 0.434633759861993,     -0.82556097279483, -0.057865517364584, 0.110326215894253,     0.477702172808992, -0.667337858778655, -0.429581468975222,     0.943981578357808, 0.031679077510835, -1.2703066531491),     A.Toddlerhood = c(1.80044313580809, 0.000283432420265667,     0.548325820591647, -1.14027294709641, 2.03941848422542, -0.4481257268348,     1.37782490304415, 0.567947485404128, -0.599436189327334,     -1.11631425200835, -0.352941536040201, 0.0449073597751439,     -0.456178095525502, -0.200260145072023, -0.719069883007849,     0.559103030586489, -0.0638069038230313, -2.11729138227766,     -0.705881551164213, 2.21072267754564, 1.08823329078015, 0.334818864660859,     -0.797366832863065, -1.13423354755659, -1.63289381916704,     0.855818202104051, -0.386998992528597, 1.32128398803406,     1.00751441459459, -1.13563673402384, -0.104901753378271,     -0.12706024032561, -1.01117218077636, -0.06639169441511,     -0.753455817300122, -1.23438224430335, -0.800688718219462,     2.23501444230896, -0.565481864425259, -0.0191377168850629,     -0.619980232693181, -0.799433047763583, -0.264979446897048,     1.39353343882279, -0.351795424097662, 0.177676305022425,     0.130875341025922, -0.552101503921946, -1.72241330261535,     -2.33988439860123)), row.names = c("1", "2", "3", "4", "5", "6", "7", "8", "9", "10", "11", "12", "13", "14", "15", "16", "17", "18", "19", "20", "21", "22", "23", "24", "25", "26", "27", "28", "29", "30", "31", "32", "33", "34", "35", "36", "37", "38", "39", "40", "41", "42", "43", "44", "45", "46", "47", "48", "49", "50"), class = "data.frame"), na.action = function (object, ...) {    if (NCOL(object) == 0)         object    else {        ok <- complete.cases(object)        if (all(ok))             object        else stop(message)    }}): object is not a matrix