Fits outcome model

Fits outcome model

getModel(
  d,
  exposure,
  exposure_time_pts,
  outcome,
  epochs,
  exp_epochs,
  int_order,
  model,
  fam,
  covariates,
  verbose
)

Arguments

d

wide data frame

exposure

name of exposure variable

exposure_time_pts

list of integers at which weights will be created/assessed that correspond to time points when exposure was measured

outcome

name of outcome variable with ".timepoint" suffix

epochs

data frame of exposure epoch labels and values

exp_epochs

list of exposure epochs

int_order

integer specification of highest order exposure main effects interaction for interaction models

model

character indicating one of the following outcome models:

"m0" (exposure main effects)

"m1" (exposure main effects & covariates)

"m2" (exposure main effects & their interactions)

"m3" (exposure main effects, their interactions, & covariates)

"covs" (covariate only model)

"int" (intercept only model)

fam

function specification for svyglm model

covariates

list of characters reflecting variable names of covariates for covariate models

verbose

TRUE or FALSE indicator for user output (default is TRUE)

Value

list of fitted model(s)

Examples

test <- data.frame(ID = 1:50,
                   A.1 = rnorm(n = 50),
                   A.2 = rnorm(n = 50),
                   A.3 = rnorm(n = 50),
                   B.1 = rnorm(n = 50),
                   B.2 = rnorm(n = 50),
                   B.3 = rnorm(n = 50),
                   C = rnorm(n = 50),
                   D.3 = rnorm(n = 50))
test[, c("A.1", "A.2", "A.3")] <- lapply(test[, c("A.1", "A.2", "A.3")], as.numeric)

f <- createFormulas(exposure = "A",
                    exposure_time_pts = c(1, 2, 3),
                    outcome = "D.3",
                    tv_confounders = c("A.1", "A.2", "A.3", "B.1", "B.2", "B.3"),
                    ti_confounders = "C",
                    type = "full",
                    save.out = FALSE)
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A - D.3 at A time point 1 is: 
#> A.1 ~ C
#> <environment: 0x55b897def1c0>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A - D.3 at A time point 2 is: 
#> A.2 ~ A.1 + B.1 + C
#> <environment: 0x55b897def1c0>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A - D.3 at A time point 3 is: 
#> A.3 ~ A.1 + A.2 + B.1 + B.2 + C
#> <environment: 0x55b897def1c0>
#> 

w <- createWeights(data = test,
                   exposure = "A",
                   outcome = "D.3",
                   formulas = f,
                   save.out = FALSE)
#> Error in weightit2cbps.cont(covs = covs, treat = treat, subset = by.factor ==     i, s.weights = s.weights, missing = missing, verbose = verbose,     ...): The package "CBPS" is required.

epochs = data.frame(epochs = c("Infancy", "Toddlerhood"),
                    values = I(list(c(1, 2), c(3))))
e <- apply(expand.grid("A", as.character(epochs[, 1])), 1,
           paste, sep = "", collapse = ".")
test$weights <- w[[1]]$weights
#> Error in eval(expr, envir, enclos): object 'w' not found

g <- getModel(d = test,
              exposure = "A",
              exposure_time_pts = c(1, 2, 3),
              outcome = "D.3",
              epochs = epochs,
              exp_epochs = e,
              fam = gaussian,
              model = "m0")
#> Error in model.frame.default(~weights, data = structure(list(ID = 1:50,     A.1 = c(0.613427478066441, 0.69145929524694, -0.382846857173246,     -1.63712408310885, -0.362923931561996, -0.285527307429434,     0.619264954210211, 0.761584886510635, -0.714005012952735,     -0.346471615048805, 0.524489426296934, 0.640883281739933,     1.37108890130879, 0.284302808737597, -0.0511822638368801,     -0.52208076348645, 1.06088649002083, 0.400012902176698, 0.726079378876575,     0.670510191175791, 1.35933865283793, 0.372359763170938, -0.491579700397398,     -0.384349039408114, 0.475948688516963, -0.712382450497326,     -0.793005432903392, -0.369070669112502, -0.681973990605204,     0.318855978370775, -0.950514586045644, -0.149708311280962,     1.18612978549752, -0.434022337012402, 0.859269999274753,     0.355504956715117, 0.198292713049603, -2.2183071381735, 0.957237036571006,     0.811972608569493, -0.59408699502715, -0.333997896470344,     -0.4154152454135, 1.20099672418018, -1.85904600327154, 0.798372081936676,     -0.147344346311423, 0.513967365164892, -0.681500361252124,     -0.395327902069249), A.2 = c(-0.296401171633966, -0.133034570311913,     -0.698606661471853, 0.1680277330977, 1.0250177512996, -2.08383934187527,     -1.63820008241476, -0.637724333393723, 0.00366385322837378,     1.2167763437375, 1.06185202067491, 0.206732559451009, 0.490677237268258,     0.850635897200315, -0.052458588165426, 1.02495379111558,     -0.363985466988984, -0.14157465761627, 1.58588253788947,     -0.902954194002647, -1.05160773274599, 1.10059865780448,     -0.0275597197092106, 0.997642055959883, 0.644323563407234,     -0.53860770959992, 0.60982401191502, -0.511532376203142,     0.60172772312359, 0.745534483649226, -0.584144819057825,     0.728687497861001, -1.39860448477273, 0.351674382898767,     -0.462763229738802, 1.26180309170946, 0.594776432891826,     0.662079367965069, 0.7046683849414, 1.59227870295322, -0.42773146917824,     0.0956255669278995, -0.634346035384888, 1.57907759851559,     1.27588771062373, 0.479548364123896, 0.160090662621025, -0.789734721132659,     -1.27498477632507, 0.152558641387651), A.3 = c(-0.870607706278175,     -0.698208021938386, 0.944165434707425, -1.01379090584771,     -1.34637040069686, 0.535172542842567, -1.54241002195025,     1.53405849999373, -0.507102649603417, -0.10045388612321,     1.88186826844102, -0.308765858055712, 0.322744669308751,     0.232593555395557, 0.747868704441134, 0.14834131879584, 0.236727191954996,     -2.59962303653058, -1.25246979488042, -1.24125341705975,     -1.85896153721855, -0.265500734498976, -0.51283230666195,     -0.467164047238227, 0.630206299899219, 1.36277057726297,     -0.672869263171041, 0.395628641487794, -0.691699450197584,     0.931458746005471, 0.183596539168729, 0.906332258024459,     1.08613452924112, -0.0828150751546846, -1.33750568838534,     -0.279623166537107, -1.29772786387416, 0.416714509357019,     -1.13952935249532, 0.631605048296824, -0.667664694947506,     1.22459606989855, -0.649876302449042, 0.14471280679046, 0.800618216377827,     0.735372229438395, 0.489558304385338, 1.33611873598614, 1.98799326048856,     -0.000821600863269169), B.1 = c(-0.0361434458069921, 2.34026509238322,     -0.592218193023797, -1.87012061168113, 0.893957725168585,     -0.818314591801952, 0.873475153078011, 1.59765858913891,     -0.354198112287919, 0.570647223807283, -0.678472479655638,     -1.40593117133615, -1.88412935325124, -0.323526539353026,     -0.791632493434386, -0.854864292939851, 0.469244619210581,     -0.477655383903718, -0.975201793166881, 0.682654517913663,     0.0518497210685737, -0.202968937736801, 1.34534553631281,     -1.16478941779913, -0.531717155109088, 0.102482740513135,     -1.4469958183774, -2.22825029945049, 0.736299299582508, 0.969078433403733,     1.68247054433157, 1.66004727322037, 0.528464924720202, 0.210371263678428,     0.597780768372602, 0.409158083246108, 1.18087266546998, -1.05500502084119,     -0.647287154490563, -1.37530565963204, 2.29079865598061,     -1.19115435712322, -0.583743717691461, -0.268211703122627,     -1.55304790031269, -0.721181800436656, 0.212858127478069,     -1.28550173336587, -0.371101072770217, -0.413344974292555    ), B.2 = c(-1.01464005195319, 0.204628586035311, 0.0353463388533074,     0.0417871658224269, 0.429973929155859, -1.69143643532566,     0.354566368239241, -1.51335331719416, -0.461257395770696,     -1.44523588502579, -0.459707829903633, -0.573909739970372,     -0.199001203846781, 1.16716694394057, 0.321095521766469,     0.282146069731189, 0.523402605929479, 0.963830265145436,     -0.75376150420671, 0.464218668905286, -0.361152264217561,     -1.45888297453573, -0.905820781584631, 0.928618281558683,     0.618482525898301, -0.603779120283219, -0.607507123955222,     -0.74024766780002, -0.699623007211536, -0.938089623093439,     0.685002510413123, 1.10782792905334, 0.585554072180063, -1.01669348500083,     0.272910567339117, 0.0287941704371473, 0.637451984117485,     -0.292302185061381, 0.961850016561611, -1.14150074840684,     0.651622359934682, -0.479097242863109, 0.321287801500135,     -1.57676918008003, -0.0264344946512149, 0.223661840870791,     0.689887400494361, 1.68884751728355, -1.59114826795481, 0.585438974039333    ), B.3 = c(0.753964268973795, 0.673348507495264, -2.29623944279174,     -0.882905039731701, 0.407698789766199, 1.72609055280319,     0.20391856184446, 0.166675663968511, -1.06500055644137, -1.50404467282019,     1.36937893011782, -0.879487872882976, -1.74623894888564,     0.626131200586922, 0.739887035280773, -1.00970993836034,     0.46443846804841, 1.12632437205808, 0.99771596200492, -0.707883632707583,     -0.147309838310233, -0.400273069059768, 0.791864295953013,     -1.03705184257833, -1.51701532803853, 0.597056759914914,     -2.18818808675701, 0.607650144893199, 0.330686280942303,     -1.76170640599654, -1.28894803194036, -1.45201713524146,     0.44911054536885, -1.3903193260462, -0.975460690052437, 1.39808568012115,     -0.484882954577414, -1.17711692119812, -0.940099869359524,     -1.6066764832219, 0.968883017934993, 0.152103514475846, -1.1304106378298,     0.262104515245678, -1.49952982968881, 0.123047448871567,     0.997871460084045, 1.85584531635656, -2.0751286251103, -0.541629039003244    ), C = c(0.746188492269561, -0.718319570344134, -1.22363560837649,     0.414036447674697, 0.511893323511778, -1.18319496185733,     0.652122599907231, 0.578479607537796, 0.502061921214377,     0.3198367083871, 0.0544221704915678, 0.795075456939868, 0.261514269544615,     -0.934871865123613, -2.13088804071348, 2.53935773428245,     -2.71938230049166, 0.891830443437701, -0.558963166250127,     -0.656248569518178, -0.138684260093054, -0.192026447982064,     -1.90414183724443, -1.04340929257884, 1.15805494475736, -0.594688467599563,     -0.100358627683198, -0.889330612973605, 0.905657817801268,     -0.825081866754652, -0.0123713508866991, 0.954560939593388,     -0.879120200552091, 0.155394305438574, 0.623496882123116,     -1.08875062681659, 2.4605223815887, -0.80956042266767, -3.43718743730003,     -0.479041711213757, -0.545369163265832, 0.551361507085472,     0.720350547112863, 0.908662256378548, 1.05154201673885, 0.0568262066805496,     -1.12410410016916, -0.576729988690141, 0.455305423227269,     0.578757928473434), D.3 = c(0.535545232971686, -1.06904333586733,     0.0994451480071824, -0.667836279173496, 0.067142466431307,     2.15036611176238, 1.90316781577672, 1.78562289669109, -0.736057860490353,     -1.67138934868369, 1.07928631971213, 0.827749770658577, 1.63196131700969,     0.472056636535478, -0.586518123661502, 0.404690541871221,     1.08580703162125, -1.34226357785351, -1.42346135982059, 1.06234046634865,     1.34800977837312, -1.16792572891863, -1.16639296309266, -0.78766739656627,     0.188869877061612, -1.73205501805457, -0.0462954986855153,     0.381740463742741, 0.0761794642522007, 1.25813149646025,     -2.12072330970691, 0.879490505058884, -0.458613596713223,     -0.561481858119989, 0.320072221890852, 0.0821455675240964,     -0.668402200288719, -0.520418504475691, -0.669777533626632,     -1.3941766798985, -1.25323124972052, 2.01997010101911, 4.38276501013402,     -0.318974871526826, -1.60570476511159, 0.889219563169225,     -0.697122717710732, 1.13588317832665, -0.670590306886899,     0.542630143179208), A.Infancy = c(0.158513153216238, 0.279212362467514,     -0.54072675932255, -0.734548175005575, 0.331046909868802,     -1.18468332465235, -0.509467564102275, 0.061930276558456,     -0.355170579862181, 0.435152364344348, 0.793170723485922,     0.423807920595471, 0.930883069288524, 0.567469352968956,     -0.0518204260011531, 0.251436513814565, 0.348450511515923,     0.129219122280214, 1.15598095838302, -0.116222001413428,     0.15386546004597, 0.736479210487709, -0.259569710053304,     0.306646508275884, 0.560136125962098, -0.625495080048623,     -0.091590710494186, -0.440301522657822, -0.040123133740807,     0.532195231010001, -0.767329702551734, 0.289489593290019,     -0.106237349637605, -0.0411739770568175, 0.198253384767975,     0.808654024212288, 0.396534572970714, -0.778113885104216,     0.830952710756203, 1.20212565576136, -0.510909232102695,     -0.119186164771222, -0.524880640399194, 1.39003716134789,     -0.291579146323905, 0.638960223030286, 0.006373158154801,     -0.137883677983883, -0.978242568788597, -0.121384630340799    ), A.Toddlerhood = c(-0.870607706278175, -0.698208021938386,     0.944165434707425, -1.01379090584771, -1.34637040069686,     0.535172542842567, -1.54241002195025, 1.53405849999373, -0.507102649603417,     -0.10045388612321, 1.88186826844102, -0.308765858055712,     0.322744669308751, 0.232593555395557, 0.747868704441134,     0.14834131879584, 0.236727191954996, -2.59962303653058, -1.25246979488042,     -1.24125341705975, -1.85896153721855, -0.265500734498976,     -0.51283230666195, -0.467164047238227, 0.630206299899219,     1.36277057726297, -0.672869263171041, 0.395628641487794,     -0.691699450197584, 0.931458746005471, 0.183596539168729,     0.906332258024459, 1.08613452924112, -0.0828150751546846,     -1.33750568838534, -0.279623166537107, -1.29772786387416,     0.416714509357019, -1.13952935249532, 0.631605048296824,     -0.667664694947506, 1.22459606989855, -0.649876302449042,     0.14471280679046, 0.800618216377827, 0.735372229438395, 0.489558304385338,     1.33611873598614, 1.98799326048856, -0.000821600863269169    )), row.names = c("1", "2", "3", "4", "5", "6", "7", "8", "9", "10", "11", "12", "13", "14", "15", "16", "17", "18", "19", "20", "21", "22", "23", "24", "25", "26", "27", "28", "29", "30", "31", "32", "33", "34", "35", "36", "37", "38", "39", "40", "41", "42", "43", "44", "45", "46", "47", "48", "49", "50"), class = "data.frame"),     na.action = function (object, ...)     {        if (NCOL(object) == 0)             object        else {            ok <- complete.cases(object)            if (all(ok))                 object            else stop(message)        }    }): object is not a matrix