Fits outcome model

Fits outcome model

getModel(
  d,
  exposure,
  exposure_time_pts,
  outcome,
  epochs,
  exp_epochs,
  int_order,
  model,
  fam,
  covariates,
  verbose
)

Arguments

d

wide data frame

exposure

name of exposure variable

exposure_time_pts

list of integers at which weights will be created/assessed that correspond to time points when exposure was measured

outcome

name of outcome variable with ".timepoint" suffix

epochs

data frame of exposure epoch labels and values

exp_epochs

list of exposure epochs

int_order

integer specification of highest order exposure main effects interaction for interaction models

model

character indicating one of the following outcome models:

"m0" (exposure main effects)

"m1" (exposure main effects & covariates)

"m2" (exposure main effects & their interactions)

"m3" (exposure main effects, their interactions, & covariates)

"covs" (covariate only model)

"int" (intercept only model)

fam

function specification for svyglm model

covariates

list of characters reflecting variable names of covariates for covariate models

verbose

TRUE or FALSE indicator for user output (default is TRUE)

Value

list of fitted model(s)

Examples

test <- data.frame(ID = 1:50,
                   A.1 = rnorm(n = 50),
                   A.2 = rnorm(n = 50),
                   A.3 = rnorm(n = 50),
                   B.1 = rnorm(n = 50),
                   B.2 = rnorm(n = 50),
                   B.3 = rnorm(n = 50),
                   C = rnorm(n = 50),
                   D.3 = rnorm(n = 50))
test[, c("A.1", "A.2", "A.3")] <- lapply(test[, c("A.1", "A.2", "A.3")], as.numeric)

f <- createFormulas(exposure = "A",
                    exposure_time_pts = c(1, 2, 3),
                    outcome = "D.3",
                    tv_confounders = c("A.1", "A.2", "A.3", "B.1", "B.2", "B.3"),
                    ti_confounders = "C",
                    type = "full",
                    save.out = FALSE)
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A-D.3 at A time point 1 is: 
#> A.1 ~ C
#> <environment: 0x561a509f7a80>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A-D.3 at A time point 2 is: 
#> A.2 ~ A.1 + B.1 + C
#> <environment: 0x561a509f7a80>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A-D.3 at A time point 3 is: 
#> A.3 ~ A.1 + A.2 + B.1 + B.2 + C
#> <environment: 0x561a509f7a80>
#> 

w <- createWeights(data = test,
                   exposure = "A",
                   outcome = "D.3",
                   formulas = f,
                   save.out = FALSE)
#> Error in weightit2cbps.cont(covs = covs, treat = treat, subset = by.factor ==     i, s.weights = s.weights, missing = missing, verbose = verbose,     ...): The package "CBPS" is required.

epochs = data.frame(epochs = c("Infancy", "Toddlerhood"),
                    values = I(list(c(1, 2), c(3))))
e <- apply(expand.grid("A", as.character(epochs[, 1])), 1,
           paste, sep = "", collapse = ".")
test$weights <- w[[1]]$weights
#> Error in eval(expr, envir, enclos): object 'w' not found

g <- getModel(d = test,
              exposure = "A",
              exposure_time_pts = c(1, 2, 3),
              outcome = "D.3",
              epochs = epochs,
              exp_epochs = e,
              fam = gaussian,
              model = "m0")
#> Error in model.frame.default(~weights, data = structure(list(ID = 1:50,     A.1 = c(1.01944942561355, -0.864834411821448, 0.275324415917066,     2.92302164542677, 0.0630986798114381, -1.80625814229797,     1.42842347982184, -1.58279883306312, -0.0322566946810782,     0.615116423987907, 0.554312183671042, -0.143791273421398,     -1.14508532356768, -0.0812559553748952, 0.164912791922886,     1.47034141745056, -0.378015182975184, -0.226339404775899,     -0.264781514054468, 1.30454784502972, 0.645681781097998,     1.69813783613586, 2.12346416756102, 1.42420701857126, 1.44473935860823,     0.886069005077565, -1.52980384921561, 1.71698882445711, -0.491642904621403,     -0.851668832605979, -0.796518071197218, -0.543034647472007,     -0.702471400294935, -0.0762970642354944, -0.212981472512892,     -0.22153329879206, -0.47957814121079, 0.395126918478579,     -1.75462682581782, -0.521839070619167, 0.0608680820834629,     1.43278160943188, -0.852189972450279, -0.673179907741431,     -0.0358278723373196, 0.59189768896251, 0.716754712300467,     -0.699446770103478, -0.273299607667414, 0.376243550476844    ), A.2 = c(-0.958024346502478, 1.08015791458822, -0.715526071912307,     -0.363558563707203, -1.89269128392294, 1.20454056271097,     -0.510127547926048, -0.0105955112539365, 0.118910998123906,     -0.262806602023153, -0.461386886635667, -0.462831987005868,     -0.556206554810123, 1.01899358013424, -1.24212980530209,     0.163964478930539, -0.221582174439158, -0.169034732431382,     -0.278888770863979, 2.07174666805339, -0.613941825656798,     -0.842827986507942, -0.342752132453232, -0.839235016511224,     1.72684823780777, 0.247848457081723, 1.27840053913385, 0.89378455005772,     -0.558467446685258, 0.66034061870683, -0.0390160367110181,     0.787335627848119, -0.0572064519003098, -1.67688542265906,     -0.0637506038087049, 0.775240999341558, -0.777341350088271,     -0.54258876979451, 0.843445348518661, 0.383907191091888,     -1.04794034604458, 0.061317251998454, 0.59488092795907, 0.538213566035803,     -1.4088756289044, -1.37594750714401, 1.67942404097699, 0.953764932422272,     -0.40601439208989, -1.52360508915304), A.3 = c(-0.704034729287968,     1.5334034110892, 0.110621725684365, 0.809246333677954, -0.170722645463786,     -0.188631688499142, 0.673884532719794, 0.122350651257708,     -0.873173164123976, -0.238881322039891, -0.334509277165625,     1.26267217950611, -1.12306158908974, -0.520916285510272,     -0.222332689206945, 0.287181304292014, 0.55686484052602,     0.751414061703106, -0.387013717431999, 0.481195440651183,     -0.566182996708152, -0.761643123686514, 0.454924463478946,     -0.454665488013569, -1.64076124873216, -1.31840332905913,     1.55237926190619, 0.342465131597603, 0.562898814056634, 0.0904801880780724,     -0.504096221443199, 0.000911015812765451, -1.20221434142805,     1.2941393833807, -0.483516970101207, 0.33476005806627, -0.537574847312403,     0.0305483213443504, -0.332839591042035, -1.11682531246795,     -1.62196761636962, 0.822016913391243, 0.443918540341889,     -0.10561214981304, -1.71423431973347, -1.40624304696484,     -2.06545835645921, 1.07070850189164, -0.83069434468693, -0.67498343087005    ), B.1 = c(1.50530793949755, 1.09948518501962, 1.96134428330094,     -0.0143975082724855, -0.669284371096881, 0.140914951921828,     -0.266466147558652, -0.289747584329002, 0.5743303764795,     -1.51114733157395, 0.618102869338724, -1.07749312186694,     -0.657088862880437, 0.12712039347694, 0.52630836313063, 1.30644672459988,     0.530767637432868, -0.431803557023124, 1.68848378403733,     -0.733823289314224, -0.633822380076261, -0.730142455010187,     -0.894072495841011, 0.861592215621451, -1.67776177791035,     0.534025326547252, -0.0428282911446073, 0.136606544137921,     -1.51082726847018, -0.180660874851709, -0.0782816181057635,     0.951970674224227, -0.669787120784422, 1.42110490465804,     0.878090682916006, -0.110884203950672, 0.674161778109789,     1.62953929348604, 1.33295441616139, -1.19483398966507, -1.28672109770725,     1.94411340699257, -0.453586765288556, 0.754525436817027,     -0.807553818131725, -0.769003567449298, -1.36060897944665,     -0.450011887213012, -0.233547104551588, -0.243383960387672    ), B.2 = c(-1.02385061074926, 2.09896872593739, -0.362678304739918,     0.0397028445893014, 1.03268652245074, -0.714255210126657,     -0.101186457119405, -1.51360000361398, -1.70239416007487,     0.938870088645952, 0.254092940311461, 0.382639911165015,     -0.658812039153918, -0.426579614709241, 0.290475274055421,     -1.42071307736438, 0.464193230653227, 0.509665737180866,     1.8293241066965, 1.45601577744937, -0.948860303470508, -0.273741396281657,     -0.382742494063132, 0.805099668760303, 1.47667534614397,     0.325918592649888, -1.23707870309647, 0.275770214503399,     -0.205385535560292, 0.169431708234552, 0.215513579276747,     -0.944807194746869, -1.51420185828947, -0.179136645829313,     -0.895681340350197, -0.319717200587488, -0.0786764372295024,     0.793785104972491, -1.39872856055956, 0.559174469079042,     0.362416277047273, 1.73952973784432, 1.32229338942182, -0.50202964382895,     -0.824344377616499, 0.638966502382127, 1.5325235578313, 1.77360216532537,     0.702034062282944, 0.0709403640484053), B.3 = c(-0.952282476123334,     0.236927817809289, 0.449249607204808, 1.00115800376555, -0.552637284714304,     -0.242831695057834, -1.327870957815, 0.0420891716090956,     0.0622197596752601, -1.16346478502799, 0.655676700986419,     1.34867312829572, 0.96920513642293, -1.02131168092845, -0.705899145180709,     -0.81566936986377, 1.72932862796934, 0.769045044251242, -0.187099413995673,     -0.51294968316202, 0.461870194000556, 0.0225345393628463,     0.104633986335789, 1.06881077090361, -0.599406024571763,     0.179758770549307, -1.09382076392209, -0.571826331708589,     0.697174709367675, -1.0507774534459, -0.731977410049014,     -1.40576309625593, -0.342429712877358, 1.04818754330643,     -1.38815547081203, -1.35214291022121, 0.588750386345172,     1.23362585419772, 1.17673808777098, -0.212323920378493, 0.232113255437012,     -0.079392718548765, 1.05194290990908, -1.11244438161386,     1.27084780418023, -0.802883667038281, -1.10225128107954,     -0.00188225198704255, 0.142395004806022, -0.910045246600125    ), C = c(0.664548819808804, 1.83601205713468, -0.052255972581841,     -0.612633544688956, 0.719339813680838, -0.139053127718567,     0.00244502594357902, 0.403660387301343, -0.156441469400953,     2.28525463625901, 0.742788653533266, -1.46467081375743, 1.8108208037248,     1.06457130270241, -1.01905860818577, 0.48669221416623, -0.494252464906771,     -1.87929168782348, 2.64724739188446, 0.843628771890112, -0.875225749684188,     -0.891481863310828, 0.890617368739252, -0.882128673808053,     0.180978626590414, -2.16115053182726, 0.155452153889219,     1.76000787504024, -0.610579473679964, -0.323595856425971,     0.793499549513836, -0.81337895749976, 0.2513814697437, 1.89971015489469,     -1.07523767687837, 0.280461662352044, -0.475205575826548,     -1.08824395658394, 0.12261621084435, 0.0962353083845985,     0.828203852187958, -0.521868756692318, -0.470613171027334,     0.408871424373095, 2.22212016580884, 0.503017259644729, 2.3129533414029,     1.65138964125276, -0.028669238955757, -2.30317606282533),     D.3 = c(-1.82565922271868, -0.469356200540584, 0.0185972764178875,     1.25771012211024, 0.0851157540997529, 1.00643365611632, -1.37294998424621,     0.0855006919597897, 1.70171320707702, 1.0262402392086, -0.983056407426257,     -1.52582518521577, -0.347531998817146, -0.96064569648708,     0.950996382669812, 0.191913994403403, -1.11640445005116,     -0.940408323550326, -0.0312928017457901, 0.191025074877111,     -1.02795740932541, -0.591311948557928, -0.0246742168155632,     0.877244654060494, -0.531471522100152, 0.956849543721199,     -0.108885766804143, 1.34128937980699, 2.24352982662966, -1.26967110792253,     -0.0585121569252629, 0.195036661640648, 0.424093911378584,     -1.11062801690936, 0.571995493769303, 0.814053425988003,     -0.941404539903536, -1.39259124944239, 0.100144650389677,     0.167493722826109, -1.69764046703002, 1.66628042229382, -0.754764753829289,     -0.139612094771878, 0.0532942083894094, 0.266090504109129,     -0.314609613710728, -0.499857810917695, -1.98368492627024,     0.059546168940862), A.Infancy = c(0.030712539555536, 0.107661751383386,     -0.22010082799762, 1.27973154085978, -0.914796302055751,     -0.3008587897935, 0.459147965947896, -0.796697172158528,     0.0433271517214139, 0.176154910982377, 0.0464626485176875,     -0.303311630213633, -0.850645939188901, 0.468868812379672,     -0.538608506689602, 0.81715294819055, -0.299798678707171,     -0.19768706860364, -0.271835142459223, 1.68814725654155,     0.0158699777206, 0.427654924813959, 0.890356017553894, 0.292486001030018,     1.585793798208, 0.566958731079644, -0.12570165504088, 1.30538668725741,     -0.52505517565333, -0.0956641069495745, -0.417767053954118,     0.122150490188056, -0.379838926097622, -0.876591243447277,     -0.138366038160798, 0.276853850274749, -0.62845974564953,     -0.0737309256579655, -0.455590738649579, -0.0689659397636395,     -0.493536131980559, 0.747049430715167, -0.128654522245605,     -0.067483170852814, -0.72235175062086, -0.39202490909075,     1.19808937663873, 0.127159081159397, -0.339656999878652,     -0.573680769338098), A.Toddlerhood = c(-0.704034729287968,     1.5334034110892, 0.110621725684365, 0.809246333677954, -0.170722645463786,     -0.188631688499142, 0.673884532719794, 0.122350651257708,     -0.873173164123976, -0.238881322039891, -0.334509277165625,     1.26267217950611, -1.12306158908974, -0.520916285510272,     -0.222332689206945, 0.287181304292014, 0.55686484052602,     0.751414061703106, -0.387013717431999, 0.481195440651183,     -0.566182996708152, -0.761643123686514, 0.454924463478946,     -0.454665488013569, -1.64076124873216, -1.31840332905913,     1.55237926190619, 0.342465131597603, 0.562898814056634, 0.0904801880780724,     -0.504096221443199, 0.000911015812765451, -1.20221434142805,     1.2941393833807, -0.483516970101207, 0.33476005806627, -0.537574847312403,     0.0305483213443504, -0.332839591042035, -1.11682531246795,     -1.62196761636962, 0.822016913391243, 0.443918540341889,     -0.10561214981304, -1.71423431973347, -1.40624304696484,     -2.06545835645921, 1.07070850189164, -0.83069434468693, -0.67498343087005    )), row.names = c(NA, -50L), class = "data.frame"), na.action = function (object,     ...) {    if (NCOL(object) == 0)         object    else {        ok <- complete.cases(object)        if (all(ok))             object        else stop(message)    }}): object is not a matrix