Fits outcome model

Fits outcome model

getModel(
  d,
  exposure,
  exposure_time_pts,
  outcome,
  epochs,
  exp_epochs,
  int_order,
  model,
  fam,
  covariates,
  verbose
)

Arguments

d

wide data frame

exposure

name of exposure variable

exposure_time_pts

list of integers at which weights will be created/assessed that correspond to time points when exposure was measured

outcome

name of outcome variable with ".timepoint" suffix

epochs

data frame of exposure epoch labels and values

exp_epochs

list of exposure epochs

int_order

integer specification of highest order exposure main effects interaction for interaction models

model

character indicating one of the following outcome models:

"m0" (exposure main effects)

"m1" (exposure main effects & covariates)

"m2" (exposure main effects & their interactions)

"m3" (exposure main effects, their interactions, & covariates)

"covs" (covariate only model)

"int" (intercept only model)

fam

function specification for svyglm model

covariates

list of characters reflecting variable names of covariates for covariate models

verbose

TRUE or FALSE indicator for user output (default is TRUE)

Value

list of fitted model(s)

Examples

test <- data.frame(ID = 1:50,
                   A.1 = rnorm(n = 50),
                   A.2 = rnorm(n = 50),
                   A.3 = rnorm(n = 50),
                   B.1 = rnorm(n = 50),
                   B.2 = rnorm(n = 50),
                   B.3 = rnorm(n = 50),
                   C = rnorm(n = 50),
                   D.3 = rnorm(n = 50))
test[, c("A.1", "A.2", "A.3")] <- lapply(test[, c("A.1", "A.2", "A.3")], as.numeric)

f <- createFormulas(exposure = "A",
                    exposure_time_pts = c(1, 2, 3),
                    outcome = "D.3",
                    tv_confounders = c("A.1", "A.2", "A.3", "B.1", "B.2", "B.3"),
                    ti_confounders = "C",
                    type = "full",
                    save.out = FALSE)
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A-D.3 at A time point 1 is: 
#> A.1 ~ C
#> <environment: 0x55a1ea921a40>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A-D.3 at A time point 2 is: 
#> A.2 ~ A.1 + B.1 + C
#> <environment: 0x55a1ea921a40>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A-D.3 at A time point 3 is: 
#> A.3 ~ A.1 + A.2 + B.1 + B.2 + C
#> <environment: 0x55a1ea921a40>
#> 

w <- createWeights(data = test,
                   exposure = "A",
                   outcome = "D.3",
                   formulas = f,
                   save.out = FALSE)
#> Error in weightit2cbps.cont(covs = covs, treat = treat, subset = by.factor ==     i, s.weights = s.weights, missing = missing, verbose = verbose,     ...): The package "CBPS" is required.

epochs = data.frame(epochs = c("Infancy", "Toddlerhood"),
                    values = I(list(c(1, 2), c(3))))
e <- apply(expand.grid("A", as.character(epochs[, 1])), 1,
           paste, sep = "", collapse = ".")
test$weights <- w[[1]]$weights
#> Error in eval(expr, envir, enclos): object 'w' not found

g <- getModel(d = test,
              exposure = "A",
              exposure_time_pts = c(1, 2, 3),
              outcome = "D.3",
              epochs = epochs,
              exp_epochs = e,
              fam = gaussian,
              model = "m0")
#> Error in model.frame.default(~weights, data = structure(list(ID = 1:50,     A.1 = c(-0.0292405722763792, -0.180020829127678, 0.265054991769942,     -0.13990752461907, 0.207529238807061, -0.240369358150526,     -0.292938618137644, 1.82834237729791, 2.08961853430101, 0.44871664227135,     -0.000774959585165107, 1.62734032511918, -1.19022087626141,     -0.349932573730981, -0.445118870802507, 2.65610764660867,     0.5515603396158, -1.03454610672777, -0.450274466929237, 0.122446465854734,     0.925421399059615, -0.0150430234213363, -0.25986563788363,     -0.607071275044532, 0.162945592455182, -0.076419296326316,     -0.354395126270259, 0.262492917847565, 1.41908676744466,     0.425369516600147, -0.18406690252491, 0.172873704198209,     -0.643989757487433, -1.1530092209746, 1.88480198490309, 0.625259034984203,     0.416756433821275, -1.16648587103873, 0.890681332932044,     -0.367827015809702, -0.335476232144336, -0.399974560951919,     0.673633640578431, -0.551379964659108, -1.3406572425919,     0.179205377694352, 1.74415612481779, 0.0957034240064859,     1.77063706075641, -0.681866098178185), A.2 = c(-0.379450161077009,     1.13517997028489, -0.455809535941295, 0.0260849424403026,     0.0339389414376357, -0.146261832280437, -0.466373814133969,     -0.0374349608878035, -0.0864830115478366, -0.898775500074559,     -1.68180593762872, -0.788812825216959, 1.06089370689949,     1.67160745691409, 1.01992967731089, 0.500891024539958, 0.0149267199468607,     1.23310045625002, -0.847980365153247, -1.75191272102653,     -0.559139569813583, -0.86401189868267, 0.663149538227796,     0.200336927403802, 0.914041866385716, 0.241449344321261,     -0.147567983801607, 0.989573947439046, -0.146292159819123,     0.843108087165451, -0.0440575440153956, 1.28410959640555,     -0.936558948455123, -0.0445337795258678, 0.232763089386765,     0.365057795600909, 0.808879494340177, -0.443815129180803,     -0.540441972148365, -2.64502089914229, -0.660040240172781,     -0.906455519205246, -0.547784153182654, -1.09273152465625,     1.55664279563489, -0.523980938216535, 0.623593377949751,     0.993393929469843, -0.328174283029778, -0.940695273138723    ), A.3 = c(0.885546482508573, -0.0773481511887553, -0.14529005290166,     -0.0901508475190523, -0.651841638647601, 0.492283775727395,     1.78128604559991, -0.32727083638747, 0.268593702801473, 0.959460609933576,     -0.159784288021602, -0.477222693926927, -0.636815764487698,     -0.119453765621129, 0.242159719154832, 1.63514996985775,     -0.839770673276361, -1.18446199561893, 0.927702434804413,     1.39760436536344, -2.5059182505645, 0.0290109501385961, -0.435128189510393,     2.22754457476816, 0.594513375417113, -0.499930405719119,     0.0607116661978825, -1.35018159358589, 1.02437518781485,     0.413245441218741, -0.69138445506946, -0.563028210094542,     -1.45577307100813, -0.660327954261445, 0.293405993658978,     -0.223527703403842, 0.0682288639390707, -0.613633312276968,     1.57242327575115, -0.0447817784888345, 0.46487390360417,     -0.896111549057451, -0.599529275651942, -1.84303240916951,     -0.299177516202533, -1.27947970757828, -0.345827979804212,     0.744243933811806, 0.0607774475793312, -0.608980740700163    ), B.1 = c(-1.94776550329897, -0.503155411504866, 1.16721600637606,     -0.608576529485991, 0.302876931079171, 1.53065799929934,     -0.288411368459999, 1.01444526500878, 0.883607865087796,     -0.919527746723871, 0.32756339543659, 0.193603814960383,     -0.0323608816196981, 1.63569729925223, -0.00312201912633131,     -0.110869527059517, 0.193411167401428, 0.488563157435312,     -0.605839807768293, -1.85348642517012, 2.49321290577955,     -1.29346486585535, 1.08314726536996, -0.367289894710142,     -0.64311507322823, -0.541904311660402, -0.170228663957142,     1.31430488007756, 0.0564143123576673, 0.223880524956477,     0.707756404593083, 1.42469738394367, 0.423213200368554, -1.85622941755401,     -0.0823446638862693, 0.2670574134909, -1.22963381199515,     -0.643109518981163, -0.6740439933061, -0.655058737159508,     -0.602353561810916, 0.375244437049707, 0.8413357390587, 0.109235761194964,     -0.723487930261921, 0.106093798815941, -1.50552443154953,     -0.987073011331045, -0.44531414941889, 0.52698307245642),     B.2 = c(0.672259803695297, -0.221639499190004, 0.801277343920999,     2.04480298102426, 1.55106646331573, -0.199181968300917, -2.82798153901616,     -0.0273999287760953, 0.389631365019467, -0.625695430557774,     -1.22391896370261, -0.0364114846388742, 0.302125440277412,     1.95452572691203, 1.36382103062636, -0.366955225274522, -1.24250471414657,     -0.0587250863480719, 1.38456197552843, -0.711048045893221,     0.663199908085332, 1.62289172159029, 0.0299899627099402,     -1.44256875564526, -1.66718050435947, 0.657957905039811,     -0.124526257457904, 0.539014037567554, 0.23897382279272,     1.23600464891808, 1.58560373259517, -0.889537466996578, -0.295138497593455,     0.358001163250749, -2.34457086186376, 1.03068786262357, 2.17914745096306,     -0.543001135347379, -1.6990815554948, 1.13737316805547, 1.3373024530134,     -0.0682579778778636, 0.695850829928245, 0.342278994223311,     2.45118279603421, 0.945915207492993, 1.34193040884437, 1.78258487040943,     -1.28307957542664, -0.82855217940297), B.3 = c(1.04819803332527,     0.582524981618041, -0.308710749644305, 1.01552697545265,     0.279087649919133, -0.0207246194494321, -1.19791881505638,     -0.583720533030584, -0.172447737703735, -0.13940753172914,     -1.25109596332548, 0.903486439089047, 0.0209597608372056,     -0.725025716114762, -1.54515211153731, 2.19991112730197,     -1.72403275512491, -0.691786307931104, -0.237186604904858,     0.080989738562883, -0.769990606340442, 0.00450605591108647,     0.701882267807591, 0.388661408345581, 1.96431063150252, 0.193067625489307,     -0.541993091702602, 0.571473045094421, 1.1168074573751, 0.503558544150788,     2.02659240311976, -0.0349563270652593, 0.551228462993741,     1.49390279860082, -1.58781361169566, -0.351642638513358,     0.103419132877193, 0.699072870243509, 1.27698403078838, -0.438876790889053,     -0.287748108868254, -0.742635178064319, -1.81374650967743,     -0.626695947683282, -1.95497578680839, -2.06860961772261,     -1.43544206578052, 0.832516190088253, -0.136715791865738,     1.2754684551008), C = c(-0.394637853022903, -1.02115496634444,     0.720731869670534, -0.998459619328586, -0.740474279143668,     -0.611470769102736, -0.185020355646327, -0.971342946126942,     -0.957987694766549, -0.771108532476996, -0.331979529066472,     -0.7537864985834, -2.27754360996695, -0.897790823142611,     -2.75786459002207, -0.591098973168322, -0.300644526437905,     1.17192976200097, -0.949310038160194, 2.12413918564256, -0.373211678405339,     -1.58067591494975, -0.922612687078681, -0.891685882506363,     -0.0400782232365225, 1.79520400080292, 1.42533375897306,     0.328133616570995, 0.376350237764399, 3.07191889575456, 0.342734834084551,     -1.14617656831808, -0.046219619177423, 0.229323489784853,     -0.905848596221253, -0.276470598262597, -0.551244793152064,     -0.918754124862957, 1.83367446409814, -0.574309841036182,     0.27754993585682, -1.17104734762867, -0.146652518081853,     -0.444375640104166, -0.558316741060997, -1.29932311520791,     -0.672675374991439, 0.308429640386524, 0.43437294966267,     1.78227767987346), D.3 = c(-2.04388189158582, 0.0412440371753059,     1.6323148613977, 0.453239314783107, 0.848258225608512, 0.167035145476194,     -0.85923778494124, 0.736310138293594, 0.811404526150849,     0.126325760643412, -0.229649980671302, 0.0393330836162446,     -1.04738492464645, 0.60136055723827, -0.808615009576094,     0.188036876867935, -1.46740571939676, 0.0973288497110929,     1.10566892524665, 0.969343115288041, 0.225658490716457, 0.605337226909577,     -0.123719095808266, -0.314750886748933, 0.129831125653099,     -1.28454573557778, -0.184253339909824, -0.303921480966007,     0.969150723829352, -0.0540692742738138, 0.0367691960523444,     0.351304533114488, -1.26750521802959, 0.141377270949529,     -0.031332811812376, 0.641351972488382, -2.47230987287108,     0.841771662806446, 0.150453704125931, -1.50841570906394,     0.266780187228234, -1.89402704687007, 0.0931723212830112,     1.2659661422328, 0.0913918320550363, -1.28559174212964, -1.52146616618455,     -0.337254103215543, 1.01107067382546, 1.80265099242497),     A.Infancy = c(-0.204345366676694, 0.477579570578606, -0.0953772720856765,     -0.0569112910893837, 0.120734090122348, -0.193315595215482,     -0.379656216135806, 0.895453708205053, 1.00156776137659,     -0.225029428901604, -0.841290448606943, 0.419263749951111,     -0.06466358468096, 0.660837441591554, 0.287405403254191,     1.57849933557431, 0.28324352978133, 0.0992771747611251, -0.649127416041242,     -0.814733127585898, 0.183140914623016, -0.439527461052003,     0.201641950172083, -0.203367173820365, 0.538493729420449,     0.0825150239974725, -0.250981555035933, 0.626033432643305,     0.636397303812768, 0.634238801882799, -0.114062223270153,     0.728491650301879, -0.790274352971278, -0.598771500250234,     1.05878253714493, 0.495158415292556, 0.612817964080726, -0.805150500109767,     0.17511968039184, -1.506423957476, -0.497758236158558, -0.653215040078583,     0.0629247436978885, -0.822055744657679, 0.107992776521495,     -0.172387780261092, 1.18387475138377, 0.544548676738164,     0.721231388863316, -0.811280685658454), A.Toddlerhood = c(0.885546482508573,     -0.0773481511887553, -0.14529005290166, -0.0901508475190523,     -0.651841638647601, 0.492283775727395, 1.78128604559991,     -0.32727083638747, 0.268593702801473, 0.959460609933576,     -0.159784288021602, -0.477222693926927, -0.636815764487698,     -0.119453765621129, 0.242159719154832, 1.63514996985775,     -0.839770673276361, -1.18446199561893, 0.927702434804413,     1.39760436536344, -2.5059182505645, 0.0290109501385961, -0.435128189510393,     2.22754457476816, 0.594513375417113, -0.499930405719119,     0.0607116661978825, -1.35018159358589, 1.02437518781485,     0.413245441218741, -0.69138445506946, -0.563028210094542,     -1.45577307100813, -0.660327954261445, 0.293405993658978,     -0.223527703403842, 0.0682288639390707, -0.613633312276968,     1.57242327575115, -0.0447817784888345, 0.46487390360417,     -0.896111549057451, -0.599529275651942, -1.84303240916951,     -0.299177516202533, -1.27947970757828, -0.345827979804212,     0.744243933811806, 0.0607774475793312, -0.608980740700163    )), row.names = c(NA, -50L), class = "data.frame"), na.action = function (object,     ...) {    if (NCOL(object) == 0)         object    else {        ok <- complete.cases(object)        if (all(ok))             object        else stop(message)    }}): object is not a matrix