Fits outcome model

Fits outcome model

getModel(
  d,
  exposure,
  exposure_time_pts,
  outcome,
  epochs,
  exp_epochs,
  int_order,
  model,
  fam,
  covariates,
  verbose
)

Arguments

d

wide data frame

exposure

name of exposure variable

exposure_time_pts

list of integers at which weights will be created/assessed that correspond to time points when exposure was measured

outcome

name of outcome variable with ".timepoint" suffix

epochs

data frame of exposure epoch labels and values

exp_epochs

list of exposure epochs

int_order

integer specification of highest order exposure main effects interaction for interaction models

model

character indicating one of the following outcome models:

"m0" (exposure main effects)

"m1" (exposure main effects & covariates)

"m2" (exposure main effects & their interactions)

"m3" (exposure main effects, their interactions, & covariates)

"covs" (covariate only model)

"int" (intercept only model)

fam

function specification for svyglm model

covariates

list of characters reflecting variable names of covariates for covariate models

verbose

TRUE or FALSE indicator for user output (default is TRUE)

Value

list of fitted model(s)

Examples

test <- data.frame(ID = 1:50,
                   A.1 = rnorm(n = 50),
                   A.2 = rnorm(n = 50),
                   A.3 = rnorm(n = 50),
                   B.1 = rnorm(n = 50),
                   B.2 = rnorm(n = 50),
                   B.3 = rnorm(n = 50),
                   C = rnorm(n = 50),
                   D.3 = rnorm(n = 50))
test[, c("A.1", "A.2", "A.3")] <- lapply(test[, c("A.1", "A.2", "A.3")], as.numeric)

f <- createFormulas(exposure = "A",
                    exposure_time_pts = c(1, 2, 3),
                    outcome = "D.3",
                    tv_confounders = c("A.1", "A.2", "A.3", "B.1", "B.2", "B.3"),
                    ti_confounders = "C",
                    type = "full",
                    save.out = FALSE)
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A - D.3 at A time point 1 is: 
#> A.1 ~ C
#> <environment: 0x55a06a080110>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A - D.3 at A time point 2 is: 
#> A.2 ~ A.1 + B.1 + C
#> <environment: 0x55a06a080110>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A - D.3 at A time point 3 is: 
#> A.3 ~ A.1 + A.2 + B.1 + B.2 + C
#> <environment: 0x55a06a080110>
#> 

w <- createWeights(data = test,
                   exposure = "A",
                   outcome = "D.3",
                   formulas = f,
                   save.out = FALSE)
#> Error in weightit2cbps.cont(covs = covs, treat = treat, subset = by.factor ==     i, s.weights = s.weights, missing = missing, verbose = verbose,     ...): The package "CBPS" is required.

epochs = data.frame(epochs = c("Infancy", "Toddlerhood"),
                    values = I(list(c(1, 2), c(3))))
e <- apply(expand.grid("A", as.character(epochs[, 1])), 1,
           paste, sep = "", collapse = ".")
test$weights <- w[[1]]$weights
#> Error in eval(expr, envir, enclos): object 'w' not found

g <- getModel(d = test,
              exposure = "A",
              exposure_time_pts = c(1, 2, 3),
              outcome = "D.3",
              epochs = epochs,
              exp_epochs = e,
              fam = gaussian,
              model = "m0")
#> Error in model.frame.default(~weights, data = structure(list(ID = 1:50,     A.1 = c(0.655281635501574, -0.454445700832604, -0.751197120830461,     0.916192701251663, -0.19012252281015, -0.170685068181368,     0.480793187096173, -0.305094584346883, 0.132239677345843,     -0.119321437060432, 0.81210229828166, 0.171174094581846,     -0.327466241987008, -0.427227511057792, 1.00822241570397,     -2.07815368315543, 1.67440124523286, -0.958342616085404,     0.699720998112179, 0.55680580463221, 1.05715842315945, -0.630904496333702,     0.370034423124218, 0.249172222659623, 0.802453889280879,     0.116607025520455, -1.51057809847544, -0.811229149407687,     -1.49946906807058, -0.833422622753152, 0.132625476127273,     0.25310639959598, -1.39203897890683, 0.153822270977243, -0.298368843558062,     -0.163392186244996, -0.514114141059857, 2.02075212543635,     1.23346384172307, 0.465013467167806, -0.609811281446099,     -0.313169446940046, 0.09745906700944, -0.361115441752595,     1.00606950375061, -0.0166450344881014, -1.16616729778274,     0.612610016409269, 0.819765858913307, 0.459821181297977),     A.2 = c(-0.157437290157178, -0.0775441568739864, -0.252046836968894,     0.424186657112071, -3.14597231539104, 1.13359198444866, 0.54341351544752,     0.8601962130977, -0.315615398966767, -1.27841968709607, -0.557969943435541,     0.391590705036936, -0.00844525337561304, 0.70193061544951,     -0.0616346608737566, -0.939814746493947, -0.532581337662509,     0.454114311620831, -0.054432714453124, 2.18896299115398,     -0.863773026966978, 0.985548172183878, -0.59284712654131,     -0.223894925936595, 1.1628931832453, -0.214877729162642,     2.88978574670319, -0.196857461574497, 1.60795713829975, -0.473643101956713,     0.259967120289034, 1.1315896589137, 0.13663757554378, -1.54756686149757,     0.660581960436078, 0.0845907877446957, -0.295210961234794,     0.0732353576517682, 0.00638777875987202, 0.455043350830653,     1.11334416066628, 0.438089038483206, 1.09494530700418, -0.313533482826417,     1.03051329825405, 0.788449403818534, 0.146258345212101, 1.25401257848798,     1.41514490003618, 1.09619510340947), A.3 = c(2.58809459031157,     -0.533917194535604, 1.42489016614415, 2.54998308787448, -0.487981933699276,     -0.286464564511637, -1.1222751889817, -0.812642128854939,     -1.32885774140453, 1.38497145520324, 0.593141514137904, 0.0909948435489904,     -1.18033672226456, -0.30202682712712, 0.54651837068708, -0.493981594639552,     2.83237444905248, -0.0877982700599692, 0.9169106665256, 0.185965750741972,     0.4394765589036, 1.50097312804871, -0.00360579510903628,     -0.871527286474029, -0.0309438027895007, -1.01852357910042,     -1.5576722476428, 1.37896306978972, 0.122275160874528, -0.377211586046019,     0.118611424349908, 1.16543931402315, 0.544060953303168, -1.08610387717408,     1.61275645010518, 1.18949289598577, -2.35275353871717, -2.44264877718641,     -0.32722267910669, 0.515972717525192, -0.170472038237158,     -0.899005559718685, -1.05271132036064, 0.666215590766605,     1.64887745707521, -0.876913229341624, 0.61991004944111, 0.229393066556964,     -0.333688825499326, 0.771781646656295), B.1 = c(-0.725780583805244,     0.776283006470353, 0.913823663794471, -0.275217142609887,     1.73860474443248, 0.0216220997701708, -0.287061027891431,     1.26404728172708, 1.89334606772081, -0.250500105469317, 0.478938908567342,     -0.400593469309167, 1.17587991173283, -0.369956483931394,     1.10469914135503, 0.209335996405413, -0.613458235704968,     -0.458024171289261, 0.712630291356807, -0.41644897762504,     0.684048845895207, 1.40561196396752, 0.640575532640222, 0.547335860757363,     -0.348417773471251, 0.495567099078241, 0.328677604688181,     -1.66154412277006, 0.504024208917523, -0.41586243982201,     0.202931976094394, 1.98751294993372, -0.720607477105061,     -2.04618541559885, -1.79992261118679, -0.10529020352657,     -1.8521076668557, -0.681216624229723, -1.59300050191745,     0.524017815266395, 1.4141371872485, 2.37105980554852, 0.0143626247818968,     -0.336680834891251, 0.562180518993143, -0.298643532460762,     -0.523700100554148, -0.572810429446, -0.377320154179964,     -0.137378188323043), B.2 = c(-1.43534364157617, -0.113933722935471,     0.328866475613941, -0.3091689368155, -0.410440906643072,     -0.158560243023277, 0.713511233861361, 1.01312696270711,     -1.49417385655041, -0.0434158640260483, -1.20003642374149,     0.110281882776583, -0.298430439133397, 0.415257422986229,     -0.3126298383074, -0.416038127187235, 0.364628791662783,     0.482505583195618, -1.10835877534286, -0.348074873247937,     -1.0062484944103, 0.441239446990809, -0.764389738357634,     1.63052835738902, 1.43730304870505, 0.430831110835369, 0.129697449777751,     -0.752449838476744, 0.210163742614703, -0.433962246122805,     0.488790918928791, -0.605453714477275, 0.742409830361534,     1.10825288798736, 0.0619447268714131, -0.434074746057192,     -0.809783801640839, -0.452835922331939, -1.13912373870214,     -0.155456340599857, -0.781471639356066, 1.33610114930745,     -0.506652501783996, -0.522003778159286, -1.07112963338087,     -0.275951648885173, 1.48259457366534, 0.692831706370325,     0.577098191591788, -1.72005996718064), B.3 = c(0.132841612217721,     -1.08973507762441, 0.29803686667397, -0.587646209072745,     0.13446550552426, -1.19192504382143, -1.85835401661691, -0.708183496009943,     1.14577893376037, -0.844412519783042, 2.70909041434886, -0.171274799396195,     -0.294046169337687, -1.87130690082551, 0.965532930899285,     0.557412557070293, 0.0572542013309684, 0.0132328742937219,     -0.585214872872575, -2.01824361152207, 0.938053335215902,     -0.884622175565079, 1.26186514246724, -0.0358245633308925,     2.56287814306763, 1.29087555508129, 1.17378872820213, 0.362266695991343,     -1.33750170252048, 0.407435583418632, 0.768844857207335,     -0.400747198492815, 0.390189221469551, -0.785331468067847,     0.299360578810857, 0.371299957419649, -1.0938632236282, -0.2601526647282,     -1.13718883902959, -0.599834496536153, -0.882526213681949,     -0.433566577502372, -1.24798528824532, 3.25683900625124,     0.54166173087264, 0.611852254624203, 0.283005858717249, -1.03123247010329,     0.421620056667061, -0.414093657689148), C = c(0.632355708729923,     2.5278611820711, 1.94641161175159, 0.155683856691139, 1.4081468747822,     0.397355247637984, 1.60652318783206, 2.58363751097113, 0.971745009318165,     0.00408849431617589, -2.41021658763664, -1.0012397086771,     0.207541162477545, -1.17072926393503, 0.278223139629761,     0.584868712087555, 0.459412851278215, 0.460825433391593,     -0.801302053685252, 2.37938825511342, -0.554917787213678,     0.448235955304673, 1.41734317364085, -0.129162262899306,     -0.370245689302666, -2.81594762987115, 0.543278930854637,     0.0227273521587254, 0.246657045037724, 1.00585773125807,     1.02742124432486, -0.015881290552229, 1.4897949706917, 1.55649799839427,     0.421126733895904, 0.189539674119204, -1.08389912419248,     0.477556478828739, -0.309072596018394, 1.33129855080004,     0.194801676607896, -0.880383118028487, -0.663099007132187,     -0.311611701925382, -2.23487575376069, -0.667586509077556,     0.955736649092957, 1.73310382105168, -0.647475753763812,     0.365973141226526), D.3 = c(-0.0758041528447576, 1.88743599666263,     -1.03085577586105, 0.462055758864988, 0.0798621191734516,     1.3430472871252, -0.931631967072739, -0.36744125756883, 0.829668836881265,     0.0658973529007773, 0.862426815777165, 2.33669091508877,     -2.49328055733657, -0.677730218516692, 1.7562970636247, -1.2298126556986,     -0.535902149680666, 0.080676655545527, 0.730126703767331,     0.552718805685303, -1.04938991845911, -0.625827021266453,     -1.21445829668542, -0.532095201130248, -0.25015691775869,     -0.562003492571177, 1.26314792931802, 0.64551093372018, 0.122425951186588,     -0.998467483837411, -1.7050293788121, 0.500044444671875,     0.120527039692448, 0.89445094099812, 2.49942942081788, -0.00833427996526145,     -2.34092932752251, -0.45765103697187, 2.21423652865358, 0.291521830772241,     0.176047229955477, -0.0286952837224967, -1.206886408304,     0.0816066885120785, 0.953978712507994, -0.796683033640327,     -0.691857777441877, -0.876681485421417, 0.487997630997378,     -1.07301872255527), A.Infancy = c(0.248922172672198, -0.265994928853295,     -0.501621978899678, 0.670189679181867, -1.66804741910059,     0.481453458133646, 0.512103351271846, 0.277550814375409,     -0.091687860810462, -0.698870562078251, 0.12706617742306,     0.281382399809391, -0.167955747681311, 0.137351552195859,     0.473293877415107, -1.50898421482469, 0.570909953785176,     -0.252114152232287, 0.322644141829527, 1.37288439789309,     0.096692698096236, 0.177321837925088, -0.111406351708546,     0.012638648361514, 0.98267353626309, -0.0491353518210935,     0.689603824113875, -0.504043305491092, 0.054244035114585,     -0.653532862354933, 0.196296298208154, 0.69234802925484,     -0.627700701681525, -0.696872295260163, 0.181106558439008,     -0.0394006992501501, -0.404662551147326, 1.04699374154406,     0.619925810241471, 0.460028408999229, 0.25176643961009, 0.06245979577158,     0.59620218700681, -0.337324462289506, 1.01829140100233, 0.385902184665216,     -0.50995447628532, 0.933311297448625, 1.11745537947474, 0.778008142353724    ), A.Toddlerhood = c(2.58809459031157, -0.533917194535604,     1.42489016614415, 2.54998308787448, -0.487981933699276, -0.286464564511637,     -1.1222751889817, -0.812642128854939, -1.32885774140453,     1.38497145520324, 0.593141514137904, 0.0909948435489904,     -1.18033672226456, -0.30202682712712, 0.54651837068708, -0.493981594639552,     2.83237444905248, -0.0877982700599692, 0.9169106665256, 0.185965750741972,     0.4394765589036, 1.50097312804871, -0.00360579510903628,     -0.871527286474029, -0.0309438027895007, -1.01852357910042,     -1.5576722476428, 1.37896306978972, 0.122275160874528, -0.377211586046019,     0.118611424349908, 1.16543931402315, 0.544060953303168, -1.08610387717408,     1.61275645010518, 1.18949289598577, -2.35275353871717, -2.44264877718641,     -0.32722267910669, 0.515972717525192, -0.170472038237158,     -0.899005559718685, -1.05271132036064, 0.666215590766605,     1.64887745707521, -0.876913229341624, 0.61991004944111, 0.229393066556964,     -0.333688825499326, 0.771781646656295)), row.names = c("1", "2", "3", "4", "5", "6", "7", "8", "9", "10", "11", "12", "13", "14", "15", "16", "17", "18", "19", "20", "21", "22", "23", "24", "25", "26", "27", "28", "29", "30", "31", "32", "33", "34", "35", "36", "37", "38", "39", "40", "41", "42", "43", "44", "45", "46", "47", "48", "49", "50"), class = "data.frame"), na.action = function (object,     ...) {    if (NCOL(object) == 0)         object    else {        ok <- complete.cases(object)        if (all(ok))             object        else stop(message)    }}): object is not a matrix