Fits outcome model

Fits outcome model

getModel(
  d,
  exposure,
  exposure_time_pts,
  outcome,
  epochs,
  exp_epochs,
  int_order,
  model,
  fam,
  covariates,
  verbose
)

Arguments

d

wide data frame

exposure

name of exposure variable

exposure_time_pts

list of integers at which weights will be created/assessed that correspond to time points when exposure was measured

outcome

name of outcome variable with ".timepoint" suffix

epochs

data frame of exposure epoch labels and values

exp_epochs

list of exposure epochs

int_order

integer specification of highest order exposure main effects interaction for interaction models

model

character indicating one of the following outcome models:

"m0" (exposure main effects)

"m1" (exposure main effects & covariates)

"m2" (exposure main effects & their interactions)

"m3" (exposure main effects, their interactions, & covariates)

"covs" (covariate only model)

"int" (intercept only model)

fam

function specification for svyglm model

covariates

list of characters reflecting variable names of covariates for covariate models

verbose

TRUE or FALSE indicator for user output (default is TRUE)

Value

list of fitted model(s)

Examples

test <- data.frame(ID = 1:50,
                   A.1 = rnorm(n = 50),
                   A.2 = rnorm(n = 50),
                   A.3 = rnorm(n = 50),
                   B.1 = rnorm(n = 50),
                   B.2 = rnorm(n = 50),
                   B.3 = rnorm(n = 50),
                   C = rnorm(n = 50),
                   D.3 = rnorm(n = 50))
test[, c("A.1", "A.2", "A.3")] <- lapply(test[, c("A.1", "A.2", "A.3")], as.numeric)

f <- createFormulas(exposure = "A",
                    exposure_time_pts = c(1, 2, 3),
                    outcome = "D.3",
                    tv_confounders = c("A.1", "A.2", "A.3", "B.1", "B.2", "B.3"),
                    ti_confounders = "C",
                    type = "full",
                    save.out = FALSE)
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A-D.3 at A time point 1 is: 
#> A.1 ~ C
#> <environment: 0x55e7875cebf8>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A-D.3 at A time point 2 is: 
#> A.2 ~ A.1 + B.1 + C
#> <environment: 0x55e7875cebf8>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A-D.3 at A time point 3 is: 
#> A.3 ~ A.1 + A.2 + B.1 + B.2 + C
#> <environment: 0x55e7875cebf8>
#> 

w <- createWeights(data = test,
                   exposure = "A",
                   outcome = "D.3",
                   formulas = f,
                   save.out = FALSE)
#> Error in weightit2cbps.cont(covs = covs, treat = treat, subset = by.factor ==     i, s.weights = s.weights, missing = missing, verbose = verbose,     ...): The package "CBPS" is required.

epochs = data.frame(epochs = c("Infancy", "Toddlerhood"),
                    values = I(list(c(1, 2), c(3))))
e <- apply(expand.grid("A", as.character(epochs[, 1])), 1,
           paste, sep = "", collapse = ".")
test$weights <- w[[1]]$weights
#> Error in eval(expr, envir, enclos): object 'w' not found

g <- getModel(d = test,
              exposure = "A",
              exposure_time_pts = c(1, 2, 3),
              outcome = "D.3",
              epochs = epochs,
              exp_epochs = e,
              fam = gaussian,
              model = "m0")
#> Error in model.frame.default(~weights, data = structure(list(ID = 1:50,     A.1 = c(0.537712364684817, -0.10850863760105, -1.01935751480281,     2.2573284742835, 1.72762528028907, -0.981630511093245, 1.85349820934815,     0.732144488738109, 1.56427242904343, 0.114917076745904, 1.98709189701062,     -0.874282929593545, 0.305959368715143, -1.51577440826996,     -1.75164929656242, -0.665964300440905, 0.654809641879815,     -0.474892294495837, 0.187468599878584, -1.4674647032428,     -0.000235033581223549, 1.57562352526419, -1.35309281526038,     1.29785910917306, 1.22635069142193, -0.759205652074006, 0.185194855355094,     -0.078893793939738, -0.299577864191145, 1.06396154561808,     0.446543773101359, 0.375612314085517, -0.752295661571132,     0.0331734227512725, 0.922782014104273, 0.587059139118877,     0.284789066712412, -0.0326543775881887, -0.528043176244212,     -0.708106339112699, -1.325452123225, 0.154364792661443, -1.18454222894013,     0.979431080726174, -0.89317998604586, 1.81816645177035, -0.859565022126142,     2.04676334585748, -0.527715067312462, -0.115705657337524),     A.2 = c(0.391793700386144, -0.140468455432452, -0.578578039852646,     0.579247145187236, 1.30233258249817, -1.03645506520965, -1.20480596096463,     -0.53281561797775, -0.0313966201643494, -0.510007460227431,     -1.01373260698299, 0.314705143872268, 1.50679411243299, -2.13907292947865,     -0.205488438290164, 1.48835864602846, -0.916204212990636,     0.079198815801058, 0.643500594150164, -0.1448217862446, 0.154525717487087,     -0.196405314869533, -0.479016183899513, 0.454877555608756,     -1.17728141553548, -0.612999043821359, -0.30902321578391,     0.761652852607084, -0.244374701366332, -1.92862853023643,     -0.299010459520407, -0.186766130179156, -1.73964696907643,     -1.90210387705333, 0.895479556479081, -0.00976870833348142,     0.887998591660068, -0.882543853650957, -0.77988263909398,     -0.661250626654724, 0.188833835485045, -0.598309961144726,     -0.553217656692243, 0.144063437459122, 1.60342625764389,     -1.09490059190431, -0.314476448469039, 1.86452598609057,     1.54524821488043, -0.768399764285422), A.3 = c(-0.762111528457018,     0.619881375516842, -0.160030460381754, 2.27734335762691,     -2.4185994026096, 0.0221985344050712, -0.41083364342638,     -2.56826004423112, -1.91690340715156, -1.25799138126109,     -2.2698696691777, 1.20239437117691, 1.27936399449541, -2.05229216086063,     -0.231622039625692, -1.58136961553311, 0.219068483400244,     -0.387688046838523, -0.867125549114664, -0.526340134639423,     1.18293555672721, -0.143533677738806, -1.53998068410919,     -1.37810238997033, -0.758708441904046, -1.23621437185067,     2.02197942354057, -0.850632623629178, 1.77803837332223, 1.24402100826994,     -0.571466143308571, -0.649973013046165, -0.793690580047647,     1.18922479516306, -1.67866980727219, -1.03113851652957, -0.773512753627595,     -0.660438539528749, 1.26883800836015, 0.670948151336315,     1.65168015856273, -0.287134858268585, -0.793143820804239,     1.34376076737448, -0.46265143316908, 1.67471085467768, 1.68677003872831,     0.221118249750863, -0.459834064884636, -1.03969065174772),     B.1 = c(1.23471368636135, 1.88958235706075, -0.273656300886683,     0.0971182187406978, -1.25015196152822, 0.913055066851299,     -1.14310994949184, 0.875387877053053, -1.76040903776453,     -0.783505481145029, 0.0992714124571499, 1.78783018324071,     1.04339468959846, -0.665818638737396, -0.324772938919484,     -0.675119544664443, -1.75930537921678, -1.18851683277703,     0.114704683348182, -0.82407464274999, -0.648400172706635,     0.0717636348135093, -0.295609408623905, -0.494235147375674,     -0.168512844890355, 0.850640275184251, -1.08702128063833,     0.0768964495502803, 0.0144032176399699, -0.635539241244584,     1.41923414837097, 0.117604467089746, -0.364582195630364,     0.929851289443079, -1.62988014600019, -1.25399174880084,     0.277532122334748, 0.665260593705873, -1.08714644398847,     -0.3405596151842, -2.60850987417837, -0.519719560948582,     0.159414609468454, -0.18263661520486, -2.05790740497674,     0.00887152283196292, 1.38570237797609, -2.02968940330577,     -1.50438991787708, 0.200604514934535), B.2 = c(0.305895952440692,     0.40039223749769, 0.485804053618723, 0.467109925902154, 0.67302191955409,     0.0339421243455569, -1.73298550327648, 1.69843638349978,     0.346619731878892, 0.790067168713381, 1.11264047786103, -1.36381758019006,     1.72066699930708, -0.424956613991061, -2.04892267786525,     0.752585322487126, -0.563415984660473, 0.129776321501771,     0.767615867821392, 0.622039861207814, -0.447847900226481,     1.07330295238072, -0.499993497652011, -0.564889004022111,     0.770311683720655, -1.15277349066911, -0.803005630250014,     0.758716122121934, 2.7450359872937, 3.0658529823372, -0.496331965460438,     -1.21806170501026, 1.98718791536751, -0.382149207841451,     0.608458096872045, 0.753731469156564, -0.472481392021715,     -1.91988288041946, 0.233412603524092, -1.2817422271803, -0.849675146865908,     0.206738897682432, -0.452955293682138, -1.28708409858893,     -0.411916626504171, 1.36272990284886, -0.189254315255142,     0.14642316242978, 1.24225150313822, -1.00082033703759), B.3 = c(0.641855000178522,     -0.966343853909548, 1.18965656455377, -0.0248986833343145,     -0.289776429480303, -0.660326018596132, -0.524654834442517,     -1.99427534265411, -0.0493657996595407, -0.644515383246922,     0.469427171123408, -0.453717310557924, -1.37599144943307,     -0.158081202160774, 0.545912719679225, 0.062784632030938,     -0.951903717572941, 2.73810150721125, 0.636285984215744,     1.59568634713355, -0.184799478609621, 0.735761789986721,     0.761739318614761, -0.105412369696413, -0.360006922622104,     -0.76064386461128, 0.421391418674201, -0.354386263546877,     -0.477541493832556, 0.559142993199626, -1.78002732991721,     0.855762994412921, 0.442230299760615, 0.500117360595182,     0.770338824920835, 0.794999124542601, 0.466708977427116,     0.617312675806659, -0.400854289942252, 0.937012498930022,     -0.774693975186211, -1.208721760897, 0.711232017118032, 0.827465972431484,     -2.41678164290646, -0.471903733887522, -0.229544637325854,     0.179820989144908, -0.363721608208414, -0.637193304434179    ), C = c(1.148674260948, 0.599792767427738, 0.786562923060788,     -0.613842354118087, 0.509773133264765, 0.486318321739929,     -0.861247180866921, 0.0786143824565467, -0.00714398743381703,     0.46093471364295, 0.883076254909982, -0.254460731069666,     -0.344192843724624, -0.238209386954689, 0.774360855754431,     -0.631292898274102, 0.185680183810154, -0.365388843299021,     -0.579592334356971, 1.05759278176415, 0.183982871221753,     -1.59737285927702, 1.73833553647317, -2.09488744038775, 1.87150462985498,     -0.0906484726869842, 0.0797187243165968, 1.38702056150369,     -0.370561588027055, -1.35443702406908, -0.195558170754872,     1.57993167104039, -1.15922228070042, 1.87687386792381, -1.09056728192798,     -0.645036102647016, 1.00464042065381, -0.0394035299342293,     1.05531537000345, -0.480544703065817, -1.56148544855552,     0.414607797444549, -0.393074296508309, -0.227654661139981,     -0.408964774771397, -0.304579506921928, 0.838786008742043,     -1.49408355710839, -1.10002662280526, 0.0275642086738134),     D.3 = c(-1.80283187185832, -0.97325924732427, -0.261343797887779,     -1.54016085674164, -1.59561516669828, -0.41670212219146,     0.745791354473921, -0.738059094848746, 1.09888515169872,     2.44353882177773, 0.0989559290934121, -0.308824036150188,     0.0206263449413991, 0.589770692338228, -0.582178654530465,     -0.576622040771369, -0.153305855483261, 0.545804155363003,     -1.03706450786286, -0.727411803509045, -1.25198048198807,     0.591316453921567, 0.770886841731501, 0.753577918533927,     -0.91309215651478, -2.30174232907099, 0.624743958188281,     -0.435166702957234, -0.345797059321548, 0.321844310225511,     1.13066421533691, -0.382915981585311, 0.809503948319019,     -0.398236503679359, 0.362621650195753, 0.0448782232508148,     -0.906887720822252, 0.491618542370662, -0.158112702266036,     -1.02661049297703, 2.83262771756463, 0.448919264871228, -0.557834982593362,     0.658350398472962, -1.90405038481917, -0.42076632321663,     0.413489765732782, -1.54502540540162, -0.276425567918445,     -0.282404613982716), A.Infancy = c(0.464753032535481, -0.124488546516751,     -0.798967777327728, 1.41828780973537, 1.51497893139362, -1.00904278815145,     0.32434612419176, 0.0996644353801795, 0.76643790443954, -0.197545191740764,     0.486679645013815, -0.279788892860639, 0.906376740574066,     -1.82742366887431, -0.978568867426292, 0.411197172793777,     -0.13069728555541, -0.197846739347389, 0.415484597014374,     -0.8061432447437, 0.0771453419529317, 0.689609105197328,     -0.916054499579947, 0.876368332390908, 0.024534637943225,     -0.686102347947682, -0.061914180214408, 0.341379529333673,     -0.271976282778739, -0.432333492309175, 0.073766656790476,     0.0944230919531805, -1.24597131532378, -0.934465227151029,     0.909130785291677, 0.288645215392698, 0.58639382918624, -0.457599115619573,     -0.653962907669096, -0.684678482883712, -0.568309143869978,     -0.221972584241642, -0.868879942816186, 0.561747259092648,     0.355123135799015, 0.36163292993302, -0.58702073529759, 1.95564466597402,     0.508766573783984, -0.442052710811473), A.Toddlerhood = c(-0.762111528457018,     0.619881375516842, -0.160030460381754, 2.27734335762691,     -2.4185994026096, 0.0221985344050712, -0.41083364342638,     -2.56826004423112, -1.91690340715156, -1.25799138126109,     -2.2698696691777, 1.20239437117691, 1.27936399449541, -2.05229216086063,     -0.231622039625692, -1.58136961553311, 0.219068483400244,     -0.387688046838523, -0.867125549114664, -0.526340134639423,     1.18293555672721, -0.143533677738806, -1.53998068410919,     -1.37810238997033, -0.758708441904046, -1.23621437185067,     2.02197942354057, -0.850632623629178, 1.77803837332223, 1.24402100826994,     -0.571466143308571, -0.649973013046165, -0.793690580047647,     1.18922479516306, -1.67866980727219, -1.03113851652957, -0.773512753627595,     -0.660438539528749, 1.26883800836015, 0.670948151336315,     1.65168015856273, -0.287134858268585, -0.793143820804239,     1.34376076737448, -0.46265143316908, 1.67471085467768, 1.68677003872831,     0.221118249750863, -0.459834064884636, -1.03969065174772)), row.names = c(NA, -50L), class = "data.frame"), na.action = function (object, ...) {    if (NCOL(object) == 0)         object    else {        ok <- complete.cases(object)        if (all(ok))             object        else stop(message)    }}): object is not a matrix