Fits outcome model

Fits outcome model

getModel(
  d,
  exposure,
  exposure_time_pts,
  outcome,
  epochs,
  exp_epochs,
  int_order,
  model,
  fam,
  covariates,
  verbose
)

Arguments

d

wide data frame

exposure

name of exposure variable

exposure_time_pts

list of integers at which weights will be created/assessed that correspond to time points when exposure was measured

outcome

name of outcome variable with ".timepoint" suffix

epochs

data frame of exposure epoch labels and values

exp_epochs

list of exposure epochs

int_order

integer specification of highest order exposure main effects interaction for interaction models

model

character indicating one of the following outcome models:

"m0" (exposure main effects)

"m1" (exposure main effects & covariates)

"m2" (exposure main effects & their interactions)

"m3" (exposure main effects, their interactions, & covariates)

"covs" (covariate only model)

"int" (intercept only model)

fam

function specification for svyglm model

covariates

list of characters reflecting variable names of covariates for covariate models

verbose

TRUE or FALSE indicator for user output (default is TRUE)

Value

list of fitted model(s)

Examples

test <- data.frame(ID = 1:50,
                   A.1 = rnorm(n = 50),
                   A.2 = rnorm(n = 50),
                   A.3 = rnorm(n = 50),
                   B.1 = rnorm(n = 50),
                   B.2 = rnorm(n = 50),
                   B.3 = rnorm(n = 50),
                   C = rnorm(n = 50),
                   D.3 = rnorm(n = 50))
test[, c("A.1", "A.2", "A.3")] <- lapply(test[, c("A.1", "A.2", "A.3")], as.numeric)

f <- createFormulas(exposure = "A",
                    exposure_time_pts = c(1, 2, 3),
                    outcome = "D.3",
                    tv_confounders = c("A.1", "A.2", "A.3", "B.1", "B.2", "B.3"),
                    ti_confounders = "C",
                    type = "full",
                    save.out = FALSE)
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A-D.3 at A time point 1 is: 
#> A.1 ~ C
#> <environment: 0x55b1c37a2510>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A-D.3 at A time point 2 is: 
#> A.2 ~ A.1 + B.1 + C
#> <environment: 0x55b1c37a2510>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A-D.3 at A time point 3 is: 
#> A.3 ~ A.1 + A.2 + B.1 + B.2 + C
#> <environment: 0x55b1c37a2510>
#> 

w <- createWeights(data = test,
                   exposure = "A",
                   outcome = "D.3",
                   formulas = f,
                   save.out = FALSE)
#> Error in weightit2cbps.cont(covs = covs, treat = treat, subset = by.factor ==     i, s.weights = s.weights, missing = missing, verbose = verbose,     ...): The package "CBPS" is required.

epochs = data.frame(epochs = c("Infancy", "Toddlerhood"),
                    values = I(list(c(1, 2), c(3))))
e <- apply(expand.grid("A", as.character(epochs[, 1])), 1,
           paste, sep = "", collapse = ".")
test$weights <- w[[1]]$weights
#> Error in eval(expr, envir, enclos): object 'w' not found

g <- getModel(d = test,
              exposure = "A",
              exposure_time_pts = c(1, 2, 3),
              outcome = "D.3",
              epochs = epochs,
              exp_epochs = e,
              fam = gaussian,
              model = "m0")
#> Error in model.frame.default(~weights, data = structure(list(ID = 1:50,     A.1 = c(-0.178165647539555, 0.615518528926471, 0.706472778664024,     -0.26201679787614, -0.113968910766884, 0.170150809871549,     1.09810502653138, -0.0947405011956105, -0.942026080539992,     -1.00535439790178, 0.133350904211974, -1.11415833182076,     0.718882066936684, 0.335299318046651, 1.31004337500606, -0.865324525711305,     0.223297952429872, 0.115448817450325, -1.2561307249801, 1.14263133597938,     0.122000072282017, 0.320000028752159, 0.349982795644132,     -1.42993128448583, -0.417853321559901, -1.60686910859262,     -0.0522190259641001, 1.04627233186536, -0.0200375940385779,     -0.130602184364075, 1.67293639859635, -0.692814107562364,     2.48540920234608, -0.40572575450167, -1.32970542231728, 0.130607428902863,     -1.39396848415242, 0.959007665399611, -1.67496784337592,     -2.06860013679189, -0.630093265577512, 0.635044580686887,     1.0684548723235, 0.411609582773023, -1.73568806871747, 0.169487391454087,     0.0712360020692474, -0.30390340298697, 0.118331850074779,     0.823835549571912), A.2 = c(0.382070423977647, 0.306128016021269,     -1.10915739878777, 0.235760698197225, -0.248422876229382,     -0.982997186543684, 1.30689080875062, 1.68030885949305, 0.0537035715459919,     -1.76570361671147, -2.31221176809996, -0.759834271778274,     -0.66633894541101, -1.49292672704674, -0.220832535411368,     -0.905863100453573, -0.969556967709064, 1.28867652619103,     0.291922904521129, -0.208774291121169, -0.20011522267149,     -0.488608177258221, 1.48234288614357, -1.50633137280282,     0.259908768065829, 1.2373542241353, 1.07427819456565, 0.243583338757794,     -1.14208688558286, -0.837725450551871, 2.59088104046123,     -1.15715129997831, 0.00564959223608285, 0.600820045603635,     0.105591220030159, 0.0128123107151138, 0.246971633161796,     -0.398413212942429, -1.175467477882, 0.626847778816832, 1.77846886876305,     -0.591456397689165, -0.158789421307341, 0.559896996495819,     0.112290040202645, -0.575953743997274, -0.750107055841892,     0.530256387182898, -0.513816911261725, -0.0692070721634601    ), A.3 = c(1.1149567668641, -0.192283134641271, -0.924067100684333,     0.791488999743996, -0.162726742314376, 0.292580210616945,     -1.01072056005563, -1.06704244853902, -0.917829420270349,     -1.49530251213482, 0.55164735767438, 0.231639906707315, 1.20511345841344,     0.638051544857534, 0.636670051056175, -0.118568968334303,     0.815017067426338, 0.721082972688589, 0.950934636381208,     -0.56467645133333, -0.576807706076571, 1.29359903142954,     1.49527604698423, 0.258152450487747, -0.876779477100404,     1.06229801279152, -0.847046602624519, 2.2525641923268, 0.689506745677353,     -0.659217870825013, 0.488761025299921, 1.05944935347842,     0.710452551481023, -0.215593968852594, 0.561959051750116,     -1.56280817774282, 2.51466096196415, 0.151438891874951, -0.661083804964376,     0.141889543613604, -0.901871993663756, 0.598514050900579,     -1.18094595225453, -0.309750132777415, 0.758740739092494,     -2.00986975249948, -1.23401335126648, 0.288250485483344,     -0.549836115647816, 0.436071573372441), B.1 = c(0.88227811598192,     0.897842384415048, -0.184777630171888, 0.118623316088924,     -0.0564824367865818, 0.545750846013, 0.268539719747025, 0.975841723357757,     -2.04629909437382, 0.412823350181239, -0.721933677654918,     0.318875039910934, 0.941837273535689, 0.640099659697978,     0.607227229176774, -0.948411438247242, 0.0738159493037895,     0.0337977915032621, 1.88198706974487, -0.0647854801027226,     -0.225400513088105, -0.120662480391292, 1.62536502869335,     1.04461890314218, 0.834814211684392, -1.60538711098302, -1.9851286173617,     1.89363388158355, -0.74695035704131, 1.9009826329358, -0.25847209817079,     -0.463029904663805, 0.584618966201008, -0.496558323521513,     1.15403771880279, 0.0441845632496315, -0.858678748791789,     0.262813078168277, 0.343773341945782, -0.508843243222405,     0.815219065496461, -0.839396305330155, 0.492457581634416,     -0.905915099681751, -1.0122397833541, 0.517244812382749,     0.211041177828764, 0.713455422441974, -1.23845438407888,     -0.327696758596643), B.2 = c(-0.153572180663021, -0.993987430000422,     0.14397289539827, -0.459324904751497, 0.397736977399803,     0.26667977798876, 1.28935846259134, -0.437323355737327, 1.52980310555637,     0.825634206262754, -0.797868139893991, -1.22647376409885,     -1.10898535069477, -1.50078232502116, -0.42749597918191,     0.515924722153957, 1.29385972817248, -0.481647074682474,     -0.719700336457863, 1.27755898341171, -0.266409503696097,     0.0784696194150443, -1.010887001423, -0.77609513652601, -0.90703497406718,     0.982243205244392, 0.49032299125996, 0.12584592172481, 0.84976562859811,     0.161984293486552, -2.4691487243806, 1.78842828872065, -1.63130848442471,     1.46460972320939, 0.0237105772487264, 1.40561346695275, -0.0123910371447213,     -0.234785678242306, -0.409054749888131, -0.0174810774517829,     0.102456010912447, -1.00544526133298, -1.54609170567956,     -1.50480428472334, -1.05284602353784, 0.26539023877268, 0.0120135977073631,     -0.481237845906985, -0.626618217211774, -0.0992340928743656    ), B.3 = c(0.806415373399649, 0.306040750760057, -0.54549241741607,     -1.09091850273044, 1.54744026990954, 1.29109430534825, 0.956841068628721,     -0.846572150287317, -0.577459079281982, 0.478056551332068,     0.298627145332155, 0.0489160922472035, 2.2970055429043, 1.02761254168648,     -0.306596095606344, -1.52586278912606, 0.2930190669126, -1.17444393448793,     0.312874687407782, 2.67613009871411, -0.256348678566626,     -0.99118185725676, 0.155504945424921, -0.0850613002264661,     -0.0669556507792918, 1.24649671913251, -2.27239020355918,     -0.161585591533009, 0.751758633049158, 0.196544347442692,     0.383164626272538, 0.311058659057445, -0.697385111612832,     -0.47409835270024, -0.559106762888743, 0.674203678476434,     -0.238018419552101, -1.32789960155193, -0.761597790116623,     0.30690960926349, 1.01758676028939, -0.0219624718968189,     1.64261294681858, 0.406692590713878, 2.3249991683034, -0.993798627881267,     -0.573653122324496, -0.246897364884703, -0.458360684998085,     -0.347776688931503), C = c(-0.367486046546633, 0.63932750273213,     -0.334820300632738, 0.911114153502134, 2.00868569935465,     0.3832291653645, 1.45862587959787, 0.716873531612031, -0.614280019814033,     1.61663165502109, -0.208099726007868, -1.71604001044217,     -0.718716737535585, 0.704405409592569, 0.0760528834749774,     0.561088178710928, 0.488590409898465, 0.971286586064149,     -0.755373682550199, -2.16209204186468, 0.689252085068228,     -2.13501275140411, 0.513017516977636, 1.04247204916763, -0.860811202332291,     -0.47872996367036, 0.836710887123723, 1.66315679827161, 0.184112141322191,     0.788450077995874, -0.0398260411391283, 0.4293552662175,     -0.160437466004755, 0.485585942418867, 1.32306248481115,     0.629730195768536, -1.70331712232134, -1.03707077971807,     0.318421923629666, 0.723023273563144, 1.36950258044347, -0.1507180580569,     1.08359944841217, -0.278920560793472, 0.407238136053788,     -0.717859047868323, -1.28723942093888, -0.768482659146387,     0.572071282937629, 0.306653942014992), D.3 = c(1.26430389975073,     -1.11283790188355, -0.66030745375527, 1.66443158199656, -0.181593058954558,     0.0368492812298079, -1.1870841988767, -0.223739363228135,     0.581651352146898, -0.0716522842185245, -0.680199159936641,     1.11275036285987, 0.354514875481286, 0.0248429348457335,     -0.667547861684412, -1.05648839110249, 0.492165835095042,     -1.40196892238784, -0.0932265078833581, 1.11315393137081,     0.336474490236605, -0.932394945266408, -1.68330036830465,     1.31835437938226, 0.846154746071986, 0.333421363791736, -1.97070441555146,     -0.5662440933059, -1.22746739531291, -0.906177456203705,     0.12625543628784, -1.25126293985218, 0.152247011419143, -1.23611481321429,     1.1316061019306, -0.280189531588507, -0.668172081262207,     1.52922266074534, 0.731426830441485, -1.38610393690203, -1.43325274361465,     0.145390483286963, 1.32588913560687, 1.50904030587281, 0.405372857964161,     1.0663091826094, -1.78493489770928, 0.648892130036079, -0.15385392971435,     0.266254755714662), A.Infancy = c(0.101952388219046, 0.46082327247387,     -0.201342310061873, -0.0131280498394575, -0.181195893498133,     -0.406423188336067, 1.202497917641, 0.79278417914872, -0.444161254497,     -1.38552900730663, -1.08943043194399, -0.936996301799517,     0.026271560762837, -0.578813704500044, 0.544605419797346,     -0.885593813082439, -0.373129507639596, 0.702062671820677,     -0.482103910229485, 0.466928522429106, -0.0390575751947365,     -0.084304074253031, 0.916162840893851, -1.46813132864433,     -0.078972276747036, -0.18475744222866, 0.511029584300775,     0.644927835311577, -0.581062239810719, -0.484163817457973,     2.13190871952879, -0.924982703770337, 1.24552939729108, 0.0975471455509825,     -0.61205710114356, 0.0717098698089884, -0.573498425495312,     0.280297226228591, -1.42521766062896, -0.720876178987529,     0.574187801592769, 0.021794091498861, 0.454832725508079,     0.485753289634421, -0.811699014257413, -0.203233176271593,     -0.339435526886322, 0.113176492097964, -0.197742530593473,     0.377314238704226), A.Toddlerhood = c(1.1149567668641, -0.192283134641271,     -0.924067100684333, 0.791488999743996, -0.162726742314376,     0.292580210616945, -1.01072056005563, -1.06704244853902,     -0.917829420270349, -1.49530251213482, 0.55164735767438,     0.231639906707315, 1.20511345841344, 0.638051544857534, 0.636670051056175,     -0.118568968334303, 0.815017067426338, 0.721082972688589,     0.950934636381208, -0.56467645133333, -0.576807706076571,     1.29359903142954, 1.49527604698423, 0.258152450487747, -0.876779477100404,     1.06229801279152, -0.847046602624519, 2.2525641923268, 0.689506745677353,     -0.659217870825013, 0.488761025299921, 1.05944935347842,     0.710452551481023, -0.215593968852594, 0.561959051750116,     -1.56280817774282, 2.51466096196415, 0.151438891874951, -0.661083804964376,     0.141889543613604, -0.901871993663756, 0.598514050900579,     -1.18094595225453, -0.309750132777415, 0.758740739092494,     -2.00986975249948, -1.23401335126648, 0.288250485483344,     -0.549836115647816, 0.436071573372441)), row.names = c("1", "2", "3", "4", "5", "6", "7", "8", "9", "10", "11", "12", "13", "14", "15", "16", "17", "18", "19", "20", "21", "22", "23", "24", "25", "26", "27", "28", "29", "30", "31", "32", "33", "34", "35", "36", "37", "38", "39", "40", "41", "42", "43", "44", "45", "46", "47", "48", "49", "50"), class = "data.frame"), na.action = function (object,     ...) {    if (NCOL(object) == 0)         object    else {        ok <- complete.cases(object)        if (all(ok))             object        else stop(message)    }}): object is not a matrix