Fits outcome model

Fits outcome model

getModel(
  d,
  exposure,
  exposure_time_pts,
  outcome,
  epochs,
  exp_epochs,
  int_order,
  model,
  fam,
  covariates,
  verbose
)

Arguments

d

wide data frame

exposure

name of exposure variable

exposure_time_pts

list of integers at which weights will be created/assessed that correspond to time points when exposure was measured

outcome

name of outcome variable with ".timepoint" suffix

epochs

data frame of exposure epoch labels and values

exp_epochs

list of exposure epochs

int_order

integer specification of highest order exposure main effects interaction for interaction models

model

character indicating one of the following outcome models:

"m0" (exposure main effects)

"m1" (exposure main effects & covariates)

"m2" (exposure main effects & their interactions)

"m3" (exposure main effects, their interactions, & covariates)

"covs" (covariate only model)

"int" (intercept only model)

fam

function specification for svyglm model

covariates

list of characters reflecting variable names of covariates for covariate models

verbose

TRUE or FALSE indicator for user output (default is TRUE)

Value

list of fitted model(s)

Examples

test <- data.frame(ID = 1:50,
                   A.1 = rnorm(n = 50),
                   A.2 = rnorm(n = 50),
                   A.3 = rnorm(n = 50),
                   B.1 = rnorm(n = 50),
                   B.2 = rnorm(n = 50),
                   B.3 = rnorm(n = 50),
                   C = rnorm(n = 50),
                   D.3 = rnorm(n = 50))
test[, c("A.1", "A.2", "A.3")] <- lapply(test[, c("A.1", "A.2", "A.3")], as.numeric)

f <- createFormulas(exposure = "A",
                    exposure_time_pts = c(1, 2, 3),
                    outcome = "D.3",
                    tv_confounders = c("A.1", "A.2", "A.3", "B.1", "B.2", "B.3"),
                    ti_confounders = "C",
                    type = "full",
                    save.out = FALSE)
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A - D.3 at A time point 1 is: 
#> A.1 ~ C
#> <environment: 0x55d91e82f078>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A - D.3 at A time point 2 is: 
#> A.2 ~ A.1 + B.1 + C
#> <environment: 0x55d91e82f078>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A - D.3 at A time point 3 is: 
#> A.3 ~ A.1 + A.2 + B.1 + B.2 + C
#> <environment: 0x55d91e82f078>
#> 

w <- createWeights(data = test,
                   exposure = "A",
                   outcome = "D.3",
                   formulas = f,
                   save.out = FALSE)
#> Error in weightit2cbps.cont(covs = covs, treat = treat, subset = by.factor ==     i, s.weights = s.weights, missing = missing, verbose = verbose,     ...): The package "CBPS" is required.

epochs = data.frame(epochs = c("Infancy", "Toddlerhood"),
                    values = I(list(c(1, 2), c(3))))
e <- apply(expand.grid("A", as.character(epochs[, 1])), 1,
           paste, sep = "", collapse = ".")
test$weights <- w[[1]]$weights
#> Error in eval(expr, envir, enclos): object 'w' not found

g <- getModel(d = test,
              exposure = "A",
              exposure_time_pts = c(1, 2, 3),
              outcome = "D.3",
              epochs = epochs,
              exp_epochs = e,
              fam = gaussian,
              model = "m0")
#> Error in model.frame.default(~weights, data = structure(list(ID = 1:50,     A.1 = c(-0.60185202651332, 0.528034290319414, 1.8921863883068,     0.52522503243583, -0.353680010152163, -1.38853092933244,     -0.0701112688431464, 0.2941234633837, 0.924571800304047,     1.45851121403679, 0.243752673437705, 0.250616866047152, 0.168485565163588,     1.30683993163584, -2.21727545641683, -1.33699317679778, 1.63281178180291,     0.337862804200482, 0.146106962392631, 0.343197095745641,     0.211108269452301, -1.51812779323677, 0.857221894755625,     -2.05696889604144, -0.287822892471928, -1.53193973398818,     -0.80323507520317, -0.622526232372035, -0.344991939847064,     -1.81892888975417, 0.718221448855369, -0.448116728154763,     0.147397964102719, -0.13192740196463, 0.220752593871431,     0.172386871935496, 2.82120312135267, -0.493110240075883,     -1.51486319287383, 0.766992881030238, -0.6566333872284, -0.521526132741258,     0.370682279482499, 0.482973049395727, 0.827140920629698,     -0.372385685202817, 0.802984577349694, 0.692128051233959,     0.694258905712898, -0.171857883476257), A.2 = c(0.270961006499067,     -2.25984642306578, 0.593229434103126, 0.407456643056402,     -1.29007610912291, 0.217906113599209, 0.647182695818277,     1.10816982393872, -1.63689108509304, 2.49699523593072, 0.119838153902833,     0.263133227362862, 0.0322430990808775, -0.874318059180091,     0.466586863751725, -0.969029433520521, -1.54562001240314,     -1.14880052721635, -0.353081722390703, -0.572405762259367,     0.982881840434135, -0.576316764580739, -0.629675776947107,     0.119720387073813, -0.394824294090142, 0.313244516146971,     -0.28107930973844, 0.308182805148049, -1.676190569932, -2.14607603988438,     -0.486038278024561, -1.39742418324164, -1.56154188936531,     0.131644889841654, -0.597301394150915, 1.12420187484902,     -0.257295258219576, -0.0307728555128605, 0.319512762192839,     -0.425148852957325, -0.0615825236399539, -0.0119578813047045,     1.04304785411878, 0.764386842512483, -0.952545755269189,     -0.702271801193377, -0.679466095131532, 1.62959030312742,     2.95877098962688, 0.334246684807173), A.3 = c(-1.10121683689032,     -0.439101066148826, -0.449107595131614, -0.847400096377347,     0.531836346239353, -0.149396561154588, -1.04645352722945,     0.24281503183219, -0.619567259878546, -0.356093100508229,     1.63605098821418, -0.876080168253935, 0.0739503684173753,     -0.85662636738569, 0.506691454748517, -1.00489671808254,     1.65487139090127, 0.701232770487843, -0.634663721246181,     -0.497351352866503, -0.984175682692811, 0.840528076642117,     1.70903251450484, -0.681313378324621, 0.680912055276409,     -0.57771930605824, 1.08703336634534, 0.0262155870497758,     0.877810608134852, -2.03600440432231, 0.390129830040549,     -2.0843899084823, -0.14374305922094, -1.09491327151627, -0.334734143382563,     -0.277414301543777, 0.316806712630429, 0.0553068274074319,     0.646309445018397, -0.616849707437418, -2.37863501172447,     -0.258221534863211, 1.59305331621686, -0.287054815671446,     -0.509649704659254, 1.6118250553722, 0.411488245931958, 0.166540348302886,     -0.143561085185634, -1.05324404285882), B.1 = c(0.733902263853273,     -0.302393165119548, -0.922502354372411, 1.55924041350366,     -0.918386217997633, 0.125995133914373, 0.0221692809002496,     -0.45988286651908, -0.105568776091082, 0.393006588463357,     -0.135545878190474, 0.562954755342478, 0.606095219577209,     0.0178822666653738, 1.87590111949591, 1.02974320445934, -0.019387137884942,     -0.735965668398314, -0.754435083616058, 0.21438572323582,     0.61742627722285, 0.812282179357373, -1.3543947430607, -0.841488413038896,     0.903007384852404, 0.527520066432669, -0.91256503526451,     0.136100489015161, -0.93475988488684, -0.0321331223757922,     -1.8783765067257, 0.428471577182344, -0.448655791004163,     0.599766443108442, 1.10869958064586, -0.625189727641844,     2.04011953607883, -0.0221682910283609, 0.0603054780250284,     0.755892799150283, -1.21489590767327, -0.211683922487947,     0.139525575861318, 2.14190916276178, -0.418783804188219,     -2.5180273455689, -0.134364135813141, 0.131747798668025,     0.897320808644343, -0.6665397711661), B.2 = c(-0.788338365126166,     2.09283577169468, -0.758438650880858, -0.416843877325216,     0.372518372461331, -0.678110305882178, -1.69919614629951,     -0.106418160498366, 0.237630994924471, 0.0538613141608765,     0.446014255286879, 1.47737397079152, 0.756235225599906, -0.0148647712504928,     -1.10618171106434, 0.888773693560104, 1.64780984488611, 0.737505662830606,     0.512476841094554, 1.18544542001686, -0.686717423118921,     1.3441555711853, 0.335936502955824, 1.26567315675659, 0.203237157961107,     -1.61568065268094, -1.42650427033954, -0.0642571974762652,     -0.443441004686428, 0.420375770499958, 0.820025785000244,     -0.307638459321226, -2.49600671386617, 0.242814509572375,     0.592523606272107, -1.86816388146961, -0.956411121908543,     0.326860540505014, -0.870290090823052, 0.119977395786093,     -1.22882818430808, 0.517027625244997, 1.47111025166769, -0.716790820780153,     0.997074264687338, -1.3493359729758, -0.0101876204488153,     -1.15899362284245, 0.217701466399245, -1.74280012633517),     B.3 = c(0.334546728475291, 1.71261452840316, -0.853586532748612,     -0.0423606994766766, -2.23088359909212, 1.18763438249573,     -0.0592872973915312, 0.0832252818676151, 0.0159287060075786,     -2.0819886175425, -0.176797067255853, 0.0267704567658655,     1.52842795293179, 0.757288223165645, -1.14981579933489, -0.806146974339466,     -0.913526621900302, -0.18033042895796, -0.51141890368307,     0.659506368653517, -1.21591621674018, 0.421735510245671,     0.382992574856304, -0.083820239775985, 0.350724240563802,     2.08125421296987, 1.15127532670214, 0.0621748618183793, -2.02209825504288,     1.35045906456161, 0.751417421505493, -1.29656186547962, 0.416151215414589,     -0.700688966508001, -0.153027095285305, 0.849171382427242,     0.253911071846921, 0.40024087285692, -1.76473753459955, -0.136188777128799,     0.104340120091198, -0.463493365793222, 0.0723937544449032,     0.290718017941978, 0.823706964473925, -1.28028435264458,     0.441371612378869, 0.0187340543918872, 0.271877110402899,     -1.22679854405383), C = c(0.443531886271057, -1.0623641733152,     -0.652224372246064, 0.214091652507112, -1.43714620830225,     0.532760007296105, -1.91959852934378, -1.15197078759096,     0.145576455467865, 0.429269572327042, -0.481834658937795,     -0.793544175208829, -1.9485310223369, 1.90643948633443, -0.486908012152127,     0.456182441769302, -0.700944919735761, 2.17766309407031,     0.117684611351793, 1.294902475015, 0.414989531132921, 0.240071557289162,     0.0532573608072097, 2.99371380525721, -0.301276951431916,     0.526737328029559, 1.1330809500643, -0.642401446885967, -0.440840359805014,     -1.02414662063633, -1.05725161044965, 0.129062548135474,     1.39460988687588, -0.404649765520257, 0.0229826702534354,     -0.109915004399314, -0.194432858077655, 0.292677508668626,     1.11872409289631, 1.34828947003229, -0.154549776034092, -0.073460726075227,     0.013978956174036, -0.356647422014649, -0.160968341020184,     -0.599015181843066, 0.656670543026364, -0.314028979575451,     -0.458049078120262, -1.69544629552375), D.3 = c(0.167965095935954,     -1.52384085239148, -0.162840606037548, 0.739958802565245,     -1.15376388496291, 2.1270537067234, 0.877191494427465, 0.983200874902224,     -0.655636472420653, -0.362351328681766, -0.35244800524216,     -1.32518316999771, 1.24429349010537, -0.144580233925726,     -1.85179639717905, -2.5359195470723, 0.222829462838939, 0.246570086337119,     -1.27803799818458, -1.97326945580516, 0.982142237092125,     0.226600741369729, 0.589038487525489, 0.566431212185551,     0.579396058423609, 0.775954399026039, -2.15082407014063,     -0.592047840237367, -0.599265295887919, 0.130101605705654,     0.738296062035396, -0.679879377809226, 1.0995242971924, -0.134988882623517,     0.903761777068103, -0.0473229563399943, -0.446235378192363,     0.429738822086036, -1.20622496044267, 0.831169569725886,     -1.42754161456518, 1.62307356796775, 0.105463220809351, 0.132488442387096,     0.343812789932879, -0.0993521341504282, 0.521196974103031,     0.837565801550919, -0.879142745070767, 0.900839284272521),     A.Infancy = c(-0.165445510007127, -0.865906066373183, 1.24270791120496,     0.466340837746116, -0.821878059637537, -0.585312407866615,     0.288535713487565, 0.70114664366121, -0.356159642394497,     1.97775322498376, 0.181795413670269, 0.256875046705007, 0.100364332122233,     0.216260936227874, -0.875344296332553, -1.15301130515915,     0.043595884699885, -0.405468861507934, -0.103487379999036,     -0.114604333256863, 0.596995054943218, -1.04722227890875,     0.113773058904259, -0.968624254483814, -0.341323593281035,     -0.609347608920604, -0.542157192470805, -0.157171713611993,     -1.01059125488953, -1.98250246481928, 0.116091585415404,     -0.922770455698201, -0.707071962631295, -0.000141256061487988,     -0.188274400139742, 0.648294373392258, 1.28195393156655,     -0.261941547794372, -0.597675215340496, 0.170922014036456,     -0.359107955434177, -0.266742007022981, 0.70686506680064,     0.623679945954105, -0.0627024173197455, -0.537328743198097,     0.061759241109081, 1.16085917718069, 1.82651494766989, 0.081194400665458    ), A.Toddlerhood = c(-1.10121683689032, -0.439101066148826,     -0.449107595131614, -0.847400096377347, 0.531836346239353,     -0.149396561154588, -1.04645352722945, 0.24281503183219,     -0.619567259878546, -0.356093100508229, 1.63605098821418,     -0.876080168253935, 0.0739503684173753, -0.85662636738569,     0.506691454748517, -1.00489671808254, 1.65487139090127, 0.701232770487843,     -0.634663721246181, -0.497351352866503, -0.984175682692811,     0.840528076642117, 1.70903251450484, -0.681313378324621,     0.680912055276409, -0.57771930605824, 1.08703336634534, 0.0262155870497758,     0.877810608134852, -2.03600440432231, 0.390129830040549,     -2.0843899084823, -0.14374305922094, -1.09491327151627, -0.334734143382563,     -0.277414301543777, 0.316806712630429, 0.0553068274074319,     0.646309445018397, -0.616849707437418, -2.37863501172447,     -0.258221534863211, 1.59305331621686, -0.287054815671446,     -0.509649704659254, 1.6118250553722, 0.411488245931958, 0.166540348302886,     -0.143561085185634, -1.05324404285882)), row.names = c("1", "2", "3", "4", "5", "6", "7", "8", "9", "10", "11", "12", "13", "14", "15", "16", "17", "18", "19", "20", "21", "22", "23", "24", "25", "26", "27", "28", "29", "30", "31", "32", "33", "34", "35", "36", "37", "38", "39", "40", "41", "42", "43", "44", "45", "46", "47", "48", "49", "50"), class = "data.frame"), na.action = function (object,     ...) {    if (NCOL(object) == 0)         object    else {        ok <- complete.cases(object)        if (all(ok))             object        else stop(message)    }}): object is not a matrix