Fits outcome model

Fits outcome model

getModel(
  d,
  exposure,
  exposure_time_pts,
  outcome,
  epochs,
  exp_epochs,
  int_order,
  model,
  fam,
  covariates,
  verbose
)

Arguments

d

wide data frame

exposure

name of exposure variable

exposure_time_pts

list of integers at which weights will be created/assessed that correspond to time points when exposure was measured

outcome

name of outcome variable with ".timepoint" suffix

epochs

data frame of exposure epoch labels and values

exp_epochs

list of exposure epochs

int_order

integer specification of highest order exposure main effects interaction for interaction models

model

character indicating one of the following outcome models:

"m0" (exposure main effects)

"m1" (exposure main effects & covariates)

"m2" (exposure main effects & their interactions)

"m3" (exposure main effects, their interactions, & covariates)

"covs" (covariate only model)

"int" (intercept only model)

fam

function specification for svyglm model

covariates

list of characters reflecting variable names of covariates for covariate models

verbose

TRUE or FALSE indicator for user output (default is TRUE)

Value

list of fitted model(s)

Examples

test <- data.frame(ID = 1:50,
                   A.1 = rnorm(n = 50),
                   A.2 = rnorm(n = 50),
                   A.3 = rnorm(n = 50),
                   B.1 = rnorm(n = 50),
                   B.2 = rnorm(n = 50),
                   B.3 = rnorm(n = 50),
                   C = rnorm(n = 50),
                   D.3 = rnorm(n = 50))
test[, c("A.1", "A.2", "A.3")] <- lapply(test[, c("A.1", "A.2", "A.3")], as.numeric)

f <- createFormulas(exposure = "A",
                    exposure_time_pts = c(1, 2, 3),
                    outcome = "D.3",
                    tv_confounders = c("A.1", "A.2", "A.3", "B.1", "B.2", "B.3"),
                    ti_confounders = "C",
                    type = "full",
                    save.out = FALSE)
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A-D.3 at A time point 1 is: 
#> A.1 ~ C
#> <environment: 0x55c8ba2c4020>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A-D.3 at A time point 2 is: 
#> A.2 ~ A.1 + B.1 + C
#> <environment: 0x55c8ba2c4020>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A-D.3 at A time point 3 is: 
#> A.3 ~ A.1 + A.2 + B.1 + B.2 + C
#> <environment: 0x55c8ba2c4020>
#> 

w <- createWeights(data = test,
                   exposure = "A",
                   outcome = "D.3",
                   formulas = f,
                   save.out = FALSE)
#> Error in weightit2cbps.cont(covs = covs, treat = treat, subset = by.factor ==     i, s.weights = s.weights, missing = missing, verbose = verbose,     ...): The package "CBPS" is required.

epochs = data.frame(epochs = c("Infancy", "Toddlerhood"),
                    values = I(list(c(1, 2), c(3))))
e <- apply(expand.grid("A", as.character(epochs[, 1])), 1,
           paste, sep = "", collapse = ".")
test$weights <- w[[1]]$weights
#> Error in eval(expr, envir, enclos): object 'w' not found

g <- getModel(d = test,
              exposure = "A",
              exposure_time_pts = c(1, 2, 3),
              outcome = "D.3",
              epochs = epochs,
              exp_epochs = e,
              fam = gaussian,
              model = "m0")
#> Error in model.frame.default(~weights, data = structure(list(ID = 1:50,     A.1 = c(-2.25450124119163, -1.14145977960781, -2.13534861157083,     -0.118810136596918, -0.30324174831627, -0.376425114624171,     -0.788075327203601, -2.67516332127397, -1.07789064734709,     -1.22467420051028, -0.178760543537634, -0.760074058794734,     0.163949229996246, 0.0246687124383226, -0.934031702852378,     -0.270510928293259, 0.0809099735710346, -0.78908854168339,     0.301774908923858, 0.40128226757549, 2.22447969013788, 0.254289899475033,     1.34808949484331, 2.29398806975202, -1.21255293474454, -0.302355598393505,     0.168199175873409, -0.387755233232884, -0.379545633425957,     -1.56484714663701, -0.0318226120058215, 0.687978959895563,     -0.0813124253006793, 2.12669700659682, 0.47622504700988,     0.732958468945113, -0.48740319839282, -0.950014794221114,     -0.390432560221552, -0.689107518735479, 0.406368245377326,     -0.326822274374307, -0.807451502644137, -0.0694317827261806,     0.191962616834976, -0.539854024883315, 1.54328047723254,     -1.99463880020722, -0.581984283828905, -1.4558654922923),     A.2 = c(0.577992203331744, -1.11085012886211, 0.84277291085755,     0.595060155753058, 0.183936428403484, 0.632173273025868,     2.44730099142639, -0.123335973220438, 1.01115237876435, 0.169050434837047,     -0.203735635617952, 0.775377216859346, -0.122473234987539,     0.375545069403844, 1.03743973828961, -0.547550233305927,     1.11455109379909, 1.66977791276059, -2.06473719000494, 1.19125802769125,     -0.532913252406135, 1.3024831073358, -0.375362513591956,     -0.515751213438881, -1.6002864276098, -0.339081085536809,     -1.05816919413866, 1.55467074520256, -0.54651510831059, -1.15849208180778,     -0.75787765803346, -0.280811308419089, -1.25703203711126,     -0.0608999313809675, 0.0891266411762748, 0.198844132451854,     -1.28684424388989, 1.1512729244014, 0.53329949181841, 0.242416194494,     1.2086916924765, -0.2848024882502, 0.287320237583405, -2.24570698525179,     0.378225579276626, -0.325206918037119, 1.23231948219825,     -0.140919197402143, 0.625508773489556, 0.354973286458446),     A.3 = c(-0.87744380003437, -0.391627442895531, -0.185259674172291,     0.109212411272259, -0.275956656748692, -0.269946159832467,     -0.021663760362375, -0.236103305832099, 0.388450224476514,     0.154217551628649, 0.302937963720788, 1.55991792246305, -0.770134377066586,     -1.90937424170836, -1.88342194832306, -1.46463927812016,     0.0799441273422186, -0.264548319200325, 0.229521424532671,     -0.408833028168225, 0.354325919380273, 0.992924103281775,     1.29832459997857, -0.695181572331244, -0.708143321920888,     0.842677015536343, -0.908495881448631, -1.09693803018811,     2.46926863392779, 2.08109137897556, -2.35424722827131, -0.0420220254900963,     -0.501130047660984, -0.923429020978674, -0.436017717684017,     -0.254325518241958, 0.341038721830311, 0.273418556936212,     0.77190325673402, 0.364029387684567, 1.82415957410635, -0.756300867408734,     -1.64115605525074, 1.58183160017081, 0.532355532286597, -1.10352957821051,     0.436722735787931, -0.532108070738554, 1.06716581261015,     0.853666117549074), B.1 = c(1.08116419393203, 0.399944362762405,     1.00036111761688, -0.958591650382271, 0.735763713007786,     0.335778534477458, 1.04180067415588, -0.793299855059838,     -1.06085399086653, 0.359066271410564, 2.97714895597826, 0.00428154962576196,     0.373468753733554, -1.83626265078192, 0.588211634480402,     -0.481663916072481, -0.947490990528462, 1.28684817658054,     0.670007604180656, -0.347960374228244, -1.35208786721267,     -1.23994521114107, 0.584412694695565, -1.05479347154314,     0.360834853749589, -0.772307880442893, -1.89435788624282,     0.555891557644142, -0.480403532464094, 1.84395922306934,     1.22450933081626, -0.801253176164783, -1.0094158156599, -1.20339856652403,     -0.0292941499041596, -0.324914412231335, -1.86319163977725,     -0.254607483445443, 0.568842291913969, 0.858548375764094,     -0.533352870734617, -1.70734371070329, 0.341270489688627,     -0.932541041450691, 2.10561859987023, -0.577594068295795,     0.0620216013096346, 0.38013475913668, 0.491191134232659,     0.269436791132285), B.2 = c(-2.02518950412056, 0.189771695089096,     -1.90217576811978, 0.806017751736168, -1.57663103422269,     0.421075422395284, 0.180285494355628, 0.517201780972531,     -0.103244517261033, 0.223056568800453, 0.198225670970054,     -1.09567903441137, 1.69205282450068, -0.123316451849542,     -0.00805885720988164, 1.00267344070503, 0.803906861012534,     -0.247151041455966, 0.259063041738012, 0.097022094603209,     -0.88814118572397, 0.716260488991964, -0.588559159628505,     0.221132238952848, -0.683130110785879, 0.685648417861377,     -2.45235170530027, -1.73937639235918, 0.734221616517775,     -0.520772116201367, 0.33391871514346, -0.122794331987274,     0.57724028154293, -2.68883812975659, -1.28645646853413, -0.542452394373652,     -1.7552308835716, -0.198719128601631, -1.00440645245911,     -1.17007134485623, 0.52603990002516, 0.850717576751906, 0.736218686316835,     0.0932020384142337, 1.15908918619169, 0.0603772052978726,     0.6631271723821, -0.766469560529952, -0.827805221729054,     0.648173849859918), B.3 = c(0.785926521447533, -0.139142518118026,     1.09626714312102, 0.60011786731798, -1.37837519815434, 1.59306183171488,     -1.12809673381621, -0.642602537703011, 1.2265267466373, -1.7848447158113,     1.39908942522199, 0.122868306561592, 0.384004343705099, 0.563194981786881,     -0.473899763911188, -1.76026079232867, 0.902580263374671,     0.0594279371068133, 1.75276833979081, -1.40330418045088,     -0.168906975872621, 0.4934279458656, -0.725979681011403,     1.34862845638355, -1.72775176696615, -0.0616867022916336,     -0.113396480595841, 0.211563545061665, 0.212656773969863,     0.075101355014399, 0.0452364081828674, -0.0284862804618119,     0.0759516915823291, 0.632805514680612, -0.110947560598769,     -0.705711008514708, 0.149965858799546, -1.11375698516547,     -0.868067953546973, -1.44963971255524, -0.191030627080072,     0.559480569785197, 0.417753634438444, -0.447412209286498,     -0.855639345153863, 0.376123923106952, 0.41796150527337,     -0.418737397565568, -0.279130785120863, -0.881053972377063    ), C = c(0.196414653936781, 0.769466294619196, -2.02845765870176,     0.595503279573125, -0.0274093402678335, 2.12178383695806,     1.18381579528186, -1.11674213417336, 0.125169078090342, -0.729937996099652,     0.323931680366577, 0.401292546576264, 1.53651454092704, 0.618587035407169,     -1.42474735748881, -0.368717375137216, 0.0842921688820445,     1.11728805885442, -0.288527803127431, 0.179461444692101,     0.813072410303759, -0.649936215530812, -1.62911915629914,     0.00283992804557151, 1.19489220158764, -1.5378124477964,     -0.204431653325616, 1.49799533322746, -0.192429234884345,     -1.04529256149843, -1.57478961412046, -0.74902736678607,     0.063163660165554, 0.138322698630399, -0.686245278000631,     0.626758630966074, -0.533677615176019, -1.72873085472469,     -0.871203519056301, 0.762306735801971, 0.303482851700858,     -0.181987871126856, 0.343239935560046, -0.631274515797172,     -0.465440142890209, 0.526849610478414, -1.87659789034738,     -0.403871440090391, -0.86487041617877, 1.443078835829), D.3 = c(-1.82920818015845,     1.19978695284054, 0.230414272882386, -0.0767371760320617,     -0.217759774620487, 0.838039492876837, -0.221970057642233,     -0.17116774197908, -1.66585467025688, -0.196887559828175,     0.485382572282835, -3.02990175300521, -0.563650141471645,     1.31514810027059, 0.346017972448907, 1.01398275922492, 2.22598176971044,     -1.23898838891325, -0.555094517063851, 0.667248673219079,     -0.801872948886422, -0.793614331309931, -1.44305122722272,     -1.63697650375539, -0.561107965743226, 0.200906811433961,     -0.104218220271806, -1.00864423849931, -1.35057575227689,     -0.128682435858916, 1.82083191790706, -0.883375466790459,     0.859495338767827, 0.996805838845873, -0.42849845039475,     -0.491895469285468, -1.97913828312347, 0.606704253505769,     1.25058870890671, -0.685377559832003, 0.12035269456375, -0.385121530291156,     -0.259339979062416, 0.632235098494262, 1.14181490607591,     -1.12987901313091, -0.368786230110257, 0.423264054375855,     1.86280024484442, -1.75231932343339), A.Infancy = c(-0.838254518929943,     -1.12615495423496, -0.64628785035664, 0.23812500957807, -0.059652659956393,     0.127874079200849, 0.829612832111395, -1.3992496472472, -0.03336913429137,     -0.527811882836617, -0.191248089577793, 0.00765157903230601,     0.0207379975043535, 0.200106890921083, 0.0517040177186159,     -0.409030580799593, 0.597730533685062, 0.4403446855386, -0.881481140540541,     0.79627014763337, 0.845783218865872, 0.778386503405416, 0.486363490625677,     0.889118428156569, -1.40641968117717, -0.320718341965157,     -0.444985009132626, 0.583457755984838, -0.463030370868273,     -1.36166961422239, -0.394850135019641, 0.203583825738237,     -0.66917223120597, 1.03289853760793, 0.282675844093077, 0.465901300698483,     -0.887123721141355, 0.100629065090143, 0.071433465798429,     -0.22334566212074, 0.807529968926913, -0.305812381312254,     -0.260065632530366, -1.15756938398899, 0.285094098055801,     -0.432530471460217, 1.38779997971539, -1.06777899880468,     0.0217622448303255, -0.550446102916927), A.Toddlerhood = c(-0.87744380003437,     -0.391627442895531, -0.185259674172291, 0.109212411272259,     -0.275956656748692, -0.269946159832467, -0.021663760362375,     -0.236103305832099, 0.388450224476514, 0.154217551628649,     0.302937963720788, 1.55991792246305, -0.770134377066586,     -1.90937424170836, -1.88342194832306, -1.46463927812016,     0.0799441273422186, -0.264548319200325, 0.229521424532671,     -0.408833028168225, 0.354325919380273, 0.992924103281775,     1.29832459997857, -0.695181572331244, -0.708143321920888,     0.842677015536343, -0.908495881448631, -1.09693803018811,     2.46926863392779, 2.08109137897556, -2.35424722827131, -0.0420220254900963,     -0.501130047660984, -0.923429020978674, -0.436017717684017,     -0.254325518241958, 0.341038721830311, 0.273418556936212,     0.77190325673402, 0.364029387684567, 1.82415957410635, -0.756300867408734,     -1.64115605525074, 1.58183160017081, 0.532355532286597, -1.10352957821051,     0.436722735787931, -0.532108070738554, 1.06716581261015,     0.853666117549074)), row.names = c(NA, -50L), class = "data.frame"),     na.action = function (object, ...)     {        if (NCOL(object) == 0)             object        else {            ok <- complete.cases(object)            if (all(ok))                 object            else stop(message)        }    }): object is not a matrix