Fits outcome model

Fits outcome model

getModel(
  d,
  exposure,
  exposure_time_pts,
  outcome,
  epochs,
  exp_epochs,
  int_order,
  model,
  fam,
  covariates,
  verbose
)

Arguments

d

wide data frame

exposure

name of exposure variable

exposure_time_pts

list of integers at which weights will be created/assessed that correspond to time points when exposure was measured

outcome

name of outcome variable with ".timepoint" suffix

epochs

data frame of exposure epoch labels and values

exp_epochs

list of exposure epochs

int_order

integer specification of highest order exposure main effects interaction for interaction models

model

character indicating one of the following outcome models:

"m0" (exposure main effects)

"m1" (exposure main effects & covariates)

"m2" (exposure main effects & their interactions)

"m3" (exposure main effects, their interactions, & covariates)

"covs" (covariate only model)

"int" (intercept only model)

fam

function specification for svyglm model

covariates

list of characters reflecting variable names of covariates for covariate models

verbose

TRUE or FALSE indicator for user output (default is TRUE)

Value

list of fitted model(s)

Examples

test <- data.frame(ID = 1:50,
                   A.1 = rnorm(n = 50),
                   A.2 = rnorm(n = 50),
                   A.3 = rnorm(n = 50),
                   B.1 = rnorm(n = 50),
                   B.2 = rnorm(n = 50),
                   B.3 = rnorm(n = 50),
                   C = rnorm(n = 50),
                   D.3 = rnorm(n = 50))
test[, c("A.1", "A.2", "A.3")] <- lapply(test[, c("A.1", "A.2", "A.3")], as.numeric)

f <- createFormulas(exposure = "A",
                    exposure_time_pts = c(1, 2, 3),
                    outcome = "D.3",
                    tv_confounders = c("A.1", "A.2", "A.3", "B.1", "B.2", "B.3"),
                    ti_confounders = "C",
                    type = "full",
                    save.out = FALSE)
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A-D.3 at A time point 1 is: 
#> A.1 ~ C
#> <environment: 0x5613dc57a8b0>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A-D.3 at A time point 2 is: 
#> A.2 ~ A.1 + B.1 + C
#> <environment: 0x5613dc57a8b0>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A-D.3 at A time point 3 is: 
#> A.3 ~ A.1 + A.2 + B.1 + B.2 + C
#> <environment: 0x5613dc57a8b0>
#> 

w <- createWeights(data = test,
                   exposure = "A",
                   outcome = "D.3",
                   formulas = f,
                   save.out = FALSE)
#> Error in weightit2cbps.cont(covs = covs, treat = treat, subset = by.factor ==     i, s.weights = s.weights, missing = missing, verbose = verbose,     ...): The package "CBPS" is required.

epochs = data.frame(epochs = c("Infancy", "Toddlerhood"),
                    values = I(list(c(1, 2), c(3))))
e <- apply(expand.grid("A", as.character(epochs[, 1])), 1,
           paste, sep = "", collapse = ".")
test$weights <- w[[1]]$weights
#> Error in eval(expr, envir, enclos): object 'w' not found

g <- getModel(d = test,
              exposure = "A",
              exposure_time_pts = c(1, 2, 3),
              outcome = "D.3",
              epochs = epochs,
              exp_epochs = e,
              fam = gaussian,
              model = "m0")
#> Error in model.frame.default(~weights, data = structure(list(ID = 1:50,     A.1 = c(-0.357142017537959, -0.78616306531271, 0.0313172712333999,     1.10909562331208, -0.890309363705783, 0.491436897854422,     0.5760753138254, -0.581449584612477, -0.282222094781882,     -0.34715085659213, 1.29064465972849, -0.034231126688156,     -1.86085374544581, 0.528733605930534, -0.932330056389807,     -1.0212088853169, 2.6781223657443, -0.119952159688731, 0.893459374066032,     0.0334774344116735, -0.111619123449652, -0.0448870643342569,     -0.491785265930654, 0.852515304892555, -1.10543661011884,     -1.85746989594195, 0.0568468977791565, 0.596288958036118,     0.738259778223259, 1.62400185431564, -0.690956735694872,     0.426598158823577, -0.177167908011971, -0.646459175975434,     -0.361455136604273, 0.779515422709386, 0.455973445217749,     -0.395161114750018, -0.299816059903269, -0.261637964724087,     -0.336145600959598, -0.0853465562718682, -0.171402771171732,     1.58200307406742, -2.03490373401177, 0.890139233460189, -0.697658929546205,     -0.98492955898211, -0.505664585364758, -0.0927078125912301    ), A.2 = c(-0.376309408107937, 0.140164630015594, -0.432003934056097,     0.407576831630274, 0.402057467134433, 2.57669852620636, -1.32696076973734,     -0.0985597867516357, -1.72692499199578, 1.73712024141833,     1.45606738078442, 0.123396150064515, -2.11898770478393, 0.709519609959134,     0.422016081400043, -0.649773698397644, 0.696666192745932,     -0.99326655443001, -0.309038750186702, -0.769392252230459,     0.0636182234465674, -0.156040950197509, -0.374847917415885,     0.700123368118552, -0.934997012803424, 0.489431054012225,     -0.841183119398004, -0.802848244205125, 1.68934393356479,     0.249099814799177, 0.220596371883958, -0.354701318369369,     0.292050648783612, -0.546679357189893, -0.0563112798623923,     -0.250992130165696, -0.409340766860882, 0.426882607278982,     -1.08901483284782, -1.16931950743188, 1.16231116134138, 0.0855867839322198,     0.440426498543838, 0.152757167106324, -0.84405452618725,     -1.03998047038667, 0.599740185888371, -0.198162721803075,     1.19520035689504, -0.428203957870027), A.3 = c(0.803738448942293,     -1.40093909680395, 0.501021644460398, 0.627764725330147,     -0.682370281933376, -0.210592644534357, -0.646559706754922,     -0.125291668003063, -2.0421862280844, -0.773716235944314,     0.487105378725234, 0.817148484050123, -1.0829292546508, -0.400303559118843,     0.515339123248743, 1.85460687527208, -0.0798665999222875,     1.43326645941174, -2.02644722374954, -0.432178276336206,     0.806593384258358, 1.06844576397558, -0.126437160864639,     -0.884901038514779, 0.88104660203705, 1.04421184946849, 0.738003812710871,     0.321110497525147, -0.445341082118405, 0.00057111385204023,     1.08814916695053, 0.186197392119608, 1.44750155870641, -1.75240413733595,     0.0190436933269138, -1.19221508821653, -0.836603216382805,     2.11619657974543, 1.01196508294399, 0.331703277061333, 0.0847712605446479,     -0.0166879466583142, -0.403346589175493, -1.23102775821589,     -0.135754336793222, -0.312212859337929, 1.76263250542903,     0.404263444181913, -0.334537096940515, 0.513035402605202),     B.1 = c(-1.75592682647925, 0.355673652296292, -1.12042438770366,     0.278134335826953, -0.277187091511802, -0.41004674483736,     1.00526145426517, 1.70496137616992, -0.282901807529879, 1.14140298719834,     -0.0523510192350488, -0.0450511948751056, -0.493330847930185,     -1.56417928122569, 1.08426270562803, -1.61375583106028, -0.869373243033515,     -0.240358375512116, -0.0482683955177868, 1.08785772093114,     -0.419736187102007, -0.453212514175446, 1.24100901142751,     -0.421771446514272, -0.151117338841854, 0.525512290650879,     1.19787992884015, 1.33300567711775, 0.76349387467574, 0.795909506233705,     -0.669756028608779, -0.710898926796782, 0.469202113564601,     0.151428399728638, 1.24114164705842, -0.349750525800942,     0.666371143738925, 0.251343343051088, 1.01802456631053, -1.65754770364228,     -3.65928591616864, 1.06238608470861, -0.17479079544569, 1.08130152541851,     -1.0933817304283, 0.153743309540845, -0.164796121790785,     0.40381663306676, -1.49656287146283, 0.023065244482323),     B.2 = c(-0.224133035580685, 1.8261898062516, -2.20164636022315,     1.78834836494998, -1.50283584085355, -0.722226778353247,     1.0698648412339, -0.464515189355982, -0.853023931184298,     1.6485456227893, -0.645583354783313, -1.2411784435248, 0.404723171317961,     0.133018418248132, -0.391781926081722, -0.498861497269687,     0.278919217448237, 0.78316017870281, 0.107658475830785, 0.0692889813787212,     0.531016349292541, -0.0162920217133038, 0.419093653496203,     -1.50586823196823, -0.660126677620326, -1.38435587404653,     -0.308818371361856, -1.9233564992279, -0.0888250706664453,     -1.11519519945588, -1.06809684920687, 0.234971719002255,     -1.76877811200658, 1.05675912840662, 1.54364918667296, 0.174594093295752,     1.31691417346497, 1.0802624254613, -0.688136050201989, -0.608545327717326,     -1.25827464858748, -1.56200870827936, -0.839529616336478,     0.232223461029391, 0.347247899761856, 3.23333379699169, 0.474892610818869,     0.0657752404140707, -1.10980677342483, -1.57520051686525),     B.3 = c(-0.42906124315501, 0.234616582279294, -1.88074989285369,     0.101687421576851, -0.922408900141491, -0.70139271702619,     -2.16377191337988, 0.052242299339456, 1.7999380382265, -0.344932954906603,     1.38708506345082, 1.03581001835713, 1.16594123347826, -3.01897414575869,     1.3876383306715, -2.06191569472044, 0.131574060957059, 1.65000835664863,     -0.833184026589261, 0.220646051747762, -0.659799708730862,     0.0553504182593478, 0.183998380412175, 1.18134675165454,     -0.534084242299431, 0.254148853276574, -0.364823975771307,     0.227378227145523, -0.474939465004527, -0.48444275285884,     -1.20465468984182, 0.0784527662719408, 0.0605776912800052,     2.87153313869633, 1.34695799656575, -1.26852003310826, 1.4117590892885,     -0.862704187349559, 0.219061358218731, 0.0791639631378284,     -0.737016664796352, 2.24956809924796, -2.44744293991557,     -0.865684184806056, 0.352386388688697, 0.343152544133139,     0.594861413760552, 1.4602536306527, 0.339842160974305, 1.56174496813271    ), C = c(-0.271553550026394, 1.81383990251671, -0.371045070692278,     -0.349091672534361, -0.719094527647106, 0.0247043652571686,     -2.13892772132318, -1.26121957068276, -0.0378586333816511,     -1.41890928251942, 0.0147890559632518, 0.326783877478917,     1.48669356473283, 0.288105923666293, 0.380087718393806, 1.15273816823025,     0.985199759333368, -1.05623309288719, -1.02952946650252,     0.216455095739656, -0.878357119629956, 1.09706116922404,     -0.160284557312773, -0.0497790994413637, 0.317934284919961,     -0.647396551247867, -0.952792958153032, -0.990581144227105,     -0.623750483491537, -0.512186815063061, 0.0180215299735512,     0.524701125122021, 1.49050225473236, 0.534117049071556, -0.683793242111486,     0.888284705298017, 0.317111311509861, 0.412699569421575,     0.670223426569304, 0.116219129220572, -1.52527122436974,     2.33953764103585, -1.05977119498842, 1.43166514112, -0.453653613732989,     -1.88648761138156, -0.960406865097774, -0.217626142657388,     -1.09200072092329, 0.994873466066566), D.3 = c(-0.892644173762487,     -0.694726750589532, 2.4739140438175, 0.711124686487899, 0.0279777312885201,     -1.57695722888513, 0.141471523454278, 1.15169791316274, 1.07397189297363,     -0.242283394935375, 0.20650870358119, 0.0281014590264013,     1.008418620638, 0.289195806748602, 0.138880485205497, -1.81885417636985,     -0.90371898671028, -0.389569017674949, 1.42566842460547,     -0.579775237826684, 0.334275911937998, -0.316851808229461,     0.388267350961333, 1.03891312702179, -0.792827843248359,     0.969277732758263, -0.308196555020419, -0.823142589896185,     0.307141491952501, -2.17086814289431, -1.62523514866211,     -0.731221787329036, 0.702378281604881, -0.427372843102943,     -1.5483881887052, -0.266979342734868, 0.132303189459118,     -0.0761904543721457, 1.2437038395549, 0.303819551774695,     -0.350829561458185, 1.01754897830714, 2.43555942746112, -1.26047076624737,     -1.45614795104904, 1.15055872150363, -1.38925042833483, 0.0924401424879437,     -1.08235925796102, 2.08668433384914), A.Infancy = c(-0.366725712822948,     -0.322999217648558, -0.200343331411349, 0.758336227471177,     -0.244125948285675, 1.53406771203039, -0.37544272795597,     -0.340004685682056, -1.00457354338883, 0.6949846924131, 1.37335602025645,     0.0445825116881795, -1.98992072511487, 0.619126607944834,     -0.255156987494882, -0.835491291857272, 1.68739427924512,     -0.556609357059371, 0.292210311939665, -0.367957408909393,     -0.0240004500015423, -0.100464007265883, -0.43331659167327,     0.776319336505554, -1.02021681146113, -0.684019420964862,     -0.392168110809424, -0.103279643084503, 1.21380185589402,     0.936550834557409, -0.235180181905457, 0.035948420227104,     0.0574413703858205, -0.596569266582664, -0.208883208233333,     0.264261646271845, 0.0233163391784335, 0.015860746264482,     -0.694415446375545, -0.715478736077984, 0.413082780190891,     0.000120113830175801, 0.134511863686053, 0.867380120586872,     -1.43947913009951, -0.0749206184632405, -0.048959371828917,     -0.591546140392592, 0.344767885765141, -0.260455885230629    ), A.Toddlerhood = c(0.803738448942293, -1.40093909680395,     0.501021644460398, 0.627764725330147, -0.682370281933376,     -0.210592644534357, -0.646559706754922, -0.125291668003063,     -2.0421862280844, -0.773716235944314, 0.487105378725234,     0.817148484050123, -1.0829292546508, -0.400303559118843,     0.515339123248743, 1.85460687527208, -0.0798665999222875,     1.43326645941174, -2.02644722374954, -0.432178276336206,     0.806593384258358, 1.06844576397558, -0.126437160864639,     -0.884901038514779, 0.88104660203705, 1.04421184946849, 0.738003812710871,     0.321110497525147, -0.445341082118405, 0.00057111385204023,     1.08814916695053, 0.186197392119608, 1.44750155870641, -1.75240413733595,     0.0190436933269138, -1.19221508821653, -0.836603216382805,     2.11619657974543, 1.01196508294399, 0.331703277061333, 0.0847712605446479,     -0.0166879466583142, -0.403346589175493, -1.23102775821589,     -0.135754336793222, -0.312212859337929, 1.76263250542903,     0.404263444181913, -0.334537096940515, 0.513035402605202)), row.names = c(NA, -50L), class = "data.frame"), na.action = function (object, ...) {    if (NCOL(object) == 0)         object    else {        ok <- complete.cases(object)        if (all(ok))             object        else stop(message)    }}): object is not a matrix