Fits outcome model

Fits outcome model

getModel(
  d,
  exposure,
  exposure_time_pts,
  outcome,
  epochs,
  exp_epochs,
  int_order,
  model,
  fam,
  covariates,
  verbose
)

Arguments

d

wide data frame

exposure

name of exposure variable

exposure_time_pts

list of integers at which weights will be created/assessed that correspond to time points when exposure was measured

outcome

name of outcome variable with ".timepoint" suffix

epochs

data frame of exposure epoch labels and values

exp_epochs

list of exposure epochs

int_order

integer specification of highest order exposure main effects interaction for interaction models

model

character indicating one of the following outcome models:

"m0" (exposure main effects)

"m1" (exposure main effects & covariates)

"m2" (exposure main effects & their interactions)

"m3" (exposure main effects, their interactions, & covariates)

"covs" (covariate only model)

"int" (intercept only model)

fam

function specification for svyglm model

covariates

list of characters reflecting variable names of covariates for covariate models

verbose

TRUE or FALSE indicator for user output (default is TRUE)

Value

list of fitted model(s)

Examples

test <- data.frame(ID = 1:50,
                   A.1 = rnorm(n = 50),
                   A.2 = rnorm(n = 50),
                   A.3 = rnorm(n = 50),
                   B.1 = rnorm(n = 50),
                   B.2 = rnorm(n = 50),
                   B.3 = rnorm(n = 50),
                   C = rnorm(n = 50),
                   D.3 = rnorm(n = 50))
test[, c("A.1", "A.2", "A.3")] <- lapply(test[, c("A.1", "A.2", "A.3")], as.numeric)

f <- createFormulas(exposure = "A",
                    exposure_time_pts = c(1, 2, 3),
                    outcome = "D.3",
                    tv_confounders = c("A.1", "A.2", "A.3", "B.1", "B.2", "B.3"),
                    ti_confounders = "C",
                    type = "full",
                    save.out = FALSE)
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A-D.3 at A time point 1 is: 
#> A.1 ~ C
#> <environment: 0x558012e91090>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A-D.3 at A time point 2 is: 
#> A.2 ~ A.1 + B.1 + C
#> <environment: 0x558012e91090>
#> 
#> USER ALERT: Please manually inspect the full balancing formula below:
#> The full formula for A-D.3 at A time point 3 is: 
#> A.3 ~ A.1 + A.2 + B.1 + B.2 + C
#> <environment: 0x558012e91090>
#> 

w <- createWeights(data = test,
                   exposure = "A",
                   outcome = "D.3",
                   formulas = f,
                   save.out = FALSE)
#> Error in weightit2cbps.cont(covs = covs, treat = treat, subset = by.factor ==     i, s.weights = s.weights, missing = missing, verbose = verbose,     ...): The package "CBPS" is required.

epochs = data.frame(epochs = c("Infancy", "Toddlerhood"),
                    values = I(list(c(1, 2), c(3))))
e <- apply(expand.grid("A", as.character(epochs[, 1])), 1,
           paste, sep = "", collapse = ".")
test$weights <- w[[1]]$weights
#> Error in eval(expr, envir, enclos): object 'w' not found

g <- getModel(d = test,
              exposure = "A",
              exposure_time_pts = c(1, 2, 3),
              outcome = "D.3",
              epochs = epochs,
              exp_epochs = e,
              fam = gaussian,
              model = "m0")
#> Error in model.frame.default(~weights, data = structure(list(ID = 1:50,     A.1 = c(-0.51906864589041, 1.45073053256013, -0.436298232040166,     -1.35858137084162, -0.0332306164493083, -0.0711999837992712,     -1.86042145819028, 0.248167154282031, -1.79254480003282,     1.70306205545543, 0.355656588630145, 1.53461019959719, -0.211905281725623,     -0.452677862210282, 1.34603739652038, 1.93822577437961, -0.335861567824205,     -0.0801559096385415, -0.225676993391929, -1.7272870981989,     1.05306328384755, -0.28901731686441, 1.40785965043586, 0.478692979603117,     0.206799350302551, 0.933885338527813, 0.845106974501039,     0.274045852001026, 1.80920346347146, 0.165897435515865, 1.15078102966404,     -1.98778595244138, -0.0544694962596185, -0.758006630308587,     -0.64719556809961, -0.219348797065578, -0.383006145012258,     -0.00565030384092437, -0.0949721843534641, 0.29403840333469,     -0.311898379464648, -0.569546562632364, -0.40402900958457,     0.512413039962603, -0.772865565685269, 0.565640769199534,     -2.00376158106516, 0.961957830229014, 0.419371256094244,     -0.194861023514806), A.2 = c(1.22924903258339, -0.353901316833796,     0.268779677546977, 0.690574355627992, 0.711591643602704,     -0.462382910341889, 1.28691604952607, -0.145218749479177,     -0.0322289599179591, -0.672219367126576, -0.201208807931549,     0.449162788361627, 1.38483282825456, 2.46393499470323, 0.941228433380207,     -0.453008359193999, 0.134363151939347, 0.00730064644669076,     0.486906987103663, -0.259281958575807, 1.13596378977322,     -1.79223585618973, 1.90369900267464, -2.08281799366453, -1.03438449093727,     -1.28693539065113, 0.87005981938886, 0.963202188090252, 1.44101420521607,     -0.831677179750809, 0.246633960854801, 0.0260408216706839,     -0.0238195009889989, 0.588141531685956, 1.08944930011807,     0.589103347393669, 1.48812835530367, -1.45702569465037, 0.416600311316826,     0.39333682349007, -0.104698029480793, 1.41396602849054, 0.285434344856576,     0.407919245768742, 0.0256741478343001, -1.72598330787747,     1.48955668149682, -0.51085794594645, 0.13388702893834, 0.451442147439716    ), A.3 = c(-1.94204439334697, 0.259362074061287, -1.17397333233112,     -2.11486937857468, 0.838479555240587, 0.349509414457403,     -0.0886946921458613, 0.355696713697221, 0.910216780705568,     -0.985911081434815, 0.804258490363155, -0.373083186570763,     0.406911901294601, -0.0938812987255075, 0.78855782360105,     -0.046606370897296, 0.30937138630558, -0.925289738544237,     -1.33281663157116, -1.63380762312902, 0.994230918390833,     -1.84803464679224, -0.719812225157536, -0.209876276638258,     1.58002412792472, -0.445279425695612, -0.772616528707414,     1.50584085455894, 0.1488350146776, 0.9473969265117, -0.834509228524809,     -0.746637918368201, -1.4372971283891, -0.136076596552874,     -1.30789170743046, -0.523595133040624, -0.137026310020157,     -0.668692048559048, -1.68703740160592, -3.20066458632307,     -0.785408432601184, -2.49388789514323, -1.04584213315791,     0.0260292140246883, 1.93290572172941, 1.50803718417912, -1.44390781754661,     1.13367780342221, -1.11030025621799, -0.939779327533531),     B.1 = c(-0.398722228236444, -1.81573855147732, 0.255558145647448,     -1.46005063510723, 0.539631949486369, -0.310603495683694,     -0.263064241933535, -1.09320601766791, -0.113490641056701,     0.90743322372528, -0.144984210682817, 1.72451863735084, 0.327103592428086,     0.593368926674714, 0.907218417262458, 0.996516888438071,     0.134092137424344, 0.394289408953857, 0.379689289230189,     -0.544218641325593, 0.207558516346934, 1.37323526856362,     0.211599973866794, 0.019229011031242, -0.185504027396305,     2.65073036974517, -1.45133560489694, 0.419526348119528, 1.25431552535327,     -2.02416041830082, 0.875658641073113, -1.05302826462704,     -0.136174118765653, 0.902672737909845, -0.0348232311148631,     -0.0805385816646736, 1.71574411088038, -1.48990354447485,     -0.145704248545277, -1.51480158390601, 1.35839761337177,     -0.609507336791397, 0.170993656732691, -1.36167104757817,     0.711102169362459, -0.958219066097715, 0.832272884666144,     0.0960140337337842, 0.236866335380994, 0.419006082403044),     B.2 = c(-0.158040984032805, 1.44975003742471, 2.04287385173585,     0.653907190236938, 0.750157440118483, 0.475491926968756,     1.07515795789141, 2.34416773914997, -2.35081332005969, 0.438644651067554,     0.240729959185643, -0.647222955529558, -0.339323882430152,     1.27469178158776, -0.584840718454189, -0.320901600419578,     -0.583760597291661, -1.43040747423139, -2.53508376629579,     -1.28568465991237, 0.589766824025867, 1.64478686558908, -1.30046306764024,     0.682314886171955, -0.63486480843859, 1.65770146942205, 0.873478594196362,     -0.894389969840778, -2.42759765451955, 1.45653246879801,     0.280030990076525, -0.839930293645539, -0.633147230152932,     -0.32073811461631, 1.12845641811702, 0.856147392020871, -0.522363190156057,     1.1427101881123, 0.279985229301809, -2.20502601180671, 2.27855423908271,     0.150736508025961, 0.273113901280612, 1.02975680377476, -0.273545087263979,     -0.587317398820152, 0.0971790506966324, -1.24426332349401,     0.0592783344801872, -1.7927085764489), B.3 = c(0.301255252456858,     1.9406952926923, 0.617019719135204, -0.675836381529908, -0.957599976151865,     -1.24527963168901, 0.324702897097685, -0.828600554682926,     1.04073849872007, 0.0548318093390499, -0.23300605765082,     -1.19713379151909, -0.345922223801025, -0.12032991117581,     -0.214923148238125, -0.777954387040858, -0.201675682087713,     1.24315352678428, -0.511041103476551, 0.350919661905557,     -1.11710616972333, 0.640715950968771, 0.974231319143665,     0.284909563248504, -0.777978424496364, 1.56028162056741,     0.415931042618493, -0.740006416130789, 0.653083958631654,     2.73788501573181, 0.0785797645948764, -1.19441724026467,     0.597151456965673, -0.810225687108043, 0.990931378136618,     2.55217772880559, -0.297382902434353, -0.300099195903576,     0.454413393333099, -1.51310908399483, 1.64419842705492, 0.171795551236165,     -0.277912782659866, -0.352454886958383, -1.31789566722671,     -0.559295574200847, -0.446527004236134, 1.79765124164318,     1.14881611757099, -1.39251176991877), C = c(0.106571649624854,     -1.80356089718952, -0.355951861012855, -0.738034763214953,     -0.950579592105532, -0.425499164832481, 0.657368152651831,     -0.895988060369519, -0.395210323604342, -1.26758835373507,     -0.243244755264833, -0.539229136947688, -0.0634578054139351,     -2.26998172098227, -0.00935616571685314, -0.640504834759005,     0.420846727307358, 0.323829163424266, 0.424500113143703,     -1.03328942000843, -0.123062631945683, -2.24796372053419,     0.625772280201061, 0.108631983205236, 0.161584944877061,     -0.553929343804957, 0.118480263413676, -0.595922356276786,     0.808362534535068, 0.641652192140041, -0.452276464456499,     1.13595237638013, 1.32241987066079, -0.957704884761951, 0.575140737392222,     -0.37547730184664, -0.685362290918448, 0.212849487047367,     1.38314128591202, -0.0455311217177282, -0.00127902555601573,     -2.13709143557811, 1.1631195161576, -0.12695702384557, -0.112991998480811,     0.550387685476434, -0.433673875822376, -0.134175067520925,     -1.69563800916128, -0.292309511105191), D.3 = c(0.502790418044403,     -0.577343682296697, 0.712616052335534, -0.309602638205243,     1.08282401905372, -1.85768420151877, -0.687307595340243,     -1.11991959090299, -1.38716538214656, -0.819853975493639,     0.302574323408928, -0.438557179973584, -1.37702047441792,     -1.05820421055916, -0.131947555054732, -1.57673001083243,     -0.25280456720032, -1.23680084235634, -0.612542654852908,     -0.569940678427403, -0.903820195152924, -0.00744536092115117,     -1.2972649533132, 0.698818215266244, -0.936158340574726,     1.17329842321868, 0.900785685634017, -0.10445004186711, 0.31182981074131,     -0.59343452258287, -0.0964149502128847, 0.0721638794426724,     -1.61323999982878, -0.686643468495236, 0.28766006535767,     -0.625664400137962, -0.897300564946241, -1.02239591519869,     -1.51675557073996, -0.60812628764911, 0.122871654098497,     -0.338562802031588, -0.597111020582886, -0.00235510053189913,     1.2278000955956, -2.28138952661145, -0.242116487001759, -1.60220232658543,     1.91476706268886, -0.47540194226014), A.Infancy = c(0.35509019334649,     0.548414607863167, -0.0837592772465945, -0.334003507606814,     0.339180513576698, -0.26679144707058, -0.286752704332105,     0.051474202401427, -0.912386879975389, 0.515421344164427,     0.077223890349298, 0.991886493979408, 0.586463773264469,     1.00562856624647, 1.14363291495029, 0.742608707592806, -0.100749207942429,     -0.0364276315959254, 0.130614996855867, -0.993284528387353,     1.09451353681039, -1.04062658652707, 1.65577932655525, -0.802062507030707,     -0.413792570317359, -0.176525026061659, 0.85758339694495,     0.618624020045639, 1.62510883434376, -0.332889872117472,     0.69870749525942, -0.980872565385348, -0.0391444986243087,     -0.0849325493113155, 0.22112686600923, 0.184877275164046,     0.552561105145706, -0.731337999245647, 0.160814063481681,     0.34368761341238, -0.20829820447272, 0.422209732929088, -0.059297332363997,     0.460166142865672, -0.373595708925484, -0.580171269338968,     -0.25710244978417, 0.225549942141282, 0.276629142516292,     0.128290561962455), A.Toddlerhood = c(-1.94204439334697,     0.259362074061287, -1.17397333233112, -2.11486937857468,     0.838479555240587, 0.349509414457403, -0.0886946921458613,     0.355696713697221, 0.910216780705568, -0.985911081434815,     0.804258490363155, -0.373083186570763, 0.406911901294601,     -0.0938812987255075, 0.78855782360105, -0.046606370897296,     0.30937138630558, -0.925289738544237, -1.33281663157116,     -1.63380762312902, 0.994230918390833, -1.84803464679224,     -0.719812225157536, -0.209876276638258, 1.58002412792472,     -0.445279425695612, -0.772616528707414, 1.50584085455894,     0.1488350146776, 0.9473969265117, -0.834509228524809, -0.746637918368201,     -1.4372971283891, -0.136076596552874, -1.30789170743046,     -0.523595133040624, -0.137026310020157, -0.668692048559048,     -1.68703740160592, -3.20066458632307, -0.785408432601184,     -2.49388789514323, -1.04584213315791, 0.0260292140246883,     1.93290572172941, 1.50803718417912, -1.44390781754661, 1.13367780342221,     -1.11030025621799, -0.939779327533531)), row.names = c(NA, -50L), class = "data.frame"), na.action = function (object, ...) {    if (NCOL(object) == 0)         object    else {        ok <- complete.cases(object)        if (all(ok))             object        else stop(message)    }}): object is not a matrix